Chen & Chen Poisson

Helper function for Chen & Chen Poisson regression estimation.

Usage

chen_poisson(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

intercept

(Logical): Do the design matrices include intercept columns? Default is TRUE.

Value

(list): A list containing the following:

est

(vector): vector of Chen & Chen Poisson regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

Details

Another look at statistical inference with machine learning-imputed data (Gronsbell et al., 2026) doi:10.48550/arXiv.2411.19908

Examples

dat <- simdat(model = "poisson")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>
  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

chen_poisson(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)
#> Warning: non-integer x = 0.272009
#> Warning: non-integer x = 0.763964
#> Warning: non-integer x = 0.587593
#> Warning: non-integer x = 0.081730
#> Warning: non-integer x = 72.166509
#> Warning: non-integer x = 5.920829
#> Warning: non-integer x = 2.314976
#> Warning: non-integer x = 2.248047
#> Warning: non-integer x = 2.542241
#> Warning: non-integer x = 17.863098
#> Warning: non-integer x = 27.400125
#> Warning: non-integer x = 5.008618
#> Warning: non-integer x = 7.535563
#> Warning: non-integer x = 1.984485
#> Warning: non-integer x = 25.691345
#> Warning: non-integer x = 8.701386
#> Warning: non-integer x = 7.963324
#> Warning: non-integer x = 1.707917
#> Warning: non-integer x = 12.124774
#> Warning: non-integer x = 8.450500
#> Warning: non-integer x = 6.701719
#> Warning: non-integer x = 17.732591
#> Warning: non-integer x = 5.787534
#> Warning: non-integer x = 14.386440
#> Warning: non-integer x = 10.032540
#> Warning: non-integer x = 0.578141
#> Warning: non-integer x = 6.382050
#> Warning: non-integer x = 5.460286
#> Warning: non-integer x = 26.307016
#> Warning: non-integer x = 46.201285
#> Warning: non-integer x = 19.308713
#> Warning: non-integer x = 51.191161
#> Warning: non-integer x = 14.781223
#> Warning: non-integer x = 7.524653
#> Warning: non-integer x = 4.351624
#> Warning: non-integer x = 1.058596
#> Warning: non-integer x = 1.033619
#> Warning: non-integer x = 0.850777
#> Warning: non-integer x = 11.353744
#> Warning: non-integer x = 7.873975
#> Warning: non-integer x = 100.566816
#> Warning: non-integer x = 0.781026
#> Warning: non-integer x = 26.539589
#> Warning: non-integer x = 7.542771
#> Warning: non-integer x = 6.616478
#> Warning: non-integer x = 4.881530
#> Warning: non-integer x = 18.060400
#> Warning: non-integer x = 22.479205
#> Warning: non-integer x = 6.060138
#> Warning: non-integer x = 1.243718
#> Warning: non-integer x = 13.120645
#> Warning: non-integer x = 0.906478
#> Warning: non-integer x = 1.143326
#> Warning: non-integer x = 15.138745
#> Warning: non-integer x = 3.382681
#> Warning: non-integer x = 9.984588
#> Warning: non-integer x = 2.712599
#> Warning: non-integer x = 7.205103
#> Warning: non-integer x = 6.067256
#> Warning: non-integer x = 7.096383
#> Warning: non-integer x = 25.205063
#> Warning: non-integer x = 71.316936
#> Warning: non-integer x = 12.926057
#> Warning: non-integer x = 16.786988
#> Warning: non-integer x = 37.587076
#> Warning: non-integer x = 0.464732
#> Warning: non-integer x = 0.933990
#> Warning: non-integer x = 1.951012
#> Warning: non-integer x = 2.847963
#> Warning: non-integer x = 1.498052
#> Warning: non-integer x = 2.111852
#> Warning: non-integer x = 32.320328
#> Warning: non-integer x = 5.420994
#> Warning: non-integer x = 11.273986
#> Warning: non-integer x = 2.347209
#> Warning: non-integer x = 4.583448
#> Warning: non-integer x = 0.970209
#> Warning: non-integer x = 1.548614
#> Warning: non-integer x = 1.536620
#> Warning: non-integer x = 10.020104
#> Warning: non-integer x = 3.434709
#> Warning: non-integer x = 0.377353
#> Warning: non-integer x = 6.276104
#> Warning: non-integer x = 13.014996
#> Warning: non-integer x = 15.693718
#> Warning: non-integer x = 1.333096
#> Warning: non-integer x = 14.586454
#> Warning: non-integer x = 6.335494
#> Warning: non-integer x = 65.597038
#> Warning: non-integer x = 0.362598
#> Warning: non-integer x = 3.841144
#> Warning: non-integer x = 4.659803
#> Warning: non-integer x = 20.184681
#> Warning: non-integer x = 1.392822
#> Warning: non-integer x = 48.275518
#> Warning: non-integer x = 19.766925
#> Warning: non-integer x = 4.156679
#> Warning: non-integer x = 5.771377
#> Warning: non-integer x = 86.555073
#> Warning: non-integer x = 2.178302
#> Warning: non-integer x = 1.699642
#> Warning: non-integer x = 5.178364
#> Warning: non-integer x = 5.132666
#> Warning: non-integer x = 0.499549
#> Warning: non-integer x = 5.659888
#> Warning: non-integer x = 0.949039
#> Warning: non-integer x = 3.093032
#> Warning: non-integer x = 261.928008
#> Warning: non-integer x = 2.730547
#> Warning: non-integer x = 10.103330
#> Warning: non-integer x = 5.005815
#> Warning: non-integer x = 1.148358
#> Warning: non-integer x = 74.766218
#> Warning: non-integer x = 26.475612
#> Warning: non-integer x = 5.997070
#> Warning: non-integer x = 0.354609
#> Warning: non-integer x = 1.533764
#> Warning: non-integer x = 3.853778
#> Warning: non-integer x = 17.336245
#> Warning: non-integer x = 10.762168
#> Warning: non-integer x = 5.679860
#> Warning: non-integer x = 0.266390
#> Warning: non-integer x = 2.883124
#> Warning: non-integer x = 6.283073
#> Warning: non-integer x = 1.510384
#> Warning: non-integer x = 1.745234
#> Warning: non-integer x = 21.367778
#> Warning: non-integer x = 13.305928
#> Warning: non-integer x = 6.314480
#> Warning: non-integer x = 1.400627
#> Warning: non-integer x = 3.135752
#> Warning: non-integer x = 11.131600
#> Warning: non-integer x = 2.477315
#> Warning: non-integer x = 34.288509
#> Warning: non-integer x = 1.617951
#> Warning: non-integer x = 3.312786
#> Warning: non-integer x = 16.172952
#> Warning: non-integer x = 3.068092
#> Warning: non-integer x = 6.241113
#> Warning: non-integer x = 6.408897
#> Warning: non-integer x = 14.775465
#> Warning: non-integer x = 5.448581
#> Warning: non-integer x = 5.065558
#> Warning: non-integer x = 5.543367
#> Warning: non-integer x = 28.788793
#> Warning: non-integer x = 0.921288
#> Warning: non-integer x = 15.079566
#> Warning: non-integer x = 7.535173
#> Warning: non-integer x = 5.862649
#> Warning: non-integer x = 1.258722
#> Warning: non-integer x = 3.227722
#> Warning: non-integer x = 61.575874
#> Warning: non-integer x = 2.534489
#> Warning: non-integer x = 2.069569
#> Warning: non-integer x = 118.709424
#> Warning: non-integer x = 14.200364
#> Warning: non-integer x = 57.345273
#> Warning: non-integer x = 1.785995
#> Warning: non-integer x = 15.535222
#> Warning: non-integer x = 2.674759
#> Warning: non-integer x = 0.096859
#> Warning: non-integer x = 18.266923
#> Warning: non-integer x = 3.843987
#> Warning: non-integer x = 3.526838
#> Warning: non-integer x = 3.219350
#> Warning: non-integer x = 63.928381
#> Warning: non-integer x = 5.049485
#> Warning: non-integer x = 1.521358
#> Warning: non-integer x = 26.676996
#> Warning: non-integer x = 2.022893
#> Warning: non-integer x = 2.741253
#> Warning: non-integer x = 6.509822
#> Warning: non-integer x = 3.116213
#> Warning: non-integer x = 8.538946
#> Warning: non-integer x = 6.851044
#> Warning: non-integer x = 3.539764
#> Warning: non-integer x = 18.619641
#> Warning: non-integer x = 2.888637
#> Warning: non-integer x = 10.053694
#> Warning: non-integer x = 2.342636
#> Warning: non-integer x = 6.093397
#> Warning: non-integer x = 12.781323
#> Warning: non-integer x = 114.173977
#> Warning: non-integer x = 4.907384
#> Warning: non-integer x = 18.918524
#> Warning: non-integer x = 11.790124
#> Warning: non-integer x = 8.913084
#> Warning: non-integer x = 5.190345
#> Warning: non-integer x = 2.346940
#> Warning: non-integer x = 18.667705
#> Warning: non-integer x = 1.130741
#> Warning: non-integer x = 79.874270
#> Warning: non-integer x = 0.599893
#> Warning: non-integer x = 0.896757
#> Warning: non-integer x = 18.815116
#> Warning: non-integer x = 0.686272
#> Warning: non-integer x = 10.917331
#> Warning: non-integer x = 7.886841
#> Warning: non-integer x = 0.939184
#> Warning: non-integer x = 2.728494
#> Warning: non-integer x = 0.376013
#> Warning: non-integer x = 2.663915
#> Warning: non-integer x = 1.537559
#> Warning: non-integer x = 5.685982
#> Warning: non-integer x = 11.020476
#> Warning: non-integer x = 11.199641
#> Warning: non-integer x = 0.689271
#> Warning: non-integer x = 0.773253
#> Warning: non-integer x = 1.738646
#> Warning: non-integer x = 1.027187
#> Warning: non-integer x = 1.297446
#> Warning: non-integer x = 32.569143
#> Warning: non-integer x = 19.355384
#> Warning: non-integer x = 4.999237
#> Warning: non-integer x = 1.056657
#> Warning: non-integer x = 12.484627
#> Warning: non-integer x = 2.455976
#> Warning: non-integer x = 11.831902
#> Warning: non-integer x = 5.588019
#> Warning: non-integer x = 2.759625
#> Warning: non-integer x = 1.744129
#> Warning: non-integer x = 6.622621
#> Warning: non-integer x = 48.592104
#> Warning: non-integer x = 2.485856
#> Warning: non-integer x = 2.486196
#> Warning: non-integer x = 20.803360
#> Warning: non-integer x = 0.453347
#> Warning: non-integer x = 3.187469
#> Warning: non-integer x = 8.223508
#> Warning: non-integer x = 11.483419
#> Warning: non-integer x = 3.073551
#> Warning: non-integer x = 11.681058
#> Warning: non-integer x = 5.395144
#> Warning: non-integer x = 14.040306
#> Warning: non-integer x = 0.416870
#> Warning: non-integer x = 5.045173
#> Warning: non-integer x = 1.245203
#> Warning: non-integer x = 4.691093
#> Warning: non-integer x = 105.185816
#> Warning: non-integer x = 17.652549
#> Warning: non-integer x = 2.776683
#> Warning: non-integer x = 21.479684
#> Warning: non-integer x = 15.730193
#> Warning: non-integer x = 7.797584
#> Warning: non-integer x = 7.882012
#> Warning: non-integer x = 2.131211
#> Warning: non-integer x = 358.935300
#> Warning: non-integer x = 2.945643
#> Warning: non-integer x = 4.208935
#> Warning: non-integer x = 3.261150
#> Warning: non-integer x = 0.872314
#> Warning: non-integer x = 29.355744
#> Warning: non-integer x = 9.120484
#> Warning: non-integer x = 26.283528
#> Warning: non-integer x = 33.462361
#> Warning: non-integer x = 0.845530
#> Warning: non-integer x = 5.720925
#> Warning: non-integer x = 47.537295
#> Warning: non-integer x = 25.088889
#> Warning: non-integer x = 34.379984
#> Warning: non-integer x = 0.661231
#> Warning: non-integer x = 21.007139
#> Warning: non-integer x = 1.009298
#> Warning: non-integer x = 33.023623
#> Warning: non-integer x = 14.349993
#> Warning: non-integer x = 2.578582
#> Warning: non-integer x = 2.792696
#> Warning: non-integer x = 1.686668
#> Warning: non-integer x = 5.651195
#> Warning: non-integer x = 0.916040
#> Warning: non-integer x = 20.955176
#> Warning: non-integer x = 2.632490
#> Warning: non-integer x = 4.696064
#> Warning: non-integer x = 0.982338
#> Warning: non-integer x = 54.721316
#> Warning: non-integer x = 10.425139
#> Warning: non-integer x = 0.220239
#> Warning: non-integer x = 32.514476
#> Warning: non-integer x = 56.186997
#> Warning: non-integer x = 2.932863
#> Warning: non-integer x = 7.013639
#> Warning: non-integer x = 21.773213
#> Warning: non-integer x = 3.712176
#> Warning: non-integer x = 0.974436
#> Warning: non-integer x = 3.719995
#> Warning: non-integer x = 15.167466
#> Warning: non-integer x = 7.509117
#> Warning: non-integer x = 3.152567
#> Warning: non-integer x = 9.068405
#> Warning: non-integer x = 10.738252
#> Warning: non-integer x = 7.199520
#> Warning: non-integer x = 3.908893
#> Warning: non-integer x = 0.696130
#> Warning: non-integer x = 2.721276
#> Warning: non-integer x = 2.607437
#> Warning: non-integer x = 5.055080
#> Warning: non-integer x = 0.544459
#> Warning: non-integer x = 37.546114
#> Warning: non-integer x = 12.583416
#> Warning: non-integer x = 14.242852
#> Warning: non-integer x = 0.272009
#> Warning: non-integer x = 0.763964
#> Warning: non-integer x = 0.587593
#> Warning: non-integer x = 0.081730
#> Warning: non-integer x = 72.166509
#> Warning: non-integer x = 5.920829
#> Warning: non-integer x = 2.314976
#> Warning: non-integer x = 2.248047
#> Warning: non-integer x = 2.542241
#> Warning: non-integer x = 17.863098
#> Warning: non-integer x = 27.400125
#> Warning: non-integer x = 5.008618
#> Warning: non-integer x = 7.535563
#> Warning: non-integer x = 1.984485
#> Warning: non-integer x = 25.691345
#> Warning: non-integer x = 8.701386
#> Warning: non-integer x = 7.963324
#> Warning: non-integer x = 1.707917
#> Warning: non-integer x = 12.124774
#> Warning: non-integer x = 8.450500
#> Warning: non-integer x = 6.701719
#> Warning: non-integer x = 17.732591
#> Warning: non-integer x = 5.787534
#> Warning: non-integer x = 14.386440
#> Warning: non-integer x = 10.032540
#> Warning: non-integer x = 0.578141
#> Warning: non-integer x = 6.382050
#> Warning: non-integer x = 5.460286
#> Warning: non-integer x = 26.307016
#> Warning: non-integer x = 46.201285
#> Warning: non-integer x = 19.308713
#> Warning: non-integer x = 51.191161
#> Warning: non-integer x = 14.781223
#> Warning: non-integer x = 7.524653
#> Warning: non-integer x = 4.351624
#> Warning: non-integer x = 1.058596
#> Warning: non-integer x = 1.033619
#> Warning: non-integer x = 0.850777
#> Warning: non-integer x = 11.353744
#> Warning: non-integer x = 7.873975
#> Warning: non-integer x = 100.566816
#> Warning: non-integer x = 0.781026
#> Warning: non-integer x = 26.539589
#> Warning: non-integer x = 7.542771
#> Warning: non-integer x = 6.616478
#> Warning: non-integer x = 4.881530
#> Warning: non-integer x = 18.060400
#> Warning: non-integer x = 22.479205
#> Warning: non-integer x = 6.060138
#> Warning: non-integer x = 1.243718
#> Warning: non-integer x = 13.120645
#> Warning: non-integer x = 0.906478
#> Warning: non-integer x = 1.143326
#> Warning: non-integer x = 15.138745
#> Warning: non-integer x = 3.382681
#> Warning: non-integer x = 9.984588
#> Warning: non-integer x = 2.712599
#> Warning: non-integer x = 7.205103
#> Warning: non-integer x = 6.067256
#> Warning: non-integer x = 7.096383
#> Warning: non-integer x = 25.205063
#> Warning: non-integer x = 71.316936
#> Warning: non-integer x = 12.926057
#> Warning: non-integer x = 16.786988
#> Warning: non-integer x = 37.587076
#> Warning: non-integer x = 0.464732
#> Warning: non-integer x = 0.933990
#> Warning: non-integer x = 1.951012
#> Warning: non-integer x = 2.847963
#> Warning: non-integer x = 1.498052
#> Warning: non-integer x = 2.111852
#> Warning: non-integer x = 32.320328
#> Warning: non-integer x = 5.420994
#> Warning: non-integer x = 11.273986
#> Warning: non-integer x = 2.347209
#> Warning: non-integer x = 4.583448
#> Warning: non-integer x = 0.970209
#> Warning: non-integer x = 1.548614
#> Warning: non-integer x = 1.536620
#> Warning: non-integer x = 10.020104
#> Warning: non-integer x = 3.434709
#> Warning: non-integer x = 0.377353
#> Warning: non-integer x = 6.276104
#> Warning: non-integer x = 13.014996
#> Warning: non-integer x = 15.693718
#> Warning: non-integer x = 1.333096
#> Warning: non-integer x = 14.586454
#> Warning: non-integer x = 6.335494
#> Warning: non-integer x = 65.597038
#> Warning: non-integer x = 0.362598
#> Warning: non-integer x = 3.841144
#> Warning: non-integer x = 4.659803
#> Warning: non-integer x = 20.184681
#> Warning: non-integer x = 1.392822
#> Warning: non-integer x = 48.275518
#> Warning: non-integer x = 19.766925
#> Warning: non-integer x = 4.156679
#> Warning: non-integer x = 5.771377
#> Warning: non-integer x = 86.555073
#> Warning: non-integer x = 2.178302
#> Warning: non-integer x = 1.699642
#> Warning: non-integer x = 5.178364
#> Warning: non-integer x = 5.132666
#> Warning: non-integer x = 0.499549
#> Warning: non-integer x = 5.659888
#> Warning: non-integer x = 0.949039
#> Warning: non-integer x = 3.093032
#> Warning: non-integer x = 261.928008
#> Warning: non-integer x = 2.730547
#> Warning: non-integer x = 10.103330
#> Warning: non-integer x = 5.005815
#> Warning: non-integer x = 1.148358
#> Warning: non-integer x = 74.766218
#> Warning: non-integer x = 26.475612
#> Warning: non-integer x = 5.997070
#> Warning: non-integer x = 0.354609
#> Warning: non-integer x = 1.533764
#> Warning: non-integer x = 3.853778
#> Warning: non-integer x = 17.336245
#> Warning: non-integer x = 10.762168
#> Warning: non-integer x = 5.679860
#> Warning: non-integer x = 0.266390
#> Warning: non-integer x = 2.883124
#> Warning: non-integer x = 6.283073
#> Warning: non-integer x = 1.510384
#> Warning: non-integer x = 1.745234
#> Warning: non-integer x = 21.367778
#> Warning: non-integer x = 13.305928
#> Warning: non-integer x = 6.314480
#> Warning: non-integer x = 1.400627
#> Warning: non-integer x = 3.135752
#> Warning: non-integer x = 11.131600
#> Warning: non-integer x = 2.477315
#> Warning: non-integer x = 34.288509
#> Warning: non-integer x = 1.617951
#> Warning: non-integer x = 3.312786
#> Warning: non-integer x = 16.172952
#> Warning: non-integer x = 3.068092
#> Warning: non-integer x = 6.241113
#> Warning: non-integer x = 6.408897
#> Warning: non-integer x = 14.775465
#> Warning: non-integer x = 5.448581
#> Warning: non-integer x = 5.065558
#> Warning: non-integer x = 5.543367
#> Warning: non-integer x = 28.788793
#> Warning: non-integer x = 0.921288
#> Warning: non-integer x = 15.079566
#> Warning: non-integer x = 7.535173
#> Warning: non-integer x = 5.862649
#> Warning: non-integer x = 1.258722
#> Warning: non-integer x = 3.227722
#> Warning: non-integer x = 61.575874
#> Warning: non-integer x = 2.534489
#> Warning: non-integer x = 2.069569
#> Warning: non-integer x = 118.709424
#> Warning: non-integer x = 14.200364
#> Warning: non-integer x = 57.345273
#> Warning: non-integer x = 1.785995
#> Warning: non-integer x = 15.535222
#> Warning: non-integer x = 2.674759
#> Warning: non-integer x = 0.096859
#> Warning: non-integer x = 18.266923
#> Warning: non-integer x = 3.843987
#> Warning: non-integer x = 3.526838
#> Warning: non-integer x = 3.219350
#> Warning: non-integer x = 63.928381
#> Warning: non-integer x = 5.049485
#> Warning: non-integer x = 1.521358
#> Warning: non-integer x = 26.676996
#> Warning: non-integer x = 2.022893
#> Warning: non-integer x = 2.741253
#> Warning: non-integer x = 6.509822
#> Warning: non-integer x = 3.116213
#> Warning: non-integer x = 8.538946
#> Warning: non-integer x = 6.851044
#> Warning: non-integer x = 3.539764
#> Warning: non-integer x = 18.619641
#> Warning: non-integer x = 2.888637
#> Warning: non-integer x = 10.053694
#> Warning: non-integer x = 2.342636
#> Warning: non-integer x = 6.093397
#> Warning: non-integer x = 12.781323
#> Warning: non-integer x = 114.173977
#> Warning: non-integer x = 4.907384
#> Warning: non-integer x = 18.918524
#> Warning: non-integer x = 11.790124
#> Warning: non-integer x = 8.913084
#> Warning: non-integer x = 5.190345
#> Warning: non-integer x = 2.346940
#> Warning: non-integer x = 18.667705
#> Warning: non-integer x = 1.130741
#> Warning: non-integer x = 79.874270
#> Warning: non-integer x = 0.599893
#> Warning: non-integer x = 0.896757
#> Warning: non-integer x = 18.815116
#> Warning: non-integer x = 0.686272
#> Warning: non-integer x = 10.917331
#> Warning: non-integer x = 7.886841
#> Warning: non-integer x = 0.939184
#> Warning: non-integer x = 2.728494
#> Warning: non-integer x = 0.376013
#> Warning: non-integer x = 2.663915
#> Warning: non-integer x = 1.537559
#> Warning: non-integer x = 5.685982
#> Warning: non-integer x = 11.020476
#> Warning: non-integer x = 11.199641
#> Warning: non-integer x = 0.689271
#> Warning: non-integer x = 0.773253
#> Warning: non-integer x = 1.738646
#> Warning: non-integer x = 1.027187
#> Warning: non-integer x = 1.297446
#> Warning: non-integer x = 32.569143
#> Warning: non-integer x = 19.355384
#> Warning: non-integer x = 4.999237
#> Warning: non-integer x = 1.056657
#> Warning: non-integer x = 12.484627
#> Warning: non-integer x = 2.455976
#> Warning: non-integer x = 11.831902
#> Warning: non-integer x = 5.588019
#> Warning: non-integer x = 2.759625
#> Warning: non-integer x = 1.744129
#> Warning: non-integer x = 6.622621
#> Warning: non-integer x = 48.592104
#> Warning: non-integer x = 2.485856
#> Warning: non-integer x = 2.486196
#> Warning: non-integer x = 20.803360
#> Warning: non-integer x = 0.453347
#> Warning: non-integer x = 3.187469
#> Warning: non-integer x = 8.223508
#> Warning: non-integer x = 11.483419
#> Warning: non-integer x = 3.073551
#> Warning: non-integer x = 11.681058
#> Warning: non-integer x = 5.395144
#> Warning: non-integer x = 14.040306
#> Warning: non-integer x = 0.416870
#> Warning: non-integer x = 5.045173
#> Warning: non-integer x = 1.245203
#> Warning: non-integer x = 4.691093
#> Warning: non-integer x = 105.185816
#> Warning: non-integer x = 17.652549
#> Warning: non-integer x = 2.776683
#> Warning: non-integer x = 21.479684
#> Warning: non-integer x = 15.730193
#> Warning: non-integer x = 7.797584
#> Warning: non-integer x = 7.882012
#> Warning: non-integer x = 2.131211
#> Warning: non-integer x = 358.935300
#> Warning: non-integer x = 2.945643
#> Warning: non-integer x = 4.208935
#> Warning: non-integer x = 3.261150
#> Warning: non-integer x = 0.872314
#> Warning: non-integer x = 29.355744
#> Warning: non-integer x = 9.120484
#> Warning: non-integer x = 26.283528
#> Warning: non-integer x = 33.462361
#> Warning: non-integer x = 0.845530
#> Warning: non-integer x = 5.720925
#> Warning: non-integer x = 47.537295
#> Warning: non-integer x = 25.088889
#> Warning: non-integer x = 34.379984
#> Warning: non-integer x = 0.661231
#> Warning: non-integer x = 21.007139
#> Warning: non-integer x = 1.009298
#> Warning: non-integer x = 33.023623
#> Warning: non-integer x = 14.349993
#> Warning: non-integer x = 2.578582
#> Warning: non-integer x = 2.792696
#> Warning: non-integer x = 1.686668
#> Warning: non-integer x = 5.651195
#> Warning: non-integer x = 0.916040
#> Warning: non-integer x = 20.955176
#> Warning: non-integer x = 2.632490
#> Warning: non-integer x = 4.696064
#> Warning: non-integer x = 0.982338
#> Warning: non-integer x = 54.721316
#> Warning: non-integer x = 10.425139
#> Warning: non-integer x = 0.220239
#> Warning: non-integer x = 32.514476
#> Warning: non-integer x = 56.186997
#> Warning: non-integer x = 2.932863
#> Warning: non-integer x = 7.013639
#> Warning: non-integer x = 21.773213
#> Warning: non-integer x = 3.712176
#> Warning: non-integer x = 0.974436
#> Warning: non-integer x = 3.719995
#> Warning: non-integer x = 15.167466
#> Warning: non-integer x = 7.509117
#> Warning: non-integer x = 3.152567
#> Warning: non-integer x = 9.068405
#> Warning: non-integer x = 10.738252
#> Warning: non-integer x = 7.199520
#> Warning: non-integer x = 3.908893
#> Warning: non-integer x = 0.696130
#> Warning: non-integer x = 2.721276
#> Warning: non-integer x = 2.607437
#> Warning: non-integer x = 5.055080
#> Warning: non-integer x = 0.544459
#> Warning: non-integer x = 37.546114
#> Warning: non-integer x = 12.583416
#> Warning: non-integer x = 14.242852
#> Warning: non-integer x = 4.763131
#> Warning: non-integer x = 17.892398
#> Warning: non-integer x = 35.162086
#> Warning: non-integer x = 1.272332
#> Warning: non-integer x = 1.673144
#> Warning: non-integer x = 18.869032
#> Warning: non-integer x = 0.713509
#> Warning: non-integer x = 5.533164
#> Warning: non-integer x = 13.667043
#> Warning: non-integer x = 32.480641
#> Warning: non-integer x = 0.593488
#> Warning: non-integer x = 0.797864
#> Warning: non-integer x = 0.674422
#> Warning: non-integer x = 0.811225
#> Warning: non-integer x = 5.786690
#> Warning: non-integer x = 27.476659
#> Warning: non-integer x = 5.014716
#> Warning: non-integer x = 28.909960
#> Warning: non-integer x = 1.238345
#> Warning: non-integer x = 59.392279
#> Warning: non-integer x = 23.388377
#> Warning: non-integer x = 3.043276
#> Warning: non-integer x = 0.749737
#> Warning: non-integer x = 0.516684
#> Warning: non-integer x = 1.738626
#> Warning: non-integer x = 11.206792
#> Warning: non-integer x = 0.869948
#> Warning: non-integer x = 10.166448
#> Warning: non-integer x = 5.519745
#> Warning: non-integer x = 7.812794
#> Warning: non-integer x = 1.272675
#> Warning: non-integer x = 1.815815
#> Warning: non-integer x = 57.264076
#> Warning: non-integer x = 3.948688
#> Warning: non-integer x = 16.212356
#> Warning: non-integer x = 5.118853
#> Warning: non-integer x = 9.963817
#> Warning: non-integer x = 2.415014
#> Warning: non-integer x = 88.852932
#> Warning: non-integer x = 0.937593
#> Warning: non-integer x = 1.551173
#> Warning: non-integer x = 2.014564
#> Warning: non-integer x = 9.824689
#> Warning: non-integer x = 0.910397
#> Warning: non-integer x = 5.759400
#> Warning: non-integer x = 4.338988
#> Warning: non-integer x = 2.400498
#> Warning: non-integer x = 23.955252
#> Warning: non-integer x = 1.138491
#> Warning: non-integer x = 6.822687
#> Warning: non-integer x = 2.265869
#> Warning: non-integer x = 20.908953
#> Warning: non-integer x = 0.698507
#> Warning: non-integer x = 11.554121
#> Warning: non-integer x = 12.320762
#> Warning: non-integer x = 11.964770
#> Warning: non-integer x = 0.381273
#> Warning: non-integer x = 4.066257
#> Warning: non-integer x = 2.541306
#> Warning: non-integer x = 3.906478
#> Warning: non-integer x = 3.621969
#> Warning: non-integer x = 3.270124
#> Warning: non-integer x = 25.940631
#> Warning: non-integer x = 0.510464
#> Warning: non-integer x = 15.994931
#> Warning: non-integer x = 2.957891
#> Warning: non-integer x = 78.522327
#> Warning: non-integer x = 0.877708
#> Warning: non-integer x = 3.608849
#> Warning: non-integer x = 1.940065
#> Warning: non-integer x = 1.399996
#> Warning: non-integer x = 10.127299
#> Warning: non-integer x = 2.851930
#> Warning: non-integer x = 7.261508
#> Warning: non-integer x = 0.951092
#> Warning: non-integer x = 2.838966
#> Warning: non-integer x = 16.079660
#> Warning: non-integer x = 4.333802
#> Warning: non-integer x = 2.105136
#> Warning: non-integer x = 97.904733
#> Warning: non-integer x = 3.918281
#> Warning: non-integer x = 5.016026
#> Warning: non-integer x = 108.544558
#> Warning: non-integer x = 8.679356
#> Warning: non-integer x = 7.976228
#> Warning: non-integer x = 0.699752
#> Warning: non-integer x = 9.714830
#> Warning: non-integer x = 4.759222
#> Warning: non-integer x = 4.315062
#> Warning: non-integer x = 2.767537
#> Warning: non-integer x = 30.413415
#> Warning: non-integer x = 1.836179
#> Warning: non-integer x = 5.495042
#> Warning: non-integer x = 11.444965
#> Warning: non-integer x = 2.959710
#> Warning: non-integer x = 4.664592
#> Warning: non-integer x = 31.404981
#> Warning: non-integer x = 4.218508
#> Warning: non-integer x = 10.495536
#> Warning: non-integer x = 7.920315
#> Warning: non-integer x = 7.941519
#> Warning: non-integer x = 3.402260
#> Warning: non-integer x = 8.906376
#> Warning: non-integer x = 14.538431
#> Warning: non-integer x = 2.846374
#> Warning: non-integer x = 70.464690
#> Warning: non-integer x = 3.900966
#> Warning: non-integer x = 14.095281
#> Warning: non-integer x = 4.938606
#> Warning: non-integer x = 5.136688
#> Warning: non-integer x = 3.085738
#> Warning: non-integer x = 2.460739
#> Warning: non-integer x = 3.777428
#> Warning: non-integer x = 2.640457
#> Warning: non-integer x = 6.435514
#> Warning: non-integer x = 3.771071
#> Warning: non-integer x = 0.263521
#> Warning: non-integer x = 35.372469
#> Warning: non-integer x = 5.778682
#> Warning: non-integer x = 15.928314
#> Warning: non-integer x = 20.263273
#> Warning: non-integer x = 3.762466
#> Warning: non-integer x = 1.243122
#> Warning: non-integer x = 2.949330
#> Warning: non-integer x = 2.514230
#> Warning: non-integer x = 1.264655
#> Warning: non-integer x = 12.630694
#> Warning: non-integer x = 7.957209
#> Warning: non-integer x = 7.864558
#> Warning: non-integer x = 3.509229
#> Warning: non-integer x = 8.803446
#> Warning: non-integer x = 0.450146
#> Warning: non-integer x = 34.200176
#> Warning: non-integer x = 9.751348
#> Warning: non-integer x = 3.385859
#> Warning: non-integer x = 2.384549
#> Warning: non-integer x = 1.807402
#> Warning: non-integer x = 21.375750
#> Warning: non-integer x = 8.370194
#> Warning: non-integer x = 13.865198
#> Warning: non-integer x = 21.168622
#> Warning: non-integer x = 83.320423
#> Warning: non-integer x = 187.896937
#> Warning: non-integer x = 1.366010
#> Warning: non-integer x = 13.830532
#> Warning: non-integer x = 2.453460
#> Warning: non-integer x = 8.274733
#> Warning: non-integer x = 2.728692
#> Warning: non-integer x = 0.707686
#> Warning: non-integer x = 3.557212
#> Warning: non-integer x = 5.082516
#> Warning: non-integer x = 28.605768
#> Warning: non-integer x = 2.395692
#> Warning: non-integer x = 19.830294
#> Warning: non-integer x = 7.826664
#> Warning: non-integer x = 1.131855
#> Warning: non-integer x = 1.217328
#> Warning: non-integer x = 0.739241
#> Warning: non-integer x = 19.699065
#> Warning: non-integer x = 66.409654
#> Warning: non-integer x = 8.106118
#> Warning: non-integer x = 37.576243
#> Warning: non-integer x = 38.488982
#> Warning: non-integer x = 38.878142
#> Warning: non-integer x = 4.378927
#> Warning: non-integer x = 4.101608
#> Warning: non-integer x = 2.115093
#> Warning: non-integer x = 8.145076
#> Warning: non-integer x = 3.803004
#> Warning: non-integer x = 9.635552
#> Warning: non-integer x = 7.892665
#> Warning: non-integer x = 12.864391
#> Warning: non-integer x = 20.745032
#> Warning: non-integer x = 4.600761
#> Warning: non-integer x = 18.995231
#> Warning: non-integer x = 14.058035
#> Warning: non-integer x = 7.229531
#> Warning: non-integer x = 27.901604
#> Warning: non-integer x = 4.245454
#> Warning: non-integer x = 4.215955
#> Warning: non-integer x = 2.854487
#> Warning: non-integer x = 22.863914
#> Warning: non-integer x = 38.532143
#> Warning: non-integer x = 97.452339
#> Warning: non-integer x = 7.666345
#> Warning: non-integer x = 1.619197
#> Warning: non-integer x = 6.416097
#> Warning: non-integer x = 4.506251
#> Warning: non-integer x = 3.567235
#> Warning: non-integer x = 1.061538
#> Warning: non-integer x = 4.115735
#> Warning: non-integer x = 0.405974
#> Warning: non-integer x = 8.352262
#> Warning: non-integer x = 25.739698
#> Warning: non-integer x = 14.153384
#> Warning: non-integer x = 3.158104
#> Warning: non-integer x = 2.726021
#> Warning: non-integer x = 111.171749
#> Warning: non-integer x = 25.637082
#> Warning: non-integer x = 6.537268
#> Warning: non-integer x = 2.644332
#> Warning: non-integer x = 17.023968
#> Warning: non-integer x = 47.888276
#> Warning: non-integer x = 23.829878
#> Warning: non-integer x = 4.763924
#> Warning: non-integer x = 0.976622
#> Warning: non-integer x = 3.080768
#> Warning: non-integer x = 1.010777
#> Warning: non-integer x = 6.416768
#> Warning: non-integer x = 96.714708
#> Warning: non-integer x = 5.307958
#> Warning: non-integer x = 1.195666
#> Warning: non-integer x = 4.639206
#> Warning: non-integer x = 87.181905
#> Warning: non-integer x = 5.357125
#> Warning: non-integer x = 3.792602
#> Warning: non-integer x = 7.071621
#> Warning: non-integer x = 14.013590
#> Warning: non-integer x = 34.637320
#> Warning: non-integer x = 25.645706
#> Warning: non-integer x = 2.349022
#> Warning: non-integer x = 35.573235
#> Warning: non-integer x = 6.012070
#> Warning: non-integer x = 43.060214
#> Warning: non-integer x = 0.480814
#> Warning: non-integer x = 0.943326
#> Warning: non-integer x = 24.485150
#> Warning: non-integer x = 8.984689
#> Warning: non-integer x = 29.440840
#> Warning: non-integer x = 2.441302
#> Warning: non-integer x = 4.570989
#> Warning: non-integer x = 8.784168
#> Warning: non-integer x = 2.339246
#> Warning: non-integer x = 14.237742
#> Warning: non-integer x = 5.641441
#> Warning: non-integer x = 2.542006
#> Warning: non-integer x = 4.241124
#> Warning: non-integer x = 41.312068
#> Warning: non-integer x = 11.040607
#> Warning: non-integer x = 32.026013
#> Warning: non-integer x = 10.950843
#> Warning: non-integer x = 4.319451
#> Warning: non-integer x = 25.996215
#> Warning: non-integer x = 1.463281
#> Warning: non-integer x = 2.910473
#> Warning: non-integer x = 2.914243
#> Warning: non-integer x = 83.696177
#> Warning: non-integer x = 10.154144
#> Warning: non-integer x = 1.229850
#> Warning: non-integer x = 14.879562
#> Warning: non-integer x = 25.665025
#> Warning: non-integer x = 2.024326
#> Warning: non-integer x = 1.391356
#> Warning: non-integer x = 0.464722
#> Warning: non-integer x = 13.696867
#> Warning: non-integer x = 8.147145
#> Warning: non-integer x = 0.745391
#> Warning: non-integer x = 1.868655
#> Warning: non-integer x = 1.307909
#> Warning: non-integer x = 5.337876
#> Warning: non-integer x = 17.236352
#> Warning: non-integer x = 9.341150
#> Warning: non-integer x = 12.446037
#> Warning: non-integer x = 14.678505
#> Warning: non-integer x = 9.033872
#> Warning: non-integer x = 20.253523
#> Warning: non-integer x = 19.524018
#> Warning: non-integer x = 6.843716
#> Warning: non-integer x = 42.155084
#> Warning: non-integer x = 1.945254
#> Warning: non-integer x = 3.637671
#> Warning: non-integer x = 1.497053
#> Warning: non-integer x = 0.293471
#> Warning: non-integer x = 34.721934
#> Warning: non-integer x = 10.536690
#> Warning: non-integer x = 20.455553
#> Warning: non-integer x = 1.621222
#> Warning: non-integer x = 1.531133
#> Warning: non-integer x = 9.969938
#> Warning: non-integer x = 6.600093
#> Warning: non-integer x = 4.708607
#> Warning: non-integer x = 9.051699
#> Warning: non-integer x = 1.804265
#> Warning: non-integer x = 0.125728
#> Warning: non-integer x = 19.291960
#> Warning: non-integer x = 5.222794
#> Warning: non-integer x = 40.859828
#> Warning: non-integer x = 0.591783
#> Warning: non-integer x = 7.336213
#> Warning: non-integer x = 66.456361
#> Warning: non-integer x = 1.858597
#> Warning: non-integer x = 124.850817
#> Warning: non-integer x = 122.818793
#> Warning: non-integer x = 6.206763
#> Warning: non-integer x = 1.675532
#> Warning: non-integer x = 45.049534
#> Warning: non-integer x = 13.630777
#> Warning: non-integer x = 13.708486
#> Warning: non-integer x = 30.655234
#> Warning: non-integer x = 21.480280
#> $est
#> [1] 6.1703531 0.1891561
#> 
#> $se
#> (Intercept)          X1 
#>  0.87236932  0.08123126 
#>