PPI Logistic Regression

Helper function for PPI logistic regression

Usage

ppi_logistic(X_l, Y_l, f_l, X_u, f_u, opts = NULL)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

opts

(list, optional): Options to pass to the optimizer. See ?optim for details.

Value

(list): A list containing the following:

est

(vector): vector of PPI logistic regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

rectifier_est

(vector): vector of the rectifier logistic regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

Prediction Powered Inference (Angelopoulos et al., 2023) https://www.science.org/doi/10.1126/science.adi6000

Examples

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_logistic(X_l, Y_l, f_l, X_u, f_u)
#> $est
#>           [,1]
#> [1,] 0.6805478
#> [2,] 0.5519743
#> 
#> $se
#> [1] 0.1933250 0.1813317
#> 
#> $rectifier_est
#>            [,1]
#> [1,] -0.1088673
#> [2,]  0.2809123
#> 
#> $var_u
#>             [,1]        [,2]
#> [1,]  0.19055011 -0.06817244
#> [2,] -0.06817244  0.12608638
#> 
#> $var_l
#>             [,1]        [,2]
#> [1,]  0.33644370 -0.08666132
#> [2,] -0.08666132  0.21637946
#> 
#> $grads
#>               [,1]         [,2]
#>   [1,] -0.36377075  0.080081219
#>   [2,] -0.52821759  0.759384713
#>   [3,] -0.51192840  0.675434226
#>   [4,] -0.41875602  0.267550287
#>   [5,]  0.75486543  0.607454723
#>   [6,] -0.35269913  0.046870260
#>   [7,]  0.69664344  0.190345232
#>   [8,]  0.43852952 -0.737018027
#>   [9,]  0.66411848  0.001388748
#>  [10,] -0.17994403 -0.272594937
#>  [11,] -0.20062557 -0.255097573
#>  [12,]  0.71340815  0.299137145
#>  [13,] -0.28532346 -0.122849220
#>  [14,] -0.34296519  0.018913537
#>  [15,] -0.32649387 -0.025752876
#>  [16,]  0.39429353 -0.792800961
#>  [17,] -0.28557832 -0.122312488
#>  [18,] -0.23519227 -0.212481082
#>  [19,] -0.42699596  0.298933995
#>  [20,] -0.51175109  0.674542359
#>  [21,] -0.26074938 -0.170784369
#>  [22,]  0.57154374 -0.406307263
#>  [23,]  0.46817465 -0.685349310
#>  [24,] -0.21813539 -0.235541000
#>  [25,] -0.25865473 -0.174517862
#>  [26,] -0.18358806 -0.269965760
#>  [27,] -0.22516575 -0.226507915
#>  [28,] -0.29679906 -0.097883456
#>  [29,] -0.13030278 -0.287467089
#>  [30,] -0.33236048 -0.010224410
#>  [31,] -0.49892806  0.611269602
#>  [32,] -0.30795648 -0.072050959
#>  [33,] -0.30609290 -0.076472220
#>  [34,]  0.68795966  0.137162000
#>  [35,]  0.79090007  0.931086753
#>  [36,]  0.81031062  1.132552715
#>  [37,] -0.51720263  0.702177877
#>  [38,] -0.59087526  1.122000849
#>  [39,]  0.61728041 -0.226488748
#>  [40,] -0.40975415  0.234258172
#>  [41,] -0.24463680 -0.198056974
#>  [42,]  0.64872438 -0.078874292
#>  [43,]  0.74188281  0.504328728
#>  [44,] -0.16950267 -0.279019357
#>  [45,]  0.48924556 -0.641342343
#>  [46,] -0.14654642 -0.287106581
#>  [47,]  0.72927300  0.410088993
#>  [48,] -0.44427025  0.367585359
#>  [49,] -0.24727823 -0.193814989
#>  [50,] -0.30739934 -0.073377233
#>  [51,]  0.67439183  0.058118564
#>  [52,]  0.68028205  0.091842897
#>  [53,]  0.47228039 -0.677257708
#>  [54,] -0.34873086  0.035332943
#>  [55,] -0.24403900 -0.199004462
#>  [56,] -0.28858183 -0.115926347
#>  [57,]  0.56693038 -0.422351789
#>  [58,]  0.58887211 -0.342717279
#>  [59,]  0.80200360  1.043694093
#>  [60,] -0.40357891  0.212015227
#>  [61,]  0.76398181  0.683860975
#>  [62,] -0.43298286  0.322286359
#>  [63,]  0.57591478 -0.390766499
#>  [64,] -0.49515270  0.593096769
#>  [65,] -0.28549046 -0.122497611
#>  [66,] -0.35324347  0.048467936
#>  [67,] -0.12114288 -0.285555798
#>  [68,]  0.50385986 -0.607131995
#>  [69,] -0.29852457 -0.093988659
#>  [70,] -0.40807556  0.228164186
#>  [71,]  0.63284119 -0.156065461
#>  [72,] -0.28555743 -0.122356509
#>  [73,]  0.64017534 -0.121108443
#>  [74,] -0.20589920 -0.249655385
#>  [75,] -0.30513934 -0.078717992
#>  [76,] -0.38947494  0.163014163
#>  [77,]  0.66988052  0.032884997
#>  [78,] -0.44962026  0.389642377
#>  [79,]  0.59137980 -0.333068916
#>  [80,] -0.16301012 -0.282165889
#>  [81,] -0.39483947  0.181358015
#>  [82,] -0.33381969 -0.006296628
#>  [83,]  0.71336527  0.298848151
#>  [84,]  0.56491439 -0.429248937
#>  [85,]  0.60225440 -0.289881689
#>  [86,]  0.49955970 -0.617518050
#>  [87,] -0.30988902 -0.067421045
#>  [88,] -0.44711136  0.379251444
#>  [89,] -0.37301731  0.108973230
#>  [90,] -0.20628688 -0.249239946
#>  [91,]  0.62973384 -0.170529679
#>  [92,]  0.68515695  0.120436622
#>  [93,]  0.58231916 -0.367391654
#>  [94,]  0.58195596 -0.368736717
#>  [95,] -0.34298284  0.018963186
#>  [96,] -0.47381656  0.494198374
#>  [97,] -0.32300679 -0.034783416
#>  [98,]  0.63316213 -0.154559860
#>  [99,]  0.67647271  0.069930940
#> [100,]  0.48519779 -0.650277905
#> [101,]  0.45601329 -0.707969377
#> [102,]  0.84779245  1.592515498
#> [103,]  0.61108591 -0.253157163
#> [104,]  0.60079202 -0.295818433
#> [105,] -0.15133980 -0.286128720
#> [106,]  0.42407497 -0.758004034
#> [107,]  0.49555127 -0.626958287
#> [108,]  0.37094830 -0.812292667
#> [109,]  0.59676184 -0.311970775
#> [110,] -0.46943044  0.474654450
#> [111,]  0.46410149 -0.693148811
#> [112,] -0.19524963 -0.260240747
#> [113,]  0.55660506 -0.456954466
#> [114,] -0.22505753 -0.226652010
#> [115,] -0.17923276 -0.273085000
#> [116,] -0.16096185 -0.283020543
#> [117,] -0.18570390 -0.268348825
#> [118,]  0.65315374 -0.056345318
#> [119,] -0.23441993 -0.213608922
#> [120,] -0.39582622  0.184771404
#> [121,] -0.13122361 -0.287572601
#> [122,] -0.20690294 -0.248575454
#> [123,] -0.44414716  0.367082329
#> [124,] -0.40612262  0.221119107
#> [125,] -0.46393497  0.450539433
#> [126,]  0.41227301 -0.773140933
#> [127,] -0.13108094 -0.287557270
#> [128,]  0.67601198  0.067306040
#> [129,] -0.37913967  0.128684320
#> [130,] -0.18898462 -0.265711574
#> [131,] -0.33773240  0.004363802
#> [132,] -0.47058458  0.479771458
#> [133,]  0.43409961 -0.743738372
#> [134,] -0.47614055  0.504661297
#> [135,]  0.57497162 -0.394147933
#> [136,] -0.32347159 -0.033588310
#> [137,] -0.51952876  0.714105406
#> [138,] -0.37165190  0.104640500
#> [139,] -0.30555589 -0.077738341
#> [140,] -0.41007238  0.235417495
#> [141,] -0.19731981 -0.258308912
#> [142,]  0.37584521 -0.808764507
#> [143,] -0.17909685 -0.273177776
#> [144,] -0.28793035 -0.117321061
#> [145,] -0.26328440 -0.166191357
#> [146,]  0.64332897 -0.105717719
#> [147,] -0.37050099  0.101006276
#> [148,] -0.23163996 -0.217603277
#> [149,] -0.35506578  0.053842931
#> [150,] -0.35834160  0.063607613
#> [151,]  0.76352313  0.679934072
#> [152,] -0.37188489  0.105378175
#> [153,] -0.21434154 -0.240137233
#> [154,] -0.30129510 -0.087658274
#> [155,] -0.21503553 -0.239311146
#> [156,] -0.36901103  0.096325718
#> [157,]  0.71655080  0.320474590
#> [158,] -0.43630070  0.335428532
#> [159,] -0.38906998  0.161643994
#> [160,] -0.43086912  0.313988442
#> [161,]  0.74616360  0.537623868
#> [162,] -0.21429364 -0.240194001
#> [163,]  0.52974239 -0.538824566
#> [164,] -0.38281151  0.140728685
#> [165,]  0.77986928  0.825621825
#> [166,]  0.37831966 -0.806868906
#> [167,]  0.65642516 -0.039416006
#> [168,] -0.26131562 -0.169765522
#> [169,] -0.17976229 -0.272720874
#> [170,] -0.30125227 -0.087756865
#> [171,] -0.31077568 -0.065281465
#> [172,] -0.27606590 -0.141796908
#> [173,] -0.39294485  0.174838328
#> [174,]  0.43650853 -0.740115764
#> [175,]  0.36266236 -0.817588813
#> [176,] -0.15452775 -0.285266414
#> [177,]  0.51077952 -0.589851090
#> [178,] -0.29248650 -0.107456896
#> [179,] -0.34600863  0.027529911
#> [180,] -0.28953643 -0.113873215
#> [181,] -0.30338452 -0.082821611
#> [182,] -0.20289410 -0.252804409
#> [183,] -0.32452488 -0.030870275
#> [184,] -0.44971637  0.390042080
#> [185,]  0.66202033 -0.009879590
#> [186,]  0.60007622 -0.298709599
#> [187,] -0.21236802 -0.242450306
#> [188,] -0.49195636  0.577871069
#> [189,] -0.22126032 -0.231608226
#> [190,]  0.57624799 -0.389568178
#> [191,]  0.69548678  0.183140338
#> [192,]  0.59042580 -0.336752989
#> [193,]  0.61037974 -0.256149267
#> [194,] -0.34956078  0.037729902
#> [195,] -0.42783167  0.302165842
#> [196,] -0.22695470 -0.224103271
#> [197,]  0.54712606 -0.487165492
#> [198,] -0.34309941  0.019291137
#> [199,] -0.39679541  0.188135937
#> [200,] -0.18531951 -0.268647482
#> [201,] -0.50994412  0.665479903
#> [202,] -0.35984494  0.068132914
#> [203,]  0.52823032 -0.543094205
#> [204,]  0.80437372  1.068628200
#> [205,] -0.17409437 -0.276398772
#> [206,] -0.20694977 -0.248524727
#> [207,]  0.50141067 -0.613080414
#> [208,] -0.12790231 -0.287118561
#> [209,] -0.40745865  0.225933527
#> [210,] -0.08959589 -0.265883520
#> [211,]  0.66611611  0.012216169
#> [212,] -0.23460759 -0.213335603
#> [213,]  0.53215061 -0.531951432
#> [214,] -0.28456886 -0.124433663
#> [215,]  0.67317787  0.051278143
#> [216,] -0.48536794  0.546946224
#> [217,] -0.48453402  0.543075699
#> [218,] -0.29707686 -0.097258893
#> [219,]  0.63063053 -0.166376637
#> [220,] -0.29849486 -0.094056033
#> [221,] -0.22673096 -0.224406350
#> [222,] -0.27141995 -0.150901965
#> [223,] -0.30447423 -0.080277804
#> [224,] -0.23465456 -0.213267122
#> [225,] -0.32498363 -0.029682227
#> [226,] -0.30461679 -0.079943926
#> [227,] -0.25473104 -0.181360836
#> [228,]  0.65232511 -0.060594099
#> [229,] -0.31549101 -0.053740994
#> [230,] -0.36387633  0.080405182
#> [231,] -0.55817025  0.924550147
#> [232,] -0.22110391 -0.231808219
#> [233,] -0.36781991  0.092603607
#> [234,] -0.10348405 -0.277198445
#> [235,] -0.34864822  0.035094733
#> [236,] -0.22075132 -0.232257828
#> [237,] -0.24124808 -0.203366408
#> [238,]  0.51648765 -0.575062142
#> [239,] -0.18914911 -0.265575205
#> [240,] -0.37166289  0.104675261
#> [241,] -0.18109943 -0.271782732
#> [242,] -0.20035032 -0.255370877
#> [243,] -0.27418050 -0.145524538
#> [244,]  0.70789247  0.262422794
#> [245,]  0.72834577  0.403377187
#> [246,]  0.59879284 -0.303869018
#> [247,] -0.24812635 -0.192433796
#> [248,] -0.18977357 -0.265053915
#> [249,]  0.50214644 -0.611302613
#> [250,]  0.65024531 -0.071188860
#> [251,]  0.53553209 -0.522148200
#> [252,]  0.58906425 -0.341982062
#> [253,]  0.66307522 -0.004227397
#> [254,] -0.22205967 -0.230581026
#> [255,] -0.28594266 -0.121543790
#> [256,] -0.18470289 -0.269122037
#> [257,] -0.22289969 -0.229492271
#> [258,]  0.59236665 -0.329240452
#> [259,] -0.26281981 -0.167039217
#> [260,] -0.31052373 -0.065890416
#> [261,] -0.30349570 -0.082562758
#> [262,] -0.48414128  0.541256259
#> [263,]  0.48500405 -0.650699781
#> [264,]  0.54594980 -0.490812501
#> [265,] -0.42540334  0.292799980
#> [266,]  0.70440704  0.239694808
#> [267,] -0.38957605  0.163356581
#> [268,]  0.75779686  0.631651334
#> [269,] -0.38608458  0.151606199
#> [270,]  0.74625829  0.538368055
#> [271,] -0.19688844 -0.258716507
#> [272,] -0.23694319 -0.209895178
#> [273,]  0.52055329 -0.564230881
#> [274,] -0.23560900 -0.211869276
#> [275,] -0.12289245 -0.286043753
#> [276,]  0.61799825 -0.223348933
#> [277,]  0.38671608 -0.799870496
#> [278,] -0.40821130  0.228655644
#> [279,] -0.27417126 -0.145542695
#> [280,]  0.68526330  0.121067451
#> [281,] -0.37410779  0.112450137
#> [282,] -0.25755331 -0.176458590
#> [283,]  0.65252691 -0.059560834
#> [284,]  0.62543080 -0.190225580
#> [285,] -0.21402473 -0.240512149
#> [286,] -0.25161013 -0.186663041
#> [287,]  0.58045375 -0.374274936
#> [288,] -0.40468758  0.215973066
#> [289,]  0.58355858 -0.362783948
#> [290,] -0.43418127  0.327016781
#> [291,]  0.55929315 -0.448118010
#> [292,] -0.19049981 -0.264440533
#> [293,] -0.30847165 -0.070821211
#> [294,]  0.60817786 -0.265416391
#> [295,]  0.77604908  0.790483548
#> [296,] -0.41249889  0.244299790
#> [297,]  0.59099171 -0.334569630
#> [298,]  0.68702575  0.131565337
#> [299,] -0.31413495 -0.057087791
#> [300,]  0.83655287  1.443200152
#> 
#> $grads_hat_unlabeled
#>               [,1]         [,2]
#>   [1,] -0.30556457 -0.077717907
#>   [2,] -0.35056455  0.040640209
#>   [3,]  0.44270053 -0.730455415
#>   [4,] -0.28136346 -0.131085086
#>   [5,] -0.25243478 -0.185274171
#>   [6,]  0.39304318 -0.794016937
#>   [7,] -0.52492912  0.742112252
#>   [8,] -0.30298805 -0.083743524
#>   [9,]  0.64013339 -0.121311701
#>  [10,]  0.59842399 -0.305346125
#>  [11,] -0.46174396  0.441039564
#>  [12,] -0.32802775 -0.021733450
#>  [13,] -0.39030384  0.165825045
#>  [14,] -0.36373274  0.079964629
#>  [15,] -0.35699419  0.059575270
#>  [16,] -0.32233462 -0.036507026
#>  [17,] -0.28472282 -0.124110953
#>  [18,] -0.27157479 -0.150602882
#>  [19,]  0.62521454 -0.191205321
#>  [20,]  0.51002606 -0.591767007
#>  [21,] -0.40757800  0.226364701
#>  [22,] -0.37265174  0.107810956
#>  [23,] -0.41114416  0.239331563
#>  [24,]  0.57844533 -0.381616947
#>  [25,] -0.16389773 -0.281774828
#>  [26,] -0.38402947  0.144760795
#>  [27,] -0.16462252 -0.281446263
#>  [28,] -0.32067350 -0.040742854
#>  [29,]  0.73185918  0.428963703
#>  [30,] -0.25629554 -0.178655900
#>  [31,]  0.69248434  0.164612938
#>  [32,] -0.42851236  0.304804903
#>  [33,] -0.25596400 -0.179231727
#>  [34,] -0.25531800 -0.180349687
#>  [35,]  0.41619949 -0.768302669
#>  [36,] -0.35212812  0.045198232
#>  [37,] -0.20913843 -0.246119927
#>  [38,]  0.47340300 -0.675004902
#>  [39,] -0.50380537  0.635052113
#>  [40,] -0.25105209 -0.187597920
#>  [41,] -0.38965744  0.163632284
#>  [42,] -0.51467536  0.689311175
#>  [43,] -0.21431781 -0.240165363
#>  [44,] -0.19012614 -0.264757095
#>  [45,] -0.22438952 -0.227537962
#>  [46,] -0.44016463  0.350915177
#>  [47,] -0.20357470 -0.252102298
#>  [48,] -0.34061221  0.012329633
#>  [49,] -0.39535771  0.183149220
#>  [50,] -0.19244426 -0.262760597
#>  [51,]  0.34822913 -0.824799034
#>  [52,] -0.33936058  0.008855059
#>  [53,]  0.64855736 -0.079715054
#>  [54,] -0.37328126  0.109813475
#>  [55,] -0.40085851  0.202369326
#>  [56,] -0.42936422  0.308116033
#>  [57,] -0.40154159  0.204782596
#>  [58,] -0.27020627 -0.153235846
#>  [59,]  0.57055258 -0.409785201
#>  [60,] -0.21576043 -0.238441241
#>  [61,] -0.39088500  0.167800951
#>  [62,] -0.40014098  0.199840669
#>  [63,] -0.46645721  0.461556270
#>  [64,] -0.48347844  0.538190423
#>  [65,] -0.36156209  0.073335796
#>  [66,] -0.17071530 -0.278358820
#>  [67,] -0.43648742  0.336172424
#>  [68,] -0.40437655  0.214861153
#>  [69,] -0.42506982  0.291519569
#>  [70,] -0.28911598 -0.114778898
#>  [71,] -0.22417248 -0.227824521
#>  [72,] -0.28285221 -0.128011734
#>  [73,]  0.64190253 -0.112706325
#>  [74,] -0.29945725 -0.091868143
#>  [75,]  0.67597527  0.067097112
#>  [76,] -0.46440635  0.452591791
#>  [77,]  0.64619712 -0.091529149
#>  [78,] -0.48349459  0.538265080
#>  [79,] -0.17511744 -0.275770887
#>  [80,] -0.42140876  0.277558578
#>  [81,]  0.56231940 -0.438025676
#>  [82,] -0.25462957 -0.181535185
#>  [83,] -0.08847693 -0.264774278
#>  [84,] -0.17083385 -0.278293023
#>  [85,]  0.45271945 -0.713752623
#>  [86,] -0.44851404  0.385050521
#>  [87,] -0.28986018 -0.113174344
#>  [88,] -0.43141192  0.316113794
#>  [89,]  0.77401869  0.772085725
#>  [90,] -0.37819937  0.125626836
#>  [91,] -0.43153651  0.316602180
#>  [92,] -0.23414966 -0.214001735
#>  [93,] -0.12906109 -0.287300118
#>  [94,] -0.17178187 -0.277759084
#>  [95,] -0.33317465 -0.008036094
#>  [96,] -0.28932027 -0.114339127
#>  [97,] -0.55577392  0.910802616
#>  [98,] -0.35335653  0.048800220
#>  [99,] -0.35038059  0.040105935
#> [100,] -0.34798807  0.033194784
#> [101,]  0.63736591 -0.134636225
#> [102,]  0.50996888 -0.591912042
#> [103,] -0.26117368 -0.170021291
#> [104,] -0.28869166 -0.115690716
#> [105,] -0.10525934 -0.278332164
#> [106,] -0.37350022  0.110511124
#> [107,] -0.35467363  0.052682845
#> [108,]  0.46650253 -0.688578560
#> [109,] -0.49797896  0.606681743
#> [110,] -0.07884772 -0.253918497
#> [111,] -0.30020441 -0.090161700
#> [112,] -0.25784652 -0.175943464
#> [113,]  0.45030094 -0.717906668
#> [114,] -0.24680080 -0.194588402
#> [115,] -0.47098439  0.481548341
#> [116,]  0.68085973  0.095198668
#> [117,] -0.40708488  0.224584347
#> [118,]  0.43544398 -0.741726013
#> [119,] -0.36702851  0.090140263
#> [120,] -0.40958538  0.233643847
#> [121,] -0.37751411  0.123405546
#> [122,] -0.19962314 -0.256087749
#> [123,] -0.33913306  0.008225503
#> [124,]  0.69060846  0.153164093
#> [125,] -0.22034918 -0.232768567
#> [126,] -0.18041889 -0.272263538
#> [127,] -0.17941233 -0.272961991
#> [128,] -0.22417355 -0.227823115
#> [129,] -0.32781528 -0.022291933
#> [130,] -0.25267049 -0.184875572
#> [131,]  0.59993569 -0.299276058
#> [132,] -0.22898812 -0.221318114
#> [133,] -0.26593635 -0.161299316
#> [134,]  0.63375562 -0.151769975
#> [135,]  0.69237340  0.163933158
#> [136,] -0.25832534 -0.175099878
#> [137,]  0.46142205 -0.698156477
#> [138,] -0.37767959  0.123941413
#> [139,] -0.36815278  0.093642021
#> [140,]  0.55449718 -0.463799902
#> [141,]  0.44624665 -0.724695740
#> [142,] -0.20125319 -0.254470401
#> [143,] -0.35882505  0.065059834
#> [144,] -0.45723674  0.421701128
#> [145,] -0.21024697 -0.244876583
#> [146,]  0.64748910 -0.085077729
#> [147,]  0.65293061 -0.057490968
#> [148,] -0.26627106 -0.160675550
#> [149,] -0.26075261 -0.170778554
#> [150,]  0.53531138 -0.522793505
#> [151,] -0.51844114  0.708518364
#> [152,] -0.28300119 -0.127702671
#> [153,] -0.30907875 -0.069367852
#> [154,]  0.70084217  0.216818635
#> [155,]  0.46650785 -0.688568345
#> [156,] -0.48071075  0.525455690
#> [157,]  0.47172904 -0.678357771
#> [158,] -0.30234267 -0.085240056
#> [159,] -0.32327404 -0.034096575
#> [160,]  0.75449883  0.604453011
#> [161,] -0.38149063  0.136376650
#> [162,] -0.27361124 -0.146641264
#> [163,] -0.44175237  0.357335617
#> [164,]  0.45968137 -0.701357576
#> [165,] -0.36632606  0.087960239
#> [166,] -0.11804068 -0.284544482
#> [167,] -0.17282961 -0.277152906
#> [168,] -0.16393552 -0.281757901
#> [169,] -0.38493426  0.147768099
#> [170,]  0.53571504 -0.521612689
#> [171,]  0.40553068 -0.780991762
#> [172,] -0.25197274 -0.186053401
#> [173,]  0.65883604 -0.026779969
#> [174,] -0.47722126  0.509552277
#> [175,]  0.66850516  0.025293001
#> [176,] -0.32369788 -0.033005492
#> [177,] -0.49051722  0.571063394
#> [178,] -0.42057202  0.274391968
#> [179,] -0.18942351 -0.265346845
#> [180,] -0.19948404 -0.256224020
#> [181,] -0.33391380 -0.006042428
#> [182,]  0.48105804 -0.659177762
#> [183,] -0.53080651  0.773100038
#> [184,] -0.27589423 -0.142138168
#> [185,] -0.18767744 -0.266781254
#> [186,] -0.25986072 -0.172375129
#> [187,] -0.30807840 -0.071760231
#> [188,] -0.39563848  0.184121025
#> [189,] -0.31641241 -0.051454086
#> [190,]  0.69158976  0.159141037
#> [191,] -0.35710160  0.059895876
#> [192,]  0.60272495 -0.287962789
#> [193,] -0.35144772  0.043211098
#> [194,] -0.38462179  0.146728383
#> [195,]  0.59418772 -0.322128468
#> [196,] -0.58650671  1.094534792
#> [197,] -0.24614228 -0.195650341
#> [198,] -0.35548464  0.055084142
#> [199,] -0.26190964 -0.168692309
#> [200,] -0.24070344 -0.204205800
#> [201,]  0.42105681 -0.762044867
#> [202,] -0.25322272 -0.183938961
#> [203,] -0.44466329  0.369192886
#> [204,] -0.21443554 -0.240025723
#> [205,]  0.66648112  0.014205039
#> [206,]  0.48735093 -0.645553631
#> [207,] -0.31381121 -0.057883457
#> [208,] -0.40641486  0.222170270
#> [209,]  0.71747077  0.326779408
#> [210,]  0.60854826 -0.263864071
#> [211,]  0.68911992  0.144147911
#> [212,] -0.25667097 -0.178002138
#> [213,] -0.46523228  0.456195118
#> [214,] -0.52991300  0.768354785
#> [215,] -0.29013145 -0.112587750
#> [216,] -0.19658728 -0.258999488
#> [217,]  0.73233857  0.432487566
#> [218,] -0.29118636 -0.110297941
#> [219,] -0.21849431 -0.235096025
#> [220,] -0.19014220 -0.264743531
#> [221,]  0.69813573  0.199696626
#> [222,]  0.66120011 -0.014255904
#> [223,]  0.58191427 -0.368890941
#> [224,]  0.51156842 -0.587836049
#> [225,]  0.38123377 -0.804539193
#> [226,]  0.64190957 -0.112671940
#> [227,]  0.62952978 -0.171472418
#> [228,]  0.61100137 -0.253515892
#> [229,]  0.68073076  0.094448712
#> [230,] -0.33167602 -0.012057818
#> [231,]  0.55912679 -0.448668374
#> [232,] -0.29390021 -0.104343930
#> [233,]  0.67655741  0.070414090
#> [234,]  0.59524118 -0.317986337
#> [235,]  0.52600304 -0.549319241
#> [236,] -0.32932704 -0.018306248
#> [237,] -0.17607529 -0.275168606
#> [238,] -0.26693888 -0.159426767
#> [239,]  0.60009893 -0.298618023
#> [240,] -0.16133838 -0.282868451
#> [241,]  0.54231683 -0.501936240
#> [242,]  0.51210441 -0.586461749
#> [243,] -0.29950089 -0.091768677
#> [244,] -0.49551395  0.594826788
#> [245,] -0.28413071 -0.125350387
#> [246,] -0.45308650  0.404136028
#> [247,] -0.30772805 -0.072595202
#> [248,]  0.72494892  0.379035327
#> [249,] -0.56355191  0.955773215
#> [250,]  0.43151195 -0.747545532
#> [251,] -0.40362307  0.212172589
#> [252,] -0.21595086 -0.238211533
#> [253,] -0.27152312 -0.150702728
#> [254,] -0.48385273  0.539920884
#> [255,]  0.75786090  0.632183775
#> [256,]  0.65377795 -0.053134264
#> [257,] -0.22413931 -0.227868255
#> [258,]  0.58776145 -0.346953905
#> [259,] -0.28045241 -0.132952197
#> [260,] -0.49091563  0.572945096
#> [261,]  0.68894185  0.143073354
#> [262,]  0.45256221 -0.714025075
#> [263,] -0.48597084  0.549750547
#> [264,] -0.23239030 -0.216535229
#> [265,]  0.61392398 -0.241033116
#> [266,]  0.68183847  0.100904102
#> [267,]  0.68851438  0.140497370
#> [268,]  0.71386183  0.302198461
#> [269,] -0.31458360 -0.055983003
#> [270,] -0.39946781  0.197474215
#> [271,] -0.46954481  0.475160696
#> [272,]  0.41262109 -0.772719956
#> [273,] -0.37845537  0.126458149
#> [274,] -0.30321474 -0.083216639
#> [275,]  0.61548736 -0.234286452
#> [276,] -0.39900182  0.195839439
#> [277,]  0.53871204 -0.512765619
#> [278,] -0.40292291  0.209680657
#> [279,]  0.59979508 -0.299842459
#> [280,] -0.41331768  0.247313883
#> [281,] -0.24090770 -0.203891446
#> [282,] -0.21587693 -0.238300773
#> [283,] -0.37482825  0.114755309
#> [284,]  0.52070412 -0.563824256
#> [285,] -0.22407687 -0.227950550
#> [286,] -0.17682921 -0.274684775
#> [287,]  0.69206810  0.162064160
#> [288,] -0.21821599 -0.235441235
#> [289,] -0.30301053 -0.083691300
#> [290,]  0.57120907 -0.407483517
#> [291,] -0.53679451  0.805221886
#> [292,]  0.67167280  0.042848867
#> [293,] -0.45201616  0.399643438
#> [294,]  0.44633538 -0.724549490
#> [295,]  0.55028715 -0.477253635
#> [296,]  0.57075669 -0.409070371
#> [297,]  0.64541083 -0.095437095
#> [298,] -0.28259738 -0.128539763
#> [299,] -0.29115839 -0.110358839
#> [300,] -0.30438835 -0.080478808
#> 
#> $grads_hat
#>               [,1]          [,2]
#>   [1,] -0.36377075  0.0800812185
#>   [2,]  0.47178241 -0.6782514646
#>   [3,] -0.51192840  0.6754342257
#>   [4,] -0.41875602  0.2675502868
#>   [5,] -0.24513457 -0.1972645007
#>   [6,] -0.35269913  0.0468702596
#>   [7,] -0.30335656 -0.0828866984
#>   [8,]  0.43852952 -0.7370180275
#>   [9,] -0.33588152 -0.0007023667
#>  [10,] -0.17994403 -0.2725949366
#>  [11,] -0.20062557 -0.2550975729
#>  [12,] -0.28659185 -0.1201700153
#>  [13,] -0.28532346 -0.1228492198
#>  [14,] -0.34296519  0.0189135372
#>  [15,]  0.67350613  0.0531241835
#>  [16,]  0.39429353 -0.7928009612
#>  [17,] -0.28557832 -0.1223124881
#>  [18,] -0.23519227 -0.2124810822
#>  [19,] -0.42699596  0.2989339952
#>  [20,] -0.51175109  0.6745423586
#>  [21,] -0.26074938 -0.1707843690
#>  [22,] -0.42845626  0.3045871670
#>  [23,]  0.46817465 -0.6853493097
#>  [24,] -0.21813539 -0.2355410000
#>  [25,] -0.25865473 -0.1745178616
#>  [26,] -0.18358806 -0.2699657597
#>  [27,] -0.22516575 -0.2265079148
#>  [28,]  0.70320094  0.2319135966
#>  [29,] -0.13030278 -0.2874670893
#>  [30,]  0.66763952  0.0205386040
#>  [31,]  0.50107194 -0.6138962150
#>  [32,] -0.30795648 -0.0720509586
#>  [33,]  0.69390710  0.1733611455
#>  [34,]  0.68795966  0.1371619998
#>  [35,] -0.20909993 -0.2461628015
#>  [36,] -0.18968938 -0.2651245294
#>  [37,]  0.48279737 -0.6554677393
#>  [38,]  0.40912474 -0.7768785474
#>  [39,] -0.38271959  0.1404251307
#>  [40,] -0.40975415  0.2342581722
#>  [41,] -0.24463680 -0.1980569737
#>  [42,]  0.64872438 -0.0788742920
#>  [43,]  0.74188281  0.5043287284
#>  [44,] -0.16950267 -0.2790193572
#>  [45,]  0.48924556 -0.6413423432
#>  [46,] -0.14654642 -0.2871065809
#>  [47,] -0.27072700 -0.1522367669
#>  [48,] -0.44427025  0.3675853589
#>  [49,] -0.24727823 -0.1938149894
#>  [50,] -0.30739934 -0.0733772329
#>  [51,]  0.67439183  0.0581185641
#>  [52,]  0.68028205  0.0918428973
#>  [53,]  0.47228039 -0.6772577076
#>  [54,]  0.65126914 -0.0659857152
#>  [55,] -0.24403900 -0.1990044623
#>  [56,] -0.28858183 -0.1159263474
#>  [57,] -0.43306962  0.3226281988
#>  [58,] -0.41112789  0.2392720462
#>  [59,] -0.19799640 -0.2576642779
#>  [60,]  0.59642109 -0.3133224971
#>  [61,] -0.23601819 -0.2112663243
#>  [62,]  0.56701714 -0.4220534084
#>  [63,]  0.57591478 -0.3907664989
#>  [64,]  0.50484730 -0.6047090174
#>  [65,] -0.28549046 -0.1224976110
#>  [66,] -0.35324347  0.0484679356
#>  [67,] -0.12114288 -0.2855557984
#>  [68,]  0.50385986 -0.6071319953
#>  [69,]  0.70147543  0.2208553007
#>  [70,] -0.40807556  0.2281641861
#>  [71,]  0.63284119 -0.1560654606
#>  [72,] -0.28555743 -0.1223565092
#>  [73,] -0.35982466  0.0680716684
#>  [74,] -0.20589920 -0.2496553849
#>  [75,] -0.30513934 -0.0787179920
#>  [76,] -0.38947494  0.1630141627
#>  [77,] -0.33011948 -0.0162058426
#>  [78,] -0.44962026  0.3896423775
#>  [79,]  0.59137980 -0.3330689164
#>  [80,] -0.16301012 -0.2821658892
#>  [81,] -0.39483947  0.1813580146
#>  [82,] -0.33381969 -0.0062966277
#>  [83,]  0.71336527  0.2988481514
#>  [84,] -0.43508561  0.3305988266
#>  [85,]  0.60225440 -0.2898816888
#>  [86,]  0.49955970 -0.6175180504
#>  [87,] -0.30988902 -0.0674210446
#>  [88,] -0.44711136  0.3792514439
#>  [89,]  0.62698269 -0.1831666478
#>  [90,] -0.20628688 -0.2492399455
#>  [91,] -0.37026616  0.1002667564
#>  [92,] -0.31484305 -0.0553429896
#>  [93,] -0.41768084  0.2635195015
#>  [94,]  0.58195596 -0.3687367167
#>  [95,]  0.65701716 -0.0363258361
#>  [96,] -0.47381656  0.4941983745
#>  [97,]  0.67699321  0.0729029158
#>  [98,]  0.63316213 -0.1545598597
#>  [99,] -0.32352729 -0.0334449078
#> [100,] -0.51480221  0.6899547130
#> [101,]  0.45601329 -0.7079693769
#> [102,] -0.15220755 -0.2859106399
#> [103,] -0.38891409  0.1611170972
#> [104,]  0.60079202 -0.2958184334
#> [105,] -0.15133980 -0.2861287195
#> [106,]  0.42407497 -0.7580040344
#> [107,]  0.49555127 -0.6269582865
#> [108,]  0.37094830 -0.8122926667
#> [109,] -0.40323816  0.2108018856
#> [110,] -0.46943044  0.4746544495
#> [111,]  0.46410149 -0.6931488110
#> [112,] -0.19524963 -0.2602407471
#> [113,] -0.44339494  0.3640126751
#> [114,] -0.22505753 -0.2266520099
#> [115,] -0.17923276 -0.2730850002
#> [116,] -0.16096185 -0.2830205428
#> [117,] -0.18570390 -0.2683488253
#> [118,]  0.65315374 -0.0563453179
#> [119,] -0.23441993 -0.2136089224
#> [120,] -0.39582622  0.1847714038
#> [121,] -0.13122361 -0.2875726012
#> [122,] -0.20690294 -0.2485754540
#> [123,] -0.44414716  0.3670823293
#> [124,] -0.40612262  0.2211191068
#> [125,] -0.46393497  0.4505394333
#> [126,] -0.58772699  1.1021720722
#> [127,] -0.13108094 -0.2875572704
#> [128,] -0.32398802 -0.0322573433
#> [129,]  0.62086033 -0.2107270607
#> [130,] -0.18898462 -0.2657115739
#> [131,]  0.66226760 -0.0085570836
#> [132,] -0.47058458  0.4797714575
#> [133,]  0.43409961 -0.7437383721
#> [134,]  0.52385945 -0.5552385535
#> [135,] -0.42502838  0.2913605636
#> [136,] -0.32347159 -0.0335883095
#> [137,]  0.48047124 -0.6604198550
#> [138,]  0.62834810 -0.1769146306
#> [139,] -0.30555589 -0.0777383409
#> [140,] -0.41007238  0.2354174952
#> [141,] -0.19731981 -0.2583089115
#> [142,]  0.37584521 -0.8087645067
#> [143,] -0.17909685 -0.2731777762
#> [144,] -0.28793035 -0.1173210613
#> [145,]  0.73671560  0.4650323529
#> [146,] -0.35667103  0.0586114558
#> [147,]  0.62949901 -0.1716145211
#> [148,] -0.23163996 -0.2176032774
#> [149,]  0.64493422 -0.0977991993
#> [150,] -0.35834160  0.0636076127
#> [151,] -0.23647687 -0.2105878271
#> [152,]  0.62811511 -0.1779841741
#> [153,] -0.21434154 -0.2401372331
#> [154,] -0.30129510 -0.0876582736
#> [155,] -0.21503553 -0.2393111457
#> [156,]  0.63098897 -0.1647117832
#> [157,] -0.28344920 -0.1267715656
#> [158,]  0.56369930 -0.4333727417
#> [159,] -0.38906998  0.1616439937
#> [160,]  0.56913088 -0.4147443233
#> [161,] -0.25383640 -0.1828935431
#> [162,] -0.21429364 -0.2401940007
#> [163,] -0.47025761  0.4783199420
#> [164,]  0.61718849 -0.2268900530
#> [165,] -0.22013072 -0.2330451111
#> [166,]  0.37831966 -0.8068689060
#> [167,] -0.34357484  0.0206304521
#> [168,] -0.26131562 -0.1697655223
#> [169,] -0.17976229 -0.2727208742
#> [170,] -0.30125227 -0.0877568652
#> [171,] -0.31077568 -0.0652814650
#> [172,] -0.27606590 -0.1417969079
#> [173,]  0.60705515 -0.2701053480
#> [174,]  0.43650853 -0.7401157636
#> [175,]  0.36266236 -0.8175888130
#> [176,] -0.15452775 -0.2852664144
#> [177,]  0.51077952 -0.5898510900
#> [178,] -0.29248650 -0.1074568961
#> [179,] -0.34600863  0.0275299110
#> [180,]  0.71046357  0.2794217308
#> [181,] -0.30338452 -0.0828216108
#> [182,] -0.20289410 -0.2528044092
#> [183,] -0.32452488 -0.0308702747
#> [184,]  0.55028363 -0.4772647514
#> [185,] -0.33797967  0.0050438035
#> [186,]  0.60007622 -0.2987095991
#> [187,] -0.21236802 -0.2424503059
#> [188,]  0.50804364 -0.5967678174
#> [189,] -0.22126032 -0.2316082264
#> [190,]  0.57624799 -0.3895681776
#> [191,]  0.69548678  0.1831403378
#> [192,] -0.40957420  0.2336031662
#> [193,]  0.61037974 -0.2561492666
#> [194,] -0.34956078  0.0377299022
#> [195,]  0.57216833 -0.4041068961
#> [196,] -0.22695470 -0.2241032713
#> [197,]  0.54712606 -0.4871654920
#> [198,] -0.34309941  0.0192911374
#> [199,] -0.39679541  0.1881359372
#> [200,] -0.18531951 -0.2686474823
#> [201,] -0.50994412  0.6654799032
#> [202,]  0.64015506 -0.1212067328
#> [203,] -0.47176968  0.4850448162
#> [204,]  0.80437372  1.0686282003
#> [205,] -0.17409437 -0.2763987717
#> [206,] -0.20694977 -0.2485247265
#> [207,]  0.50141067 -0.6130804138
#> [208,] -0.12790231 -0.2871185608
#> [209,] -0.40745865  0.2259335268
#> [210,] -0.08959589 -0.2658835200
#> [211,]  0.66611611  0.0122161686
#> [212,] -0.23460759 -0.2133356029
#> [213,] -0.46784939  0.4676742815
#> [214,] -0.28456886 -0.1244336631
#> [215,]  0.67317787  0.0512781432
#> [216,]  0.51463206 -0.5799230524
#> [217,]  0.51546598 -0.5777448786
#> [218,] -0.29707686 -0.0972588927
#> [219,]  0.63063053 -0.1663766369
#> [220,] -0.29849486 -0.0940560331
#> [221,] -0.22673096 -0.2244063502
#> [222,] -0.27141995 -0.1509019648
#> [223,]  0.69552577  0.1833826196
#> [224,] -0.23465456 -0.2132671218
#> [225,]  0.67501637  0.0616523045
#> [226,] -0.30461679 -0.0799439258
#> [227,] -0.25473104 -0.1813608355
#> [228,]  0.65232511 -0.0605940988
#> [229,] -0.31549101 -0.0537409942
#> [230,] -0.36387633  0.0804051818
#> [231,]  0.44182975 -0.7318443801
#> [232,] -0.22110391 -0.2318082193
#> [233,] -0.36781991  0.0926036066
#> [234,] -0.10348405 -0.2771984452
#> [235,] -0.34864822  0.0350947335
#> [236,] -0.22075132 -0.2322578279
#> [237,] -0.24124808 -0.2033664076
#> [238,]  0.51648765 -0.5750621418
#> [239,] -0.18914911 -0.2655752049
#> [240,]  0.62833711 -0.1769650768
#> [241,] -0.18109943 -0.2717827319
#> [242,] -0.20035032 -0.2553708772
#> [243,] -0.27418050 -0.1455245381
#> [244,]  0.70789247  0.2624227935
#> [245,]  0.72834577  0.4033771873
#> [246,]  0.59879284 -0.3038690184
#> [247,] -0.24812635 -0.1924337962
#> [248,] -0.18977357 -0.2650539149
#> [249,]  0.50214644 -0.6113026135
#> [250,]  0.65024531 -0.0711888595
#> [251,] -0.46446791  0.4528600466
#> [252,]  0.58906425 -0.3419820622
#> [253,] -0.33692478  0.0021480441
#> [254,] -0.22205967 -0.2305810256
#> [255,] -0.28594266 -0.1215437900
#> [256,] -0.18470289 -0.2691220370
#> [257,] -0.22289969 -0.2294922710
#> [258,]  0.59236665 -0.3292404522
#> [259,] -0.26281981 -0.1670392170
#> [260,] -0.31052373 -0.0658904165
#> [261,] -0.30349570 -0.0825627579
#> [262,] -0.48414128  0.5412562586
#> [263,]  0.48500405 -0.6506997808
#> [264,] -0.45405020  0.4081941540
#> [265,] -0.42540334  0.2927999797
#> [266,]  0.70440704  0.2396948080
#> [267,]  0.61042395 -0.2559622638
#> [268,] -0.24220314 -0.2018851547
#> [269,] -0.38608458  0.1516061992
#> [270,] -0.25374171 -0.1830551601
#> [271,] -0.19688844 -0.2587165074
#> [272,] -0.23694319 -0.2098951785
#> [273,]  0.52055329 -0.5642308806
#> [274,] -0.23560900 -0.2118692763
#> [275,] -0.12289245 -0.2860437530
#> [276,] -0.38200175  0.1380581338
#> [277,] -0.61328392  1.2684957806
#> [278,] -0.40821130  0.2286556444
#> [279,] -0.27417126 -0.1455426946
#> [280,] -0.31473670 -0.0556054436
#> [281,]  0.62589221 -0.1881320461
#> [282,] -0.25755331 -0.1764585896
#> [283,] -0.34747309  0.0317163734
#> [284,] -0.37456920  0.1139257015
#> [285,] -0.21402473 -0.2405121488
#> [286,] -0.25161013 -0.1866630410
#> [287,] -0.41954625  0.2705222346
#> [288,] -0.40468758  0.2159730659
#> [289,] -0.41644142  0.2588913380
#> [290,]  0.56581873 -0.4261635268
#> [291,] -0.44070685  0.3531040507
#> [292,] -0.19049981 -0.2644405333
#> [293,] -0.30847165 -0.0708212108
#> [294,]  0.60817786 -0.2654163912
#> [295,] -0.22395092 -0.2281163965
#> [296,] -0.41249889  0.2442997901
#> [297,] -0.40900829  0.2315459769
#> [298,] -0.31297425 -0.0599345272
#> [299,]  0.68586505  0.1246423563
#> [300,] -0.16344713 -0.2819749127
#> 
#> $inv_hessian
#>           [,1]      [,2]
#> [1,] 4.8276069 0.9017989
#> [2,] 0.9017989 5.6725735
#>