PPI++ Logistic Regression

Helper function for PPI++ logistic regression

Usage

ppi_plusplus_logistic(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  opts = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see doi:10.48550/arXiv.2311.01453 ). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

opts

(list, optional): Options to pass to the optimizer. See ?optim for details.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ logistic regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier logistic regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) doi:10.48550/arXiv.2311.01453

Examples

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_plusplus_logistic(X_l, Y_l, f_l, X_u, f_u)
#> $est
#>           [,1]
#> [1,] 0.4331649
#> [2,] 0.4562385
#> 
#> $se
#> [1] 0.1206594 0.1315849
#> 
#> $lambda
#> [1] 0.08520502
#> 
#> $rectifier_est
#>            [,1]
#> [1,] 0.01802753
#> [2,] 0.01284002
#> 
#> $var_u
#>               [,1]          [,2]
#> [1,]  0.0015617817 -0.0004155139
#> [2,] -0.0004155139  0.0011808486
#> 
#> $var_l
#>             [,1]        [,2]
#> [1,]  0.22295397 -0.03402186
#> [2,] -0.03402186  0.24167983
#> 
#> $grads
#>              [,1]          [,2]
#>   [1,]  0.4208420 -0.6941025536
#>   [2,] -0.3827019 -0.0376888608
#>   [3,] -0.2797615 -0.3142483946
#>   [4,]  0.6771436  0.4564037614
#>   [5,] -0.6373903  1.3931707983
#>   [6,] -0.3192480 -0.2267609240
#>   [7,]  0.6709248  0.4105882945
#>   [8,] -0.2024092 -0.4162034859
#>   [9,]  0.6749323  0.4399768624
#>  [10,]  0.5636427 -0.2189223140
#>  [11,]  0.4750320 -0.5550803408
#>  [12,] -0.3571012 -0.1211657700
#>  [13,] -0.4366675  0.1708111676
#>  [14,] -0.3902460 -0.0112165923
#>  [15,] -0.2988507 -0.2748589838
#>  [16,] -0.3030083 -0.2655572110
#>  [17,] -0.3172439 -0.2317604075
#>  [18,] -0.5829657  0.9814832388
#>  [19,] -0.2880319 -0.2978543363
#>  [20,]  0.5048576 -0.4578235419
#>  [21,]  0.3083749 -0.8387282858
#>  [22,]  0.7279141  0.8789438122
#>  [23,] -0.2473796 -0.3684202583
#>  [24,] -0.3768771 -0.0575425147
#>  [25,] -0.3938338  0.0016741655
#>  [26,]  0.6822005  0.4945435512
#>  [27,]  0.4459834 -0.6354655702
#>  [28,] -0.5904301  1.0338882583
#>  [29,]  0.5632977 -0.2205199505
#>  [30,]  0.6295188  0.1338247250
#>  [31,] -0.3569972 -0.1214849498
#>  [32,]  0.5511273 -0.2753453634
#>  [33,] -0.3539493 -0.1307681417
#>  [34,] -0.3359155 -0.1828789008
#>  [35,]  0.4021872 -0.7312496358
#>  [36,] -0.5061231  0.5076980798
#>  [37,]  0.4216117 -0.6924545714
#>  [38,] -0.2514493 -0.3624999911
#>  [39,] -0.3634574 -0.1013518857
#>  [40,] -0.3729853 -0.0705245484
#>  [41,] -0.2946404 -0.2840161073
#>  [42,]  0.5448970 -0.3022740067
#>  [43,] -0.3487525 -0.1462790160
#>  [44,]  0.6142873  0.0433616329
#>  [45,] -0.3379398 -0.1772693535
#>  [46,] -0.5441704  0.7279343131
#>  [47,]  0.6907790  0.5611142909
#>  [48,] -0.2268996 -0.3942490051
#>  [49,]  0.7072840  0.6961622052
#>  [50,]  0.5970275 -0.0524352114
#>  [51,] -0.6016679  1.1151203092
#>  [52,]  0.6022367 -0.0242502543
#>  [53,] -0.3578729 -0.1187921125
#>  [54,] -0.4585592  0.2683788117
#>  [55,]  0.5561766 -0.2529595540
#>  [56,]  0.7573225  1.1700672749
#>  [57,] -0.3799204 -0.0472332443
#>  [58,] -0.4650345  0.2987243203
#>  [59,] -0.2584570 -0.3517023281
#>  [60,] -0.2808853 -0.3120811472
#>  [61,]  0.6056589 -0.0053949637
#>  [62,] -0.2720340 -0.3286327468
#>  [63,] -0.5583629  0.8171399559
#>  [64,] -0.3217223 -0.2205066661
#>  [65,] -0.5026407  0.4888573307
#>  [66,] -0.6594676  1.5814431398
#>  [67,] -0.4754064  0.3487732710
#>  [68,]  0.6413325  0.2080254814
#>  [69,]  0.3595598 -0.7963238894
#>  [70,] -0.3700991 -0.0800056105
#>  [71,]  0.5100380 -0.4393506890
#>  [72,] -0.5478276  0.7505420629
#>  [73,] -0.3694810 -0.0820199009
#>  [74,]  0.4885760 -0.5128101476
#>  [75,]  0.4894074 -0.5101138084
#>  [76,] -0.4873242  0.4085092177
#>  [77,] -0.3993069  0.0217150805
#>  [78,] -0.3707588 -0.0778497380
#>  [79,]  0.6562396  0.3069713895
#>  [80,]  0.7262197  0.8633071377
#>  [81,] -0.4341883  0.1602437242
#>  [82,] -0.2321011 -0.3883219385
#>  [83,] -0.2592914 -0.3503662118
#>  [84,]  0.5259710 -0.3795010363
#>  [85,]  0.5895186 -0.0919857018
#>  [86,]  0.5299463 -0.3638412763
#>  [87,] -0.3746951 -0.0648491298
#>  [88,] -0.3407210 -0.1694633926
#>  [89,] -0.2697240 -0.3327566728
#>  [90,] -0.3428233 -0.1634868679
#>  [91,] -0.4586902  0.2689857310
#>  [92,] -0.4807507  0.3752630353
#>  [93,] -0.3327751 -0.1914610099
#>  [94,] -0.3940758  0.0025507197
#>  [95,]  0.4299174 -0.6740832199
#>  [96,] -0.4868457  0.4060640862
#>  [97,] -0.3394506 -0.1730431058
#>  [98,] -0.2855854 -0.3028111154
#>  [99,]  0.2151173 -0.8145268159
#> [100,] -0.3645515 -0.0978806345
#> [101,] -0.3769082 -0.0574380868
#> [102,]  0.6978330  0.6176803481
#> [103,]  0.6419010  0.2116882037
#> [104,] -0.2877117 -0.2985082242
#> [105,]  0.5391583 -0.3264104025
#> [106,] -0.3931557 -0.0007770719
#> [107,]  0.7382555  0.9769588979
#> [108,] -0.2912244 -0.2912509439
#> [109,]  0.7949798  1.6066284871
#> [110,]  0.6999856  0.6352795120
#> [111,]  0.5963956 -0.0558119997
#> [112,] -0.3229294 -0.2174224375
#> [113,]  0.3039307 -0.8405781299
#> [114,] -0.3882619 -0.0182618981
#> [115,] -0.3298654 -0.1992828878
#> [116,] -0.3654656 -0.0949669847
#> [117,]  0.5816885 -0.1319028185
#> [118,] -0.3527932 -0.1342534989
#> [119,]  0.6545755  0.2956215133
#> [120,] -0.3040787 -0.2631205300
#> [121,]  0.6931317  0.5797938273
#> [122,]  0.4125069 -0.7113618284
#> [123,]  0.6018969 -0.0261077218
#> [124,]  0.5640762 -0.2169109421
#> [125,] -0.4205541  0.1038535964
#> [126,]  0.5082415 -0.4458111433
#> [127,] -0.2393846 -0.3792952246
#> [128,] -0.2219257 -0.3995077150
#> [129,]  0.6065539 -0.0004191150
#> [130,] -0.2852583 -0.3034670721
#> [131,] -0.2453011 -0.3713442290
#> [132,]  0.6476879  0.2494661194
#> [133,] -0.3411734 -0.1681827762
#> [134,] -0.5013942  0.4821656913
#> [135,] -0.5036993  0.4945624368
#> [136,] -0.3457636 -0.1550186570
#> [137,]  0.5343626 -0.3460969257
#> [138,]  0.5355029 -0.3414554988
#> [139,] -0.3775786 -0.0551786024
#> [140,] -0.6181045  1.2391876446
#> [141,]  0.5312873 -0.3584918310
#> [142,] -0.3263090 -0.2086727213
#> [143,] -0.4401029  0.1856151267
#> [144,]  0.4554347 -0.6108223579
#> [145,] -0.3343172 -0.1872648683
#> [146,] -0.3529840 -0.1336796134
#> [147,]  0.5208083 -0.3994014591
#> [148,]  0.5793057 -0.1437872002
#> [149,] -0.3291722 -0.2011276419
#> [150,]  0.5110110 -0.4358278220
#> [151,]  0.2512218 -0.8398708741
#> [152,] -0.2652261 -0.3405532290
#> [153,] -0.4464133  0.2132971170
#> [154,] -0.3817104 -0.0411042137
#> [155,]  0.5574910 -0.2470485087
#> [156,]  0.4736235 -0.5592987333
#> [157,]  0.5825828 -0.1274111098
#> [158,] -0.5013513  0.4819356511
#> [159,] -0.5000872  0.4751780797
#> [160,]  0.5042378 -0.4600015726
#> [161,] -0.2656459 -0.3398385828
#> [162,] -0.4425961  0.1964763078
#> [163,]  0.4318347 -0.6696889203
#> [164,] -0.2781491 -0.3173244382
#> [165,]  0.8112524  1.8225925299
#> [166,] -0.4590693  0.2707445996
#> [167,]  0.5916319 -0.0809818691
#> [168,] -0.1942882 -0.4212566337
#> [169,]  0.4787929 -0.5436538676
#> [170,] -0.3346024 -0.1864849450
#> [171,]  0.5802699 -0.1389929577
#> [172,]  0.3965080 -0.7415006827
#> [173,] -0.4059690  0.0467242869
#> [174,] -0.3591152 -0.1149524399
#> [175,] -0.3204252 -0.2237966202
#> [176,] -0.2776781 -0.3182155751
#> [177,]  0.5691866 -0.1929129073
#> [178,]  0.5347565 -0.3444963858
#> [179,] -0.4424147  0.1956826345
#> [180,] -0.2064969 -0.4132337640
#> [181,] -0.3921602 -0.0043629735
#> [182,]  0.4679812 -0.5758652675
#> [183,] -0.4532435  0.2439784888
#> [184,]  0.5711016 -0.1837810248
#> [185,]  0.5458688 -0.2981235962
#> [186,] -0.3822084 -0.0393908503
#> [187,]  0.4847622 -0.5250280159
#> [188,] -0.4333810  0.1568239428
#> [189,]  0.6083431  0.0095843172
#> [190,]  0.6167026  0.0573273928
#> [191,] -0.3105363 -0.2480598838
#> [192,]  0.5672492 -0.2020739634
#> [193,]  0.7376266  0.9708693442
#> [194,] -0.2743495 -0.3244175785
#> [195,] -0.3189249 -0.2275710618
#> [196,] -0.5283538  0.6331159752
#> [197,]  0.5871189 -0.1043613366
#> [198,] -0.6586332  1.5740810922
#> [199,] -0.4203463  0.1030164582
#> [200,]  0.6026692 -0.0218824096
#> [201,]  0.6965117  0.6069559225
#> [202,]  0.4812060 -0.5361971296
#> [203,] -0.4101692  0.0628406636
#> [204,] -0.3658721 -0.0936670629
#> [205,] -0.5788005  0.9527656783
#> [206,] -0.6289850  1.3249064151
#> [207,]  0.5594335 -0.2382497906
#> [208,] -0.3708789 -0.0774564096
#> [209,] -0.2945486 -0.2842127451
#> [210,]  0.6243316  0.1023725710
#> [211,] -0.3636834 -0.1006362190
#> [212,] -0.4054432  0.0447257893
#> [213,]  0.4014878 -0.7325385979
#> [214,] -0.6907110  1.8721358118
#> [215,]  0.3939003 -0.7460434898
#> [216,] -0.3367508 -0.1805714609
#> [217,] -0.3333951 -0.1897783658
#> [218,] -0.4368339  0.1715238350
#> [219,] -0.3971208  0.0136556850
#> [220,] -0.3173154 -0.2315829413
#> [221,] -0.5923182  1.0473385700
#> [222,] -0.3079343 -0.2542029609
#> [223,] -0.4292904  0.1396515934
#> [224,] -0.3546634 -0.1286054598
#> [225,] -0.5000359  0.4749044808
#> [226,] -0.3615915 -0.1072307442
#> [227,]  0.6853057  0.5183649285
#> [228,]  0.4444009 -0.6394514932
#> [229,]  0.4296602 -0.6746683781
#> [230,] -0.4895423  0.4198934013
#> [231,] -0.3913345 -0.0073260971
#> [232,]  0.3605416 -0.7951310141
#> [233,]  0.5820541 -0.1300689506
#> [234,] -0.1900287 -0.4234503477
#> [235,] -0.3493903 -0.1443970659
#> [236,] -0.3834399 -0.0351371165
#> [237,]  0.7124274  0.7402212925
#> [238,]  0.5310103 -0.3595997612
#> [239,] -0.3560081 -0.1245127175
#> [240,] -0.6714931  1.6897922810
#> [241,] -0.3490111 -0.1455164913
#> [242,] -0.3263788 -0.2084901166
#> [243,]  0.6847651  0.5141953242
#> [244,] -0.5122346  0.5412850526
#> [245,] -0.2630598 -0.3441978700
#> [246,] -0.4196645  0.1002751815
#> [247,] -0.3463800 -0.1532271253
#> [248,] -0.3838564 -0.0336932649
#> [249,]  0.5569874 -0.2493176001
#> [250,] -0.2849082 -0.3041671583
#> [251,]  0.6524505  0.2812415514
#> [252,] -0.2735319 -0.3259151805
#> [253,] -0.5622785  0.8424585687
#> [254,] -0.2136722 -0.4073412414
#> [255,] -0.3757462 -0.0613384302
#> [256,] -0.3923936 -0.0035235433
#> [257,]  0.6716019  0.4155193392
#> [258,]  0.6158402  0.0523244808
#> [259,]  0.4995301 -0.4763249221
#> [260,]  0.5400355 -0.3227622155
#> [261,] -0.4377291  0.1753657538
#> [262,]  0.4733299 -0.5601736839
#> [263,] -0.3753549 -0.0626473664
#> [264,]  0.6835442  0.5048131914
#> [265,]  0.5364182 -0.3377122420
#> [266,] -0.2857487 -0.3024829638
#> [267,] -0.5246294  0.6114740537
#> [268,] -0.3462174 -0.1537001062
#> [269,] -0.4860270  0.4018901201
#> [270,] -0.3758478 -0.0609982066
#> [271,] -0.2128190 -0.4080869542
#> [272,] -0.5414815  0.7114780760
#> [273,]  0.6075433  0.0051033997
#> [274,] -0.2664091 -0.3385326124
#> [275,] -0.2523104 -0.3612140886
#> [276,]  0.7353812  0.9492645569
#> [277,] -0.2479357 -0.3676266443
#> [278,]  0.4301773 -0.6734909700
#> [279,] -0.3399838 -0.1715436494
#> [280,] -0.3950381  0.0060452518
#> [281,] -0.2635479 -0.3433828910
#> [282,] -0.4813832  0.3784303112
#> [283,] -0.3697892 -0.0810163044
#> [284,]  0.7039039  0.6677311401
#> [285,]  0.5628351 -0.2226585355
#> [286,] -0.4915318  0.4301767711
#> [287,] -0.5316351  0.6523979605
#> [288,] -0.3373837 -0.1788163577
#> [289,] -0.3652486 -0.0956596966
#> [290,]  0.7120252  0.7367408788
#> [291,]  0.4136250 -0.7091090488
#> [292,]  0.6479793  0.2513920854
#> [293,] -0.2489427 -0.3661770325
#> [294,] -0.5197430  0.5834689206
#> [295,]  0.7245434  0.8479499577
#> [296,]  0.4861378 -0.5206498311
#> [297,] -0.3656995 -0.0942195244
#> [298,]  0.6501431  0.2657689471
#> [299,]  0.5802483 -0.1391004936
#> [300,]  0.6002464 -0.0350918536
#> 
#> $grads_hat_unlabeled
#>              [,1]         [,2]
#>   [1,]  0.4855106 -0.522650112
#>   [2,] -0.4973148  0.460455964
#>   [3,]  0.6024585 -0.023036762
#>   [4,] -0.3483239 -0.147540498
#>   [5,] -0.2399460 -0.378564451
#>   [6,]  0.5501102 -0.279793390
#>   [7,] -0.4117643  0.069032130
#>   [8,] -0.4225057  0.111747203
#>   [9,]  0.5492841 -0.283390911
#>  [10,] -0.2543706 -0.358091184
#>  [11,] -0.2102214 -0.410282538
#>  [12,] -0.3890928 -0.015318790
#>  [13,] -0.2150984 -0.406067905
#>  [14,] -0.4634174  0.291081612
#>  [15,] -0.3477644 -0.149182950
#>  [16,]  0.4874020 -0.516597817
#>  [17,]  0.4263533 -0.682098481
#>  [18,] -0.5024650  0.487912411
#>  [19,]  0.4617852 -0.593450974
#>  [20,] -0.4650353  0.298728365
#>  [21,] -0.3687829 -0.084288196
#>  [22,] -0.5228673  0.601324655
#>  [23,] -0.4238401  0.117178822
#>  [24,]  0.4951547 -0.491147905
#>  [25,] -0.3827879 -0.037392070
#>  [26,] -0.3224578 -0.218629835
#>  [27,]  0.3379674 -0.818942607
#>  [28,] -0.4973815  0.460808334
#>  [29,]  0.5785984 -0.147291253
#>  [30,]  0.4764574 -0.550777544
#>  [31,] -0.6124542  1.195815848
#>  [32,]  0.5744376 -0.167689482
#>  [33,] -0.4510063  0.233846290
#>  [34,] -0.4276502  0.132839610
#>  [35,]  0.5313943 -0.358063714
#>  [36,] -0.4119672  0.069822640
#>  [37,]  0.5009210 -0.471542750
#>  [38,]  0.5885638 -0.096924605
#>  [39,] -0.2853935 -0.303196086
#>  [40,] -0.1635180 -0.429769118
#>  [41,] -0.3402910 -0.170677758
#>  [42,] -0.5122903  0.541594132
#>  [43,]  0.6431128  0.219524523
#>  [44,]  0.4873906 -0.516634278
#>  [45,] -0.4819300  0.381173622
#>  [46,]  0.5337313 -0.348656192
#>  [47,] -0.5836286  0.986088084
#>  [48,]  0.8035387  1.717890421
#>  [49,]  0.3439192 -0.813397575
#>  [50,] -0.3292428 -0.200940110
#>  [51,] -0.2350310 -0.384792413
#>  [52,] -0.3789857 -0.050414228
#>  [53,]  0.5514612 -0.273880234
#>  [54,] -0.2776661 -0.318238142
#>  [55,] -0.4204200  0.103313273
#>  [56,] -0.4349154  0.163332981
#>  [57,] -0.3228659 -0.217585098
#>  [58,] -0.3967210  0.012189546
#>  [59,] -0.3234507 -0.216083712
#>  [60,] -0.3480041 -0.148479772
#>  [61,] -0.3754560 -0.062309281
#>  [62,] -0.3819646 -0.040230012
#>  [63,] -0.3622597 -0.105131384
#>  [64,] -0.4768799  0.356029072
#>  [65,] -0.4146185  0.080208417
#>  [66,]  0.6855657  0.520373345
#>  [67,] -0.3357281 -0.183394889
#>  [68,]  0.5416942 -0.315823028
#>  [69,]  0.6515839  0.275412905
#>  [70,]  0.5366794 -0.336640986
#>  [71,] -0.4120709  0.070226604
#>  [72,] -0.3884496 -0.017598109
#>  [73,]  0.3800579 -0.768437985
#>  [74,]  0.6083631  0.009696698
#>  [75,]  0.7014469  0.647318835
#>  [76,]  0.4518214 -0.620413723
#>  [77,] -0.4563550  0.258205294
#>  [78,]  0.4835001 -0.529016451
#>  [79,]  0.5794065 -0.143286534
#>  [80,]  0.5389409 -0.327312334
#>  [81,] -0.4308040  0.145975048
#>  [82,] -0.2405881 -0.377722594
#>  [83,] -0.3176643 -0.230716608
#>  [84,] -0.2240970 -0.397261513
#>  [85,] -0.3960862  0.009866649
#>  [86,]  0.5900410 -0.089274511
#>  [87,]  0.4830532 -0.530422285
#>  [88,] -0.5566230  0.805989081
#>  [89,]  0.3842703 -0.761928643
#>  [90,] -0.2801984 -0.313408102
#>  [91,] -0.3220229 -0.219740578
#>  [92,]  0.5036809 -0.461952856
#>  [93,] -0.3116819 -0.245323223
#>  [94,]  0.5903583 -0.087625027
#>  [95,] -0.2343728 -0.385597240
#>  [96,]  0.6027325 -0.021535602
#>  [97,] -0.5170110  0.568001148
#>  [98,] -0.3378774 -0.177443027
#>  [99,]  0.4555469 -0.610521027
#> [100,] -0.3876373 -0.020467444
#> [101,]  0.5521457 -0.270870751
#> [102,] -0.2664696 -0.338428726
#> [103,] -0.3266837 -0.207692210
#> [104,]  0.5634447 -0.219839188
#> [105,]  0.4968841 -0.485328842
#> [106,]  0.6153895  0.049717021
#> [107,] -0.3240866 -0.214445513
#> [108,]  0.7107608  0.725838804
#> [109,] -0.3548045 -0.128177442
#> [110,]  0.5290467 -0.367410876
#> [111,]  0.6848265  0.514668026
#> [112,]  0.5145714 -0.422791359
#> [113,]  0.6544879  0.295026028
#> [114,] -0.4016313  0.030363910
#> [115,] -0.2823947 -0.309140418
#> [116,] -0.3971346  0.013706467
#> [117,]  0.6302499  0.138311726
#> [118,] -0.3475384 -0.149844977
#> [119,] -0.3594594 -0.113884595
#> [120,]  0.5345200 -0.345457930
#> [121,]  0.5434931 -0.308237212
#> [122,] -0.4572917  0.262518984
#> [123,]  0.5626091 -0.223701597
#> [124,] -0.3979180  0.016586358
#> [125,]  0.4287610 -0.676705600
#> [126,]  0.5675284 -0.200758510
#> [127,] -0.3360760 -0.182436120
#> [128,]  0.5816480 -0.132106060
#> [129,] -0.3383554 -0.176110100
#> [130,] -0.2172785 -0.404056300
#> [131,]  0.4558650 -0.609665922
#> [132,] -0.4468355  0.215171873
#> [133,] -0.3254904 -0.210807455
#> [134,]  0.5585248 -0.242375660
#> [135,]  0.4091823 -0.717947248
#> [136,] -0.4054184  0.044631999
#> [137,] -0.3780234 -0.053675792
#> [138,] -0.2083791 -0.411771484
#> [139,] -0.3546365 -0.128687258
#> [140,]  0.6637076  0.358878085
#> [141,]  0.4548262 -0.612452355
#> [142,] -0.3981567  0.017465748
#> [143,] -0.3241987 -0.214156061
#> [144,] -0.4393451  0.182333432
#> [145,] -0.2703532 -0.331641452
#> [146,] -0.2883322 -0.297239673
#> [147,]  0.6464898  0.241570578
#> [148,] -0.5291215  0.637609524
#> [149,]  0.5851772 -0.114282968
#> [150,] -0.3696184 -0.081572707
#> [151,] -0.3909350 -0.008756331
#> [152,]  0.5552360 -0.257167952
#> [153,]  0.5809647 -0.135525738
#> [154,]  0.5934832 -0.071260571
#> [155,] -0.4458721  0.210898211
#> [156,]  0.6749263  0.439932414
#> [157,] -0.3007576 -0.270624617
#> [158,] -0.2868754 -0.300208735
#> [159,] -0.2596229 -0.349832247
#> [160,] -0.4427891  0.197321262
#> [161,] -0.4444520  0.204625572
#> [162,] -0.3352429 -0.184729095
#> [163,] -0.3912253 -0.007717223
#> [164,] -0.3850789 -0.029440614
#> [165,]  0.6035677 -0.016949292
#> [166,] -0.3455083 -0.155759019
#> [167,] -0.5733704  0.915883157
#> [168,] -0.6335200  1.361512242
#> [169,] -0.3039125 -0.263499999
#> [170,]  0.4711990 -0.566482108
#> [171,]  0.6343714  0.163865602
#> [172,]  0.4548140 -0.612484880
#> [173,] -0.3249734 -0.212150842
#> [174,] -0.3501192 -0.142238494
#> [175,] -0.3034709 -0.264506203
#> [176,]  0.5731268 -0.174039881
#> [177,]  0.4685458 -0.574231425
#> [178,] -0.4379441  0.176290493
#> [179,] -0.2666874 -0.338054090
#> [180,]  0.5490670 -0.284334611
#> [181,]  0.5421872 -0.313750349
#> [182,] -0.2738958 -0.325249902
#> [183,]  0.5623420 -0.224933226
#> [184,] -0.2985061 -0.275618545
#> [185,]  0.5232143 -0.390188148
#> [186,]  0.6446746  0.229682093
#> [187,]  0.5533374 -0.265609039
#> [188,] -0.4013518  0.029319597
#> [189,] -0.2235147 -0.397871427
#> [190,] -0.2819984 -0.309915891
#> [191,] -0.4682117  0.313866021
#> [192,] -0.2983216 -0.276024450
#> [193,] -0.3489273 -0.145763880
#> [194,] -0.3531839 -0.133077928
#> [195,] -0.3189968 -0.227390970
#> [196,] -0.4822419  0.382740541
#> [197,]  0.4933869 -0.497042408
#> [198,]  0.5071659 -0.449651162
#> [199,] -0.4162509  0.086657079
#> [200,]  0.4044926 -0.726948167
#> [201,] -0.6472967  1.476005147
#> [202,]  0.4242069 -0.686829902
#> [203,] -0.2862182 -0.301537584
#> [204,] -0.3407665 -0.169334827
#> [205,]  0.4343334 -0.663875394
#> [206,]  0.5349598 -0.343669443
#> [207,] -0.6167217  1.228501974
#> [208,]  0.6172303  0.060397303
#> [209,] -0.3488497 -0.145992420
#> [210,]  0.4180910 -0.699917771
#> [211,]  0.5215889 -0.396423983
#> [212,] -0.2827788 -0.308386624
#> [213,] -0.3452572 -0.156486183
#> [214,] -0.3422203 -0.165207806
#> [215,] -0.5470753  0.745870544
#> [216,]  0.6648038  0.366632698
#> [217,]  0.6292277  0.132041735
#> [218,]  0.6046290 -0.011098386
#> [219,] -0.6135914  1.204483266
#> [220,] -0.3143893 -0.238778713
#> [221,]  0.3174090 -0.834063720
#> [222,]  0.5851760 -0.114288940
#> [223,] -0.2019275 -0.416534747
#> [224,]  0.5778058 -0.151204968
#> [225,]  0.6212295  0.083883468
#> [226,] -0.4167308  0.088560541
#> [227,] -0.5094012  0.525630503
#> [228,]  0.5493372 -0.283160059
#> [229,]  0.7471177  1.064632344
#> [230,] -0.4528367  0.242130097
#> [231,] -0.3905584 -0.010101742
#> [232,] -0.3950546  0.006105191
#> [233,]  0.5686137 -0.195630045
#> [234,]  0.5112235 -0.435055898
#> [235,] -0.3269015 -0.207121324
#> [236,]  0.5891739 -0.093771004
#> [237,] -0.2687190 -0.334525495
#> [238,]  0.5019179 -0.468094172
#> [239,]  0.4791555 -0.542539767
#> [240,]  0.5930476 -0.073554563
#> [241,]  0.6646709  0.365690865
#> [242,] -0.4390139  0.180902036
#> [243,] -0.2662355 -0.338830519
#> [244,]  0.5959322 -0.058282851
#> [245,]  0.4723778 -0.563001697
#> [246,]  0.5561953 -0.252875509
#> [247,]  0.3182324 -0.833578366
#> [248,] -0.3259387 -0.209639552
#> [249,]  0.6030521 -0.019782286
#> [250,] -0.3781869 -0.053122514
#> [251,] -0.2588079 -0.351141805
#> [252,]  0.3800658 -0.768426029
#> [253,] -0.3722808 -0.072850385
#> [254,] -0.4204041  0.103249172
#> [255,] -0.2364724 -0.383005683
#> [256,] -0.3348447 -0.185821604
#> [257,]  0.6902296  0.556778382
#> [258,] -0.2032201 -0.415636899
#> [259,]  0.4261467 -0.682556976
#> [260,] -0.4810673  0.376847495
#> [261,] -0.2944436 -0.284437719
#> [262,]  0.6725009  0.422087915
#> [263,] -0.5471507  0.746338288
#> [264,]  0.6080345  0.007853573
#> [265,] -0.4103670  0.063606546
#> [266,]  0.5812298 -0.134199995
#> [267,] -0.3313634 -0.195271572
#> [268,] -0.3869998 -0.022712580
#> [269,]  0.4940442 -0.494857017
#> [270,]  0.4700430 -0.569873008
#> [271,] -0.3250800 -0.211874226
#> [272,] -0.4097193  0.061101659
#> [273,] -0.4800047  0.371536452
#> [274,] -0.3269218 -0.207067997
#> [275,] -0.3688586 -0.084042642
#> [276,] -0.4341486  0.160075248
#> [277,] -0.3879080 -0.019512518
#> [278,] -0.2923357 -0.288916497
#> [279,] -0.4611741  0.280550517
#> [280,]  0.5596228 -0.237388408
#> [281,] -0.2709025 -0.330662862
#> [282,]  0.4451671 -0.637526600
#> [283,] -0.3252162 -0.211520563
#> [284,] -0.4950622  0.448592441
#> [285,]  0.4571777 -0.606119974
#> [286,] -0.5066589  0.510616454
#> [287,] -0.3040848 -0.263106656
#> [288,] -0.4228368  0.113092605
#> [289,]  0.5284750 -0.369671156
#> [290,] -0.4169630  0.089482798
#> [291,]  0.4897603 -0.508965573
#> [292,]  0.7042744  0.670827251
#> [293,]  0.5915720 -0.081295081
#> [294,]  0.4527698 -0.617916682
#> [295,]  0.4763161 -0.551205637
#> [296,]  0.4923491 -0.500477749
#> [297,]  0.5984329 -0.044892116
#> [298,] -0.4373143  0.173584106
#> [299,] -0.3326782 -0.191723635
#> [300,]  0.5462392 -0.296536773
#> 
#> $grads_hat
#>              [,1]          [,2]
#>   [1,]  0.4208420 -0.6941025536
#>   [2,] -0.3827019 -0.0376888608
#>   [3,] -0.2797615 -0.3142483946
#>   [4,] -0.3228564 -0.2176095033
#>   [5,] -0.6373903  1.3931707983
#>   [6,] -0.3192480 -0.2267609240
#>   [7,] -0.3290752 -0.2013853410
#>   [8,] -0.2024092 -0.4162034859
#>   [9,] -0.3250677 -0.2119060831
#>  [10,] -0.4363573  0.1694839022
#>  [11,]  0.4750320 -0.5550803408
#>  [12,] -0.3571012 -0.1211657700
#>  [13,]  0.5633325 -0.2203587184
#>  [14,] -0.3902460 -0.0112165923
#>  [15,] -0.2988507 -0.2748589838
#>  [16,]  0.6969917  0.6108453056
#>  [17,] -0.3172439 -0.2317604075
#>  [18,]  0.4170343 -0.7021204402
#>  [19,] -0.2880319 -0.2978543363
#>  [20,] -0.4951424  0.4490134997
#>  [21,]  0.3083749 -0.8387282858
#>  [22,]  0.7279141  0.8789438122
#>  [23,] -0.2473796 -0.3684202583
#>  [24,] -0.3768771 -0.0575425147
#>  [25,]  0.6061662 -0.0025767792
#>  [26,] -0.3177995 -0.2303804573
#>  [27,] -0.5540166  0.7893980728
#>  [28,] -0.5904301  1.0338882583
#>  [29,] -0.4367023  0.1709603575
#>  [30,]  0.6295188  0.1338247250
#>  [31,]  0.6430028  0.2188117345
#>  [32,] -0.4488727  0.2242586087
#>  [33,] -0.3539493 -0.1307681417
#>  [34,] -0.3359155 -0.1828789008
#>  [35,] -0.5978128  1.0869324308
#>  [36,]  0.4938769 -0.4954138624
#>  [37,]  0.4216117 -0.6924545714
#>  [38,] -0.2514493 -0.3624999911
#>  [39,] -0.3634574 -0.1013518857
#>  [40,] -0.3729853 -0.0705245484
#>  [41,]  0.7053596  0.6799253691
#>  [42,] -0.4551030  0.2524620688
#>  [43,]  0.6512475  0.2731560235
#>  [44,]  0.6142873  0.0433616329
#>  [45,] -0.3379398 -0.1772693535
#>  [46,]  0.4558296 -0.6097612275
#>  [47,] -0.3092210 -0.2511777457
#>  [48,] -0.2268996 -0.3942490051
#>  [49,] -0.2927160 -0.2881132047
#>  [50,]  0.5970275 -0.0524352114
#>  [51,]  0.3983321 -0.7382615989
#>  [52,]  0.6022367 -0.0242502543
#>  [53,] -0.3578729 -0.1187921125
#>  [54,]  0.5414408 -0.3168865217
#>  [55,]  0.5561766 -0.2529595540
#>  [56,] -0.2426775 -0.3749380239
#>  [57,] -0.3799204 -0.0472332443
#>  [58,]  0.5349655 -0.3436459600
#>  [59,] -0.2584570 -0.3517023281
#>  [60,] -0.2808853 -0.3120811472
#>  [61,]  0.6056589 -0.0053949637
#>  [62,] -0.2720340 -0.3286327468
#>  [63,] -0.5583629  0.8171399559
#>  [64,] -0.3217223 -0.2205066661
#>  [65,]  0.4973593 -0.4837208272
#>  [66,]  0.3405324 -0.8166172130
#>  [67,]  0.5245936 -0.3848585378
#>  [68,]  0.6413325  0.2080254814
#>  [69,]  0.3595598 -0.7963238894
#>  [70,]  0.6299009  0.1361678453
#>  [71,]  0.5100380 -0.4393506890
#>  [72,]  0.4521724 -0.6194912584
#>  [73,]  0.6305190  0.1399668604
#>  [74,] -0.5114240  0.5367913422
#>  [75,]  0.4894074 -0.5101138084
#>  [76,] -0.4873242  0.4085092177
#>  [77,] -0.3993069  0.0217150805
#>  [78,] -0.3707588 -0.0778497380
#>  [79,]  0.6562396  0.3069713895
#>  [80,] -0.2737803 -0.3254613440
#>  [81,] -0.4341883  0.1602437242
#>  [82,] -0.2321011 -0.3883219385
#>  [83,] -0.2592914 -0.3503662118
#>  [84,]  0.5259710 -0.3795010363
#>  [85,] -0.4104814  0.0640495881
#>  [86,] -0.4700537  0.3227213168
#>  [87,] -0.3746951 -0.0648491298
#>  [88,] -0.3407210 -0.1694633926
#>  [89,] -0.2697240 -0.3327566728
#>  [90,]  0.6571767  0.3133968389
#>  [91,] -0.4586902  0.2689857310
#>  [92,]  0.5192493 -0.4053142348
#>  [93,] -0.3327751 -0.1914610099
#>  [94,] -0.3940758  0.0025507197
#>  [95,]  0.4299174 -0.6740832199
#>  [96,]  0.5131543 -0.4280073410
#>  [97,] -0.3394506 -0.1730431058
#>  [98,] -0.2855854 -0.3028111154
#>  [99,]  0.2151173 -0.8145268159
#> [100,]  0.6354485  0.1706153839
#> [101,] -0.3769082 -0.0574380868
#> [102,] -0.3021670 -0.2674602590
#> [103,]  0.6419010  0.2116882037
#> [104,]  0.7122883  0.7390174322
#> [105,]  0.5391583 -0.3264104025
#> [106,] -0.3931557 -0.0007770719
#> [107,] -0.2617445 -0.3463755460
#> [108,] -0.2912244 -0.2912509439
#> [109,] -0.2050202 -0.4143392127
#> [110,]  0.6999856  0.6352795120
#> [111,] -0.4036044  0.0377701771
#> [112,] -0.3229294 -0.2174224375
#> [113,]  0.3039307 -0.8405781299
#> [114,]  0.6117381  0.0287730981
#> [115,] -0.3298654 -0.1992828878
#> [116,] -0.3654656 -0.0949669847
#> [117,]  0.5816885 -0.1319028185
#> [118,] -0.3527932 -0.1342534989
#> [119,] -0.3454245 -0.1560017430
#> [120,] -0.3040787 -0.2631205300
#> [121,] -0.3068683 -0.2566905605
#> [122,]  0.4125069 -0.7113618284
#> [123,]  0.6018969 -0.0261077218
#> [124,] -0.4359238  0.1676309477
#> [125,] -0.4205541  0.1038535964
#> [126,]  0.5082415 -0.4458111433
#> [127,] -0.2393846 -0.3792952246
#> [128,] -0.2219257 -0.3995077150
#> [129,] -0.3934461  0.0002718623
#> [130,] -0.2852583 -0.3034670721
#> [131,] -0.2453011 -0.3713442290
#> [132,]  0.6476879  0.2494661194
#> [133,] -0.3411734 -0.1681827762
#> [134,]  0.4986058 -0.4794842054
#> [135,]  0.4963007 -0.4872979499
#> [136,] -0.3457636 -0.1550186570
#> [137,] -0.4656374  0.3015848412
#> [138,]  0.5355029 -0.3414554988
#> [139,] -0.3775786 -0.0551786024
#> [140,] -0.6181045  1.2391876446
#> [141,] -0.4687127  0.3162689163
#> [142,] -0.3263090 -0.2086727213
#> [143,] -0.4401029  0.1856151267
#> [144,]  0.4554347 -0.6108223579
#> [145,] -0.3343172 -0.1872648683
#> [146,] -0.3529840 -0.1336796134
#> [147,] -0.4791917  0.3674862583
#> [148,]  0.5793057 -0.1437872002
#> [149,] -0.3291722 -0.2011276419
#> [150,]  0.5110110 -0.4358278220
#> [151,]  0.2512218 -0.8398708741
#> [152,] -0.2652261 -0.3405532290
#> [153,] -0.4464133  0.2132971170
#> [154,] -0.3817104 -0.0411042137
#> [155,] -0.4425090  0.1960949458
#> [156,]  0.4736235 -0.5592987333
#> [157,]  0.5825828 -0.1274111098
#> [158,] -0.5013513  0.4819356511
#> [159,] -0.5000872  0.4751780797
#> [160,]  0.5042378 -0.4600015726
#> [161,] -0.2656459 -0.3398385828
#> [162,] -0.4425961  0.1964763078
#> [163,]  0.4318347 -0.6696889203
#> [164,] -0.2781491 -0.3173244382
#> [165,] -0.1887476 -0.4240479212
#> [166,] -0.4590693  0.2707445996
#> [167,] -0.4083681  0.0558969501
#> [168,] -0.1942882 -0.4212566337
#> [169,]  0.4787929 -0.5436538676
#> [170,] -0.3346024 -0.1864849450
#> [171,] -0.4197301  0.1005386206
#> [172,]  0.3965080 -0.7415006827
#> [173,] -0.4059690  0.0467242869
#> [174,]  0.6408848  0.2051466007
#> [175,] -0.3204252 -0.2237966202
#> [176,] -0.2776781 -0.3182155751
#> [177,] -0.4308134  0.1460144645
#> [178,]  0.5347565 -0.3444963858
#> [179,] -0.4424147  0.1956826345
#> [180,] -0.2064969 -0.4132337640
#> [181,] -0.3921602 -0.0043629735
#> [182,] -0.5320188  0.6546654742
#> [183,] -0.4532435  0.2439784888
#> [184,] -0.4288984  0.1380199299
#> [185,] -0.4541312  0.2480215791
#> [186,] -0.3822084 -0.0393908503
#> [187,]  0.4847622 -0.5250280159
#> [188,] -0.4333810  0.1568239428
#> [189,]  0.6083431  0.0095843172
#> [190,] -0.3832974 -0.0356305254
#> [191,] -0.3105363 -0.2480598838
#> [192,]  0.5672492 -0.2020739634
#> [193,] -0.2623734 -0.3453376746
#> [194,] -0.2743495 -0.3244175785
#> [195,] -0.3189249 -0.2275710618
#> [196,]  0.4716462 -0.5651644726
#> [197,]  0.5871189 -0.1043613366
#> [198,]  0.3413668 -0.8158397045
#> [199,] -0.4203463  0.1030164582
#> [200,]  0.6026692 -0.0218824096
#> [201,]  0.6965117  0.6069559225
#> [202,]  0.4812060 -0.5361971296
#> [203,] -0.4101692  0.0628406636
#> [204,] -0.3658721 -0.0936670629
#> [205,]  0.4211995 -0.6933381819
#> [206,]  0.3710150 -0.7815132986
#> [207,] -0.4405665  0.1876270507
#> [208,] -0.3708789 -0.0774564096
#> [209,] -0.2945486 -0.2842127451
#> [210,] -0.3756684 -0.0615988904
#> [211,]  0.6363166  0.1760775576
#> [212,] -0.4054432  0.0447257893
#> [213,]  0.4014878 -0.7325385979
#> [214,]  0.3092890 -0.8383114153
#> [215,]  0.3939003 -0.7460434898
#> [216,] -0.3367508 -0.1805714609
#> [217,] -0.3333951 -0.1897783658
#> [218,] -0.4368339  0.1715238350
#> [219,] -0.3971208  0.0136556850
#> [220,]  0.6826846  0.4982363853
#> [221,]  0.4076818 -0.7208638835
#> [222,] -0.3079343 -0.2542029609
#> [223,] -0.4292904  0.1396515934
#> [224,] -0.3546634 -0.1286054598
#> [225,]  0.4999641 -0.4748363454
#> [226,]  0.6384085  0.1893214563
#> [227,] -0.3146943 -0.2380345978
#> [228,]  0.4444009 -0.6394514932
#> [229,]  0.4296602 -0.6746683781
#> [230,]  0.5104577 -0.4378331298
#> [231,] -0.3913345 -0.0073260971
#> [232,]  0.3605416 -0.7951310141
#> [233,] -0.4179459  0.0933964524
#> [234,] -0.1900287 -0.4234503477
#> [235,]  0.6506097  0.2688859761
#> [236,] -0.3834399 -0.0351371165
#> [237,]  0.7124274  0.7402212925
#> [238,]  0.5310103 -0.3595997612
#> [239,] -0.3560081 -0.1245127175
#> [240,]  0.3285069 -0.8266778971
#> [241,] -0.3490111 -0.1455164913
#> [242,] -0.3263788 -0.2084901166
#> [243,] -0.3152349 -0.2367122747
#> [244,]  0.4877654 -0.5154280351
#> [245,] -0.2630598 -0.3441978700
#> [246,] -0.4196645  0.1002751815
#> [247,] -0.3463800 -0.1532271253
#> [248,]  0.6161436  0.0540824363
#> [249,]  0.5569874 -0.2493176001
#> [250,] -0.2849082 -0.3041671583
#> [251,] -0.3475495 -0.1498126515
#> [252,] -0.2735319 -0.3259151805
#> [253,]  0.4377215 -0.6558354447
#> [254,] -0.2136722 -0.4073412414
#> [255,] -0.3757462 -0.0613384302
#> [256,]  0.6076064  0.0054560703
#> [257,] -0.3283981 -0.2031795432
#> [258,]  0.6158402  0.0523244808
#> [259,]  0.4995301 -0.4763249221
#> [260,] -0.4599645  0.2749063311
#> [261,]  0.5622709 -0.2252605079
#> [262,] -0.5266701  0.6233003097
#> [263,]  0.6246451  0.1042543401
#> [264,] -0.3164558 -0.2337099635
#> [265,] -0.4635818  0.2918566956
#> [266,] -0.2857487 -0.3024829638
#> [267,] -0.5246294  0.6114740537
#> [268,] -0.3462174 -0.1537001062
#> [269,] -0.4860270  0.4018901201
#> [270,] -0.3758478 -0.0609982066
#> [271,] -0.2128190 -0.4080869542
#> [272,] -0.5414815  0.7114780760
#> [273,]  0.6075433  0.0051033997
#> [274,] -0.2664091 -0.3385326124
#> [275,] -0.2523104 -0.3612140886
#> [276,] -0.2646188 -0.3415822926
#> [277,] -0.2479357 -0.3676266443
#> [278,]  0.4301773 -0.6734909700
#> [279,] -0.3399838 -0.1715436494
#> [280,] -0.3950381  0.0060452518
#> [281,] -0.2635479 -0.3433828910
#> [282,] -0.4813832  0.3784303112
#> [283,] -0.3697892 -0.0810163044
#> [284,] -0.2960961 -0.2808800690
#> [285,]  0.5628351 -0.2226585355
#> [286,] -0.4915318  0.4301767711
#> [287,]  0.4683649 -0.5747555916
#> [288,] -0.3373837 -0.1788163577
#> [289,]  0.6347514  0.1662432611
#> [290,] -0.2879748 -0.2979709484
#> [291,]  0.4136250 -0.7091090488
#> [292,] -0.3520207 -0.1365710803
#> [293,] -0.2489427 -0.3661770325
#> [294,] -0.5197430  0.5834689206
#> [295,] -0.2754566 -0.3223733072
#> [296,]  0.4861378 -0.5206498311
#> [297,]  0.6343005  0.1634224346
#> [298,]  0.6501431  0.2657689471
#> [299,]  0.5802483 -0.1391004936
#> [300,] -0.3997536  0.0233705563
#> 
#> $inv_hessian
#>           [,1]      [,2]
#> [1,] 4.4584276 0.7193295
#> [2,] 0.7193295 4.6758765
#>