PPI++ OLS

Helper function for PPI++ OLS estimation

Usage

ppi_plusplus_ols(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see doi:10.48550/arXiv.2311.01453 ). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ OLS regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier OLS regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) doi:10.48550/arXiv.2311.01453

Examples

dat <- simdat(model = "ols")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_plusplus_ols(X_l, Y_l, f_l, X_u, f_u)
#> $est
#> X(Intercept)          XX1 
#>    0.7902755    0.8052522 
#> 
#> $se
#> [1] 0.09437302 0.08358099
#> 
#> $lambda
#> [1] 0.2424469
#> 
#> $rectifier_est
#> X(Intercept)          XX1 
#>   0.04119700  -0.01345125 
#> 
#> $var_u
#>            [,1]       [,2]
#> [1,] 0.12050278 0.02834766
#> [2,] 0.02834766 0.13909185
#> 
#> $var_l
#>           [,1]      [,2]
#> [1,] 2.5509137 0.1794944
#> [2,] 0.1794944 1.7944488
#> 
#> $grads
#>                [,1]         [,2]
#>   [1,] -1.662479234  1.233381647
#>   [2,] -3.409893598  2.866000171
#>   [3,] -0.205979353  0.069524600
#>   [4,] -3.685904762 -0.609426393
#>   [5,]  0.729522432  0.254872078
#>   [6,] -0.696294305 -0.360288740
#>   [7,]  0.912351674  1.206697000
#>   [8,]  2.256726103 -0.712224324
#>   [9,] 10.617883958  5.973640355
#>  [10,] -0.863745811 -0.106554706
#>  [11,] -4.681675945  8.188416161
#>  [12,]  1.489138925 -1.096330680
#>  [13,]  2.397138837  2.233118985
#>  [14,] -0.689502986  1.827634235
#>  [15,] -2.619264238 -3.833182927
#>  [16,] -3.458192927 -0.531307013
#>  [17,] -2.182559194 -0.838674163
#>  [18,]  0.222626391 -0.128808599
#>  [19,]  0.010874889  0.008532800
#>  [20,] -1.058498427 -0.103243578
#>  [21,]  1.618760426 -1.833943776
#>  [22,]  0.839151641  0.188316138
#>  [23,] -0.644479734  0.050484298
#>  [24,]  3.238400794  1.940981065
#>  [25,] -0.678491129  0.089932447
#>  [26,]  0.409011325 -0.685410833
#>  [27,] -0.327088717  0.316510331
#>  [28,]  1.316913161  2.032440726
#>  [29,] -2.817862751  2.299338269
#>  [30,]  1.688682247 -1.036282635
#>  [31,] -0.298899659 -0.486086486
#>  [32,]  0.649531057  0.936587105
#>  [33,] -0.961027211 -1.605524202
#>  [34,]  0.447981371 -0.090887008
#>  [35,] -0.316603159  0.329444551
#>  [36,]  0.253184843 -0.001813353
#>  [37,]  1.025366668 -0.791292985
#>  [38,]  0.296925206  0.023525812
#>  [39,]  0.394175693 -0.553385554
#>  [40,]  1.936237032 -1.993703953
#>  [41,] -0.072539652 -0.087237496
#>  [42,]  1.343967973  0.315783075
#>  [43,]  0.246227078 -0.019741823
#>  [44,] -1.223925104 -1.831778317
#>  [45,]  0.711464550  0.203239312
#>  [46,]  0.079858238 -0.015487709
#>  [47,]  1.312867977  1.771877710
#>  [48,] -0.953242213  1.409688216
#>  [49,]  0.522742973 -0.490609084
#>  [50,]  1.629478207  0.241819406
#>  [51,] -0.369616476 -0.244561390
#>  [52,]  1.127466258 -1.939097126
#>  [53,] -1.484633213  0.659530753
#>  [54,] -1.367766257 -1.499343753
#>  [55,]  0.580001793  0.062526209
#>  [56,]  3.224244340 -1.319501592
#>  [57,]  0.538294990  0.153161033
#>  [58,]  0.239485935 -0.190704924
#>  [59,]  0.757148736 -0.160383458
#>  [60,] -0.181186080 -0.023721306
#>  [61,] -0.318465363  0.473613243
#>  [62,] -0.526627538  0.315040339
#>  [63,]  1.541168269  1.813026770
#>  [64,] -2.395624488  0.057963625
#>  [65,]  0.920813858  0.344710293
#>  [66,]  1.767509680 -3.080483342
#>  [67,] -0.254660827 -0.352980449
#>  [68,]  2.382675510 -2.444487243
#>  [69,]  2.036687200 -0.314321983
#>  [70,]  0.074791608  0.030592272
#>  [71,] -1.046879659  1.307301043
#>  [72,]  2.894295642 -3.690091641
#>  [73,]  1.567786361  2.217119671
#>  [74,]  1.588637660  0.318093898
#>  [75,]  2.489560436 -0.491806379
#>  [76,]  0.451682277 -0.606069390
#>  [77,]  0.291983042  0.176681315
#>  [78,]  0.752886774  0.695891173
#>  [79,] -1.022092442  0.360111531
#>  [80,] -3.687195034 -3.107425828
#>  [81,]  0.018232150  0.008700286
#>  [82,] -1.728607622  0.411869575
#>  [83,]  1.390577580  1.627849620
#>  [84,] -0.434664415  0.327375179
#>  [85,]  0.628390812  0.867926784
#>  [86,]  1.661610795  1.283422849
#>  [87,]  0.794013464 -0.710093664
#>  [88,]  0.986130227 -0.693956214
#>  [89,] -1.957621281  0.040166450
#>  [90,]  1.091842790  0.738630189
#>  [91,]  0.044581274 -0.005822144
#>  [92,] -1.876911453  0.519324391
#>  [93,] -0.021999495  0.018793416
#>  [94,]  1.523192761  0.422901813
#>  [95,] -1.579936499 -2.198316098
#>  [96,]  0.694057057 -0.961608527
#>  [97,]  0.937129139  0.263649833
#>  [98,]  0.849567886 -0.473781974
#>  [99,]  2.053137180 -5.932843998
#> [100,]  3.658414767  1.001553395
#> [101,] -0.123344867  0.066527629
#> [102,] -1.694160849  0.142492041
#> [103,] -0.305795077  0.068957119
#> [104,] -0.295004244  0.041434859
#> [105,]  2.319574024  1.449761550
#> [106,]  1.231510830  1.707115205
#> [107,] -2.944512294  1.668248208
#> [108,] -3.103430663 -2.311844730
#> [109,] -0.969822671  0.043741950
#> [110,] -0.246124738  0.505658599
#> [111,] -1.257755649  0.361506131
#> [112,] -1.426432361 -1.668505598
#> [113,] -0.891980130 -0.129673950
#> [114,]  0.542752451 -0.152040792
#> [115,] -0.856960194  0.121550072
#> [116,]  1.433726003 -0.249140015
#> [117,] -0.983735530 -0.614547554
#> [118,]  0.230107508  0.259713184
#> [119,]  0.308646523 -0.232896702
#> [120,]  0.438959747  0.376279710
#> [121,] -0.927967788  0.920028205
#> [122,]  1.355228174  1.092301897
#> [123,] -0.924247940  1.376660915
#> [124,]  0.930117158 -1.283411255
#> [125,]  0.742948961 -0.623628371
#> [126,] -1.710451676 -1.442892300
#> [127,] -0.492888912  0.002225347
#> [128,] -0.267406946 -0.443529899
#> [129,]  0.303054692  0.603337008
#> [130,]  0.412518031  0.388818506
#> [131,]  1.703602953 -1.625700182
#> [132,]  0.870398990  0.646373674
#> [133,]  0.889362950 -1.291582167
#> [134,] -0.823791717 -0.306516998
#> [135,]  1.841289323 -1.064432756
#> [136,] -0.447692645  0.980852858
#> [137,] -5.991200511  2.140838171
#> [138,] -3.448261352  1.201204633
#> [139,]  2.425422835  0.880752522
#> [140,] -0.555008277  0.200799123
#> [141,] -2.173593813  0.017810348
#> [142,] -0.808687467 -0.755492003
#> [143,] -0.987970315  1.149498442
#> [144,]  0.743498389 -1.169873006
#> [145,]  2.042550622  3.072538936
#> [146,]  0.827045858  1.112654543
#> [147,]  0.257789294  0.012787423
#> [148,]  0.029406559  0.026895529
#> [149,] -1.234112858 -1.005715349
#> [150,]  0.054500996 -0.052517315
#> [151,]  1.780941928 -1.646356205
#> [152,] -1.250148019 -0.063101508
#> [153,] -0.302040110  0.152874569
#> [154,] -2.172908268 -1.704459991
#> [155,] -0.990192276  0.325399051
#> [156,] -1.086724795 -0.608387190
#> [157,]  0.090462506  0.023575377
#> [158,]  0.773159310 -1.309516032
#> [159,] -0.008258778  0.005010082
#> [160,]  1.069814536 -1.275135339
#> [161,] -0.535515202  0.263382114
#> [162,] -0.926038958 -0.767996675
#> [163,] -2.612589443 -4.196851876
#> [164,] -0.578946991 -0.449322923
#> [165,] -0.421614597 -0.036966286
#> [166,]  0.350960804 -0.040933677
#> [167,]  0.839480128 -0.536207906
#> [168,]  0.684481918  0.678553136
#> [169,]  1.358853306  1.868964783
#> [170,]  0.787032088 -0.409572330
#> [171,]  0.405047911 -0.322269614
#> [172,] -0.041044684  0.040296661
#> [173,] -0.814826391 -0.889629252
#> [174,]  0.831358851 -0.111196107
#> [175,]  2.331094014  1.072662496
#> [176,] -0.335959553  0.223147322
#> [177,]  3.298039986  2.708486581
#> [178,] -1.305120925  1.118730186
#> [179,] -2.056418067 -0.736731990
#> [180,]  1.443472292 -0.609961808
#> [181,] -1.443299074 -0.235432267
#> [182,] -0.901315762 -0.682670372
#> [183,] -0.711323071  0.543865672
#> [184,]  0.583910412  0.337405043
#> [185,]  1.444155194 -1.300254073
#> [186,] -0.710740984  0.621210162
#> [187,]  5.792969226 -3.996490449
#> [188,]  1.953240266  1.226507231
#> [189,]  1.763901442  2.005433914
#> [190,]  0.269888039  0.004181772
#> [191,] -0.059560843  0.071958806
#> [192,] -2.698577947  2.386566558
#> [193,]  0.304879455  0.223885644
#> [194,] -1.007161889 -1.319825794
#> [195,]  0.330035131  0.250759958
#> [196,] -1.132062162 -0.445137309
#> [197,]  0.644169292 -0.326528114
#> [198,] -0.580752019 -0.887658042
#> [199,] -4.167240476  2.661272945
#> [200,]  2.334431461 -0.676383452
#> [201,] -2.999799402  2.584704061
#> [202,] -1.409413859 -0.940863242
#> [203,]  0.409864619  0.541348323
#> [204,] -0.580018067  0.599511317
#> [205,] -1.211075224  0.684569569
#> [206,]  1.495806905  0.335748938
#> [207,]  1.837500889  1.745414311
#> [208,] -0.406289533 -0.029898582
#> [209,]  0.243981384 -0.043342374
#> [210,]  0.701149215 -0.034210445
#> [211,]  1.445454846  0.652791138
#> [212,] -0.306265748  0.142476353
#> [213,]  0.363192675  0.006554793
#> [214,] -0.510686669  0.051751504
#> [215,] -4.789488031 -1.943165281
#> [216,]  0.474369727  0.421535225
#> [217,] -0.980186996  1.001952106
#> [218,] -3.242144917 -1.366102564
#> [219,] -0.060380252 -0.160231967
#> [220,] -1.726984876  0.097086262
#> [221,] -2.766529981 -2.698173946
#> [222,]  2.192144591  0.081681891
#> [223,] -1.765037097 -0.308923992
#> [224,]  0.920111782  0.770914860
#> [225,] -2.558959464 -4.669332991
#> [226,] -0.696572984 -0.743119739
#> [227,]  0.600485049  0.252436248
#> [228,]  1.526966393  2.565918637
#> [229,]  0.947689259 -0.689270192
#> [230,] -2.726805904  3.283023902
#> [231,]  1.686281377  0.430389078
#> [232,] -0.113447646 -0.082055178
#> [233,]  0.639303226 -0.442733976
#> [234,] -1.076919096 -1.655280006
#> [235,]  0.016997879 -0.007252758
#> [236,]  1.650621301  0.118604578
#> [237,]  1.395720818 -0.381048711
#> [238,] -0.069529351  0.023006642
#> [239,] -1.030871210  0.124312299
#> [240,]  0.831469960 -0.025289898
#> [241,] -0.706467750  0.678083639
#> [242,] -3.253668013 -0.212483518
#> [243,]  3.036704566 -1.525435896
#> [244,]  0.345755628  0.190956275
#> [245,]  0.615836132 -0.619355035
#> [246,] -0.122812445 -0.025402062
#> [247,] -0.927365674  1.083794394
#> [248,] -0.540936439  0.389557907
#> [249,] -0.081533849  0.047564165
#> [250,]  0.194154444 -0.090946107
#> [251,] -1.988086877  2.515574318
#> [252,] -0.746562705  0.575785397
#> [253,] -0.081264695  0.099847257
#> [254,]  0.486793971 -0.266672781
#> [255,] -1.301500299 -1.161020807
#> [256,]  0.383908019  0.232909553
#> [257,]  1.172542031 -1.060233046
#> [258,]  4.504167092  2.881087984
#> [259,]  0.375899899 -0.902931167
#> [260,] -0.919389608  0.535335860
#> [261,] -0.869823503  0.937658971
#> [262,]  1.227625271 -0.916589887
#> [263,]  2.042269294  0.064034419
#> [264,] -0.604053745  0.008164309
#> [265,] -1.220728604 -0.078814155
#> [266,]  1.500841363  0.686727896
#> [267,]  0.366950461  0.015313705
#> [268,] -6.977431818 -7.487334140
#> [269,]  1.294404091  0.596402827
#> [270,] -1.678383692  1.918336181
#> [271,] -0.352185874 -0.205960382
#> [272,]  2.587301430  2.827432523
#> [273,] -0.522774575  0.495861433
#> [274,]  1.508250714  3.453552409
#> [275,] -0.997327817  0.711110026
#> [276,]  1.961913892 -1.217780826
#> [277,]  0.768372500  0.428357275
#> [278,] -2.009508389  0.368686234
#> [279,]  0.207077107 -0.222675102
#> [280,]  0.060671739  0.025091826
#> [281,] -0.403941959 -0.130849800
#> [282,]  1.520095462 -0.001178182
#> [283,]  0.921358824  1.205615833
#> [284,]  0.975362772  0.881128815
#> [285,]  1.143004277 -0.396280584
#> [286,] -0.312813967  0.616119548
#> [287,]  1.026724204  0.439665598
#> [288,] -0.061868363  0.172830519
#> [289,] -1.922820180 -1.821879757
#> [290,]  0.153795309 -0.043793807
#> [291,] -1.380615131 -3.177487795
#> [292,] -0.707943704  0.877428043
#> [293,]  0.633432486 -0.234993228
#> [294,]  1.616728659 -1.933258056
#> [295,]  1.331757024 -0.848103354
#> [296,] -1.538786465 -1.318897323
#> [297,] -0.700925196  0.125650852
#> [298,] -0.191436182  0.112482659
#> [299,] -0.111870907 -0.099326214
#> [300,] -2.262761170  0.688525925
#> 
#> $grads_hat_unlabeled
#>                 [,1]          [,2]
#>   [1,] -0.7152083691  7.924121e-02
#>   [2,]  1.2927561764  9.406701e-02
#>   [3,]  0.0316313915 -4.912748e-02
#>   [4,] -1.3734820419  5.334131e-01
#>   [5,] -0.1124244453  7.934850e-03
#>   [6,]  1.4004964892  7.193990e-01
#>   [7,]  0.3050915957  6.074250e-01
#>   [8,] -1.6990855812 -2.475408e+00
#>   [9,]  0.1106200595  5.039720e-02
#>  [10,] -0.6717415616  4.050419e-01
#>  [11,]  1.2751007722 -2.453139e+00
#>  [12,] -0.9870589846  9.286497e-02
#>  [13,]  1.7580244390  3.948864e+00
#>  [14,] -0.0714558457 -6.333456e-03
#>  [15,]  1.8558549932 -1.456590e-01
#>  [16,] -1.8170829763  2.808655e+00
#>  [17,]  1.6599132647  1.885773e-01
#>  [18,] -0.1145203519  1.698241e-01
#>  [19,] -0.7250055667  1.583040e-01
#>  [20,] -1.7289909634  1.685745e+00
#>  [21,]  0.9669679049  7.977955e-01
#>  [22,] -1.4906291972  7.259836e-02
#>  [23,] -1.0091958410 -1.284544e-01
#>  [24,] -2.5311385761 -6.564521e-01
#>  [25,] -0.0025542383 -2.338397e-03
#>  [26,]  1.2234674025 -5.039276e-01
#>  [27,] -0.6847323144 -1.035842e+00
#>  [28,]  0.0574827271 -2.468208e-02
#>  [29,]  2.3220953752  1.351838e+00
#>  [30,] -1.5618944728 -1.235322e-01
#>  [31,]  1.0946415024 -1.357638e+00
#>  [32,] -1.1424220090  1.509788e+00
#>  [33,] -0.2412971680  1.437709e-01
#>  [34,]  0.6959351871 -1.375390e+00
#>  [35,] -1.4441795654  4.621269e-01
#>  [36,]  0.6016255094  3.398848e-01
#>  [37,]  0.7947660400  8.425494e-01
#>  [38,]  0.0073551747 -1.417412e-03
#>  [39,]  2.2969686642  1.466142e+00
#>  [40,] -1.8855930412 -2.262291e+00
#>  [41,]  0.6638759576 -7.480051e-01
#>  [42,]  0.0621187072  1.245032e-03
#>  [43,]  1.4996615086  4.920381e-01
#>  [44,]  0.2255516237 -1.940983e-01
#>  [45,]  2.3883748637  5.819397e+00
#>  [46,] -0.9109791289 -4.665872e-01
#>  [47,]  0.4789624981  1.224131e-02
#>  [48,]  0.3527036897  3.215883e-01
#>  [49,] -0.2697879580  1.566313e-01
#>  [50,]  0.6829356303  8.752126e-01
#>  [51,] -0.6270193530 -8.642353e-02
#>  [52,]  0.6642108710  4.492452e-01
#>  [53,] -1.9752350138 -5.372464e-01
#>  [54,] -1.7276809203  3.420094e+00
#>  [55,] -1.1993487399  1.062444e+00
#>  [56,]  0.2104406850  1.036829e-01
#>  [57,] -1.2106142117  7.777746e-01
#>  [58,]  2.8752349826 -2.505515e+00
#>  [59,] -2.1581059589  3.474790e-01
#>  [60,] -0.0095728733  7.537132e-03
#>  [61,]  0.0415784664  2.599814e-02
#>  [62,]  1.8200916139  2.618665e+00
#>  [63,] -1.4755425626 -1.765778e+00
#>  [64,] -0.2097877369 -5.697112e-02
#>  [65,]  2.3538812302 -6.613747e+00
#>  [66,]  2.4392400760  2.845993e+00
#>  [67,]  0.9675782842 -5.461617e-01
#>  [68,]  1.9774570446  1.362806e+00
#>  [69,]  1.3748509581  8.808512e-01
#>  [70,]  0.9044443979  5.789568e-01
#>  [71,]  0.3955606169  2.888734e-01
#>  [72,] -0.4345665001 -1.922262e-01
#>  [73,]  0.7114894710 -4.240982e-01
#>  [74,] -0.9989644438 -2.532740e-01
#>  [75,]  1.5297922458 -5.179354e-01
#>  [76,]  0.0078901676 -6.344337e-03
#>  [77,]  0.7004217694 -1.038957e+00
#>  [78,] -2.2245925784 -3.249561e+00
#>  [79,]  0.7210977371 -6.749519e-02
#>  [80,] -1.2874129135 -8.832721e-01
#>  [81,]  0.0572848744 -1.603977e-02
#>  [82,]  2.1244557059 -1.881486e+00
#>  [83,] -0.1145889500 -1.010635e-01
#>  [84,]  0.8341218435  1.572788e+00
#>  [85,]  0.0701487011  1.149852e-01
#>  [86,] -1.6647995922  9.183391e-02
#>  [87,] -1.4302056440 -2.255549e-02
#>  [88,] -3.6711565660 -1.089879e+01
#>  [89,] -0.7186383137 -2.260075e+00
#>  [90,] -1.9645687728 -5.125517e-01
#>  [91,]  0.9336339352  1.129077e+00
#>  [92,] -0.3514725070 -1.202269e-02
#>  [93,]  0.0007149599 -7.283326e-04
#>  [94,] -1.5866745522 -1.522375e+00
#>  [95,] -2.5366778892 -7.387193e-01
#>  [96,] -1.4711855569 -6.084916e-01
#>  [97,]  1.5124762013  1.644966e+00
#>  [98,] -1.8722725640 -2.658704e+00
#>  [99,]  1.2307095394 -1.497914e+00
#> [100,]  0.1242406919  5.136567e-02
#> [101,]  2.3726140270  1.942993e+00
#> [102,]  0.7373477582 -1.636484e-01
#> [103,] -1.5163620769 -2.717595e+00
#> [104,]  0.3431088253 -2.294925e-01
#> [105,]  1.8864039923  2.943281e+00
#> [106,]  0.4130638501  4.083199e-01
#> [107,] -0.8546789036  1.430968e+00
#> [108,] -1.9695013977  4.244889e+00
#> [109,] -2.8908169457 -1.706466e+00
#> [110,] -0.8738678783 -1.493271e+00
#> [111,]  0.1327074013  8.056850e-02
#> [112,]  1.2659148640  2.652725e+00
#> [113,]  1.4230838542 -2.804153e-01
#> [114,] -1.2115763442  5.440380e-02
#> [115,] -0.7192032281  2.699971e-02
#> [116,] -0.6005790419  1.648222e-01
#> [117,]  0.9884343519 -8.557295e-01
#> [118,] -0.9928272470 -1.079919e+00
#> [119,] -0.0904296658  3.528654e-02
#> [120,]  1.3325351499 -2.939314e-01
#> [121,]  1.1920689123  9.264066e-01
#> [122,]  0.0859661798 -1.275151e-01
#> [123,] -1.8236043338 -1.172824e+00
#> [124,]  0.3678308951  2.722202e-01
#> [125,] -3.0200347776  4.424492e-02
#> [126,] -0.5550985819  1.252653e-01
#> [127,] -0.5869490722 -1.487040e-01
#> [128,]  0.4855868578  5.943693e-01
#> [129,] -1.7459610477  7.242545e-01
#> [130,]  0.1440045571  8.630342e-02
#> [131,]  0.0701772495 -5.483225e-02
#> [132,]  0.7741510214  3.804039e-02
#> [133,] -1.5377860227  2.611580e-04
#> [134,]  0.1982564295 -2.459601e-01
#> [135,]  0.9538042467  1.973782e+00
#> [136,] -1.3087014321  9.432175e-01
#> [137,]  2.2197897685  3.228820e+00
#> [138,]  0.7696741219  1.295598e+00
#> [139,]  2.9268660472  4.536748e-01
#> [140,]  0.8878373377  3.316321e-01
#> [141,] -0.8638966716 -1.005624e+00
#> [142,]  2.3136498988  1.567515e+00
#> [143,]  0.1783843148  9.002622e-02
#> [144,] -1.5018466192  1.203121e+00
#> [145,] -0.6380794171 -2.761515e-01
#> [146,]  0.6785542339 -6.975549e-01
#> [147,]  0.2561379059  2.092228e-01
#> [148,]  1.5326182217  1.816696e+00
#> [149,] -0.0806913473  7.973574e-02
#> [150,] -0.2508892968 -3.237260e-01
#> [151,] -2.0126894867  3.252081e-01
#> [152,] -0.0958580665 -4.861016e-02
#> [153,]  0.4379059577 -3.178903e-01
#> [154,]  0.8299997472 -1.402382e+00
#> [155,] -2.1736569233  5.496867e-01
#> [156,] -2.9993476712  1.741557e+00
#> [157,] -0.3851371533 -5.539901e-02
#> [158,] -0.9931891252 -2.782527e+00
#> [159,] -0.3538954880  2.474152e-01
#> [160,] -2.4280914079 -5.809874e+00
#> [161,] -1.1070927889 -3.127223e-01
#> [162,]  0.5905135713  1.323106e-01
#> [163,]  0.5729911289  8.217643e-02
#> [164,]  2.3819389627 -3.416494e+00
#> [165,] -0.7634757951 -3.988370e-02
#> [166,]  1.2830509930  1.432443e+00
#> [167,]  0.8357111180 -1.488206e-01
#> [168,]  0.9684366620 -3.135222e-01
#> [169,]  0.1772572822  6.543676e-02
#> [170,]  2.2938178809  2.166262e+00
#> [171,] -0.3047880218 -4.185909e-01
#> [172,]  0.3512906797  1.949537e-02
#> [173,]  1.5417717442 -2.150193e+00
#> [174,] -1.3618810381  1.906853e-01
#> [175,]  0.3598708256 -2.977212e-01
#> [176,] -2.4170795511  1.096997e+00
#> [177,] -0.5771810978 -2.195663e-01
#> [178,]  1.6237486513  2.608518e+00
#> [179,]  1.7788213141 -5.993049e-01
#> [180,]  1.2189152485 -2.099258e-01
#> [181,]  2.9103173144 -5.291296e-01
#> [182,]  0.0659307314  1.040222e-01
#> [183,]  0.6949617928 -3.944485e-01
#> [184,]  3.6298686648 -2.136664e-01
#> [185,] -1.8897968047 -8.745625e-01
#> [186,] -1.6657014693  1.093156e+00
#> [187,] -0.5514899614  8.731342e-01
#> [188,]  1.4566491039  2.714207e+00
#> [189,] -0.8132430922  2.712425e-01
#> [190,]  2.0789023838 -1.158493e+00
#> [191,]  1.3222537275  2.611895e-01
#> [192,]  1.7725267448 -3.186899e+00
#> [193,]  0.5122133849  1.113702e-01
#> [194,]  0.3644392089  1.065808e-01
#> [195,] -1.9878446013 -1.779884e+00
#> [196,]  1.5489256846  2.371930e-01
#> [197,] -0.2028340211  3.263834e-01
#> [198,] -0.2249041942 -6.550155e-02
#> [199,] -0.1103052537  1.677223e-02
#> [200,] -2.7255070679 -1.455082e+00
#> [201,] -1.5102662568 -3.246101e+00
#> [202,] -1.0479216094  6.013778e-01
#> [203,]  0.6404736174  7.229812e-01
#> [204,] -0.0229682537  3.720878e-02
#> [205,]  1.3332591351 -5.642480e-01
#> [206,]  2.0974612590  2.385495e+00
#> [207,]  3.1768548642 -8.073519e-01
#> [208,]  0.7984273287  1.260486e+00
#> [209,]  1.6784447268 -1.137217e+00
#> [210,]  0.5056955212  5.893565e-01
#> [211,] -0.4459910329 -4.902377e-01
#> [212,]  0.1371448993  2.509777e-01
#> [213,]  1.7477644426  1.996372e+00
#> [214,] -0.5742323461  2.860751e-01
#> [215,]  0.7090712413 -4.476198e-01
#> [216,] -0.3893031518 -1.337103e-01
#> [217,] -1.0252500876  2.913284e-01
#> [218,]  1.2140482314  2.869991e-01
#> [219,]  0.3937312835  4.570013e-01
#> [220,] -0.3307845828  1.305228e-01
#> [221,]  1.0375589863 -4.488494e-01
#> [222,] -0.9461631843  1.500217e+00
#> [223,]  0.4867801806  1.340504e-01
#> [224,]  1.0088734819 -7.199792e-02
#> [225,] -2.4083786222 -1.380041e+00
#> [226,]  3.5554928170  1.185541e-01
#> [227,] -2.7332385498 -2.266349e+00
#> [228,]  1.2121813979  5.237679e-01
#> [229,] -1.1082447584  6.178598e-01
#> [230,] -2.7160890543  2.230958e+00
#> [231,]  0.1801822732  3.362626e-01
#> [232,] -1.4116524822 -9.233721e-01
#> [233,] -0.1657490373 -1.012830e-03
#> [234,]  3.5768234886  1.916209e+00
#> [235,] -0.9939495220  2.202987e-01
#> [236,] -1.7155574280 -1.356623e-01
#> [237,] -0.6282963462  1.119146e-01
#> [238,]  1.7051616068 -2.818961e-01
#> [239,] -2.1913782424  2.011444e-01
#> [240,] -0.9267215727 -9.707404e-01
#> [241,]  0.1555146722 -1.420891e-01
#> [242,]  1.8306835416 -1.164806e+00
#> [243,]  0.1427464842  5.731739e-02
#> [244,]  2.5259316382 -1.659202e+00
#> [245,]  0.8327128001 -2.646586e-01
#> [246,] -3.5734512263 -4.605375e+00
#> [247,] -0.3417688665  4.387447e-01
#> [248,]  1.7853196571  3.404700e+00
#> [249,]  2.1392519560 -2.969416e-01
#> [250,] -0.5588791038 -1.563795e-01
#> [251,] -2.8661676714  3.563014e+00
#> [252,] -2.4655099004 -8.285475e-01
#> [253,]  1.7307377536 -2.902963e+00
#> [254,]  1.8017578108  2.555030e+00
#> [255,] -0.8318665758 -1.365718e+00
#> [256,] -0.3795369589 -2.225131e-01
#> [257,]  1.5715191173  8.923620e-01
#> [258,] -0.0267495839  8.469099e-03
#> [259,] -0.0456064653 -3.890790e-02
#> [260,]  0.7997118092  2.082357e-01
#> [261,] -0.5548405478  4.369077e-01
#> [262,] -2.3921625575 -3.945476e-01
#> [263,]  1.2849649510  2.052284e+00
#> [264,] -2.9172778927  9.839758e-01
#> [265,] -0.0280954903 -4.374733e-02
#> [266,] -2.7089216038 -1.159625e+00
#> [267,] -0.0132051817  7.679982e-03
#> [268,]  1.1878639802 -2.765347e-02
#> [269,] -0.7458296446 -7.993590e-01
#> [270,] -0.6502966655 -1.366955e+00
#> [271,] -1.1809548881  2.597845e+00
#> [272,]  1.1310327387  1.006647e+00
#> [273,]  0.5369943990 -3.285678e-01
#> [274,] -0.9484834729  7.827290e-01
#> [275,] -2.3254173691 -3.309274e+00
#> [276,] -0.8982262045  4.758512e-01
#> [277,]  1.3707928836 -4.712640e-01
#> [278,] -0.6058210491 -5.661501e-01
#> [279,] -0.2924220363 -3.145575e-01
#> [280,]  1.9367920517  2.426673e+00
#> [281,]  1.2476733665  2.124589e-01
#> [282,] -0.1881280188 -1.238794e-01
#> [283,] -0.4316127393  2.734300e-01
#> [284,] -1.5529825516 -1.437055e-01
#> [285,] -2.4492410105 -2.478083e+00
#> [286,]  0.8882400461 -1.067622e+00
#> [287,]  1.3985823817 -8.675783e-01
#> [288,] -1.6705827712  3.848751e-01
#> [289,] -0.1452689831  2.344314e-01
#> [290,] -1.4385112605 -3.462637e-01
#> [291,]  0.5940791706  2.849702e-01
#> [292,] -0.6763908071 -1.162364e-01
#> [293,]  0.9863836669 -1.730088e-01
#> [294,]  0.7431536409  1.083554e-01
#> [295,]  1.7726195443 -1.735968e+00
#> [296,] -1.8118544840 -1.131389e+00
#> [297,]  2.2858779605 -2.756175e+00
#> [298,] -0.2599984932 -4.422362e-01
#> [299,]  1.4799560494 -1.534739e+00
#> [300,]  0.7134533268 -1.901782e+00
#> 
#> $grads_hat
#>                [,1]         [,2]
#>   [1,]  0.482575428 -0.358019315
#>   [2,] -2.555427961  2.147825662
#>   [3,]  0.291159918 -0.098275758
#>   [4,] -1.580022734 -0.261240487
#>   [5,]  1.803913758  0.630230445
#>   [6,] -1.650675525 -0.854121312
#>   [7,] -0.310693145 -0.410929794
#>   [8,]  2.626982346 -0.829077451
#>   [9,]  3.723130087  2.094639593
#>  [10,]  0.812157158  0.100190550
#>  [11,] -3.015355342  5.273962723
#>  [12,]  1.005628318 -0.740361533
#>  [13,]  2.276391403  2.120633473
#>  [14,] -0.983430068  2.606733395
#>  [15,] -2.024712439 -2.963081403
#>  [16,] -0.433556519 -0.066610401
#>  [17,] -2.192224314 -0.842388099
#>  [18,]  0.330384834 -0.191156167
#>  [19,] -0.201337407 -0.157976032
#>  [20,]  0.296245718  0.028895147
#>  [21,]  0.810463756 -0.918199468
#>  [22,]  2.423113354  0.543776985
#>  [23,] -1.079384045  0.084551838
#>  [24,]  1.320409284  0.791405876
#>  [25,] -0.304333075  0.040338653
#>  [26,]  0.587961906 -0.985291691
#>  [27,] -1.257789083  1.217110890
#>  [28,]  0.522603989  0.806554041
#>  [29,]  0.066930944 -0.054614754
#>  [30,]  1.912802156 -1.173816840
#>  [31,]  1.678245484  2.729251858
#>  [32,]  1.620013191  2.335967538
#>  [33,]  0.162283721  0.271116613
#>  [34,]  1.150731960 -0.233461907
#>  [35,]  0.967490706 -1.006732030
#>  [36,]  0.613416514 -0.004393394
#>  [37,]  1.802983273 -1.391393012
#>  [38,] -1.791653505 -0.141955287
#>  [39,]  1.457156083 -2.045709924
#>  [40,]  0.789257366 -0.812682282
#>  [41,]  0.847462662  1.019173905
#>  [42,]  1.415011379  0.332475664
#>  [43,] -0.431580036  0.034602923
#>  [44,]  1.457237173  2.180963074
#>  [45,]  0.439590759  0.125574947
#>  [46,] -1.602549672  0.310798529
#>  [47,] -0.813018159 -1.097268559
#>  [48,] -0.931042959  1.376859176
#>  [49,]  0.844661493 -0.792738732
#>  [50,]  1.129880691  0.167677651
#>  [51,] -0.891109281 -0.589613664
#>  [52,] -0.428953101  0.737744229
#>  [53,] -0.133353198  0.059240582
#>  [54,] -0.074674896 -0.081858533
#>  [55,] -1.473286805 -0.158825439
#>  [56,]  2.401921149 -0.982971030
#>  [57,]  1.781084906  0.506771955
#>  [58,]  0.594253820 -0.473209960
#>  [59,]  1.208536293 -0.255998882
#>  [60,] -0.094655283 -0.012392491
#>  [61,]  0.419912283 -0.624482412
#>  [62,] -0.024068920  0.014398565
#>  [63,] -0.973882055 -1.145672586
#>  [64,] -1.877271281  0.045421747
#>  [65,]  0.868888905  0.325271982
#>  [66,] -1.316290946  2.294082108
#>  [67,] -1.570024396 -2.176180458
#>  [68,] -0.663896103  0.681118998
#>  [69,]  0.313347730 -0.048358962
#>  [70,]  0.570828767  0.233488079
#>  [71,] -2.068632601  2.583224858
#>  [72,] -0.566214733  0.721897314
#>  [73,] -1.052625219 -1.488593177
#>  [74,]  0.409209768  0.081936324
#>  [75,] -0.912488165  0.180259734
#>  [76,]  0.563035369 -0.755483489
#>  [77,]  0.365654110  0.221260278
#>  [78,]  0.709709100  0.655982168
#>  [79,] -0.500501491  0.176340565
#>  [80,] -1.875556970 -1.580647109
#>  [81,] -1.597768810 -0.762446881
#>  [82,]  2.473995389 -0.589470633
#>  [83,]  0.790374799  0.925235193
#>  [84,] -1.251047380  0.942248423
#>  [85,] -1.679080723 -2.319128645
#>  [86,]  1.419168843  1.096161462
#>  [87,]  0.896216268 -0.801494587
#>  [88,] -0.872113348  0.613720644
#>  [89,] -0.780820423  0.016020864
#>  [90,] -0.539058598 -0.364672421
#>  [91,]  0.694785946 -0.090736383
#>  [92,] -1.984259064  0.549026502
#>  [93,] -0.156495790  0.133688997
#>  [94,]  0.530102601  0.147178582
#>  [95,]  0.270947036  0.376994412
#>  [96,] -0.993083181  1.375905980
#>  [97,] -0.712761515 -0.200526744
#>  [98,] -0.643697559  0.358973433
#>  [99,] -0.292848578  0.846229343
#> [100,]  2.517130098  0.689107265
#> [101,] -0.046510279  0.025085913
#> [102,] -2.099028781  0.176544568
#> [103,]  0.541799561 -0.122176385
#> [104,] -0.051906225  0.007290496
#> [105,] -0.371367529 -0.232109154
#> [106,]  0.519967193  0.720776367
#> [107,] -1.223130946  0.692979280
#> [108,] -2.126757090 -1.584289357
#> [109,]  0.158963464 -0.007169735
#> [110,]  0.845473311 -1.737008856
#> [111,] -2.023129157  0.581491002
#> [112,]  1.231756502  1.440792200
#> [113,]  0.184608233  0.026837906
#> [114,] -0.452561377  0.126775641
#> [115,]  1.237900665 -0.175582152
#> [116,] -0.109382477  0.019007503
#> [117,] -1.025261704 -0.640489291
#> [118,]  1.248248168  1.408848016
#> [119,]  0.795446230 -0.600223199
#> [120,] -0.441299349 -0.378285234
#> [121,]  1.447185262 -1.434803316
#> [122,]  0.797202850  0.642538431
#> [123,] -0.780393601  1.162390872
#> [124,]  0.186871677 -0.257852692
#> [125,]  1.565618180 -1.314173604
#> [126,] -0.334262118 -0.281974781
#> [127,]  0.738390343 -0.003333762
#> [128,]  1.302713636  2.160723402
#> [129,] -0.862881880 -1.717870024
#> [130,] -0.045861181 -0.043226416
#> [131,]  2.565337837 -2.448029444
#> [132,] -0.272718861 -0.202525846
#> [133,]  1.802215418 -2.617277116
#> [134,]  0.556905267  0.207213701
#> [135,]  1.343517808 -0.776675531
#> [136,] -0.682563301  1.495432575
#> [137,] -0.169482508  0.060561255
#> [138,] -2.927114303  1.019662637
#> [139,]  0.048984803  0.017788028
#> [140,] -1.343303453  0.486000239
#> [141,] -2.244687925  0.018392891
#> [142,]  1.872309371  1.749148856
#> [143,]  0.188741733 -0.219600047
#> [144,] -0.551344911  0.867525119
#> [145,]  2.185653275  3.287803356
#> [146,] -0.240920758 -0.324119362
#> [147,]  1.140104795  0.056553947
#> [148,]  0.826138789  0.755594693
#> [149,]  0.264889870  0.215866650
#> [150,]  1.154261971 -1.112250140
#> [151,]  0.663478192 -0.613339167
#> [152,] -1.274803191 -0.064345984
#> [153,] -0.345081667  0.174659621
#> [154,] -1.635681749 -1.283051908
#> [155,] -2.606194774  0.856453163
#> [156,] -0.300771090 -0.168382353
#> [157,] -0.314306978 -0.081911344
#> [158,] -0.197960701  0.335290163
#> [159,]  0.468024556 -0.283921100
#> [160,]  0.473917430 -0.564872548
#> [161,]  0.656467297 -0.322869909
#> [162,] -1.777895235 -1.474471043
#> [163,] -1.235679857 -1.984990538
#> [164,] -1.941835441 -1.507065739
#> [165,] -1.184896220 -0.103889223
#> [166,]  0.495543464 -0.057796813
#> [167,] -0.168757198  0.107791645
#> [168,]  1.087072117  1.077656216
#> [169,]  1.870309185  2.572420427
#> [170,]  1.172088041 -0.609955854
#> [171,]  0.094430656 -0.075132177
#> [172,]  2.228468442 -2.187855500
#> [173,]  0.409617430  0.447221214
#> [174,]  0.412903610 -0.055226782
#> [175,]  2.150983363  0.989783839
#> [176,] -0.004537264  0.003013691
#> [177,]  2.293955206  1.883890711
#> [178,]  0.382847512 -0.328171175
#> [179,] -0.137306230 -0.049191307
#> [180,]  0.096549151 -0.040798355
#> [181,] -1.324698716 -0.216086068
#> [182,] -0.926162754 -0.701489865
#> [183,]  0.312697328 -0.239083124
#> [184,] -0.264050170 -0.152577959
#> [185,] -0.016485714  0.014843014
#> [186,] -0.208441696  0.182184653
#> [187,]  2.705585474 -1.866546513
#> [188,] -0.200280660 -0.125763165
#> [189,] -0.279649705 -0.317942369
#> [190,] -0.313697660 -0.004860578
#> [191,]  1.744615458 -2.107768120
#> [192,] -0.807744891  0.714352886
#> [193,] -1.260985603 -0.925994092
#> [194,] -1.768310809 -2.317266213
#> [195,] -1.372095837 -1.042515364
#> [196,] -0.673642920 -0.264882625
#> [197,]  0.327258052 -0.165886446
#> [198,] -0.954785951 -1.459355112
#> [199,]  0.055639262 -0.035532209
#> [200,] -0.773214555  0.224032934
#> [201,] -0.119952143  0.103353841
#> [202,] -1.411413861 -0.942198356
#> [203,]  0.143325601  0.189304151
#> [204,]  2.211184443 -2.285497941
#> [205,] -0.105754039  0.059778283
#> [206,]  0.506898158  0.113778401
#> [207,]  0.121594489  0.115500767
#> [208,] -0.327673476 -0.024113277
#> [209,] -0.551746247  0.098015643
#> [210,]  0.571532706 -0.027886202
#> [211,]  0.022767269  0.010282072
#> [212,] -0.090001348  0.041869076
#> [213,] -0.348252896 -0.006285165
#> [214,]  0.169784799 -0.017205498
#> [215,]  0.436395307  0.177051953
#> [216,]  0.089315536  0.079367721
#> [217,]  0.593292575 -0.606466671
#> [218,] -2.556643378 -1.077261246
#> [219,]  1.637369991  4.345112917
#> [220,] -0.577570297  0.032469388
#> [221,]  0.793538932  0.773931996
#> [222,]  1.199675658  0.044701329
#> [223,] -1.764334684 -0.308801053
#> [224,] -1.463114465 -1.225869188
#> [225,] -0.702433711 -1.281730699
#> [226,]  0.437799361  0.467054213
#> [227,]  0.049484605  0.020802696
#> [228,]  2.289187141  3.846756534
#> [229,] -1.902039569  1.383385077
#> [230,]  0.429761497 -0.517424897
#> [231,] -0.116338933 -0.029693150
#> [232,] -0.664664843 -0.480743268
#> [233,] -1.592379663  1.102763997
#> [234,] -1.696390810 -2.607439872
#> [235,]  1.633091117 -0.696817255
#> [236,]  0.853542929  0.061330906
#> [237,] -0.060395243  0.016488634
#> [238,] -0.490411841  0.162272906
#> [239,]  0.661885677 -0.079816498
#> [240,]  0.923249880 -0.028081466
#> [241,] -1.399824170  1.343582727
#> [242,]  0.818898170  0.053478832
#> [243,] -0.246829102  0.123990320
#> [244,]  1.417792918  0.783028336
#> [245,] -1.140867567  1.147386512
#> [246,]  0.351336102  0.072669032
#> [247,]  1.008246530 -1.178318291
#> [248,] -0.617690661  0.444832819
#> [249,]  1.063868213 -0.620625720
#> [250,]  2.053241902 -0.961782562
#> [251,] -1.881770950  2.381050209
#> [252,]  0.803082642 -0.619376317
#> [253,]  0.398509955 -0.489636070
#> [254,] -1.609841131  0.881894266
#> [255,] -2.407112292 -2.147296822
#> [256,] -0.941520861 -0.571202454
#> [257,]  1.278234591 -1.155802111
#> [258,]  3.009734478  1.925174991
#> [259,]  0.343798056 -0.825820866
#> [260,] -0.981763747  0.571654644
#> [261,]  0.876003457 -0.944320884
#> [262,]  1.664898159 -1.243073804
#> [263,]  0.036692731  0.001150484
#> [264,] -0.555745050  0.007511375
#> [265,] -1.089380111 -0.070333875
#> [266,]  2.882966445  1.319135739
#> [267,] -0.746327957 -0.031146020
#> [268,] -3.322521844 -3.565327743
#> [269,]  0.694630369  0.320054238
#> [270,]  0.572244159 -0.654055851
#> [271,]  0.465770219  0.272385179
#> [272,]  2.282587381  2.494437533
#> [273,]  1.082333214 -1.026613237
#> [274,]  1.982827374  4.540225437
#> [275,]  0.463331775 -0.330362660
#> [276,] -0.583196255  0.361996120
#> [277,]  0.111021056  0.061892737
#> [278,] -1.162715852  0.213324478
#> [279,]  1.337458407 -1.438201893
#> [280,] -0.887191285 -0.366912994
#> [281,]  1.007611030  0.326397639
#> [282,] -0.991240854  0.000768282
#> [283,] -0.307522295 -0.402398869
#> [284,]  0.639164986  0.577412531
#> [285,]  0.046896316 -0.016258994
#> [286,] -0.372678158  0.734028280
#> [287,] -0.572331371 -0.245084720
#> [288,] -0.219068059  0.611971035
#> [289,] -1.389748443 -1.316792169
#> [290,] -1.378037167  0.392401398
#> [291,] -1.991861773 -4.584272858
#> [292,] -1.125867855  1.395404781
#> [293,]  2.243890093 -0.832447006
#> [294,] -0.261508174  0.312707257
#> [295,]  0.801533085 -0.510440632
#> [296,] -0.210591709 -0.180498625
#> [297,]  0.638076564 -0.114384337
#> [298,] -0.379296733  0.222864375
#> [299,]  1.200358303  1.065755598
#> [300,]  1.842071313 -0.560516006
#> 
#> $inv_hessian
#>             [,1]        [,2]
#> [1,]  1.00347000 -0.06022278
#> [2,] -0.06022278  1.04518213
#>