Chen & Chen Poisson

Helper function for Chen & Chen Poisson regression estimation.

Usage

chen_poisson(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

intercept

(Logical): Do the design matrices include intercept columns? Default is TRUE.

Value

(list): A list containing the following:

est

(vector): vector of Chen & Chen Poisson regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

Details

Another look at inference after prediction (Gronsbell et al., 2025) https://arxiv.org/pdf/2411.19908

Examples

dat <- simdat(model = "poisson")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>
  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

chen_poisson(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)
#> Warning: non-integer x = 1.379666
#> Warning: non-integer x = 8.992770
#> Warning: non-integer x = 8.759763
#> Warning: non-integer x = 34.109530
#> Warning: non-integer x = 2.525449
#> Warning: non-integer x = 9.747778
#> Warning: non-integer x = 8.118187
#> Warning: non-integer x = 15.612430
#> Warning: non-integer x = 5.927597
#> Warning: non-integer x = 0.890923
#> Warning: non-integer x = 5.434579
#> Warning: non-integer x = 6.232813
#> Warning: non-integer x = 1.672638
#> Warning: non-integer x = 5.433659
#> Warning: non-integer x = 14.759053
#> Warning: non-integer x = 24.889429
#> Warning: non-integer x = 33.311369
#> Warning: non-integer x = 1.722837
#> Warning: non-integer x = 7.261753
#> Warning: non-integer x = 6.367081
#> Warning: non-integer x = 6.507715
#> Warning: non-integer x = 3.967287
#> Warning: non-integer x = 5.186078
#> Warning: non-integer x = 1.137085
#> Warning: non-integer x = 12.745448
#> Warning: non-integer x = 4.220578
#> Warning: non-integer x = 52.203049
#> Warning: non-integer x = 26.223081
#> Warning: non-integer x = 1.442262
#> Warning: non-integer x = 0.465134
#> Warning: non-integer x = 4.286391
#> Warning: non-integer x = 2.344476
#> Warning: non-integer x = 2.881268
#> Warning: non-integer x = 3.492148
#> Warning: non-integer x = 24.593063
#> Warning: non-integer x = 42.147159
#> Warning: non-integer x = 1.091511
#> Warning: non-integer x = 3.968655
#> Warning: non-integer x = 8.886063
#> Warning: non-integer x = 1.067032
#> Warning: non-integer x = 16.470294
#> Warning: non-integer x = 90.142466
#> Warning: non-integer x = 6.730750
#> Warning: non-integer x = 4.026099
#> Warning: non-integer x = 7.312671
#> Warning: non-integer x = 5.184773
#> Warning: non-integer x = 4.208718
#> Warning: non-integer x = 31.089427
#> Warning: non-integer x = 6.309495
#> Warning: non-integer x = 19.420207
#> Warning: non-integer x = 0.308941
#> Warning: non-integer x = 2.834251
#> Warning: non-integer x = 3.125445
#> Warning: non-integer x = 3.097088
#> Warning: non-integer x = 3.343681
#> Warning: non-integer x = 7.133640
#> Warning: non-integer x = 0.918634
#> Warning: non-integer x = 11.294755
#> Warning: non-integer x = 4.518292
#> Warning: non-integer x = 4.525765
#> Warning: non-integer x = 6.669693
#> Warning: non-integer x = 2.576901
#> Warning: non-integer x = 2.434697
#> Warning: non-integer x = 0.568859
#> Warning: non-integer x = 1.390007
#> Warning: non-integer x = 2.337041
#> Warning: non-integer x = 6.926049
#> Warning: non-integer x = 32.412105
#> Warning: non-integer x = 23.710329
#> Warning: non-integer x = 51.559349
#> Warning: non-integer x = 3.614930
#> Warning: non-integer x = 25.851226
#> Warning: non-integer x = 9.591226
#> Warning: non-integer x = 5.781760
#> Warning: non-integer x = 0.587468
#> Warning: non-integer x = 23.891779
#> Warning: non-integer x = 0.440312
#> Warning: non-integer x = 32.379587
#> Warning: non-integer x = 3.023443
#> Warning: non-integer x = 0.829132
#> Warning: non-integer x = 7.436117
#> Warning: non-integer x = 0.681257
#> Warning: non-integer x = 10.264559
#> Warning: non-integer x = 9.097521
#> Warning: non-integer x = 27.283833
#> Warning: non-integer x = 6.938711
#> Warning: non-integer x = 1.549468
#> Warning: non-integer x = 8.944017
#> Warning: non-integer x = 2.605657
#> Warning: non-integer x = 50.046441
#> Warning: non-integer x = 9.262039
#> Warning: non-integer x = 1.822786
#> Warning: non-integer x = 1.994296
#> Warning: non-integer x = 0.595799
#> Warning: non-integer x = 2.214162
#> Warning: non-integer x = 8.308474
#> Warning: non-integer x = 11.430020
#> Warning: non-integer x = 32.445425
#> Warning: non-integer x = 5.541763
#> Warning: non-integer x = 17.844194
#> Warning: non-integer x = 21.367848
#> Warning: non-integer x = 3.918709
#> Warning: non-integer x = 1.840330
#> Warning: non-integer x = 3.384189
#> Warning: non-integer x = 0.574850
#> Warning: non-integer x = 14.353827
#> Warning: non-integer x = 4.111037
#> Warning: non-integer x = 10.503800
#> Warning: non-integer x = 0.930664
#> Warning: non-integer x = 4.322300
#> Warning: non-integer x = 5.365941
#> Warning: non-integer x = 1.068578
#> Warning: non-integer x = 2.285779
#> Warning: non-integer x = 5.211691
#> Warning: non-integer x = 4.620198
#> Warning: non-integer x = 3.395709
#> Warning: non-integer x = 3.418190
#> Warning: non-integer x = 2.977589
#> Warning: non-integer x = 2.032596
#> Warning: non-integer x = 3.007671
#> Warning: non-integer x = 9.798787
#> Warning: non-integer x = 37.496746
#> Warning: non-integer x = 18.159342
#> Warning: non-integer x = 7.377741
#> Warning: non-integer x = 0.554142
#> Warning: non-integer x = 29.213664
#> Warning: non-integer x = 3.659216
#> Warning: non-integer x = 25.855361
#> Warning: non-integer x = 1.056937
#> Warning: non-integer x = 8.483707
#> Warning: non-integer x = 1.146274
#> Warning: non-integer x = 30.186454
#> Warning: non-integer x = 1.500087
#> Warning: non-integer x = 2.863375
#> Warning: non-integer x = 37.707195
#> Warning: non-integer x = 4.301492
#> Warning: non-integer x = 15.368839
#> Warning: non-integer x = 30.338450
#> Warning: non-integer x = 0.485228
#> Warning: non-integer x = 2.888844
#> Warning: non-integer x = 8.020378
#> Warning: non-integer x = 2.459590
#> Warning: non-integer x = 1.980464
#> Warning: non-integer x = 8.632915
#> Warning: non-integer x = 10.188272
#> Warning: non-integer x = 6.778494
#> Warning: non-integer x = 22.452214
#> Warning: non-integer x = 19.260306
#> Warning: non-integer x = 1.803888
#> Warning: non-integer x = 2.453469
#> Warning: non-integer x = 3.901051
#> Warning: non-integer x = 7.805106
#> Warning: non-integer x = 33.680919
#> Warning: non-integer x = 2.214593
#> Warning: non-integer x = 0.714661
#> Warning: non-integer x = 0.930203
#> Warning: non-integer x = 3.906258
#> Warning: non-integer x = 2.160568
#> Warning: non-integer x = 11.842329
#> Warning: non-integer x = 7.172202
#> Warning: non-integer x = 11.255095
#> Warning: non-integer x = 19.497638
#> Warning: non-integer x = 13.654405
#> Warning: non-integer x = 46.423392
#> Warning: non-integer x = 2.907548
#> Warning: non-integer x = 221.764410
#> Warning: non-integer x = 5.872531
#> Warning: non-integer x = 89.880158
#> Warning: non-integer x = 1.862678
#> Warning: non-integer x = 2.223235
#> Warning: non-integer x = 2.808808
#> Warning: non-integer x = 3.160776
#> Warning: non-integer x = 3.592084
#> Warning: non-integer x = 0.562477
#> Warning: non-integer x = 5.324549
#> Warning: non-integer x = 25.058758
#> Warning: non-integer x = 1.969118
#> Warning: non-integer x = 6.414099
#> Warning: non-integer x = 3.436304
#> Warning: non-integer x = 0.923121
#> Warning: non-integer x = 6.468439
#> Warning: non-integer x = 6.948440
#> Warning: non-integer x = 6.474924
#> Warning: non-integer x = 0.634992
#> Warning: non-integer x = 5.019471
#> Warning: non-integer x = 5.117912
#> Warning: non-integer x = 26.707128
#> Warning: non-integer x = 18.239353
#> Warning: non-integer x = 2.674946
#> Warning: non-integer x = 3.160395
#> Warning: non-integer x = 1.447988
#> Warning: non-integer x = 18.381746
#> Warning: non-integer x = 5.620461
#> Warning: non-integer x = 15.334152
#> Warning: non-integer x = 8.074570
#> Warning: non-integer x = 17.348318
#> Warning: non-integer x = 1.669264
#> Warning: non-integer x = 43.740414
#> Warning: non-integer x = 6.736525
#> Warning: non-integer x = 1.521056
#> Warning: non-integer x = 13.830425
#> Warning: non-integer x = 16.412711
#> Warning: non-integer x = 7.158482
#> Warning: non-integer x = 11.033936
#> Warning: non-integer x = 3.974806
#> Warning: non-integer x = 5.794421
#> Warning: non-integer x = 29.850769
#> Warning: non-integer x = 0.120463
#> Warning: non-integer x = 6.367720
#> Warning: non-integer x = 3.351470
#> Warning: non-integer x = 9.375087
#> Warning: non-integer x = 2.476457
#> Warning: non-integer x = 8.895316
#> Warning: non-integer x = 2.029732
#> Warning: non-integer x = 3.155250
#> Warning: non-integer x = 28.687961
#> Warning: non-integer x = 2.636979
#> Warning: non-integer x = 2.328582
#> Warning: non-integer x = 1.950025
#> Warning: non-integer x = 3.504351
#> Warning: non-integer x = 16.278301
#> Warning: non-integer x = 2.536992
#> Warning: non-integer x = 67.976378
#> Warning: non-integer x = 5.494885
#> Warning: non-integer x = 116.834949
#> Warning: non-integer x = 1.130737
#> Warning: non-integer x = 3.854832
#> Warning: non-integer x = 93.805855
#> Warning: non-integer x = 2.140814
#> Warning: non-integer x = 3.182215
#> Warning: non-integer x = 28.626763
#> Warning: non-integer x = 2.655845
#> Warning: non-integer x = 14.855996
#> Warning: non-integer x = 3.994964
#> Warning: non-integer x = 0.974428
#> Warning: non-integer x = 8.097073
#> Warning: non-integer x = 3.785956
#> Warning: non-integer x = 19.972730
#> Warning: non-integer x = 23.737286
#> Warning: non-integer x = 5.310943
#> Warning: non-integer x = 12.312290
#> Warning: non-integer x = 3.170563
#> Warning: non-integer x = 12.575392
#> Warning: non-integer x = 5.001687
#> Warning: non-integer x = 4.454498
#> Warning: non-integer x = 4.206505
#> Warning: non-integer x = 3.446934
#> Warning: non-integer x = 1.037048
#> Warning: non-integer x = 10.317880
#> Warning: non-integer x = 14.367131
#> Warning: non-integer x = 5.897746
#> Warning: non-integer x = 6.439230
#> Warning: non-integer x = 1.640016
#> Warning: non-integer x = 22.073353
#> Warning: non-integer x = 0.858079
#> Warning: non-integer x = 0.737533
#> Warning: non-integer x = 1.729382
#> Warning: non-integer x = 5.517710
#> Warning: non-integer x = 2.466484
#> Warning: non-integer x = 4.571721
#> Warning: non-integer x = 4.649715
#> Warning: non-integer x = 15.177833
#> Warning: non-integer x = 20.747174
#> Warning: non-integer x = 31.377376
#> Warning: non-integer x = 1.584559
#> Warning: non-integer x = 1.554617
#> Warning: non-integer x = 11.049988
#> Warning: non-integer x = 3.842616
#> Warning: non-integer x = 15.541789
#> Warning: non-integer x = 0.204645
#> Warning: non-integer x = 24.248745
#> Warning: non-integer x = 9.127369
#> Warning: non-integer x = 14.935240
#> Warning: non-integer x = 31.349502
#> Warning: non-integer x = 0.560190
#> Warning: non-integer x = 6.501434
#> Warning: non-integer x = 14.015350
#> Warning: non-integer x = 1.709711
#> Warning: non-integer x = 9.037297
#> Warning: non-integer x = 1.403620
#> Warning: non-integer x = 6.791662
#> Warning: non-integer x = 2.748809
#> Warning: non-integer x = 28.915186
#> Warning: non-integer x = 4.712848
#> Warning: non-integer x = 2.064238
#> Warning: non-integer x = 2.711968
#> Warning: non-integer x = 5.682559
#> Warning: non-integer x = 24.155549
#> Warning: non-integer x = 20.770016
#> Warning: non-integer x = 0.402289
#> Warning: non-integer x = 1.841405
#> Warning: non-integer x = 5.050823
#> Warning: non-integer x = 3.224044
#> Warning: non-integer x = 4.252084
#> Warning: non-integer x = 4.437292
#> Warning: non-integer x = 1.315416
#> Warning: non-integer x = 4.263496
#> Warning: non-integer x = 3.784158
#> Warning: non-integer x = 3.310777
#> Warning: non-integer x = 4.716711
#> Warning: non-integer x = 1.379666
#> Warning: non-integer x = 8.992770
#> Warning: non-integer x = 8.759763
#> Warning: non-integer x = 34.109530
#> Warning: non-integer x = 2.525449
#> Warning: non-integer x = 9.747778
#> Warning: non-integer x = 8.118187
#> Warning: non-integer x = 15.612430
#> Warning: non-integer x = 5.927597
#> Warning: non-integer x = 0.890923
#> Warning: non-integer x = 5.434579
#> Warning: non-integer x = 6.232813
#> Warning: non-integer x = 1.672638
#> Warning: non-integer x = 5.433659
#> Warning: non-integer x = 14.759053
#> Warning: non-integer x = 24.889429
#> Warning: non-integer x = 33.311369
#> Warning: non-integer x = 1.722837
#> Warning: non-integer x = 7.261753
#> Warning: non-integer x = 6.367081
#> Warning: non-integer x = 6.507715
#> Warning: non-integer x = 3.967287
#> Warning: non-integer x = 5.186078
#> Warning: non-integer x = 1.137085
#> Warning: non-integer x = 12.745448
#> Warning: non-integer x = 4.220578
#> Warning: non-integer x = 52.203049
#> Warning: non-integer x = 26.223081
#> Warning: non-integer x = 1.442262
#> Warning: non-integer x = 0.465134
#> Warning: non-integer x = 4.286391
#> Warning: non-integer x = 2.344476
#> Warning: non-integer x = 2.881268
#> Warning: non-integer x = 3.492148
#> Warning: non-integer x = 24.593063
#> Warning: non-integer x = 42.147159
#> Warning: non-integer x = 1.091511
#> Warning: non-integer x = 3.968655
#> Warning: non-integer x = 8.886063
#> Warning: non-integer x = 1.067032
#> Warning: non-integer x = 16.470294
#> Warning: non-integer x = 90.142466
#> Warning: non-integer x = 6.730750
#> Warning: non-integer x = 4.026099
#> Warning: non-integer x = 7.312671
#> Warning: non-integer x = 5.184773
#> Warning: non-integer x = 4.208718
#> Warning: non-integer x = 31.089427
#> Warning: non-integer x = 6.309495
#> Warning: non-integer x = 19.420207
#> Warning: non-integer x = 0.308941
#> Warning: non-integer x = 2.834251
#> Warning: non-integer x = 3.125445
#> Warning: non-integer x = 3.097088
#> Warning: non-integer x = 3.343681
#> Warning: non-integer x = 7.133640
#> Warning: non-integer x = 0.918634
#> Warning: non-integer x = 11.294755
#> Warning: non-integer x = 4.518292
#> Warning: non-integer x = 4.525765
#> Warning: non-integer x = 6.669693
#> Warning: non-integer x = 2.576901
#> Warning: non-integer x = 2.434697
#> Warning: non-integer x = 0.568859
#> Warning: non-integer x = 1.390007
#> Warning: non-integer x = 2.337041
#> Warning: non-integer x = 6.926049
#> Warning: non-integer x = 32.412105
#> Warning: non-integer x = 23.710329
#> Warning: non-integer x = 51.559349
#> Warning: non-integer x = 3.614930
#> Warning: non-integer x = 25.851226
#> Warning: non-integer x = 9.591226
#> Warning: non-integer x = 5.781760
#> Warning: non-integer x = 0.587468
#> Warning: non-integer x = 23.891779
#> Warning: non-integer x = 0.440312
#> Warning: non-integer x = 32.379587
#> Warning: non-integer x = 3.023443
#> Warning: non-integer x = 0.829132
#> Warning: non-integer x = 7.436117
#> Warning: non-integer x = 0.681257
#> Warning: non-integer x = 10.264559
#> Warning: non-integer x = 9.097521
#> Warning: non-integer x = 27.283833
#> Warning: non-integer x = 6.938711
#> Warning: non-integer x = 1.549468
#> Warning: non-integer x = 8.944017
#> Warning: non-integer x = 2.605657
#> Warning: non-integer x = 50.046441
#> Warning: non-integer x = 9.262039
#> Warning: non-integer x = 1.822786
#> Warning: non-integer x = 1.994296
#> Warning: non-integer x = 0.595799
#> Warning: non-integer x = 2.214162
#> Warning: non-integer x = 8.308474
#> Warning: non-integer x = 11.430020
#> Warning: non-integer x = 32.445425
#> Warning: non-integer x = 5.541763
#> Warning: non-integer x = 17.844194
#> Warning: non-integer x = 21.367848
#> Warning: non-integer x = 3.918709
#> Warning: non-integer x = 1.840330
#> Warning: non-integer x = 3.384189
#> Warning: non-integer x = 0.574850
#> Warning: non-integer x = 14.353827
#> Warning: non-integer x = 4.111037
#> Warning: non-integer x = 10.503800
#> Warning: non-integer x = 0.930664
#> Warning: non-integer x = 4.322300
#> Warning: non-integer x = 5.365941
#> Warning: non-integer x = 1.068578
#> Warning: non-integer x = 2.285779
#> Warning: non-integer x = 5.211691
#> Warning: non-integer x = 4.620198
#> Warning: non-integer x = 3.395709
#> Warning: non-integer x = 3.418190
#> Warning: non-integer x = 2.977589
#> Warning: non-integer x = 2.032596
#> Warning: non-integer x = 3.007671
#> Warning: non-integer x = 9.798787
#> Warning: non-integer x = 37.496746
#> Warning: non-integer x = 18.159342
#> Warning: non-integer x = 7.377741
#> Warning: non-integer x = 0.554142
#> Warning: non-integer x = 29.213664
#> Warning: non-integer x = 3.659216
#> Warning: non-integer x = 25.855361
#> Warning: non-integer x = 1.056937
#> Warning: non-integer x = 8.483707
#> Warning: non-integer x = 1.146274
#> Warning: non-integer x = 30.186454
#> Warning: non-integer x = 1.500087
#> Warning: non-integer x = 2.863375
#> Warning: non-integer x = 37.707195
#> Warning: non-integer x = 4.301492
#> Warning: non-integer x = 15.368839
#> Warning: non-integer x = 30.338450
#> Warning: non-integer x = 0.485228
#> Warning: non-integer x = 2.888844
#> Warning: non-integer x = 8.020378
#> Warning: non-integer x = 2.459590
#> Warning: non-integer x = 1.980464
#> Warning: non-integer x = 8.632915
#> Warning: non-integer x = 10.188272
#> Warning: non-integer x = 6.778494
#> Warning: non-integer x = 22.452214
#> Warning: non-integer x = 19.260306
#> Warning: non-integer x = 1.803888
#> Warning: non-integer x = 2.453469
#> Warning: non-integer x = 3.901051
#> Warning: non-integer x = 7.805106
#> Warning: non-integer x = 33.680919
#> Warning: non-integer x = 2.214593
#> Warning: non-integer x = 0.714661
#> Warning: non-integer x = 0.930203
#> Warning: non-integer x = 3.906258
#> Warning: non-integer x = 2.160568
#> Warning: non-integer x = 11.842329
#> Warning: non-integer x = 7.172202
#> Warning: non-integer x = 11.255095
#> Warning: non-integer x = 19.497638
#> Warning: non-integer x = 13.654405
#> Warning: non-integer x = 46.423392
#> Warning: non-integer x = 2.907548
#> Warning: non-integer x = 221.764410
#> Warning: non-integer x = 5.872531
#> Warning: non-integer x = 89.880158
#> Warning: non-integer x = 1.862678
#> Warning: non-integer x = 2.223235
#> Warning: non-integer x = 2.808808
#> Warning: non-integer x = 3.160776
#> Warning: non-integer x = 3.592084
#> Warning: non-integer x = 0.562477
#> Warning: non-integer x = 5.324549
#> Warning: non-integer x = 25.058758
#> Warning: non-integer x = 1.969118
#> Warning: non-integer x = 6.414099
#> Warning: non-integer x = 3.436304
#> Warning: non-integer x = 0.923121
#> Warning: non-integer x = 6.468439
#> Warning: non-integer x = 6.948440
#> Warning: non-integer x = 6.474924
#> Warning: non-integer x = 0.634992
#> Warning: non-integer x = 5.019471
#> Warning: non-integer x = 5.117912
#> Warning: non-integer x = 26.707128
#> Warning: non-integer x = 18.239353
#> Warning: non-integer x = 2.674946
#> Warning: non-integer x = 3.160395
#> Warning: non-integer x = 1.447988
#> Warning: non-integer x = 18.381746
#> Warning: non-integer x = 5.620461
#> Warning: non-integer x = 15.334152
#> Warning: non-integer x = 8.074570
#> Warning: non-integer x = 17.348318
#> Warning: non-integer x = 1.669264
#> Warning: non-integer x = 43.740414
#> Warning: non-integer x = 6.736525
#> Warning: non-integer x = 1.521056
#> Warning: non-integer x = 13.830425
#> Warning: non-integer x = 16.412711
#> Warning: non-integer x = 7.158482
#> Warning: non-integer x = 11.033936
#> Warning: non-integer x = 3.974806
#> Warning: non-integer x = 5.794421
#> Warning: non-integer x = 29.850769
#> Warning: non-integer x = 0.120463
#> Warning: non-integer x = 6.367720
#> Warning: non-integer x = 3.351470
#> Warning: non-integer x = 9.375087
#> Warning: non-integer x = 2.476457
#> Warning: non-integer x = 8.895316
#> Warning: non-integer x = 2.029732
#> Warning: non-integer x = 3.155250
#> Warning: non-integer x = 28.687961
#> Warning: non-integer x = 2.636979
#> Warning: non-integer x = 2.328582
#> Warning: non-integer x = 1.950025
#> Warning: non-integer x = 3.504351
#> Warning: non-integer x = 16.278301
#> Warning: non-integer x = 2.536992
#> Warning: non-integer x = 67.976378
#> Warning: non-integer x = 5.494885
#> Warning: non-integer x = 116.834949
#> Warning: non-integer x = 1.130737
#> Warning: non-integer x = 3.854832
#> Warning: non-integer x = 93.805855
#> Warning: non-integer x = 2.140814
#> Warning: non-integer x = 3.182215
#> Warning: non-integer x = 28.626763
#> Warning: non-integer x = 2.655845
#> Warning: non-integer x = 14.855996
#> Warning: non-integer x = 3.994964
#> Warning: non-integer x = 0.974428
#> Warning: non-integer x = 8.097073
#> Warning: non-integer x = 3.785956
#> Warning: non-integer x = 19.972730
#> Warning: non-integer x = 23.737286
#> Warning: non-integer x = 5.310943
#> Warning: non-integer x = 12.312290
#> Warning: non-integer x = 3.170563
#> Warning: non-integer x = 12.575392
#> Warning: non-integer x = 5.001687
#> Warning: non-integer x = 4.454498
#> Warning: non-integer x = 4.206505
#> Warning: non-integer x = 3.446934
#> Warning: non-integer x = 1.037048
#> Warning: non-integer x = 10.317880
#> Warning: non-integer x = 14.367131
#> Warning: non-integer x = 5.897746
#> Warning: non-integer x = 6.439230
#> Warning: non-integer x = 1.640016
#> Warning: non-integer x = 22.073353
#> Warning: non-integer x = 0.858079
#> Warning: non-integer x = 0.737533
#> Warning: non-integer x = 1.729382
#> Warning: non-integer x = 5.517710
#> Warning: non-integer x = 2.466484
#> Warning: non-integer x = 4.571721
#> Warning: non-integer x = 4.649715
#> Warning: non-integer x = 15.177833
#> Warning: non-integer x = 20.747174
#> Warning: non-integer x = 31.377376
#> Warning: non-integer x = 1.584559
#> Warning: non-integer x = 1.554617
#> Warning: non-integer x = 11.049988
#> Warning: non-integer x = 3.842616
#> Warning: non-integer x = 15.541789
#> Warning: non-integer x = 0.204645
#> Warning: non-integer x = 24.248745
#> Warning: non-integer x = 9.127369
#> Warning: non-integer x = 14.935240
#> Warning: non-integer x = 31.349502
#> Warning: non-integer x = 0.560190
#> Warning: non-integer x = 6.501434
#> Warning: non-integer x = 14.015350
#> Warning: non-integer x = 1.709711
#> Warning: non-integer x = 9.037297
#> Warning: non-integer x = 1.403620
#> Warning: non-integer x = 6.791662
#> Warning: non-integer x = 2.748809
#> Warning: non-integer x = 28.915186
#> Warning: non-integer x = 4.712848
#> Warning: non-integer x = 2.064238
#> Warning: non-integer x = 2.711968
#> Warning: non-integer x = 5.682559
#> Warning: non-integer x = 24.155549
#> Warning: non-integer x = 20.770016
#> Warning: non-integer x = 0.402289
#> Warning: non-integer x = 1.841405
#> Warning: non-integer x = 5.050823
#> Warning: non-integer x = 3.224044
#> Warning: non-integer x = 4.252084
#> Warning: non-integer x = 4.437292
#> Warning: non-integer x = 1.315416
#> Warning: non-integer x = 4.263496
#> Warning: non-integer x = 3.784158
#> Warning: non-integer x = 3.310777
#> Warning: non-integer x = 4.716711
#> Warning: non-integer x = 1.104265
#> Warning: non-integer x = 7.919213
#> Warning: non-integer x = 50.868674
#> Warning: non-integer x = 8.949183
#> Warning: non-integer x = 2.139805
#> Warning: non-integer x = 6.102525
#> Warning: non-integer x = 8.002045
#> Warning: non-integer x = 0.718477
#> Warning: non-integer x = 14.931544
#> Warning: non-integer x = 1.928922
#> Warning: non-integer x = 5.426006
#> Warning: non-integer x = 6.494737
#> Warning: non-integer x = 2.484975
#> Warning: non-integer x = 1.761061
#> Warning: non-integer x = 0.259656
#> Warning: non-integer x = 13.075359
#> Warning: non-integer x = 7.165198
#> Warning: non-integer x = 4.513789
#> Warning: non-integer x = 7.021639
#> Warning: non-integer x = 1.743629
#> Warning: non-integer x = 23.283677
#> Warning: non-integer x = 23.156794
#> Warning: non-integer x = 5.281936
#> Warning: non-integer x = 5.144274
#> Warning: non-integer x = 12.840140
#> Warning: non-integer x = 0.486697
#> Warning: non-integer x = 0.789096
#> Warning: non-integer x = 6.770614
#> Warning: non-integer x = 21.342376
#> Warning: non-integer x = 10.803288
#> Warning: non-integer x = 12.517683
#> Warning: non-integer x = 28.857352
#> Warning: non-integer x = 2.847297
#> Warning: non-integer x = 5.274391
#> Warning: non-integer x = 2.024934
#> Warning: non-integer x = 2.749401
#> Warning: non-integer x = 5.751182
#> Warning: non-integer x = 3.347953
#> Warning: non-integer x = 6.908365
#> Warning: non-integer x = 46.316172
#> Warning: non-integer x = 4.990218
#> Warning: non-integer x = 5.384568
#> Warning: non-integer x = 18.103424
#> Warning: non-integer x = 2.976855
#> Warning: non-integer x = 8.949281
#> Warning: non-integer x = 44.048560
#> Warning: non-integer x = 18.223202
#> Warning: non-integer x = 26.284345
#> Warning: non-integer x = 8.417836
#> Warning: non-integer x = 27.242677
#> Warning: non-integer x = 1.664811
#> Warning: non-integer x = 1.646878
#> Warning: non-integer x = 8.123590
#> Warning: non-integer x = 5.918330
#> Warning: non-integer x = 1.758534
#> Warning: non-integer x = 1.162916
#> Warning: non-integer x = 4.101683
#> Warning: non-integer x = 8.989270
#> Warning: non-integer x = 2.267531
#> Warning: non-integer x = 7.398008
#> Warning: non-integer x = 13.235489
#> Warning: non-integer x = 20.147807
#> Warning: non-integer x = 0.399203
#> Warning: non-integer x = 4.050189
#> Warning: non-integer x = 9.785063
#> Warning: non-integer x = 10.276437
#> Warning: non-integer x = 8.889298
#> Warning: non-integer x = 1.173682
#> Warning: non-integer x = 5.339098
#> Warning: non-integer x = 4.220216
#> Warning: non-integer x = 1.303209
#> Warning: non-integer x = 28.561019
#> Warning: non-integer x = 6.059364
#> Warning: non-integer x = 6.647814
#> Warning: non-integer x = 2.886577
#> Warning: non-integer x = 4.110663
#> Warning: non-integer x = 4.647538
#> Warning: non-integer x = 25.415880
#> Warning: non-integer x = 1.289413
#> Warning: non-integer x = 3.803754
#> Warning: non-integer x = 1.748924
#> Warning: non-integer x = 4.344645
#> Warning: non-integer x = 1.232111
#> Warning: non-integer x = 2.430880
#> Warning: non-integer x = 4.787903
#> Warning: non-integer x = 5.195788
#> Warning: non-integer x = 3.429225
#> Warning: non-integer x = 2.518457
#> Warning: non-integer x = 3.833162
#> Warning: non-integer x = 14.555944
#> Warning: non-integer x = 9.904466
#> Warning: non-integer x = 15.254395
#> Warning: non-integer x = 31.823928
#> Warning: non-integer x = 2.086693
#> Warning: non-integer x = 6.929372
#> Warning: non-integer x = 2.721435
#> Warning: non-integer x = 4.524350
#> Warning: non-integer x = 1.883623
#> Warning: non-integer x = 1.799379
#> Warning: non-integer x = 3.490504
#> Warning: non-integer x = 1.542476
#> Warning: non-integer x = 3.854532
#> Warning: non-integer x = 8.074735
#> Warning: non-integer x = 1.403057
#> Warning: non-integer x = 8.213901
#> Warning: non-integer x = 2.681960
#> Warning: non-integer x = 5.564036
#> Warning: non-integer x = 4.939675
#> Warning: non-integer x = 1.257270
#> Warning: non-integer x = 3.871754
#> Warning: non-integer x = 3.234345
#> Warning: non-integer x = 13.347719
#> Warning: non-integer x = 22.053908
#> Warning: non-integer x = 9.045686
#> Warning: non-integer x = 1.667867
#> Warning: non-integer x = 15.782353
#> Warning: non-integer x = 3.554150
#> Warning: non-integer x = 0.893476
#> Warning: non-integer x = 15.917308
#> Warning: non-integer x = 121.318591
#> Warning: non-integer x = 4.988757
#> Warning: non-integer x = 2.623824
#> Warning: non-integer x = 82.031416
#> Warning: non-integer x = 24.577137
#> Warning: non-integer x = 20.234383
#> Warning: non-integer x = 2.241246
#> Warning: non-integer x = 6.282080
#> Warning: non-integer x = 2.684431
#> Warning: non-integer x = 5.925638
#> Warning: non-integer x = 1.422502
#> Warning: non-integer x = 13.415673
#> Warning: non-integer x = 18.979782
#> Warning: non-integer x = 4.517712
#> Warning: non-integer x = 13.681782
#> Warning: non-integer x = 0.937208
#> Warning: non-integer x = 2.711473
#> Warning: non-integer x = 17.514493
#> Warning: non-integer x = 87.639255
#> Warning: non-integer x = 25.565151
#> Warning: non-integer x = 2.667981
#> Warning: non-integer x = 0.904149
#> Warning: non-integer x = 4.931854
#> Warning: non-integer x = 26.165571
#> Warning: non-integer x = 7.649673
#> Warning: non-integer x = 2.490384
#> Warning: non-integer x = 20.197073
#> Warning: non-integer x = 9.240964
#> Warning: non-integer x = 3.695444
#> Warning: non-integer x = 11.516838
#> Warning: non-integer x = 1.012107
#> Warning: non-integer x = 2.112012
#> Warning: non-integer x = 5.916117
#> Warning: non-integer x = 39.843872
#> Warning: non-integer x = 1.016778
#> Warning: non-integer x = 12.944056
#> Warning: non-integer x = 3.282890
#> Warning: non-integer x = 1.722230
#> Warning: non-integer x = 10.698143
#> Warning: non-integer x = 5.777307
#> Warning: non-integer x = 1.520508
#> Warning: non-integer x = 66.582897
#> Warning: non-integer x = 16.569916
#> Warning: non-integer x = 7.713537
#> Warning: non-integer x = 23.358753
#> Warning: non-integer x = 1.448476
#> Warning: non-integer x = 9.859528
#> Warning: non-integer x = 10.409080
#> Warning: non-integer x = 16.746821
#> Warning: non-integer x = 18.610737
#> Warning: non-integer x = 11.940038
#> Warning: non-integer x = 84.663227
#> Warning: non-integer x = 23.461690
#> Warning: non-integer x = 0.966060
#> Warning: non-integer x = 16.835959
#> Warning: non-integer x = 2.140078
#> Warning: non-integer x = 2.018502
#> Warning: non-integer x = 1.677705
#> Warning: non-integer x = 4.096655
#> Warning: non-integer x = 11.806068
#> Warning: non-integer x = 1.148113
#> Warning: non-integer x = 7.245077
#> Warning: non-integer x = 76.839427
#> Warning: non-integer x = 15.801467
#> Warning: non-integer x = 19.228485
#> Warning: non-integer x = 1.089101
#> Warning: non-integer x = 9.598470
#> Warning: non-integer x = 1.543240
#> Warning: non-integer x = 2.758809
#> Warning: non-integer x = 3.212038
#> Warning: non-integer x = 1.468083
#> Warning: non-integer x = 1.364066
#> Warning: non-integer x = 3.500614
#> Warning: non-integer x = 22.601345
#> Warning: non-integer x = 6.202327
#> Warning: non-integer x = 3.825080
#> Warning: non-integer x = 4.368577
#> Warning: non-integer x = 12.928195
#> Warning: non-integer x = 3.485382
#> Warning: non-integer x = 21.787328
#> Warning: non-integer x = 4.555680
#> Warning: non-integer x = 5.576395
#> Warning: non-integer x = 3.400998
#> Warning: non-integer x = 8.549199
#> Warning: non-integer x = 8.387405
#> Warning: non-integer x = 0.703785
#> Warning: non-integer x = 4.828384
#> Warning: non-integer x = 8.182086
#> Warning: non-integer x = 16.759511
#> Warning: non-integer x = 42.079714
#> Warning: non-integer x = 4.981468
#> Warning: non-integer x = 1.931663
#> Warning: non-integer x = 25.341919
#> Warning: non-integer x = 0.734310
#> Warning: non-integer x = 19.832904
#> Warning: non-integer x = 7.073000
#> Warning: non-integer x = 5.258676
#> Warning: non-integer x = 17.158414
#> Warning: non-integer x = 2.191771
#> Warning: non-integer x = 12.526533
#> Warning: non-integer x = 3.438589
#> Warning: non-integer x = 1.385116
#> Warning: non-integer x = 1.204226
#> Warning: non-integer x = 5.043795
#> Warning: non-integer x = 2.924918
#> Warning: non-integer x = 7.024478
#> Warning: non-integer x = 2.964529
#> Warning: non-integer x = 23.039724
#> Warning: non-integer x = 2.684376
#> Warning: non-integer x = 2.047380
#> Warning: non-integer x = 16.264687
#> Warning: non-integer x = 1.256923
#> Warning: non-integer x = 0.438650
#> Warning: non-integer x = 7.160958
#> Warning: non-integer x = 1.038537
#> Warning: non-integer x = 1.329141
#> Warning: non-integer x = 2.027007
#> Warning: non-integer x = 17.924501
#> Warning: non-integer x = 6.723155
#> Warning: non-integer x = 13.847878
#> Warning: non-integer x = 0.433201
#> Warning: non-integer x = 4.033492
#> Warning: non-integer x = 5.664656
#> Warning: non-integer x = 21.169146
#> Warning: non-integer x = 0.901422
#> Warning: non-integer x = 41.138724
#> Warning: non-integer x = 3.859549
#> Warning: non-integer x = 5.273733
#> Warning: non-integer x = 8.766063
#> Warning: non-integer x = 23.695136
#> Warning: non-integer x = 2.054392
#> Warning: non-integer x = 9.929483
#> Warning: non-integer x = 10.176348
#> Warning: non-integer x = 20.900526
#> Warning: non-integer x = 16.956871
#> Warning: non-integer x = 18.731321
#> Warning: non-integer x = 24.691589
#> Warning: non-integer x = 1.459265
#> Warning: non-integer x = 10.792331
#> Warning: non-integer x = 20.205193
#> Warning: non-integer x = 9.288893
#> Warning: non-integer x = 5.612942
#> Warning: non-integer x = 10.902820
#> Warning: non-integer x = 1.324159
#> Warning: non-integer x = 15.524681
#> Warning: non-integer x = 1.404709
#> Warning: non-integer x = 0.767350
#> Warning: non-integer x = 11.613074
#> Warning: non-integer x = 3.298842
#> Warning: non-integer x = 15.819110
#> Warning: non-integer x = 10.080630
#> Warning: non-integer x = 3.618375
#> Warning: non-integer x = 3.998164
#> Warning: non-integer x = 27.544211
#> Warning: non-integer x = 3.667604
#> Warning: non-integer x = 4.516977
#> Warning: non-integer x = 3.490485
#> Warning: non-integer x = 17.046035
#> Warning: non-integer x = 2.603819
#> Warning: non-integer x = 2.021155
#> Warning: non-integer x = 38.694601
#> Warning: non-integer x = 3.985524
#> Warning: non-integer x = 1.704037
#> Warning: non-integer x = 7.905456
#> Warning: non-integer x = 6.193858
#> Warning: non-integer x = 6.058184
#> Warning: non-integer x = 20.197876
#> Warning: non-integer x = 52.424237
#> Warning: non-integer x = 1.035137
#> Warning: non-integer x = 1.451650
#> Warning: non-integer x = 17.330004
#> Warning: non-integer x = 1.957635
#> Warning: non-integer x = 2.394202
#> Warning: non-integer x = 11.494718
#> Warning: non-integer x = 40.081222
#> Warning: non-integer x = 9.245043
#> Warning: non-integer x = 32.514300
#> Warning: non-integer x = 1.150090
#> Warning: non-integer x = 1.073191
#> Warning: non-integer x = 0.573021
#> Warning: non-integer x = 1.084781
#> $est
#> [1] 3.45649961 0.03720655
#> 
#> $se
#> (Intercept)          X1 
#>   0.4518837   0.2745312 
#>