PPI Logistic Regression

Helper function for PPI logistic regression

Usage

ppi_logistic(X_l, Y_l, f_l, X_u, f_u, opts = NULL)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

opts

(list, optional): Options to pass to the optimizer. See ?optim for details.

Value

(list): A list containing the following:

est

(vector): vector of PPI logistic regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

rectifier_est

(vector): vector of the rectifier logistic regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

Prediction Powered Inference (Angelopoulos et al., 2023) doi:10.1126/science.adi6000

Examples

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_logistic(X_l, Y_l, f_l, X_u, f_u)
#> $est
#>           [,1]
#> [1,] 0.6376333
#> [2,] 0.7992929
#> 
#> $se
#> [1] 0.1858541 0.1867031
#> 
#> $rectifier_est
#>             [,1]
#> [1,] 0.031445852
#> [2,] 0.001260279
#> 
#> $var_u
#>             [,1]        [,2]
#> [1,]  0.14690720 -0.05846497
#> [2,] -0.05846497  0.11895860
#> 
#> $var_l
#>             [,1]        [,2]
#> [1,]  0.28284281 -0.06386218
#> [2,] -0.06386218  0.20712022
#> 
#> $grads
#>               [,1]          [,2]
#>   [1,]  0.30716167 -0.5576287251
#>   [2,] -0.11726528 -0.2026013263
#>   [3,]  0.64764090 -0.0234553741
#>   [4,] -0.10224864 -0.1963442443
#>   [5,] -0.09008143 -0.1887757184
#>   [6,] -0.36429402  0.0368528687
#>   [7,]  0.56833806 -0.2577983684
#>   [8,] -0.35195073  0.0119582998
#>   [9,] -0.27265114 -0.1172003877
#>  [10,]  0.75609345  0.4670609605
#>  [11,] -0.36274530  0.0336584138
#>  [12,] -0.37190341  0.0528446639
#>  [13,]  0.68212499  0.1074626522
#>  [14,] -0.16878350 -0.2020098926
#>  [15,]  0.19635290 -0.5028329369
#>  [16,] -0.17445278 -0.2000908122
#>  [17,] -0.35262428  0.0132834362
#>  [18,]  0.75300328  0.4494331102
#>  [19,] -0.29513482 -0.0860122775
#>  [20,] -0.45590130  0.2628194266
#>  [21,]  0.69461579  0.1600375712
#>  [22,] -0.12367595 -0.2043135840
#>  [23,] -0.57371590  0.6708807082
#>  [24,] -0.13259993 -0.2057997135
#>  [25,] -0.18111311 -0.1974046000
#>  [26,] -0.18428158 -0.1959657077
#>  [27,] -0.62585230  0.9021005417
#>  [28,] -0.19859027 -0.1882053413
#>  [29,]  0.72353631  0.2936879023
#>  [30,] -0.54246367  0.5483038008
#>  [31,] -0.32691403 -0.0345782490
#>  [32,] -0.20610240 -0.1833221596
#>  [33,] -0.40312498  0.1236530568
#>  [34,] -0.30150807 -0.0763843214
#>  [35,]  0.21566833 -0.5204156870
#>  [36,]  0.40489286 -0.5180892526
#>  [37,]  0.66555358  0.0420578804
#>  [38,]  0.73421469  0.3476636114
#>  [39,] -0.38037784  0.0712356774
#>  [40,] -0.35609110  0.0201648277
#>  [41,] -0.74688794  1.6069642887
#>  [42,]  0.63524193 -0.0658547108
#>  [43,]  0.58864897 -0.2056571727
#>  [44,]  0.30381575 -0.5575489015
#>  [45,]  0.22521302 -0.5277956440
#>  [46,] -0.21590345 -0.1761352240
#>  [47,] -0.54618475  0.5623165450
#>  [48,]  0.40720236 -0.5161658306
#>  [49,] -0.16151432 -0.2039649932
#>  [50,] -0.22301422 -0.1703531314
#>  [51,]  0.48545665 -0.4226134405
#>  [52,] -0.30792999 -0.0663333971
#>  [53,]  0.48502254 -0.4232901439
#>  [54,] -0.28558159 -0.0997946510
#>  [55,]  0.39430538 -0.5263125585
#>  [56,]  0.72582674  0.3050423013
#>  [57,]  0.55002490 -0.3006219002
#>  [58,] -0.09522456 -0.1922635321
#>  [59,] -0.43390036  0.2017673871
#>  [60,] -0.08014149 -0.1807578931
#>  [61,] -0.46284356  0.2830087788
#>  [62,] -0.23353497 -0.1609380664
#>  [63,] -0.45189083  0.2513601405
#>  [64,]  0.83246444  1.0056334438
#>  [65,] -0.16822253 -0.2021811454
#>  [66,]  0.80958564  0.8201132946
#>  [67,]  0.60056029 -0.1726883950
#>  [68,]  0.49997442 -0.3989169332
#>  [69,] -0.40615761  0.1309800450
#>  [70,] -0.73811877  1.5457384493
#>  [71,] -0.10567445 -0.1980601627
#>  [72,] -0.21410000 -0.1775262026
#>  [73,] -0.36932973  0.0473806732
#>  [74,] -0.47189939  0.3100242038
#>  [75,] -0.37918459  0.0686089594
#>  [76,]  0.80069516  0.7543325211
#>  [77,] -0.33647795 -0.0174265324
#>  [78,]  0.74434064  0.4013864869
#>  [79,] -0.50633457  0.4199788109
#>  [80,]  0.48052312 -0.4301962756
#>  [81,] -0.43126365  0.1947453033
#>  [82,] -0.28028270 -0.1071018618
#>  [83,]  0.75414291  0.4559035635
#>  [84,]  0.20119289 -0.5075775939
#>  [85,] -0.32080502 -0.0451296123
#>  [86,] -0.29809207 -0.0815878357
#>  [87,]  0.68646355  0.1253940699
#>  [88,] -0.61018241  0.8288357504
#>  [89,] -0.41095808  0.1427425611
#>  [90,]  0.16359689 -0.4644814381
#>  [91,]  0.33783770 -0.5539430641
#>  [92,] -0.42105024  0.1681317057
#>  [93,] -0.35670410  0.0213921916
#>  [94,]  0.53893232 -0.3247162429
#>  [95,] -0.06852129 -0.1690540139
#>  [96,] -0.48438892  0.3485647132
#>  [97,]  0.64608037 -0.0289208546
#>  [98,]  0.64093884 -0.0466631147
#>  [99,] -0.26618953 -0.1253559361
#> [100,] -0.12731314 -0.2050429952
#> [101,] -0.60134907  0.7890078612
#> [102,] -0.52791365  0.4949632748
#> [103,] -0.24679964 -0.1476306529
#> [104,] -0.59499534  0.7609935261
#> [105,] -0.19271252 -0.1916405925
#> [106,]  0.70653164  0.2129999124
#> [107,]  0.24207623 -0.5387821322
#> [108,] -0.10296815 -0.1967192641
#> [109,]  0.39931926 -0.5225396241
#> [110,]  0.77627697  0.5890038462
#> [111,]  0.54004452 -0.3223616492
#> [112,]  0.60483377 -0.1604122440
#> [113,] -0.34458892 -0.0022750942
#> [114,] -0.43916018  0.2159624239
#> [115,] -0.67583126  1.1603401414
#> [116,] -0.24991880 -0.1442719676
#> [117,] -0.52778268  0.4944934796
#> [118,]  0.40911141 -0.5145407046
#> [119,] -0.44656935  0.2363835459
#> [120,] -0.27340260 -0.1162283819
#> [121,] -0.36854928  0.0457349199
#> [122,] -0.14440825 -0.2062376366
#> [123,] -0.11339706 -0.2012973083
#> [124,]  0.59212947 -0.1962110192
#> [125,]  0.31984722 -0.5570694059
#> [126,] -0.36220378  0.0325462550
#> [127,]  0.52824990 -0.3466489290
#> [128,] -0.13637634 -0.2061250379
#> [129,] -0.40268262  0.1225909982
#> [130,]  0.25966545 -0.5475150705
#> [131,]  0.51932386 -0.3640427482
#> [132,] -0.12225408 -0.2039812580
#> [133,] -0.33409100 -0.0217795703
#> [134,] -0.47259086  0.3121188793
#> [135,] -0.18936970 -0.1934411250
#> [136,] -0.10787118 -0.1990687627
#> [137,] -0.09315776 -0.1909140045
#> [138,] -0.22925980 -0.1648867315
#> [139,] -0.50408309  0.4124310018
#> [140,] -0.30250258 -0.0748507513
#> [141,] -0.26682107 -0.1245748656
#> [142,] -0.14482237 -0.2062224959
#> [143,] -0.36025799  0.0285706102
#> [144,]  0.80357439  0.7752834348
#> [145,] -0.43556296  0.2062273847
#> [146,] -0.33854351 -0.0136207773
#> [147,] -0.25735588 -0.1359151828
#> [148,]  0.19027666 -0.4965488170
#> [149,]  0.91567294  2.0017398257
#> [150,] -0.37715026  0.0641588864
#> [151,] -0.14322226 -0.2062697554
#> [152,]  0.46871071 -0.4474014203
#> [153,] -0.16029330 -0.2042367138
#> [154,] -0.07586006 -0.1767527377
#> [155,] -0.33685213 -0.0167397864
#> [156,] -0.18659859 -0.1948485912
#> [157,]  0.76470701  0.5176173751
#> [158,] -0.15905353 -0.2044957213
#> [159,] -0.31423205 -0.0561299352
#> [160,] -0.39539579  0.1053407785
#> [161,] -0.17287490 -0.2006591763
#> [162,] -0.14362990 -0.2062605991
#> [163,] -0.27033487 -0.1201656363
#> [164,]  0.66674468  0.0466008049
#> [165,]  0.51312328 -0.3756355581
#> [166,] -0.21719137 -0.1751230998
#> [167,]  0.78398747  0.6389494380
#> [168,]  0.65413346 -0.0003067711
#> [169,] -0.26309876 -0.1291282632
#> [170,] -0.41832784  0.1611943532
#> [171,] -0.35267009  0.0133737069
#> [172,] -0.10894530 -0.1995362633
#> [173,] -0.27917476 -0.1085991735
#> [174,] -0.38499390  0.0815120016
#> [175,] -0.29745948 -0.0825405448
#> [176,]  0.69436808  0.1589662939
#> [177,] -0.11894624 -0.2031038389
#> [178,] -0.20360736 -0.1850047970
#> [179,]  0.59654686 -0.1840003366
#> [180,]  0.67054682  0.0612633374
#> [181,] -0.45778461  0.2682520763
#> [182,] -0.26147028 -0.1310822180
#> [183,] -0.46372816  0.2856136981
#> [184,] -0.33942062 -0.0119938137
#> [185,] -0.42932196  0.1896140595
#> [186,] -0.36886321  0.0463962808
#> [187,] -0.40722044  0.1335668932
#> [188,]  0.52202882 -0.3588602041
#> [189,] -0.12190379 -0.2038952751
#> [190,] -0.26819094 -0.1228686953
#> [191,] -0.18685770 -0.1947202748
#> [192,] -0.57987736  0.6963981833
#> [193,] -0.30296819 -0.0741298671
#> [194,] -0.44561119  0.2337145115
#> [195,] -0.46718195  0.2958545462
#> [196,] -0.21824888 -0.1742803869
#> [197,]  0.41257814 -0.5115079237
#> [198,]  0.53489862 -0.3331427575
#> [199,]  0.63396679 -0.0700841789
#> [200,] -0.08413224 -0.1841861720
#> [201,] -0.39806393  0.1116036342
#> [202,]  0.54900679 -0.3028903642
#> [203,] -0.13149664 -0.2056710663
#> [204,]  0.59627504 -0.1847589265
#> [205,] -0.30267252 -0.0745878427
#> [206,] -0.46550926  0.2908808389
#> [207,]  0.34849183 -0.5508026438
#> [208,]  0.68677298  0.1266861826
#> [209,]  0.61156067 -0.1405967630
#> [210,]  0.39380395 -0.5266779060
#> [211,] -0.27466857 -0.1145797869
#> [212,]  0.45841120 -0.4613251326
#> [213,] -0.33922229 -0.0123622721
#> [214,]  0.51009435 -0.3811544276
#> [215,]  0.42453433 -0.5002357556
#> [216,] -0.16398768 -0.2033643260
#> [217,] -0.60239659  0.7936769219
#> [218,] -0.20937821 -0.1810219909
#> [219,]  0.53857661 -0.3254664402
#> [220,] -0.47670899  0.3246884797
#> [221,]  0.67457684  0.0770761002
#> [222,]  0.56378856 -0.2687990005
#> [223,]  0.38595618 -0.5321132215
#> [224,] -0.43485840  0.2043343453
#> [225,]  0.81709865  0.8783194614
#> [226,] -0.27741707 -0.1109528412
#> [227,] -0.46522480  0.2900376252
#> [228,]  0.76289338  0.5067946477
#> [229,]  0.80969210  0.8209209108
#> [230,]  0.59743945 -0.1815025800
#> [231,]  0.73234088  0.3379981570
#> [232,]  0.55983821 -0.2781549950
#> [233,] -0.26219959 -0.1302100214
#> [234,] -0.19163857 -0.1922312387
#> [235,]  0.62916594 -0.0857941022
#> [236,]  0.76257545  0.5049073293
#> [237,] -0.27245601 -0.1174519834
#> [238,] -0.24941779 -0.1448173035
#> [239,]  0.52696142 -0.3492116860
#> [240,] -0.36423917  0.0367393698
#> [241,] -0.29665301 -0.0837501770
#> [242,]  0.64558182 -0.0306590123
#> [243,] -0.63307365  0.9370324496
#> [244,]  0.55699528 -0.2847766676
#> [245,] -0.08438384 -0.1843927346
#> [246,] -0.15266327 -0.2055577170
#> [247,] -0.24171800 -0.1529161313
#> [248,] -0.53402488  0.5170883261
#> [249,] -0.37510759  0.0597261633
#> [250,]  0.18069230 -0.4858813050
#> [251,] -0.22871205 -0.1653805424
#> [252,] -0.63369522  0.9400746312
#> [253,] -0.45809013  0.2691364992
#> [254,]  0.52273939 -0.3574860924
#> [255,] -0.31209594 -0.0596261473
#> [256,] -0.16079982 -0.2041259940
#> [257,] -0.45022086  0.2466322268
#> [258,]  0.51220443 -0.3773196898
#> [259,] -0.27944945 -0.1082289299
#> [260,] -0.36704911  0.0425860096
#> [261,] -0.25397403 -0.1397759603
#> [262,]  0.49438481 -0.4082870119
#> [263,]  0.51168840 -0.3782617128
#> [264,] -0.35105332  0.0101986919
#> [265,] -0.24464427 -0.1499007564
#> [266,] -0.42919199  0.1892717835
#> [267,] -0.21012110 -0.1804860611
#> [268,]  0.81405574  0.8544438631
#> [269,]  0.61186801 -0.1396768910
#> [270,] -0.09203637 -0.1901526672
#> [271,] -0.44840196  0.2415118564
#> [272,] -0.38185370  0.0745013800
#> [273,] -0.35467220  0.0173361201
#> [274,] -0.39805386  0.1115798894
#> [275,]  0.70893957  0.2240510930
#> [276,] -0.13781931 -0.2062028426
#> [277,] -0.31130857 -0.0609051009
#> [278,] -0.71785341  1.4113554885
#> [279,]  0.48646020 -0.4210420611
#> [280,] -0.40814965  0.1358366305
#> [281,]  0.47634967 -0.4364274298
#> [282,]  0.44008343 -0.4836702954
#> [283,] -0.62632375  0.9043581410
#> [284,] -0.18896577 -0.1936511055
#> [285,] -0.08187239 -0.1822802316
#> [286,]  0.68566166  0.1220537694
#> [287,]  0.40583707 -0.5173085100
#> [288,]  0.55255819 -0.2949268383
#> [289,]  0.71491581  0.2520044276
#> [290,] -0.49725459  0.3898512263
#> [291,] -0.47436070  0.3175010502
#> [292,] -0.12858169 -0.2052570627
#> [293,] -0.37764767  0.0652436930
#> [294,] -0.54751430  0.5673605987
#> [295,] -0.29254502 -0.0898256838
#> [296,]  0.51806874 -0.3664215437
#> [297,] -0.15949276 -0.2044059103
#> [298,] -0.05165174 -0.1468571903
#> [299,] -0.06701923 -0.1673417556
#> [300,] -0.22137895 -0.1717246880
#> 
#> $grads_hat_unlabeled
#>               [,1]         [,2]
#>   [1,] -0.32815882 -0.032389608
#>   [2,] -0.25517978 -0.138411045
#>   [3,] -0.50132152  0.403243039
#>   [4,] -0.36994756  0.048687203
#>   [5,] -0.32760423 -0.033366325
#>   [6,]  0.49964863 -0.399471607
#>   [7,] -0.33286246 -0.024001283
#>   [8,]  0.68484637  0.118669816
#>   [9,] -0.36341158  0.035030200
#>  [10,]  0.43669966 -0.487460416
#>  [11,] -0.38771652  0.087658916
#>  [12,] -0.38441798  0.080219895
#>  [13,] -0.19421307 -0.190796097
#>  [14,] -0.04616287 -0.138073061
#>  [15,] -0.28064304 -0.106612609
#>  [16,]  0.53783664 -0.327022650
#>  [17,] -0.39865023  0.112988116
#>  [18,] -0.25880735 -0.134227218
#>  [19,]  0.31295490 -0.557542369
#>  [20,] -0.43522963  0.205331212
#>  [21,] -0.12762645 -0.205097790
#>  [22,]  0.41013496 -0.513656251
#>  [23,]  0.41445446 -0.509822449
#>  [24,] -0.37627577  0.062256808
#>  [25,] -0.10265248 -0.196555698
#>  [26,] -0.28437319 -0.101482290
#>  [27,] -0.11836994 -0.202935893
#>  [28,] -0.33797670 -0.014668698
#>  [29,] -0.22919758 -0.164942959
#>  [30,] -0.15286930 -0.205530682
#>  [31,]  0.67807731  0.091041681
#>  [32,]  0.19084985 -0.497157339
#>  [33,] -0.29918674 -0.079931170
#>  [34,] -0.26318867 -0.129019710
#>  [35,] -0.23514383 -0.159408853
#>  [36,] -0.30505013 -0.070883968
#>  [37,]  0.55079386 -0.298900842
#>  [38,]  0.50323438 -0.393307978
#>  [39,]  0.47094385 -0.444250720
#>  [40,] -0.08845481 -0.187581199
#>  [41,]  0.62252934 -0.106963637
#>  [42,] -0.42295441  0.173023064
#>  [43,]  0.43483968 -0.489499704
#>  [44,] -0.25309936 -0.140758045
#>  [45,] -0.17657089 -0.199286688
#>  [46,]  0.56885814 -0.256525304
#>  [47,] -0.51185294  0.438696306
#>  [48,]  0.56916515 -0.255772304
#>  [49,]  0.48149623 -0.428719266
#>  [50,] -0.45731239  0.266886828
#>  [51,] -0.22959851 -0.164579997
#>  [52,]  0.58637710 -0.211740905
#>  [53,]  0.33069889 -0.555510801
#>  [54,] -0.56011611  0.616158784
#>  [55,] -0.22862463 -0.165459096
#>  [56,]  0.38465748 -0.532961554
#>  [57,]  0.44606436 -0.476716893
#>  [58,] -0.46117754  0.278122703
#>  [59,] -0.36531870  0.038977607
#>  [60,] -0.28728601 -0.097392965
#>  [61,] -0.61313254  0.842370283
#>  [62,]  0.37654894 -0.537931045
#>  [63,] -0.05863096 -0.156862056
#>  [64,] -0.10576081 -0.198101154
#>  [65,] -0.25020727 -0.143956969
#>  [66,] -0.41310460  0.148067478
#>  [67,] -0.23203863 -0.162339149
#>  [68,] -0.23024958 -0.163987437
#>  [69,]  0.55634647 -0.286274864
#>  [70,] -0.40752933  0.134320564
#>  [71,] -0.21220809 -0.178952330
#>  [72,]  0.24052478 -0.537879264
#>  [73,] -0.45427441  0.258152944
#>  [74,] -0.28518895 -0.100344382
#>  [75,]  0.30122063 -0.557421498
#>  [76,] -0.15566754 -0.205115751
#>  [77,]  0.59062239 -0.200320049
#>  [78,] -0.28878398 -0.095261622
#>  [79,]  0.41950773 -0.505128825
#>  [80,]  0.66254222  0.030678495
#>  [81,]  0.42434310 -0.500426017
#>  [82,] -0.23287835 -0.161555406
#>  [83,] -0.33874188 -0.013253377
#>  [84,] -0.22312240 -0.170261523
#>  [85,] -0.41575476  0.154697720
#>  [86,] -0.45191318  0.251423615
#>  [87,]  0.64832427 -0.021050079
#>  [88,]  0.24794284 -0.542000878
#>  [89,] -0.27037271 -0.120117566
#>  [90,]  0.33039900 -0.555567239
#>  [91,]  0.43575819 -0.488496544
#>  [92,]  0.55778938 -0.282936401
#>  [93,]  0.47610938 -0.436781100
#>  [94,] -0.37458670  0.058601495
#>  [95,] -0.14512791 -0.206210026
#>  [96,] -0.22259232 -0.170709374
#>  [97,] -0.15107801 -0.205749486
#>  [98,] -0.23460448 -0.159924126
#>  [99,] -0.09473544 -0.191950397
#> [100,] -0.64479783  0.995386068
#> [101,] -0.17789822 -0.198758861
#> [102,]  0.69683667  0.169695626
#> [103,]  0.37125798 -0.540870532
#> [104,]  0.37791832 -0.537131306
#> [105,] -0.30207876 -0.075505327
#> [106,] -0.13013690 -0.205491385
#> [107,] -0.46915807  0.301764540
#> [108,] -0.27026193 -0.120258260
#> [109,] -0.09259940 -0.190537499
#> [110,] -0.36929116  0.047299223
#> [111,] -0.57478848  0.675289771
#> [112,] -0.38216202  0.075185966
#> [113,]  0.37326270 -0.539784886
#> [114,] -0.65498849  1.047821327
#> [115,]  0.27399740 -0.552616554
#> [116,] -0.14326730 -0.206268841
#> [117,] -0.43340813  0.200451738
#> [118,] -0.30896185 -0.064685795
#> [119,]  0.76757486  0.534929912
#> [120,] -0.39813845  0.111779436
#> [121,]  0.45762103 -0.462352379
#> [122,] -0.25950921 -0.133404345
#> [123,] -0.53273252  0.512376079
#> [124,] -0.12445852 -0.204485113
#> [125,] -0.35162841  0.011325526
#> [126,] -0.26122705 -0.131372069
#> [127,]  0.44497511 -0.478007523
#> [128,] -0.52066693  0.469241521
#> [129,] -0.32414922 -0.039392649
#> [130,] -0.32531248 -0.037374913
#> [131,] -0.20969502 -0.180794089
#> [132,] -0.08667019 -0.186218814
#> [133,] -0.13442478 -0.205979034
#> [134,] -0.54775364  0.568270668
#> [135,] -0.19139776 -0.192362095
#> [136,] -0.16214189 -0.203818942
#> [137,] -0.20671528 -0.182899648
#> [138,] -0.26093574 -0.131718525
#> [139,]  0.25075048 -0.543432509
#> [140,] -0.34966044  0.007481108
#> [141,]  0.38399136 -0.533391053
#> [142,] -0.16537752 -0.202997450
#> [143,]  0.29855390 -0.557230900
#> [144,] -0.51657318  0.454954409
#> [145,] -0.31051322 -0.062191687
#> [146,] -0.50153804  0.403960643
#> [147,] -0.34342562 -0.004482298
#> [148,]  0.66969874  0.057971451
#> [149,] -0.19804018 -0.188541295
#> [150,] -0.36640893  0.041248083
#> [151,] -0.32749934 -0.033550764
#> [152,] -0.45851416  0.270365415
#> [153,]  0.41058441 -0.513264973
#> [154,] -0.47701779  0.325637566
#> [155,]  0.45686131 -0.463334571
#> [156,] -0.30290265 -0.074231443
#> [157,] -0.31745819 -0.050776444
#> [158,] -0.33900545 -0.012764729
#> [159,] -0.22165922 -0.171491380
#> [160,] -0.24314025 -0.151460180
#> [161,] -0.34419085 -0.003031663
#> [162,] -0.29773020 -0.082133250
#> [163,] -0.32067787 -0.045345883
#> [164,] -0.18340892 -0.196372270
#> [165,] -0.54348370  0.552129788
#> [166,] -0.27428285 -0.115083559
#> [167,] -0.45059192  0.247680539
#> [168,] -0.36702309  0.042531559
#> [169,] -0.13365288 -0.205908253
#> [170,] -0.33809352 -0.014452945
#> [171,] -0.44538121  0.233075132
#> [172,]  0.75612344  0.467233337
#> [173,] -0.28894632 -0.095029482
#> [174,] -0.34252943 -0.006174915
#> [175,] -0.40009501  0.116412517
#> [176,] -0.41260005  0.146812200
#> [177,] -0.32849334 -0.031799201
#> [178,]  0.44268685 -0.480683593
#> [179,]  0.30282693 -0.557507108
#> [180,] -0.09443101 -0.191753551
#> [181,] -0.29148778 -0.091365964
#> [182,] -0.53214357  0.510234585
#> [183,] -0.17742816 -0.198947886
#> [184,] -0.09642694 -0.193017010
#> [185,]  0.60159252 -0.169745147
#> [186,] -0.26664960 -0.124787290
#> [187,]  0.10914205 -0.373755489
#> [188,] -0.30385735 -0.072748101
#> [189,] -0.74354798  1.583413940
#> [190,]  0.69722523  0.171395244
#> [191,] -0.37204066  0.053137628
#> [192,] -0.42611343  0.181208866
#> [193,] -0.21002342 -0.180556830
#> [194,] -0.12325510 -0.204218005
#> [195,] -0.16421879 -0.203304779
#> [196,]  0.63308719 -0.072987723
#> [197,]  0.74980033  0.431436275
#> [198,]  0.44672049 -0.475934221
#> [199,] -0.18417974 -0.196013555
#> [200,] -0.25488564 -0.138745196
#> [201,]  0.91585565  2.004853238
#> [202,] -0.46825413  0.299056532
#> [203,] -0.35061150  0.009334904
#> [204,]  0.69476334  0.160676261
#> [205,] -0.45847902  0.270263507
#> [206,]  0.28149011 -0.554573672
#> [207,] -0.26601316 -0.125573449
#> [208,] -0.61254619  0.839670803
#> [209,]  0.53464040 -0.333676164
#> [210,] -0.28778908 -0.096679325
#> [211,]  0.46750060 -0.449089044
#> [212,]  0.39310410 -0.527184197
#> [213,] -0.30673518 -0.068229901
#> [214,] -0.18822101 -0.194033955
#> [215,] -0.25164399 -0.142377098
#> [216,] -0.22861502 -0.165467723
#> [217,] -0.16039434 -0.204214853
#> [218,] -0.21168779 -0.179338576
#> [219,] -0.37267923  0.054502827
#> [220,] -0.24860890 -0.145693019
#> [221,] -0.37000037  0.048799033
#> [222,]  0.61609216 -0.126903340
#> [223,] -0.29397587 -0.087725874
#> [224,] -0.30884144 -0.064878517
#> [225,] -0.23892861 -0.155718707
#> [226,] -0.15813225 -0.204677122
#> [227,]  0.32969063 -0.555697536
#> [228,]  0.27633532 -0.553279280
#> [229,] -0.33482770 -0.020441212
#> [230,]  0.51765473 -0.367202617
#> [231,]  0.26118862 -0.548142506
#> [232,] -0.39141459  0.096110350
#> [233,] -0.35647740  0.020937915
#> [234,]  0.37530662 -0.538642707
#> [235,] -0.20218990 -0.185933883
#> [236,]  0.44471187 -0.478317801
#> [237,]  0.53053516 -0.342060097
#> [238,] -0.47171161  0.309456135
#> [239,]  0.48742387 -0.419523880
#> [240,] -0.23978825 -0.154862487
#> [241,] -0.31831771 -0.049335292
#> [242,] -0.40253561  0.122238435
#> [243,] -0.19692111 -0.189215558
#> [244,] -0.41595681  0.155205751
#> [245,] -0.36363722  0.035495632
#> [246,] -0.38023079  0.070911325
#> [247,] -0.39083656  0.094781584
#> [248,] -0.53616173  0.524919364
#> [249,] -0.37478194  0.059022777
#> [250,]  0.46624024 -0.450832040
#> [251,] -0.11796112 -0.202814013
#> [252,] -0.33331966 -0.023175971
#> [253,] -0.41257852  0.146758691
#> [254,] -0.19760393 -0.188805608
#> [255,]  0.19748841 -0.503966729
#> [256,] -0.43275334  0.198704961
#> [257,] -0.11339485 -0.201296504
#> [258,]  0.63422340 -0.069234974
#> [259,] -0.43890491  0.215267742
#> [260,] -0.62069371  0.877604023
#> [261,] -0.19594083 -0.189796069
#> [262,] -0.25469731 -0.138958744
#> [263,] -0.54837266  0.570627491
#> [264,] -0.38680278  0.085588836
#> [265,] -0.35137820  0.010834925
#> [266,] -0.37037177  0.049586160
#> [267,]  0.52568185 -0.351739297
#> [268,]  0.59156096 -0.197764424
#> [269,] -0.11207206 -0.200802612
#> [270,] -0.32454656 -0.038704723
#> [271,]  0.41850522 -0.506077918
#> [272,] -0.17303475 -0.200602793
#> [273,] -0.28656265 -0.098415290
#> [274,] -0.56524474  0.636539828
#> [275,]  0.80747617  0.804210449
#> [276,] -0.17021028 -0.201559125
#> [277,] -0.58070000  0.699840059
#> [278,] -0.63929099  0.967720164
#> [279,] -0.28298041 -0.103411869
#> [280,] -0.12055637 -0.203549270
#> [281,] -0.22533595 -0.168363161
#> [282,] -0.32045857 -0.045718544
#> [283,] -0.49690570  0.388710077
#> [284,] -0.18011977 -0.197834456
#> [285,] -0.26102565 -0.131611673
#> [286,] -0.41276542  0.147223387
#> [287,] -0.10346534 -0.196973833
#> [288,] -0.08814382 -0.187347716
#> [289,]  0.76403270  0.513582212
#> [290,] -0.21851685 -0.174065176
#> [291,] -0.24124810 -0.153393169
#> [292,] -0.33628671 -0.017777067
#> [293,] -0.31127468 -0.060960039
#> [294,]  0.74596301  0.410234599
#> [295,] -0.14687992 -0.206117306
#> [296,] -0.12943061 -0.205388794
#> [297,]  0.52290747 -0.357160264
#> [298,] -0.49837830  0.393534930
#> [299,] -0.14260299 -0.206279880
#> [300,] -0.25299702 -0.140872508
#> 
#> $grads_hat
#>               [,1]          [,2]
#>   [1,]  0.30716167 -0.5576287251
#>   [2,] -0.11726528 -0.2026013263
#>   [3,]  0.64764090 -0.0234553741
#>   [4,] -0.10224864 -0.1963442443
#>   [5,] -0.09008143 -0.1887757184
#>   [6,] -0.36429402  0.0368528687
#>   [7,] -0.43166194  0.1958020279
#>   [8,] -0.35195073  0.0119582998
#>   [9,]  0.72734886  0.3126543668
#>  [10,] -0.24390655 -0.1506681861
#>  [11,] -0.36274530  0.0336584138
#>  [12,] -0.37190341  0.0528446639
#>  [13,]  0.68212499  0.1074626522
#>  [14,] -0.16878350 -0.2020098926
#>  [15,]  0.19635290 -0.5028329369
#>  [16,] -0.17445278 -0.2000908122
#>  [17,] -0.35262428  0.0132834362
#>  [18,] -0.24699672 -0.1474210071
#>  [19,]  0.70486518  0.2054215757
#>  [20,]  0.54409870 -0.3136637474
#>  [21,] -0.30538421 -0.0703596849
#>  [22,] -0.12367595 -0.2043135840
#>  [23,]  0.42628410 -0.4984797807
#>  [24,] -0.13259993 -0.2057997135
#>  [25,] -0.18111311 -0.1974046000
#>  [26,] -0.18428158 -0.1959657077
#>  [27,] -0.62585230  0.9021005417
#>  [28,] -0.19859027 -0.1882053413
#>  [29,]  0.72353631  0.2936879023
#>  [30,]  0.45753633 -0.4624621438
#>  [31,] -0.32691403 -0.0345782490
#>  [32,] -0.20610240 -0.1833221596
#>  [33,] -0.40312498  0.1236530568
#>  [34,] -0.30150807 -0.0763843214
#>  [35,]  0.21566833 -0.5204156870
#>  [36,]  0.40489286 -0.5180892526
#>  [37,] -0.33444642 -0.0211344482
#>  [38,] -0.26578531 -0.1258540372
#>  [39,] -0.38037784  0.0712356774
#>  [40,] -0.35609110  0.0201648277
#>  [41,] -0.74688794  1.6069642887
#>  [42,] -0.36475807  0.0378139985
#>  [43,]  0.58864897 -0.2056571727
#>  [44,]  0.30381575 -0.5575489015
#>  [45,]  0.22521302 -0.5277956440
#>  [46,] -0.21590345 -0.1761352240
#>  [47,] -0.54618475  0.5623165450
#>  [48,]  0.40720236 -0.5161658306
#>  [49,] -0.16151432 -0.2039649932
#>  [50,] -0.22301422 -0.1703531314
#>  [51,]  0.48545665 -0.4226134405
#>  [52,] -0.30792999 -0.0663333971
#>  [53,] -0.51497746  0.4494324751
#>  [54,] -0.28558159 -0.0997946510
#>  [55,] -0.60569462  0.8084715627
#>  [56,] -0.27417326 -0.1152264528
#>  [57,] -0.44997510  0.2459386300
#>  [58,] -0.09522456 -0.1922635321
#>  [59,] -0.43390036  0.2017673871
#>  [60,] -0.08014149 -0.1807578931
#>  [61,]  0.53715644 -0.3284478883
#>  [62,] -0.23353497 -0.1609380664
#>  [63,] -0.45189083  0.2513601405
#>  [64,] -0.16753556 -0.2023862570
#>  [65,] -0.16822253 -0.2021811454
#>  [66,] -0.19041436 -0.1928904634
#>  [67,]  0.60056029 -0.1726883950
#>  [68,] -0.50002558  0.3989577599
#>  [69,] -0.40615761  0.1309800450
#>  [70,]  0.26188123 -0.5484210798
#>  [71,] -0.10567445 -0.1980601627
#>  [72,] -0.21410000 -0.1775262026
#>  [73,] -0.36932973  0.0473806732
#>  [74,] -0.47189939  0.3100242038
#>  [75,]  0.62081541 -0.1123291956
#>  [76,] -0.19930484 -0.1877645003
#>  [77,]  0.66352205  0.0343644757
#>  [78,] -0.25565936 -0.1378645846
#>  [79,]  0.49366543 -0.4094704045
#>  [80,]  0.48052312 -0.4301962756
#>  [81,] -0.43126365  0.1947453033
#>  [82,] -0.28028270 -0.1071018618
#>  [83,] -0.24585709 -0.1486284901
#>  [84,] -0.79880711  2.0152629884
#>  [85,]  0.67919498  0.0955465283
#>  [86,] -0.29809207 -0.0815878357
#>  [87,]  0.68646355  0.1253940699
#>  [88,]  0.38981759 -0.5295051857
#>  [89,] -0.41095808  0.1427425611
#>  [90,]  0.16359689 -0.4644814381
#>  [91,]  0.33783770 -0.5539430641
#>  [92,] -0.42105024  0.1681317057
#>  [93,] -0.35670410  0.0213921916
#>  [94,]  0.53893232 -0.3247162429
#>  [95,] -0.06852129 -0.1690540139
#>  [96,]  0.51561108 -0.3710320827
#>  [97,] -0.35391963  0.0158427008
#>  [98,] -0.35906116  0.0261412030
#>  [99,] -0.26618953 -0.1253559361
#> [100,] -0.12731314 -0.2050429952
#> [101,] -0.60134907  0.7890078612
#> [102,] -0.52791365  0.4949632748
#> [103,]  0.75320036  0.4505495384
#> [104,]  0.40500466 -0.5179972106
#> [105,] -0.19271252 -0.1916405925
#> [106,] -0.29346836 -0.0884726610
#> [107,]  0.24207623 -0.5387821322
#> [108,] -0.10296815 -0.1967192641
#> [109,] -0.60068074  0.7860364408
#> [110,] -0.22372303 -0.1697509105
#> [111,]  0.54004452 -0.3223616492
#> [112,] -0.39516623  0.1048048339
#> [113,] -0.34458892 -0.0022750942
#> [114,] -0.43916018  0.2159624239
#> [115,]  0.32416874 -0.5565679344
#> [116,] -0.24991880 -0.1442719676
#> [117,] -0.52778268  0.4944934796
#> [118,]  0.40911141 -0.5145407046
#> [119,] -0.44656935  0.2363835459
#> [120,] -0.27340260 -0.1162283819
#> [121,] -0.36854928  0.0457349199
#> [122,] -0.14440825 -0.2062376366
#> [123,] -0.11339706 -0.2012973083
#> [124,] -0.40787053  0.1351540442
#> [125,]  0.31984722 -0.5570694059
#> [126,] -0.36220378  0.0325462550
#> [127,] -0.47175010  0.3095725443
#> [128,] -0.13637634 -0.2061250379
#> [129,] -0.40268262  0.1225909982
#> [130,]  0.25966545 -0.5475150705
#> [131,] -0.48067614  0.3369509380
#> [132,] -0.12225408 -0.2039812580
#> [133,] -0.33409100 -0.0217795703
#> [134,]  0.52740914 -0.3483231817
#> [135,] -0.18936970 -0.1934411250
#> [136,] -0.10787118 -0.1990687627
#> [137,] -0.09315776 -0.1909140045
#> [138,]  0.77074020  0.5543267197
#> [139,] -0.50408309  0.4124310018
#> [140,] -0.30250258 -0.0748507513
#> [141,] -0.26682107 -0.1245748656
#> [142,] -0.14482237 -0.2062224959
#> [143,] -0.36025799  0.0285706102
#> [144,] -0.19642561 -0.1895101725
#> [145,] -0.43556296  0.2062273847
#> [146,]  0.66145649  0.0266126843
#> [147,] -0.25735588 -0.1359151828
#> [148,]  0.19027666 -0.4965488170
#> [149,] -0.08432706 -0.1843462160
#> [150,] -0.37715026  0.0641588864
#> [151,] -0.14322226 -0.2062697554
#> [152,]  0.46871071 -0.4474014203
#> [153,] -0.16029330 -0.2042367138
#> [154,] -0.07586006 -0.1767527377
#> [155,] -0.33685213 -0.0167397864
#> [156,] -0.18659859 -0.1948485912
#> [157,] -0.23529299 -0.1592658847
#> [158,] -0.15905353 -0.2044957213
#> [159,] -0.31423205 -0.0561299352
#> [160,] -0.39539579  0.1053407785
#> [161,] -0.17287490 -0.2006591763
#> [162,] -0.14362990 -0.2062605991
#> [163,] -0.27033487 -0.1201656363
#> [164,]  0.66674468  0.0466008049
#> [165,]  0.51312328 -0.3756355581
#> [166,] -0.21719137 -0.1751230998
#> [167,] -0.21601253 -0.1760501109
#> [168,] -0.34586654  0.0001622022
#> [169,] -0.26309876 -0.1291282632
#> [170,] -0.41832784  0.1611943532
#> [171,] -0.35267009  0.0133737069
#> [172,] -0.10894530 -0.1995362633
#> [173,] -0.27917476 -0.1085991735
#> [174,] -0.38499390  0.0815120016
#> [175,] -0.29745948 -0.0825405448
#> [176,] -0.30563192 -0.0699703435
#> [177,] -0.11894624 -0.2031038389
#> [178,] -0.20360736 -0.1850047970
#> [179,] -0.40345314  0.1244420493
#> [180,] -0.32945318 -0.0300999139
#> [181,] -0.45778461  0.2682520763
#> [182,] -0.26147028 -0.1310822180
#> [183,] -0.46372816  0.2856136981
#> [184,] -0.33942062 -0.0119938137
#> [185,] -0.42932196  0.1896140595
#> [186,]  0.63113679 -0.0793855259
#> [187,]  0.59277956 -0.1944296452
#> [188,] -0.47797118  0.3285734947
#> [189,] -0.12190379 -0.2038952751
#> [190,] -0.26819094 -0.1228686953
#> [191,] -0.18685770 -0.1947202748
#> [192,]  0.42012264 -0.5045422785
#> [193,] -0.30296819 -0.0741298671
#> [194,]  0.55438881 -0.2907662863
#> [195,] -0.46718195  0.2958545462
#> [196,] -0.21824888 -0.1742803869
#> [197,] -0.58742186  0.7282764252
#> [198,] -0.46510138  0.2896720075
#> [199,] -0.36603321  0.0404644807
#> [200,] -0.08413224 -0.1841861720
#> [201,] -0.39806393  0.1116036342
#> [202,]  0.54900679 -0.3028903642
#> [203,] -0.13149664 -0.2056710663
#> [204,]  0.59627504 -0.1847589265
#> [205,] -0.30267252 -0.0745878427
#> [206,] -0.46550926  0.2908808389
#> [207,]  0.34849183 -0.5508026438
#> [208,] -0.31322702 -0.0577796988
#> [209,] -0.38843933  0.0893015444
#> [210,]  0.39380395 -0.5266779060
#> [211,] -0.27466857 -0.1145797869
#> [212,]  0.45841120 -0.4613251326
#> [213,] -0.33922229 -0.0123622721
#> [214,] -0.48990565  0.3660689557
#> [215,]  0.42453433 -0.5002357556
#> [216,] -0.16398768 -0.2033643260
#> [217,] -0.60239659  0.7936769219
#> [218,] -0.20937821 -0.1810219909
#> [219,]  0.53857661 -0.3254664402
#> [220,]  0.52329101 -0.3564156890
#> [221,] -0.32542316 -0.0371823436
#> [222,] -0.43621144  0.2079737113
#> [223,]  0.38595618 -0.5321132215
#> [224,] -0.43485840  0.2043343453
#> [225,] -0.18290135 -0.1966051604
#> [226,] -0.27741707 -0.1109528412
#> [227,] -0.46522480  0.2900376252
#> [228,] -0.23710662 -0.1575113515
#> [229,] -0.19030790 -0.1929470847
#> [230,]  0.59743945 -0.1815025800
#> [231,]  0.73234088  0.3379981570
#> [232,] -0.44016179  0.2186939010
#> [233,] -0.26219959 -0.1302100214
#> [234,] -0.19163857 -0.1922312387
#> [235,] -0.37083406  0.0505675414
#> [236,]  0.76257545  0.5049073293
#> [237,] -0.27245601 -0.1174519834
#> [238,] -0.24941779 -0.1448173035
#> [239,] -0.47303858  0.3134775916
#> [240,] -0.36423917  0.0367393698
#> [241,] -0.29665301 -0.0837501770
#> [242,] -0.35441818  0.0168315015
#> [243,]  0.36692635 -0.5430993579
#> [244,]  0.55699528 -0.2847766676
#> [245,] -0.08438384 -0.1843927346
#> [246,] -0.15266327 -0.2055577170
#> [247,] -0.24171800 -0.1529161313
#> [248,] -0.53402488  0.5170883261
#> [249,] -0.37510759  0.0597261633
#> [250,]  0.18069230 -0.4858813050
#> [251,] -0.22871205 -0.1653805424
#> [252,] -0.63369522  0.9400746312
#> [253,] -0.45809013  0.2691364992
#> [254,]  0.52273939 -0.3574860924
#> [255,] -0.31209594 -0.0596261473
#> [256,] -0.16079982 -0.2041259940
#> [257,] -0.45022086  0.2466322268
#> [258,]  0.51220443 -0.3773196898
#> [259,] -0.27944945 -0.1082289299
#> [260,]  0.63295089 -0.0734366382
#> [261,] -0.25397403 -0.1397759603
#> [262,] -0.50561519  0.4175616023
#> [263,]  0.51168840 -0.3782617128
#> [264,] -0.35105332  0.0101986919
#> [265,]  0.75535573  0.4628287269
#> [266,] -0.42919199  0.1892717835
#> [267,] -0.21012110 -0.1804860611
#> [268,] -0.18594426 -0.1951695979
#> [269,] -0.38813199  0.0886025570
#> [270,] -0.09203637 -0.1901526672
#> [271,] -0.44840196  0.2415118564
#> [272,] -0.38185370  0.0745013800
#> [273,] -0.35467220  0.0173361201
#> [274,] -0.39805386  0.1115798894
#> [275,] -0.29106043 -0.0919858497
#> [276,] -0.13781931 -0.2062028426
#> [277,] -0.31130857 -0.0609051009
#> [278,]  0.28214659 -0.5547220744
#> [279,]  0.48646020 -0.4210420611
#> [280,] -0.40814965  0.1358366305
#> [281,]  0.47634967 -0.4364274298
#> [282,]  0.44008343 -0.4836702954
#> [283,]  0.37367625 -0.5395566677
#> [284,] -0.18896577 -0.1936511055
#> [285,] -0.08187239 -0.1822802316
#> [286,] -0.31433834 -0.0559549714
#> [287,]  0.40583707 -0.5173085100
#> [288,]  0.55255819 -0.2949268383
#> [289,] -0.28508419 -0.1004908246
#> [290,] -0.49725459  0.3898512263
#> [291,] -0.47436070  0.3175010502
#> [292,] -0.12858169 -0.2052570627
#> [293,] -0.37764767  0.0652436930
#> [294,]  0.45248570 -0.4688873952
#> [295,] -0.29254502 -0.0898256838
#> [296,]  0.51806874 -0.3664215437
#> [297,] -0.15949276 -0.2044059103
#> [298,] -0.05165174 -0.1468571903
#> [299,] -0.06701923 -0.1673417556
#> [300,] -0.22137895 -0.1717246880
#> 
#> $inv_hessian
#>          [,1]     [,2]
#> [1,] 5.133838 1.058397
#> [2,] 1.058397 5.942456
#>