PPI Logistic Regression

Helper function for PPI logistic regression

Usage

ppi_logistic(X_l, Y_l, f_l, X_u, f_u, opts = NULL)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

opts

(list, optional): Options to pass to the optimizer. See ?optim for details.

Value

(list): A list containing the following:

est

(vector): vector of PPI logistic regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

rectifier_est

(vector): vector of the rectifier logistic regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

Prediction Powered Inference (Angelopoulos et al., 2023) https://www.science.org/doi/10.1126/science.adi6000

Examples

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_logistic(X_l, Y_l, f_l, X_u, f_u)
#> $est
#>           [,1]
#> [1,] 0.1793968
#> [2,] 0.6494671
#> 
#> $se
#> [1] 0.1865305 0.1836512
#> 
#> $rectifier_est
#>             [,1]
#> [1,]  0.08682910
#> [2,] -0.03653341
#> 
#> $var_u
#>             [,1]        [,2]
#> [1,]  0.18022109 -0.03753627
#> [2,] -0.03753627  0.16889304
#> 
#> $var_l
#>             [,1]        [,2]
#> [1,]  0.35240803 -0.05586951
#> [2,] -0.05586951  0.26862173
#> 
#> $grads
#>               [,1]          [,2]
#>   [1,]  0.42955822 -0.3062621661
#>   [2,]  0.47026499 -0.2161209596
#>   [3,] -0.28823465 -0.3215683446
#>   [4,] -0.52869416  0.2395726190
#>   [5,]  0.40663873 -0.3489162529
#>   [6,]  0.54168405 -0.0102359802
#>   [7,]  0.47473951 -0.2050546250
#>   [8,] -0.49655107  0.1266103949
#>   [9,] -0.41601339 -0.1023359986
#>  [10,]  0.70275582  0.7369394570
#>  [11,]  0.47337208 -0.2084614412
#>  [12,]  0.45496839 -0.2521981094
#>  [13,] -0.69834289  1.0954872655
#>  [14,] -0.39974297 -0.1398028443
#>  [15,]  0.56786388  0.0819656549
#>  [16,]  0.47268275 -0.2101705193
#>  [17,] -0.40977186 -0.1170459478
#>  [18,] -0.21500360 -0.3693242050
#>  [19,] -0.30122878 -0.3070678449
#>  [20,]  0.62388697  0.3138158094
#>  [21,] -0.54481122  0.3012542763
#>  [22,] -0.40371378 -0.1309216389
#>  [23,] -0.41552125 -0.1035111881
#>  [24,]  0.22746660 -0.4910545416
#>  [25,] -0.52789747  0.2366130323
#>  [26,]  0.41543254 -0.3332255917
#>  [27,]  0.26242852 -0.4900453414
#>  [28,]  0.20417069 -0.4840694030
#>  [29,] -0.43903895 -0.0443904858
#>  [30,] -0.22540331 -0.3661659456
#>  [31,] -0.74325784  1.4217799995
#>  [32,]  0.49657155 -0.1476492728
#>  [33,] -0.44540390 -0.0273371619
#>  [34,] -0.44912130 -0.0171669152
#>  [35,]  0.33128806 -0.4497798060
#>  [36,] -0.36520854 -0.2099879166
#>  [37,] -0.31999497 -0.2830081595
#>  [38,] -0.25536180 -0.3502586231
#>  [39,] -0.53421251  0.2603017665
#>  [40,] -0.22984003 -0.3644424100
#>  [41,] -0.59161992  0.5010695125
#>  [42,] -0.34911156 -0.2384237556
#>  [43,] -0.49317424  0.1154914473
#>  [44,] -0.35058332 -0.2359360554
#>  [45,]  0.56037491  0.0546053985
#>  [46,]  0.66150950  0.4997287493
#>  [47,] -0.54218112  0.2909504267
#>  [48,]  0.26076961 -0.4903958012
#>  [49,] -0.67535260  0.9482368605
#>  [50,]  0.68646300  0.6386575895
#>  [51,]  0.33000641 -0.4509822331
#>  [52,]  0.62635518  0.3252149320
#>  [53,] -0.39511183 -0.1499480837
#>  [54,]  0.41858905 -0.3273892949
#>  [55,] -0.58982794  0.4928204598
#>  [56,]  0.59601155  0.1922355336
#>  [57,]  0.34957822 -0.4307594905
#>  [58,]  0.49684555 -0.1468922539
#>  [59,]  0.49457837 -0.1531284444
#>  [60,]  0.61757229  0.2851337592
#>  [61,]  0.41418110 -0.3355095322
#>  [62,] -0.24145761 -0.3588802707
#>  [63,] -0.33308289 -0.2640578702
#>  [64,] -0.63218013  0.7017874325
#>  [65,] -0.18954372 -0.3716881706
#>  [66,] -0.34088551 -0.2519130923
#>  [67,] -0.70552996  1.1440796185
#>  [68,] -0.60056273  0.5429909155
#>  [69,]  0.47248708 -0.2106546114
#>  [70,] -0.28752708 -0.3223069547
#>  [71,] -0.59455166  0.5146734180
#>  [72,]  0.59110275  0.1721330470
#>  [73,] -0.37138205 -0.1983638148
#>  [74,] -0.54485573  0.3014294512
#>  [75,]  0.52145637 -0.0750859870
#>  [76,] -0.60849592  0.5812540974
#>  [77,] -0.47273295  0.0511118399
#>  [78,]  0.36182016 -0.4160833974
#>  [79,] -0.26440101 -0.3435234949
#>  [80,] -0.20169622 -0.3715279138
#>  [81,] -0.46012310  0.0138502514
#>  [82,] -0.61909550  0.6340046305
#>  [83,] -0.41891766 -0.0953472839
#>  [84,] -0.64963888  0.7970597728
#>  [85,]  0.26252643 -0.4900237273
#>  [86,]  0.28754790 -0.4811391526
#>  [87,] -0.44245135 -0.0353041666
#>  [88,] -0.44578022 -0.0263146907
#>  [89,]  0.56812600  0.0829379073
#>  [90,]  0.69219368  0.6725097114
#>  [91,]  0.45621248 -0.2493643481
#>  [92,]  0.30723463 -0.4694974136
#>  [93,]  0.27801414 -0.4853102841
#>  [94,] -0.42355218 -0.0840047125
#>  [95,] -0.51094759  0.1755909041
#>  [96,] -0.31979084 -0.2832896153
#>  [97,]  0.30003983 -0.4742232099
#>  [98,]  0.82888116  1.7846052481
#>  [99,]  0.75398206  1.0919272518
#> [100,] -0.41198298 -0.1118830005
#> [101,] -0.48049046  0.0749581037
#> [102,] -0.49324381  0.1157191217
#> [103,] -0.34327921 -0.2480604410
#> [104,]  0.61063699  0.2544155626
#> [105,] -0.42616493 -0.0775067330
#> [106,] -0.45148135 -0.0106292084
#> [107,]  0.36769072 -0.4084904174
#> [108,] -0.26860303 -0.3400926154
#> [109,] -0.55805524  0.3545869697
#> [110,]  0.34661621 -0.4340758440
#> [111,]  0.42590837 -0.3134382525
#> [112,] -0.24745693 -0.3554205574
#> [113,]  0.60577066  0.2333401490
#> [114,]  0.75703818  1.1156387946
#> [115,]  0.56953006  0.0881630152
#> [116,] -0.30445939 -0.3031879209
#> [117,] -0.64949728  0.7962639524
#> [118,] -0.38061728 -0.1802278473
#> [119,]  0.49296603 -0.1575252761
#> [120,] -0.51437549  0.1876354783
#> [121,]  0.33075713 -0.4502799838
#> [122,] -0.44681935 -0.0234830652
#> [123,] -0.34791327 -0.2404325618
#> [124,]  0.59469699  0.1868150114
#> [125,]  0.30391882 -0.4717426050
#> [126,] -0.31148262 -0.2943768481
#> [127,] -0.45184258 -0.0096229792
#> [128,]  0.58898109  0.1635609539
#> [129,] -0.29937325 -0.3092469314
#> [130,]  0.27836567 -0.4851735889
#> [131,]  0.54492168  0.0006508340
#> [132,]  0.66034705  0.4935765218
#> [133,] -0.45345428 -0.0051154411
#> [134,] -0.28929718 -0.3204492809
#> [135,]  0.54488787  0.0005364103
#> [136,]  0.50557353 -0.1222948810
#> [137,] -0.23705804 -0.3611641047
#> [138,]  0.34289990 -0.4381076022
#> [139,]  0.56644077  0.0767043037
#> [140,] -0.16282806 -0.3655199389
#> [141,]  0.47879869 -0.1948119501
#> [142,]  0.77028187  1.2222028642
#> [143,]  0.31640329 -0.4626918878
#> [144,]  0.41245150 -0.3386382307
#> [145,] -0.41384562 -0.1074927542
#> [146,] -0.42694749 -0.0755459236
#> [147,] -0.20972712 -0.3704476015
#> [148,]  0.21786471 -0.4889373901
#> [149,]  0.59889784  0.2042331666
#> [150,]  0.67728300  0.5859827886
#> [151,] -0.29420254 -0.3151288477
#> [152,] -0.73629479  1.3674525172
#> [153,]  0.63793927  0.3801645635
#> [154,] -0.58317793  0.4626417233
#> [155,] -0.24159786 -0.3588039340
#> [156,] -0.41645367 -0.1012824189
#> [157,] -0.24246869 -0.3583250512
#> [158,] -0.57556025  0.4288966387
#> [159,] -0.41000981 -0.1164928746
#> [160,]  0.26686894 -0.4889613550
#> [161,] -0.26505930 -0.3429985410
#> [162,] -0.33408950 -0.2625266303
#> [163,] -0.22603145 -0.3659355052
#> [164,]  0.57572884  0.1115766511
#> [165,]  0.40273809 -0.3556094386
#> [166,]  0.47951526 -0.1929836045
#> [167,] -0.65544531  0.8300293862
#> [168,] -0.59101801  0.4982931331
#> [169,]  0.61269125  0.2634301443
#> [170,]  0.45396275 -0.2544758690
#> [171,]  0.36273205 -0.4149276126
#> [172,]  0.33416059 -0.4470228029
#> [173,] -0.43324539 -0.0595202715
#> [174,] -0.24509359 -0.3568308533
#> [175,] -0.79288471  1.8578479292
#> [176,] -0.60007789  0.5406855059
#> [177,]  0.55650686  0.0407871539
#> [178,]  0.41788800 -0.3286948659
#> [179,] -0.46257699  0.0209572067
#> [180,] -0.17307921 -0.3689779013
#> [181,] -0.54814230  0.3144398232
#> [182,]  0.52101565 -0.0764392920
#> [183,]  0.69245056  0.6740450905
#> [184,] -0.41499487 -0.1047652369
#> [185,]  0.50616297 -0.1205996788
#> [186,]  0.32093053 -0.4590085208
#> [187,] -0.45544864  0.0005031620
#> [188,]  0.46517009 -0.2284372571
#> [189,] -0.47576536  0.0603489001
#> [190,] -0.54307882  0.2944567625
#> [191,]  0.44700193 -0.2699274190
#> [192,] -0.57829563  0.4409137578
#> [193,]  0.34314244 -0.4378488553
#> [194,] -0.22595568 -0.3659635405
#> [195,]  0.77072818  1.2259062384
#> [196,]  0.64156509  0.3978662776
#> [197,]  0.45122455 -0.2606195462
#> [198,] -0.09197210 -0.2988561902
#> [199,]  0.70570878  0.7554410477
#> [200,] -0.34438549 -0.2462597536
#> [201,] -0.25094091 -0.3532296809
#> [202,]  0.63370554  0.3598010008
#> [203,] -0.32183614 -0.2804499536
#> [204,]  0.29120411 -0.4792847853
#> [205,]  0.33798675 -0.4432173443
#> [206,] -0.58516202  0.4715748464
#> [207,] -0.57110855  0.4095765499
#> [208,] -0.09801533 -0.3078813149
#> [209,]  0.65439382  0.4624945604
#> [210,]  0.69405418  0.6836647971
#> [211,]  0.48124264 -0.1885510664
#> [212,] -0.46277692  0.0215392939
#> [213,]  0.51099347 -0.1065436901
#> [214,] -0.27369036 -0.3356857579
#> [215,] -0.33608777 -0.2594556885
#> [216,] -0.35097745 -0.2352660361
#> [217,]  0.42048135 -0.3238385263
#> [218,]  0.35023853 -0.4300077457
#> [219,] -0.48541228  0.0904574690
#> [220,]  0.30395449 -0.4717190574
#> [221,] -0.27473091 -0.3347503499
#> [222,]  0.51017142 -0.1089563876
#> [223,]  0.39698531 -0.3651832847
#> [224,]  0.69078335  0.6641079815
#> [225,]  0.47868937 -0.1950903574
#> [226,] -0.36680190 -0.2070260044
#> [227,] -0.26380540 -0.3439944395
#> [228,] -0.32806045 -0.2715405446
#> [229,]  0.74103406  0.9948879209
#> [230,] -0.44976367 -0.0153936843
#> [231,]  0.65057197  0.4429096402
#> [232,]  0.76146124  1.1505278211
#> [233,] -0.30275985 -0.3052426618
#> [234,]  0.30363270 -0.4719309545
#> [235,] -0.42534093 -0.0795641162
#> [236,] -0.45846478  0.0090866013
#> [237,] -0.39950067 -0.1403393239
#> [238,]  0.50389900 -0.1270871318
#> [239,]  0.45700496 -0.2475500234
#> [240,]  0.57965516  0.1267015693
#> [241,]  0.38072634 -0.3903379562
#> [242,] -0.40188498 -0.1350329601
#> [243,] -0.36907960 -0.2027457945
#> [244,] -0.28356331 -0.3263468619
#> [245,]  0.30205507 -0.4729541209
#> [246,] -0.42986554 -0.0681749853
#> [247,] -0.52907328  0.2409839127
#> [248,]  0.53192141 -0.0422096547
#> [249,]  0.18733815 -0.4750168494
#> [250,] -0.48914719  0.1024126325
#> [251,] -0.71760466  1.2286716281
#> [252,] -0.49705567  0.1282838703
#> [253,]  0.45228900 -0.2582413525
#> [254,] -0.53272085  0.2546588250
#> [255,]  0.58842100  0.1613096931
#> [256,] -0.39485190 -0.1505107194
#> [257,] -0.47484038  0.0575198587
#> [258,] -0.66465306  0.8836867295
#> [259,] -0.50449371  0.1533149205
#> [260,]  0.52052423 -0.0779453463
#> [261,] -0.23931740 -0.3600181473
#> [262,] -0.37831715 -0.1848287198
#> [263,] -0.52178553  0.2141832404
#> [264,]  0.49618248 -0.1487226246
#> [265,] -0.34988277 -0.2371230104
#> [266,] -0.44655588 -0.0242021540
#> [267,]  0.35023332 -0.4300136955
#> [268,] -0.60228856  0.5512279095
#> [269,]  0.35194221 -0.4280471702
#> [270,] -0.68023645  0.9785169545
#> [271,]  0.52585015 -0.0614566848
#> [272,] -0.58915245  0.4897239267
#> [273,] -0.42596872 -0.0779973119
#> [274,] -0.45276051 -0.0070593614
#> [275,]  0.23088074 -0.4915543461
#> [276,] -0.41418865 -0.1066801359
#> [277,] -0.19189578 -0.3717978396
#> [278,]  0.61622773  0.2791149497
#> [279,] -0.27182178 -0.3373365423
#> [280,]  0.70927075  0.7780510559
#> [281,] -0.51824411  0.2014078191
#> [282,] -0.55011358  0.3223146234
#> [283,]  0.45347770 -0.2555703592
#> [284,] -0.65718457  0.8400341991
#> [285,] -0.51578848  0.1926436409
#> [286,] -0.47975096  0.0726542434
#> [287,]  0.59294321  0.1796257595
#> [288,] -0.45847254  0.0091088070
#> [289,] -0.52745884  0.2349871501
#> [290,]  0.66891045  0.5395476083
#> [291,] -0.28129361 -0.3285853522
#> [292,] -0.49031180  0.1061748189
#> [293,]  0.40226631 -0.3564079037
#> [294,] -0.46565799  0.0299786942
#> [295,]  0.64885268  0.4341918333
#> [296,] -0.24047138 -0.3594109327
#> [297,] -0.54213365  0.2907653427
#> [298,] -0.43937970 -0.0434889957
#> [299,]  0.61256470  0.2628727753
#> [300,] -0.23207536 -0.3634897193
#> 
#> $grads_hat_unlabeled
#>              [,1]         [,2]
#>   [1,] -0.5371873  0.271643785
#>   [2,] -0.2665503 -0.341792304
#>   [3,] -0.3268345 -0.273327274
#>   [4,]  0.3024587 -0.472694813
#>   [5,] -0.7514194  1.487404139
#>   [6,] -0.2863687 -0.323504727
#>   [7,]  0.1762252 -0.467117095
#>   [8,] -0.2410241 -0.359114880
#>   [9,] -0.3694411 -0.202061493
#>  [10,]  0.4029928 -0.355177306
#>  [11,]  0.5910499  0.171918628
#>  [12,]  0.4692606 -0.218572857
#>  [13,]  0.3916991 -0.373669391
#>  [14,] -0.5769333  0.434914990
#>  [15,]  0.4272007 -0.310914145
#>  [16,] -0.4493036 -0.016664231
#>  [17,]  0.5436846 -0.003525789
#>  [18,] -0.3079500 -0.298873160
#>  [19,]  0.5523318  0.026108208
#>  [20,] -0.2402483 -0.359529460
#>  [21,] -0.3541993 -0.229728165
#>  [22,] -0.3652297 -0.209948765
#>  [23,] -0.5690053  0.400549922
#>  [24,] -0.3333112 -0.263711426
#>  [25,] -0.2646930 -0.343291222
#>  [26,] -0.7599091  1.558012649
#>  [27,]  0.5953090  0.189335193
#>  [28,] -0.6094767  0.586056406
#>  [29,] -0.5392802  0.279694433
#>  [30,] -0.6433790  0.762244359
#>  [31,] -0.3939221 -0.152517552
#>  [32,] -0.4878337  0.098188974
#>  [33,]  0.4386250 -0.287799203
#>  [34,] -0.2844447 -0.325462923
#>  [35,] -0.3612474 -0.217236214
#>  [36,]  0.4198984 -0.324936465
#>  [37,]  0.3087711 -0.468418112
#>  [38,] -0.6195974  0.636549800
#>  [39,] -0.4182388 -0.096989136
#>  [40,] -0.2420912 -0.358533686
#>  [41,]  0.3833104 -0.386527203
#>  [42,] -0.5607127  0.365583974
#>  [43,] -0.5370684  0.271188026
#>  [44,]  0.2935477 -0.478022149
#>  [45,]  0.8293826  1.790204431
#>  [46,] -0.2974156 -0.311506852
#>  [47,] -0.3596760 -0.220066246
#>  [48,] -0.1885525 -0.371621486
#>  [49,] -0.3035500 -0.304291069
#>  [50,]  0.5210774 -0.076249826
#>  [51,] -0.6064898  0.571481229
#>  [52,]  0.6029373  0.221247832
#>  [53,]  0.5104144 -0.108244046
#>  [54,]  0.4695439 -0.217882517
#>  [55,] -0.2988668 -0.309835395
#>  [56,] -0.5448917  0.301570902
#>  [57,] -0.3764591 -0.188504542
#>  [58,] -0.4121867 -0.111404577
#>  [59,]  0.4324356 -0.300501536
#>  [60,]  0.4489897 -0.265570873
#>  [61,] -0.1464398 -0.357021133
#>  [62,]  0.4144113 -0.335090722
#>  [63,]  0.8682466  2.280879053
#>  [64,] -0.3464839 -0.242809278
#>  [65,] -0.5988711  0.534963838
#>  [66,] -0.1574185 -0.363126889
#>  [67,] -0.2106229 -0.370279844
#>  [68,] -0.6515004  0.807557778
#>  [69,] -0.2217520 -0.367416351
#>  [70,] -0.2021159 -0.371490459
#>  [71,] -0.5175263  0.198837843
#>  [72,] -0.5466110  0.308359657
#>  [73,] -0.2678214 -0.340745188
#>  [74,]  0.4751806 -0.203950973
#>  [75,] -0.5764576  0.432826612
#>  [76,]  0.5873858  0.157161315
#>  [77,] -0.3899056 -0.161080318
#>  [78,] -0.3507880 -0.235588336
#>  [79,] -0.1884554 -0.371614296
#>  [80,] -0.5392887  0.279726993
#>  [81,]  0.4671832 -0.223606924
#>  [82,] -0.6333333  0.707907026
#>  [83,] -0.2437254 -0.357619170
#>  [84,]  0.2848032 -0.482438577
#>  [85,] -0.5051118  0.155425620
#>  [86,]  0.3535505 -0.426168640
#>  [87,] -0.2512248 -0.353045276
#>  [88,]  0.3448477 -0.436012407
#>  [89,] -0.3972999 -0.145183364
#>  [90,] -0.4357771 -0.052954698
#>  [91,] -0.3196176 -0.283528070
#>  [92,] -0.2406286 -0.359327042
#>  [93,] -0.2248462 -0.366366569
#>  [94,]  0.2490575 -0.492020677
#>  [95,] -0.4717960  0.048279717
#>  [96,] -0.2325236 -0.363291922
#>  [97,] -0.3342137 -0.262336897
#>  [98,]  0.2027365 -0.483431580
#>  [99,] -0.3788345 -0.183798802
#> [100,]  0.2047241 -0.484308958
#> [101,] -0.6091424  0.584417545
#> [102,] -0.4623832  0.020393337
#> [103,]  0.4951618 -0.151529763
#> [104,]  0.5872291  0.156534916
#> [105,]  0.4998811 -0.138444063
#> [106,] -0.4652902  0.028895964
#> [107,] -0.4640279  0.025191807
#> [108,] -0.2359031 -0.361727907
#> [109,]  0.7093651  0.778654197
#> [110,] -0.4951735  0.122057724
#> [111,] -0.2146205 -0.369416716
#> [112,] -0.3761448 -0.189122847
#> [113,] -0.5626079  0.373487942
#> [114,] -0.3352097 -0.260810234
#> [115,]  0.7166056  0.825643514
#> [116,] -0.5015376  0.143285067
#> [117,] -0.1628415 -0.365525411
#> [118,]  0.3878696 -0.379631594
#> [119,]  0.4334237 -0.298502029
#> [120,] -0.1876918 -0.371553700
#> [121,] -0.4274740 -0.074222977
#> [122,] -0.4543598 -0.002570022
#> [123,] -0.4676068  0.035741733
#> [124,] -0.6453987  0.773395982
#> [125,] -0.3091351 -0.297379268
#> [126,]  0.3482953 -0.432207050
#> [127,] -0.4138186 -0.107556797
#> [128,] -0.3212867 -0.281217045
#> [129,] -0.5211963  0.212046547
#> [130,]  0.4213489 -0.322197624
#> [131,]  0.3511423 -0.428971457
#> [132,]  0.6226966  0.308356434
#> [133,] -0.1844389 -0.371214139
#> [134,] -0.4524541 -0.007916166
#> [135,]  0.4024154 -0.356155889
#> [136,] -0.4149344 -0.104909208
#> [137,]  0.5484429  0.012653564
#> [138,]  0.4100810 -0.342873668
#> [139,] -0.3890536 -0.162874830
#> [140,] -0.5717781  0.412463640
#> [141,] -0.4021835 -0.134364371
#> [142,]  0.2971110 -0.475991770
#> [143,]  0.1890628 -0.476103763
#> [144,] -0.3080192 -0.298786222
#> [145,]  0.5127718 -0.101295273
#> [146,] -0.2748497 -0.334642816
#> [147,] -0.5389523  0.278428935
#> [148,] -0.4902147  0.105860390
#> [149,]  0.7643717  1.173863448
#> [150,]  0.3415599 -0.439526091
#> [151,] -0.5066796  0.160801308
#> [152,] -0.3262818 -0.274127570
#> [153,] -0.4365403 -0.050961400
#> [154,]  0.2589446 -0.490747008
#> [155,] -0.5053280  0.156165102
#> [156,] -0.5386716  0.277347064
#> [157,] -0.4832619  0.083650112
#> [158,]  0.2479179 -0.492098859
#> [159,] -0.2336495 -0.362785073
#> [160,] -0.1825243 -0.370952268
#> [161,] -0.2808232 -0.329042475
#> [162,] -0.3214925 -0.280930153
#> [163,] -0.5771382  0.435815252
#> [164,] -0.2728197 -0.336459601
#> [165,] -0.5219803  0.214890703
#> [166,] -0.4051906 -0.127575466
#> [167,]  0.4083252 -0.345971769
#> [168,] -0.3612276 -0.217272134
#> [169,] -0.5193912  0.205528119
#> [170,] -0.1983851 -0.371749235
#> [171,] -0.6492628  0.794947134
#> [172,]  0.4822333 -0.185992953
#> [173,] -0.2018607 -0.371513488
#> [174,]  0.4473377 -0.269194744
#> [175,] -0.1246349 -0.339641197
#> [176,]  0.3218481 -0.458236001
#> [177,]  0.3409757 -0.440138659
#> [178,] -0.4299339 -0.068001204
#> [179,] -0.3962753 -0.147420892
#> [180,]  0.5437814 -0.003199841
#> [181,] -0.6441145  0.766296438
#> [182,]  0.4261853 -0.312898822
#> [183,]  0.4518854 -0.259144533
#> [184,]  0.5690267  0.086286610
#> [185,]  0.4326369 -0.300094997
#> [186,] -0.3928169 -0.154890786
#> [187,] -0.4607511  0.015662521
#> [188,]  0.6441425  0.410598572
#> [189,] -0.3820469 -0.177340174
#> [190,]  0.4463770 -0.271287752
#> [191,] -0.4954137  0.122849929
#> [192,] -0.3909949 -0.158774952
#> [193,] -0.2377028 -0.360842878
#> [194,] -0.2713769 -0.337724053
#> [195,] -0.2076491 -0.370800414
#> [196,] -0.6550389  0.827700303
#> [197,] -0.2889765 -0.320788241
#> [198,]  0.5084756 -0.113906762
#> [199,] -0.3898070 -0.161288565
#> [200,]  0.3021836 -0.472871738
#> [201,] -0.6029094  0.554202570
#> [202,] -0.1941850 -0.371839300
#> [203,]  0.3142594 -0.464361701
#> [204,]  0.4356214 -0.294016411
#> [205,] -0.4888978  0.101609268
#> [206,] -0.5054544  0.156597681
#> [207,]  0.5832820  0.140878678
#> [208,] -0.2858176 -0.324069667
#> [209,] -0.3913148 -0.158095589
#> [210,] -0.5923286  0.504345591
#> [211,]  0.4676792 -0.222409398
#> [212,] -0.3005091 -0.307917293
#> [213,]  0.3989209 -0.362001558
#> [214,] -0.5896037  0.491791796
#> [215,]  0.3053779 -0.470768814
#> [216,] -0.3586776 -0.221851042
#> [217,] -0.4601415  0.013903426
#> [218,]  0.4480185 -0.267704944
#> [219,] -0.5673488  0.393485808
#> [220,]  0.4985691 -0.142109421
#> [221,] -0.3621240 -0.215646259
#> [222,] -0.4747136  0.057132961
#> [223,] -0.4329309 -0.060330884
#> [224,]  0.4410227 -0.282763980
#> [225,]  0.3080236 -0.468946297
#> [226,] -0.2846293 -0.325276729
#> [227,]  0.5422292 -0.008412900
#> [228,]  0.3093395 -0.468012572
#> [229,] -0.3252572 -0.275602940
#> [230,] -0.6333739  0.708122582
#> [231,] -0.5680595  0.396511588
#> [232,] -0.1808146 -0.370679326
#> [233,]  0.5724167  0.098995874
#> [234,] -0.1536342 -0.361213685
#> [235,] -0.3221445 -0.280018025
#> [236,]  0.4800322 -0.191660784
#> [237,]  0.5318309 -0.042500058
#> [238,]  0.2183543 -0.489070954
#> [239,] -0.7049472  1.140091816
#> [240,]  0.4699155 -0.216975460
#> [241,] -0.3882240 -0.164614688
#> [242,] -0.3603105 -0.218926714
#> [243,]  0.3421271 -0.438927928
#> [244,] -0.4579667  0.007662013
#> [245,]  0.4077855 -0.346917409
#> [246,] -0.4972516  0.128934529
#> [247,] -0.2233641 -0.366883051
#> [248,]  0.4871672 -0.173078221
#> [249,] -0.3830042 -0.175394420
#> [250,] -0.5947103  0.515413533
#> [251,] -0.4420163 -0.036469872
#> [252,] -0.3763759 -0.188668432
#> [253,] -0.3212964 -0.281203535
#> [254,]  0.3236088 -0.456729033
#> [255,]  0.3134017 -0.465016315
#> [256,] -0.2061199 -0.371027470
#> [257,] -0.4112641 -0.113567291
#> [258,] -0.2583373 -0.348139209
#> [259,]  0.3899170 -0.376463356
#> [260,] -0.3387227 -0.255342855
#> [261,] -0.5452636  0.303035949
#> [262,]  0.4735912 -0.207917117
#> [263,] -0.6208480  0.642909713
#> [264,]  0.4432916 -0.277940220
#> [265,] -0.1904687 -0.371739436
#> [266,] -0.3610599 -0.217575323
#> [267,]  0.4285525 -0.308254211
#> [268,] -0.2872203 -0.322625513
#> [269,] -0.4276999 -0.073654210
#> [270,] -0.3313174 -0.266717993
#> [271,] -0.3030196 -0.304930534
#> [272,] -0.3662575 -0.208040913
#> [273,] -0.3231825 -0.278556916
#> [274,] -0.6102077  0.589645734
#> [275,]  0.4220828 -0.320802931
#> [276,] -0.3366689 -0.258554821
#> [277,]  0.4326564 -0.300055546
#> [278,]  0.6373453  0.377287730
#> [279,]  0.3623158 -0.415456259
#> [280,]  0.4429070 -0.278761965
#> [281,] -0.1617601 -0.365079051
#> [282,] -0.3226141 -0.279358403
#> [283,] -0.4544104 -0.002427478
#> [284,]  0.4660762 -0.226268965
#> [285,] -0.2126759 -0.369859887
#> [286,] -0.1521160 -0.360389943
#> [287,] -0.4898854  0.104795612
#> [288,] -0.3964954 -0.146941181
#> [289,]  0.8292105  1.788281338
#> [290,] -0.5876438  0.482833731
#> [291,] -0.3907903 -0.159208994
#> [292,] -0.5191140  0.204530833
#> [293,] -0.4286527 -0.071249970
#> [294,] -0.2708843 -0.338150714
#> [295,] -0.5380001  0.274763271
#> [296,]  0.3337495 -0.447422639
#> [297,] -0.4172275 -0.099425527
#> [298,] -0.4762981  0.061982933
#> [299,] -0.5322688  0.252954362
#> [300,]  0.4135180 -0.336712940
#> 
#> $grads_hat
#>               [,1]          [,2]
#>   [1,]  0.42955822 -0.3062621661
#>   [2,] -0.52973501  0.2434517576
#>   [3,] -0.28823465 -0.3215683446
#>   [4,] -0.52869416  0.2395726190
#>   [5,]  0.40663873 -0.3489162529
#>   [6,]  0.54168405 -0.0102359802
#>   [7,] -0.52526049  0.2268761906
#>   [8,]  0.50344893 -0.1283692101
#>   [9,] -0.41601339 -0.1023359986
#>  [10,] -0.29724418 -0.3117028140
#>  [11,]  0.47337208 -0.2084614412
#>  [12,] -0.54503161  0.3021219582
#>  [13,] -0.69834289  1.0954872655
#>  [14,] -0.39974297 -0.1398028443
#>  [15,] -0.43213612 -0.0623746657
#>  [16,]  0.47268275 -0.2101705193
#>  [17,] -0.40977186 -0.1170459478
#>  [18,] -0.21500360 -0.3693242050
#>  [19,] -0.30122878 -0.3070678449
#>  [20,] -0.37611303 -0.1891852565
#>  [21,] -0.54481122  0.3012542763
#>  [22,] -0.40371378 -0.1309216389
#>  [23,] -0.41552125 -0.1035111881
#>  [24,]  0.22746660 -0.4910545416
#>  [25,] -0.52789747  0.2366130323
#>  [26,] -0.58456746  0.4688916240
#>  [27,]  0.26242852 -0.4900453414
#>  [28,] -0.79582931  1.8868360742
#>  [29,] -0.43903895 -0.0443904858
#>  [30,] -0.22540331 -0.3661659456
#>  [31,] -0.74325784  1.4217799995
#>  [32,] -0.50342845  0.1496880858
#>  [33,] -0.44540390 -0.0273371619
#>  [34,] -0.44912130 -0.0171669152
#>  [35,]  0.33128806 -0.4497798060
#>  [36,] -0.36520854 -0.2099879166
#>  [37,] -0.31999497 -0.2830081595
#>  [38,] -0.25536180 -0.3502586231
#>  [39,] -0.53421251  0.2603017665
#>  [40,] -0.22984003 -0.3644424100
#>  [41,] -0.59161992  0.5010695125
#>  [42,] -0.34911156 -0.2384237556
#>  [43,] -0.49317424  0.1154914473
#>  [44,] -0.35058332 -0.2359360554
#>  [45,] -0.43962509 -0.0428390037
#>  [46,] -0.33849050 -0.2557082429
#>  [47,]  0.45781888 -0.2456791564
#>  [48,]  0.26076961 -0.4903958012
#>  [49,]  0.32464740 -0.4558250499
#>  [50,] -0.31353700 -0.2917022239
#>  [51,] -0.66999359  0.9156040383
#>  [52,] -0.37364482 -0.1940031476
#>  [53,]  0.60488817  0.2295598786
#>  [54,] -0.58141095  0.4547365098
#>  [55,]  0.41017206 -0.3427121211
#>  [56,] -0.40398845 -0.1303010591
#>  [57,] -0.65042178  0.8014668339
#>  [58,] -0.50315445  0.1487574773
#>  [59,]  0.49457837 -0.1531284444
#>  [60,]  0.61757229  0.2851337592
#>  [61,] -0.58581890  0.4745456129
#>  [62,] -0.24145761 -0.3588802707
#>  [63,] -0.33308289 -0.2640578702
#>  [64,]  0.36781987 -0.4083193219
#>  [65,] -0.18954372 -0.3716881706
#>  [66,] -0.34088551 -0.2519130923
#>  [67,]  0.29447004 -0.4775093856
#>  [68,]  0.39943727 -0.3611459674
#>  [69,]  0.47248708 -0.2106546114
#>  [70,] -0.28752708 -0.3223069547
#>  [71,] -0.59455166  0.5146734180
#>  [72,] -0.40889725 -0.1190735936
#>  [73,] -0.37138205 -0.1983638148
#>  [74,] -0.54485573  0.3014294512
#>  [75,] -0.47854363  0.0689068606
#>  [76,]  0.39150408 -0.3739767963
#>  [77,] -0.47273295  0.0511118399
#>  [78,]  0.36182016 -0.4160833974
#>  [79,] -0.26440101 -0.3435234949
#>  [80,] -0.20169622 -0.3715279138
#>  [81,]  0.53987690 -0.0162509355
#>  [82,]  0.38090450 -0.3900774851
#>  [83,] -0.41891766 -0.0953472839
#>  [84,] -0.64963888  0.7970597728
#>  [85,]  0.26252643 -0.4900237273
#>  [86,]  0.28754790 -0.4811391526
#>  [87,] -0.44245135 -0.0353041666
#>  [88,] -0.44578022 -0.0263146907
#>  [89,]  0.56812600  0.0829379073
#>  [90,] -0.30780632 -0.2990532134
#>  [91,] -0.54378752  0.2972326024
#>  [92,]  0.30723463 -0.4694974136
#>  [93,] -0.72198586  1.2603213739
#>  [94,] -0.42355218 -0.0840047125
#>  [95,] -0.51094759  0.1755909041
#>  [96,]  0.68020916  0.6025694540
#>  [97,]  0.30003983 -0.4742232099
#>  [98,]  0.82888116  1.7846052481
#>  [99,] -0.24601794 -0.3562865805
#> [100,] -0.41198298 -0.1118830005
#> [101,] -0.48049046  0.0749581037
#> [102,] -0.49324381  0.1157191217
#> [103,] -0.34327921 -0.2480604410
#> [104,] -0.38936301 -0.1622240549
#> [105,] -0.42616493 -0.0775067330
#> [106,] -0.45148135 -0.0106292084
#> [107,]  0.36769072 -0.4084904174
#> [108,] -0.26860303 -0.3400926154
#> [109,]  0.44194476 -0.2808106494
#> [110,]  0.34661621 -0.4340758440
#> [111,]  0.42590837 -0.3134382525
#> [112,] -0.24745693 -0.3554205574
#> [113,] -0.39422934 -0.1518553757
#> [114,] -0.24296182 -0.3580501484
#> [115,] -0.43046994 -0.0666365668
#> [116,] -0.30445939 -0.3031879209
#> [117,] -0.64949728  0.7962639524
#> [118,]  0.61938272  0.2932867748
#> [119,] -0.50703397  0.1620206293
#> [120,]  0.48562451 -0.1771476088
#> [121,]  0.33075713 -0.4502799838
#> [122,] -0.44681935 -0.0234830652
#> [123,]  0.65208673  0.4506378314
#> [124,] -0.40530301 -0.1273197749
#> [125,]  0.30391882 -0.4717426050
#> [126,] -0.31148262 -0.2943768481
#> [127,]  0.54815742  0.0116742150
#> [128,]  0.58898109  0.1635609539
#> [129,] -0.29937325 -0.3092469314
#> [130,]  0.27836567 -0.4851735889
#> [131,] -0.45507832 -0.0005435284
#> [132,] -0.33965295 -0.2538736609
#> [133,] -0.45345428 -0.0051154411
#> [134,] -0.28929718 -0.3204492809
#> [135,] -0.45511213 -0.0004480313
#> [136,]  0.50557353 -0.1222948810
#> [137,] -0.23705804 -0.3611641047
#> [138,]  0.34289990 -0.4381076022
#> [139,]  0.56644077  0.0767043037
#> [140,] -0.16282806 -0.3655199389
#> [141,] -0.52120131  0.2120645787
#> [142,]  0.77028187  1.2222028642
#> [143,]  0.31640329 -0.4626918878
#> [144,]  0.41245150 -0.3386382307
#> [145,] -0.41384562 -0.1074927542
#> [146,] -0.42694749 -0.0755459236
#> [147,] -0.20972712 -0.3704476015
#> [148,]  0.21786471 -0.4889373901
#> [149,]  0.59889784  0.2042331666
#> [150,] -0.32271700 -0.2792135728
#> [151,]  0.70579746  0.7560000566
#> [152,] -0.73629479  1.3674525172
#> [153,] -0.36206073 -0.2157613799
#> [154,] -0.58317793  0.4626417233
#> [155,] -0.24159786 -0.3588039340
#> [156,] -0.41645367 -0.1012824189
#> [157,] -0.24246869 -0.3583250512
#> [158,]  0.42443975 -0.3162844977
#> [159,] -0.41000981 -0.1164928746
#> [160,]  0.26686894 -0.4889613550
#> [161,] -0.26505930 -0.3429985410
#> [162,] -0.33408950 -0.2625266303
#> [163,] -0.22603145 -0.3659355052
#> [164,] -0.42427116 -0.0822240465
#> [165,]  0.40273809 -0.3556094386
#> [166,]  0.47951526 -0.1929836045
#> [167,]  0.34455469 -0.4363300958
#> [168,]  0.40898199 -0.3448167691
#> [169,] -0.38730875 -0.1665256329
#> [170,] -0.54603725  0.3060896655
#> [171,]  0.36273205 -0.4149276126
#> [172,]  0.33416059 -0.4470228029
#> [173,] -0.43324539 -0.0595202715
#> [174,] -0.24509359 -0.3568308533
#> [175,]  0.20711529 -0.4853022321
#> [176,]  0.39992211 -0.3603400366
#> [177,]  0.55650686  0.0407871539
#> [178,] -0.58211200  0.4578672432
#> [179,]  0.53742301 -0.0243481308
#> [180,] -0.17307921 -0.3689779013
#> [181,] -0.54814230  0.3144398232
#> [182,] -0.47898435  0.0702727928
#> [183,] -0.30754944 -0.2993747115
#> [184,]  0.58500513  0.1476842390
#> [185,] -0.49383703  0.1176628686
#> [186,]  0.32093053 -0.4590085208
#> [187,] -0.45544864  0.0005031620
#> [188,]  0.46517009 -0.2284372571
#> [189,]  0.52423464 -0.0664970300
#> [190,] -0.54307882  0.2944567625
#> [191,]  0.44700193 -0.2699274190
#> [192,]  0.42170437 -0.3215228455
#> [193,]  0.34314244 -0.4378488553
#> [194,] -0.22595568 -0.3659635405
#> [195,] -0.22927182 -0.3646755930
#> [196,] -0.35843491 -0.2222832354
#> [197,]  0.45122455 -0.2606195462
#> [198,] -0.09197210 -0.2988561902
#> [199,] -0.29429122 -0.3150303337
#> [200,] -0.34438549 -0.2462597536
#> [201,] -0.25094091 -0.3532296809
#> [202,]  0.63370554  0.3598010008
#> [203,] -0.32183614 -0.2804499536
#> [204,]  0.29120411 -0.4792847853
#> [205,] -0.66201325  0.8681279739
#> [206,] -0.58516202  0.4715748464
#> [207,]  0.42889145 -0.3075840548
#> [208,] -0.09801533 -0.3078813149
#> [209,] -0.34560618 -0.2442580757
#> [210,] -0.30594582 -0.3013660751
#> [211,]  0.48124264 -0.1885510664
#> [212,] -0.46277692  0.0215392939
#> [213,] -0.48900653  0.1019593478
#> [214,] -0.27369036 -0.3356857579
#> [215,] -0.33608777 -0.2594556885
#> [216,] -0.35097745 -0.2352660361
#> [217,] -0.57951865  0.4463229310
#> [218,] -0.64976147  0.7977490814
#> [219,] -0.48541228  0.0904574690
#> [220,] -0.69604551  1.0802206759
#> [221,] -0.27473091 -0.3347503499
#> [222,] -0.48982858  0.1046118029
#> [223,]  0.39698531 -0.3651832847
#> [224,] -0.30921665 -0.2972759065
#> [225,] -0.52131063  0.2124606967
#> [226,] -0.36680190 -0.2070260044
#> [227,] -0.26380540 -0.3439944395
#> [228,] -0.32806045 -0.2715405446
#> [229,] -0.25896594 -0.3476791473
#> [230,] -0.44976367 -0.0153936843
#> [231,]  0.65057197  0.4429096402
#> [232,] -0.23853876 -0.3604194982
#> [233,] -0.30275985 -0.3052426618
#> [234,]  0.30363270 -0.4719309545
#> [235,] -0.42534093 -0.0795641162
#> [236,] -0.45846478  0.0090866013
#> [237,]  0.60049933  0.2109475053
#> [238,] -0.49610100  0.1251204181
#> [239,] -0.54299504  0.2941290474
#> [240,] -0.42034484 -0.0918793689
#> [241,] -0.61927366  0.6349075104
#> [242,] -0.40188498 -0.1350329601
#> [243,] -0.36907960 -0.2027457945
#> [244,] -0.28356331 -0.3263468619
#> [245,]  0.30205507 -0.4729541209
#> [246,] -0.42986554 -0.0681749853
#> [247,] -0.52907328  0.2409839127
#> [248,]  0.53192141 -0.0422096547
#> [249,]  0.18733815 -0.4750168494
#> [250,] -0.48914719  0.1024126325
#> [251,]  0.28239534 -0.4835129463
#> [252,] -0.49705567  0.1282838703
#> [253,]  0.45228900 -0.2582413525
#> [254,]  0.46727915 -0.2233754484
#> [255,] -0.41157900 -0.1128302382
#> [256,] -0.39485190 -0.1505107194
#> [257,] -0.47484038  0.0575198587
#> [258,]  0.33534694 -0.4458591464
#> [259,]  0.49550629 -0.1505836578
#> [260,] -0.47947577  0.0717985896
#> [261,] -0.23931740 -0.3600181473
#> [262,] -0.37831715 -0.1848287198
#> [263,]  0.47821447 -0.1962981302
#> [264,]  0.49618248 -0.1487226246
#> [265,] -0.34988277 -0.2371230104
#> [266,] -0.44655588 -0.0242021540
#> [267,]  0.35023332 -0.4300136955
#> [268,] -0.60228856  0.5512279095
#> [269,]  0.35194221 -0.4280471702
#> [270,] -0.68023645  0.9785169545
#> [271,]  0.52585015 -0.0614566848
#> [272,] -0.58915245  0.4897239267
#> [273,]  0.57403128  0.1051084153
#> [274,] -0.45276051 -0.0070593614
#> [275,]  0.23088074 -0.4915543461
#> [276,] -0.41418865 -0.1066801359
#> [277,] -0.19189578 -0.3717978396
#> [278,]  0.61622773  0.2791149497
#> [279,]  0.72817822  0.9036844907
#> [280,] -0.29072925 -0.3189222126
#> [281,] -0.51824411  0.2014078191
#> [282,]  0.44988642 -0.2635909721
#> [283,] -0.54652230  0.3080083056
#> [284,]  0.34281543 -0.4381975749
#> [285,] -0.51578848  0.1926436409
#> [286,] -0.47975096  0.0726542434
#> [287,] -0.40705679 -0.1233134723
#> [288,] -0.45847254  0.0091088070
#> [289,] -0.52745884  0.2349871501
#> [290,] -0.33108955 -0.2670590233
#> [291,] -0.28129361 -0.3285853522
#> [292,] -0.49031180  0.1061748189
#> [293,]  0.40226631 -0.3564079037
#> [294,] -0.46565799  0.0299786942
#> [295,] -0.35114732 -0.2349767584
#> [296,] -0.24047138 -0.3594109327
#> [297,] -0.54213365  0.2907653427
#> [298,] -0.43937970 -0.0434889957
#> [299,] -0.38743530 -0.1662619326
#> [300,] -0.23207536 -0.3634897193
#> 
#> $inv_hessian
#>           [,1]      [,2]
#> [1,] 4.4654436 0.2523231
#> [2,] 0.2523231 4.8551413
#>