PPI++ Logistic Regression

Helper function for PPI++ logistic regression

Usage

ppi_plusplus_logistic(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  opts = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see doi:10.48550/arXiv.2311.01453 ). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

opts

(list, optional): Options to pass to the optimizer. See ?optim for details.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ logistic regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier logistic regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) doi:10.48550/arXiv.2311.01453

Examples

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_plusplus_logistic(X_l, Y_l, f_l, X_u, f_u)
#> $est
#>           [,1]
#> [1,] 0.4867633
#> [2,] 0.7831499
#> 
#> $se
#> [1] 0.1287374 0.1380527
#> 
#> $lambda
#> [1] 0.09444776
#> 
#> $rectifier_est
#>             [,1]
#> [1,] -0.02243142
#> [2,]  0.02812663
#> 
#> $var_u
#>               [,1]          [,2]
#> [1,]  0.0014843434 -0.0005694134
#> [2,] -0.0005694134  0.0009791453
#> 
#> $var_l
#>              [,1]         [,2]
#> [1,]  0.209228546 -0.003635673
#> [2,] -0.003635673  0.141135168
#> 
#> $grads
#>               [,1]         [,2]
#>   [1,] -0.16720832 -0.238867804
#>   [2,] -0.44081164  0.140095311
#>   [3,]  0.47130316 -0.362091929
#>   [4,] -0.31116082 -0.122348484
#>   [5,] -0.28883430 -0.152794344
#>   [6,] -0.21026140 -0.224608348
#>   [7,]  0.72476840  0.445583352
#>   [8,] -0.52491663  0.393117779
#>   [9,]  0.40998966 -0.445390471
#>  [10,]  0.46747278 -0.368329219
#>  [11,] -0.44392837  0.148247895
#>  [12,] -0.17551201 -0.237622398
#>  [13,] -0.27678238 -0.167421414
#>  [14,]  0.59966885 -0.063307529
#>  [15,] -0.60863221  0.721460063
#>  [16,]  0.32109234 -0.506564481
#>  [17,]  0.24943438 -0.505905983
#>  [18,]  0.72164947  0.429308505
#>  [19,] -0.45506480  0.178118986
#>  [20,]  0.69796502  0.312698774
#>  [21,]  0.51790029 -0.274528215
#>  [22,] -0.34844334 -0.061900450
#>  [23,] -0.18238969 -0.236031141
#>  [24,]  0.46039219 -0.379487619
#>  [25,] -0.43354375  0.121433335
#>  [26,]  0.55999630 -0.175629456
#>  [27,] -0.58682348  0.627632934
#>  [28,] -0.34686794 -0.064697277
#>  [29,] -0.57312097  0.571810711
#>  [30,]  0.71078672  0.374334778
#>  [31,]  0.25526151 -0.507657390
#>  [32,]  0.57939337 -0.123165722
#>  [33,]  0.41300581 -0.442098792
#>  [34,]  0.76600335  0.683809815
#>  [35,]  0.21855748 -0.491410696
#>  [36,] -0.09152391 -0.211341801
#>  [37,] -0.22394216 -0.216200360
#>  [38,] -0.31227564 -0.120714937
#>  [39,] -0.32508180 -0.101180039
#>  [40,] -0.46124384  0.195197235
#>  [41,] -0.44662265  0.155368246
#>  [42,] -0.30435968 -0.132081321
#>  [43,] -0.23558674 -0.207645498
#>  [44,]  0.62302495  0.012445606
#>  [45,] -0.09753506 -0.216473276
#>  [46,] -0.33865647 -0.078928571
#>  [47,] -0.39754535  0.036071689
#>  [48,] -0.22886210 -0.212740907
#>  [49,] -0.47943357  0.247599414
#>  [50,] -0.28633827 -0.155928302
#>  [51,]  0.58981795 -0.093050221
#>  [52,]  0.48896836 -0.331471482
#>  [53,] -0.52665914  0.399122537
#>  [54,] -0.40195856  0.045912150
#>  [55,] -0.20529629 -0.227211852
#>  [56,] -0.11833712 -0.229906175
#>  [57,]  0.36004718 -0.488210853
#>  [58,] -0.48866734  0.275438880
#>  [59,] -0.33175270 -0.090443649
#>  [60,] -0.42100440  0.090373213
#>  [61,]  0.35539978 -0.491088616
#>  [62,] -0.45937076  0.189981912
#>  [63,] -0.33840571 -0.079354011
#>  [64,] -0.57706418  0.587635384
#>  [65,] -0.26710121 -0.178243723
#>  [66,]  0.80440016  0.952423317
#>  [67,] -0.49000809  0.279551694
#>  [68,] -0.32541581 -0.100651587
#>  [69,] -0.52682266  0.399687720
#>  [70,] -0.45161335  0.168736746
#>  [71,]  0.50033723 -0.310120566
#>  [72,]  0.49453374 -0.321182843
#>  [73,] -0.23418449 -0.208742785
#>  [74,]  0.60753271 -0.038640735
#>  [75,]  0.78184714  0.788386306
#>  [76,]  0.20502205 -0.482209684
#>  [77,] -0.37576757 -0.009974738
#>  [78,] -0.35160330 -0.056225949
#>  [79,]  0.49916330 -0.312385890
#>  [80,]  0.42584157 -0.427174542
#>  [81,]  0.78657102  0.821187347
#>  [82,] -0.62079529  0.776589570
#>  [83,]  0.89661248  1.915814742
#>  [84,] -0.32772171 -0.096977138
#>  [85,] -0.30330398 -0.133556155
#>  [86,] -0.28704393 -0.155047826
#>  [87,]  0.21862929 -0.491454793
#>  [88,]  0.53807288 -0.229600077
#>  [89,] -0.35597612 -0.048231163
#>  [90,]  0.49717671 -0.316187408
#>  [91,] -0.61132313  0.733479107
#>  [92,] -0.14013209 -0.237522284
#>  [93,]  0.47765668 -0.351432014
#>  [94,] -0.42819836  0.108017477
#>  [95,]  0.73711923  0.512294692
#>  [96,]  0.75223914  0.599204726
#>  [97,] -0.31231978 -0.120650028
#>  [98,]  0.75947124  0.642973025
#>  [99,] -0.18262590 -0.235967630
#> [100,]  0.43357230 -0.417464167
#> [101,] -0.35333890 -0.053072542
#> [102,]  0.58142745 -0.117397147
#> [103,] -0.41781370  0.082697651
#> [104,] -0.39974058  0.040944982
#> [105,] -0.42824499  0.108133395
#> [106,]  0.72304140  0.436544035
#> [107,] -0.17681295 -0.237359891
#> [108,]  0.41956611 -0.434657018
#> [109,]  0.83152989  1.178301469
#> [110,] -0.21997805 -0.218821110
#> [111,]  0.80239975  0.937078463
#> [112,]  0.74929781  0.581821717
#> [113,]  0.23509022 -0.500274672
#> [114,]  0.61382024 -0.018312417
#> [115,] -0.37871084 -0.003994503
#> [116,] -0.37787533 -0.005699808
#> [117,] -0.44292032  0.145601232
#> [118,]  0.67970233  0.230551960
#> [119,] -0.58814170  0.633127778
#> [120,] -0.20560658 -0.227056232
#> [121,]  0.68070561  0.234901138
#> [122,] -0.30378982 -0.132878629
#> [123,] -0.13850433 -0.237164086
#> [124,] -0.42580137  0.102086253
#> [125,]  0.54339844 -0.216992782
#> [126,]  0.65553893  0.131177543
#> [127,]  0.63003157  0.036675608
#> [128,] -0.30721430 -0.128044969
#> [129,]  0.51054180 -0.289831762
#> [130,]  0.29400656 -0.511603853
#> [131,] -0.44415126  0.148834369
#> [132,]  0.39855494 -0.457132460
#> [133,] -0.54228186  0.454443319
#> [134,]  0.79836630  0.906631487
#> [135,] -0.11527526 -0.228326521
#> [136,] -0.31640384 -0.114572294
#> [137,] -0.18203414 -0.236125636
#> [138,] -0.49568764  0.297174282
#> [139,] -0.08754039 -0.207606548
#> [140,] -0.14471235 -0.238355794
#> [141,] -0.13528364 -0.236358111
#> [142,]  0.59120535 -0.088937731
#> [143,]  0.47797643 -0.350885113
#> [144,]  0.46161825 -0.377589778
#> [145,]  0.45391328 -0.389279271
#> [146,]  0.45905882 -0.381535294
#> [147,] -0.44726062  0.157064190
#> [148,]  0.45894336 -0.381711823
#> [149,] -0.65783443  0.957937655
#> [150,]  0.84745686  1.328866755
#> [151,] -0.40041744  0.042456178
#> [152,]  0.68330743  0.246266527
#> [153,] -0.64143106  0.875016544
#> [154,]  0.60634268 -0.042427186
#> [155,]  0.69100549  0.280638345
#> [156,] -0.43016286  0.112917596
#> [157,]  0.64096935  0.075960868
#> [158,]  0.47639568 -0.353579036
#> [159,] -0.30074226 -0.137094750
#> [160,]  0.63490797  0.053966817
#> [161,] -0.48149290  0.253735132
#> [162,] -0.30453657 -0.131833259
#> [163,] -0.37569140 -0.010128502
#> [164,] -0.40645334  0.056112306
#> [165,]  0.54388553 -0.215823807
#> [166,] -0.20994590 -0.224780951
#> [167,] -0.26379247 -0.181751598
#> [168,] -0.45893547  0.188774713
#> [169,] -0.59998260  0.683493720
#> [170,]  0.42061594 -0.433430109
#> [171,]  0.53881707 -0.227857389
#> [172,] -0.35746040 -0.045479878
#> [173,] -0.68973343  1.132282122
#> [174,] -0.43681337  0.129768579
#> [175,]  0.70242965  0.333775269
#> [176,] -0.13150476 -0.235245337
#> [177,] -0.30907787 -0.125371851
#> [178,]  0.56004515 -0.175503005
#> [179,]  0.62493180  0.018968882
#> [180,]  0.47430102 -0.357111131
#> [181,] -0.19197678 -0.232988515
#> [182,]  0.54905734 -0.203245966
#> [183,] -0.23934027 -0.204618649
#> [184,] -0.32620083 -0.099405807
#> [185,]  0.58973178 -0.093304822
#> [186,] -0.31471519 -0.117102750
#> [187,] -0.49410317  0.292225252
#> [188,] -0.17348231 -0.237995754
#> [189,] -0.22201891 -0.217490508
#> [190,] -0.28768099 -0.154249197
#> [191,]  0.42846276 -0.423943293
#> [192,]  0.63570928  0.056842350
#> [193,] -0.38948364  0.018540476
#> [194,] -0.46504125  0.205873351
#> [195,]  0.55875581 -0.178831283
#> [196,]  0.67801052  0.223259662
#> [197,] -0.27687598 -0.167312725
#> [198,] -0.39696822  0.034797571
#> [199,] -0.39456610  0.029526132
#> [200,] -0.41423738  0.074203509
#> [201,] -0.17427169 -0.237855808
#> [202,]  0.65549967  0.131024162
#> [203,] -0.55875231  0.515739699
#> [204,]  0.38947087 -0.465634591
#> [205,] -0.45292262  0.172282577
#> [206,] -0.41878533  0.085025253
#> [207,] -0.42747292  0.106216889
#> [208,] -0.34722170 -0.064071119
#> [209,] -0.22978706 -0.212065103
#> [210,] -0.43876968  0.134802890
#> [211,]  0.88270251  1.726194682
#> [212,]  0.37929591 -0.474296065
#> [213,] -0.26105086 -0.184584154
#> [214,] -0.17582502 -0.237560889
#> [215,] -0.48476419  0.263567903
#> [216,]  0.50971048 -0.291524958
#> [217,] -0.29463133 -0.145305755
#> [218,] -0.29919456 -0.139205034
#> [219,] -0.57371383  0.574177743
#> [220,]  0.40957831 -0.445833087
#> [221,] -0.14061075 -0.237621392
#> [222,]  0.60591057 -0.043797277
#> [223,] -0.27172116 -0.173182784
#> [224,] -0.35935076 -0.041948291
#> [225,]  0.50090568 -0.309018583
#> [226,] -0.26504187 -0.180438478
#> [227,] -0.29052581 -0.150639570
#> [228,] -0.34186352 -0.073439634
#> [229,]  0.52952520 -0.249177402
#> [230,]  0.73084478  0.477944083
#> [231,]  0.38876193 -0.466267516
#> [232,]  0.68011370  0.232333004
#> [233,] -0.49195737  0.285563359
#> [234,]  0.18986396 -0.469758573
#> [235,]  0.76431715  0.673146221
#> [236,] -0.25464600 -0.190938731
#> [237,] -0.34870416 -0.061435355
#> [238,]  0.59653452 -0.072908870
#> [239,] -0.55201143  0.490275855
#> [240,] -0.19349298 -0.232420920
#> [241,]  0.45906049 -0.381532743
#> [242,] -0.32309376 -0.104305449
#> [243,] -0.18766712 -0.234473631
#> [244,]  0.39491634 -0.460625418
#> [245,] -0.35811071 -0.044268441
#> [246,] -0.44153622  0.141982599
#> [247,] -0.34437107 -0.069086140
#> [248,]  0.52883748 -0.250717709
#> [249,] -0.21958141 -0.219075101
#> [250,] -0.23407054 -0.208831152
#> [251,]  0.62489390  0.018838706
#> [252,] -0.19156146 -0.233139909
#> [253,] -0.26831110 -0.176936670
#> [254,] -0.59528756  0.663304015
#> [255,] -0.15839429 -0.239358340
#> [256,] -0.36121355 -0.038437930
#> [257,]  0.77073786  0.714217184
#> [258,] -0.22802586 -0.213344988
#> [259,] -0.16495127 -0.239076010
#> [260,] -0.16093692 -0.239306269
#> [261,] -0.23521910 -0.207934949
#> [262,] -0.10379840 -0.221202237
#> [263,] -0.51238989  0.350905521
#> [264,]  0.77337960  0.731489323
#> [265,]  0.73154101  0.481708227
#> [266,]  0.31921869 -0.507117387
#> [267,] -0.19366855 -0.232353684
#> [268,] -0.30776647 -0.127256066
#> [269,]  0.49073444 -0.328240251
#> [270,]  0.22728030 -0.496408125
#> [271,]  0.44968402 -0.395457166
#> [272,]  0.26789783 -0.510405988
#> [273,] -0.24660262 -0.198394092
#> [274,] -0.16278707 -0.239222542
#> [275,] -0.47705162  0.240554448
#> [276,] -0.14187437 -0.237869507
#> [277,] -0.41708202  0.080950455
#> [278,] -0.44967079  0.163505512
#> [279,]  0.46157851 -0.377651515
#> [280,] -0.21958365 -0.219073674
#> [281,]  0.56494778 -0.162668095
#> [282,]  0.66051252  0.150814075
#> [283,] -0.18698272 -0.234691914
#> [284,] -0.31886223 -0.110844365
#> [285,] -0.19135499 -0.233214517
#> [286,] -0.46212975  0.197675575
#> [287,] -0.33178043 -0.090398216
#> [288,] -0.28917327 -0.152364547
#> [289,]  0.50130164 -0.308249014
#> [290,] -0.52822658  0.404552081
#> [291,] -0.26938850 -0.175761792
#> [292,] -0.35490242 -0.050209503
#> [293,] -0.58980291  0.640084071
#> [294,] -0.20370922 -0.227992977
#> [295,] -0.35143071 -0.056538093
#> [296,]  0.30894267 -0.509610664
#> [297,] -0.42517668  0.100549076
#> [298,] -0.44826821  0.159750409
#> [299,] -0.28473704 -0.157910064
#> [300,]  0.67774310  0.222111798
#> 
#> $grads_hat_unlabeled
#>               [,1]         [,2]
#>   [1,]  0.40405612 -0.451628702
#>   [2,] -0.55879882  0.515917232
#>   [3,] -0.21887250 -0.219525323
#>   [4,] -0.14855347 -0.238859466
#>   [5,] -0.29888669 -0.139622334
#>   [6,] -0.44529132  0.151841413
#>   [7,] -0.35304527 -0.053607859
#>   [8,] -0.14996107 -0.239000082
#>   [9,]  0.49606963 -0.318288508
#>  [10,] -0.33842653 -0.079318703
#>  [11,]  0.74765096  0.572191750
#>  [12,] -0.25594237 -0.189682333
#>  [13,] -0.52403994  0.390109031
#>  [14,] -0.20988235 -0.224815603
#>  [15,]  0.44147780 -0.406966850
#>  [16,]  0.36907988 -0.482082294
#>  [17,] -0.10622758 -0.222873018
#>  [18,] -0.27248196 -0.172331201
#>  [19,]  0.38392110 -0.470471344
#>  [20,]  0.30452798 -0.510401314
#>  [21,] -0.22114169 -0.218067296
#>  [22,] -0.47250738  0.227268030
#>  [23,]  0.36586444 -0.484345072
#>  [24,] -0.47027648  0.220818841
#>  [25,] -0.39600306  0.032673399
#>  [26,] -0.27937906 -0.164377759
#>  [27,] -0.37462607 -0.012273762
#>  [28,]  0.48206194 -0.343808952
#>  [29,] -0.43258090  0.118997455
#>  [30,] -0.33901565 -0.078318236
#>  [31,] -0.49152531  0.284227560
#>  [32,] -0.27990210 -0.163757496
#>  [33,] -0.52215117  0.383654999
#>  [34,] -0.20133826 -0.229113503
#>  [35,]  0.32632319 -0.504860591
#>  [36,]  0.72532198  0.448495498
#>  [37,]  0.65359547  0.123615283
#>  [38,] -0.42821180  0.108050901
#>  [39,] -0.17953303 -0.236752830
#>  [40,]  0.36811014 -0.482774182
#>  [41,] -0.45048860  0.165703515
#>  [42,] -0.58146609  0.605528206
#>  [43,]  0.55022607 -0.200361385
#>  [44,] -0.23308451 -0.209590748
#>  [45,]  0.80907841  0.988963010
#>  [46,] -0.16666395 -0.238923171
#>  [47,] -0.19685498 -0.231079283
#>  [48,] -0.35696555 -0.046399257
#>  [49,]  0.62372670  0.014840148
#>  [50,] -0.56876449  0.554548461
#>  [51,]  0.27008684 -0.510737298
#>  [52,] -0.51383333  0.355684866
#>  [53,]  0.51206514 -0.286710337
#>  [54,]  0.23024794 -0.497944528
#>  [55,] -0.38543889  0.009960228
#>  [56,]  0.57313911 -0.140581026
#>  [57,] -0.44399103  0.148412733
#>  [58,] -0.44915600  0.162125130
#>  [59,] -0.25670722 -0.188933966
#>  [60,] -0.16885150 -0.238680918
#>  [61,] -0.34897671 -0.060948704
#>  [62,]  0.45400594 -0.389142035
#>  [63,]  0.48312242 -0.341945296
#>  [64,] -0.36699358 -0.027354262
#>  [65,] -0.18664372 -0.234798245
#>  [66,] -0.50124624  0.314737939
#>  [67,]  0.35675831 -0.490266720
#>  [68,] -0.17494396 -0.237731345
#>  [69,] -0.49772703  0.303581678
#>  [70,] -0.39867626  0.038576924
#>  [71,] -0.17846913 -0.236999615
#>  [72,]  0.56866193 -0.152753693
#>  [73,] -0.35877991 -0.043018016
#>  [74,] -0.08592405 -0.206012214
#>  [75,] -0.09000099 -0.209945981
#>  [76,] -0.19405521 -0.232204514
#>  [77,]  0.62537931  0.020507443
#>  [78,] -0.42509017  0.100336482
#>  [79,]  0.71550517  0.397893192
#>  [80,] -0.26662280 -0.178756960
#>  [81,] -0.51388675  0.355862188
#>  [82,] -0.27558985 -0.168799296
#>  [83,] -0.62489394  0.795638932
#>  [84,]  0.69761064  0.311043085
#>  [85,] -0.18663223 -0.234801829
#>  [86,] -0.70578359  1.227230401
#>  [87,]  0.48554677 -0.337642960
#>  [88,] -0.30011796 -0.137948494
#>  [89,] -0.53900763  0.442625406
#>  [90,] -0.21798580 -0.220081687
#>  [91,]  0.46626247 -0.370270625
#>  [92,] -0.39244836  0.024921009
#>  [93,] -0.20906284 -0.225258908
#>  [94,] -0.52662624  0.399008866
#>  [95,] -0.21862891 -0.219678912
#>  [96,]  0.42305728 -0.430538425
#>  [97,] -0.43478430  0.124584292
#>  [98,] -0.56616140  0.544343464
#>  [99,]  0.51343309 -0.283886620
#> [100,]  0.55302897 -0.193379537
#> [101,]  0.28474948 -0.511864280
#> [102,] -0.06839590 -0.185570843
#> [103,] -0.29723048 -0.141853079
#> [104,] -0.18951556 -0.233859974
#> [105,] -0.35366515 -0.052476880
#> [106,] -0.23769507 -0.205961375
#> [107,]  0.49982240 -0.311115776
#> [108,] -0.38665362  0.012521983
#> [109,]  0.52310100 -0.263366320
#> [110,] -0.61741600  0.761064473
#> [111,] -0.22190975 -0.217562685
#> [112,] -0.56466014  0.538494968
#> [113,] -0.31161118 -0.121689862
#> [114,] -0.26778603 -0.177505508
#> [115,] -0.52537586  0.394697112
#> [116,]  0.63010447  0.036931470
#> [117,]  0.41308195 -0.442014637
#> [118,] -0.35230260 -0.054958518
#> [119,] -0.22003442 -0.218784888
#> [120,]  0.48548293 -0.337757001
#> [121,]  0.37397431 -0.478465436
#> [122,] -0.23751457 -0.206107170
#> [123,] -0.62048985  0.775179621
#> [124,]  0.41117672 -0.444104692
#> [125,] -0.36950147 -0.022454987
#> [126,]  0.63075121  0.039205056
#> [127,]  0.32303505 -0.505959213
#> [128,] -0.25442626 -0.191150190
#> [129,] -0.19031830 -0.233582550
#> [130,]  0.43506588 -0.415525043
#> [131,]  0.36023926 -0.488087891
#> [132,]  0.44724457 -0.398944255
#> [133,] -0.21706692 -0.220650301
#> [134,]  0.39215495 -0.463198225
#> [135,]  0.48732912 -0.334442715
#> [136,] -0.26693313 -0.178424270
#> [137,] -0.26267294 -0.182916385
#> [138,] -0.24172049 -0.202631878
#> [139,]  0.45645782 -0.385481110
#> [140,] -0.18973450 -0.233784963
#> [141,] -0.50634795  0.331136332
#> [142,] -0.15626098 -0.239345291
#> [143,]  0.33476179 -0.501614561
#> [144,]  0.56141402 -0.171948010
#> [145,] -0.25307567 -0.192440324
#> [146,]  0.47775492 -0.351264086
#> [147,]  0.39822204 -0.457456905
#> [148,] -0.16480952 -0.239087203
#> [149,] -0.54702389  0.471772928
#> [150,]  0.37395984 -0.478476435
#> [151,] -0.31939421 -0.110030795
#> [152,] -0.38761501  0.014558637
#> [153,] -0.40847199  0.060751894
#> [154,] -0.19851884 -0.230373194
#> [155,]  0.33589765 -0.501130767
#> [156,]  0.26830281 -0.510470492
#> [157,] -0.19864276 -0.230319497
#> [158,]  0.68073538  0.235030500
#> [159,]  0.40287242 -0.452835654
#> [160,] -0.33048478 -0.092513811
#> [161,]  0.54839027 -0.204885392
#> [162,] -0.20287369 -0.228394210
#> [163,] -0.16031164 -0.239325858
#> [164,]  0.48306378 -0.342048626
#> [165,]  0.29569154 -0.511475999
#> [166,]  0.48018066 -0.347087765
#> [167,] -0.47414620  0.232036393
#> [168,]  0.56111288 -0.172731980
#> [169,] -0.32737166 -0.097537900
#> [170,] -0.49870220  0.306660422
#> [171,] -0.29275251 -0.147764925
#> [172,] -0.43034700  0.113378728
#> [173,] -0.38737912  0.014058156
#> [174,] -0.59468295  0.660725105
#> [175,]  0.41785061 -0.436640452
#> [176,]  0.24486236 -0.504315897
#> [177,]  0.43567444 -0.414729062
#> [178,] -0.31584690 -0.115409594
#> [179,]  0.58088155 -0.118950393
#> [180,]  0.35390166 -0.491976464
#> [181,] -0.32645628 -0.098999280
#> [182,]  0.64014907  0.072951518
#> [183,]  0.42191370 -0.431899666
#> [184,] -0.36212239 -0.036714349
#> [185,] -0.41813439  0.083464932
#> [186,]  0.71530304  0.396874000
#> [187,] -0.22564592 -0.215028200
#> [188,]  0.50002429 -0.310725822
#> [189,] -0.36613405 -0.029020816
#> [190,] -0.26393243 -0.181605194
#> [191,] -0.33242496 -0.089340411
#> [192,] -0.62250579  0.784510044
#> [193,]  0.37241971 -0.479636839
#> [194,]  0.34554217 -0.496574724
#> [195,] -0.40608534  0.055270429
#> [196,] -0.43329845  0.120811942
#> [197,] -0.28545107 -0.157029114
#> [198,]  0.52347960 -0.262542481
#> [199,] -0.52018977  0.376993218
#> [200,] -0.32305149 -0.104371538
#> [201,] -0.36221309 -0.036541958
#> [202,] -0.40029466  0.042181755
#> [203,]  0.17254096 -0.452637990
#> [204,] -0.34352069 -0.070568611
#> [205,] -0.21948871 -0.219134251
#> [206,] -0.14235982 -0.237959633
#> [207,]  0.54363869 -0.216416528
#> [208,]  0.35421307 -0.491793511
#> [209,]  0.59335063 -0.082529726
#> [210,] -0.34298915 -0.071492083
#> [211,]  0.25775908 -0.508314171
#> [212,] -0.26977356 -0.175339399
#> [213,] -0.41265350  0.070478390
#> [214,] -0.21368095 -0.222675516
#> [215,] -0.29556333 -0.144074467
#> [216,] -0.49250203  0.287249933
#> [217,]  0.42046226 -0.433610335
#> [218,]  0.41069068 -0.444632678
#> [219,] -0.31547084 -0.115973454
#> [220,] -0.50408124  0.323817387
#> [221,] -0.35238586 -0.054807332
#> [222,]  0.71682371  0.404563638
#> [223,] -0.26807279 -0.177195142
#> [224,] -0.20764301 -0.226011412
#> [225,] -0.29996006 -0.138163875
#> [226,] -0.46011643  0.192054137
#> [227,] -0.34052080 -0.075748522
#> [228,]  0.57275875 -0.141624653
#> [229,] -0.15563525 -0.239331575
#> [230,]  0.32624975 -0.504886146
#> [231,] -0.40788805  0.059406052
#> [232,]  0.56268124 -0.168637255
#> [233,]  0.54266180 -0.218755596
#> [234,] -0.38074938  0.000191842
#> [235,] -0.42398262  0.097620734
#> [236,] -0.38418464  0.007328722
#> [237,] -0.82027865  2.100055956
#> [238,]  0.26347612 -0.509606683
#> [239,] -0.32192031 -0.106134244
#> [240,] -0.25833719 -0.187321540
#> [241,]  0.49053052 -0.328614950
#> [242,] -0.45501232  0.177975477
#> [243,] -0.09172617 -0.211524219
#> [244,] -0.58818061  0.633290303
#> [245,] -0.25034997 -0.194993737
#> [246,]  0.30643956 -0.510079682
#> [247,]  0.74457117  0.554377248
#> [248,] -0.32478366 -0.101650920
#> [249,] -0.15889950 -0.239353828
#> [250,] -0.47401441  0.231651974
#> [251,] -0.26153472 -0.184089139
#> [252,] -0.46566881  0.207650991
#> [253,]  0.55212741 -0.195635134
#> [254,] -0.48913570  0.276873575
#> [255,] -0.57024585  0.560392402
#> [256,] -0.11015626 -0.225387874
#> [257,]  0.62431462  0.016851755
#> [258,]  0.44076330 -0.407939340
#> [259,] -0.22711945 -0.213992329
#> [260,] -0.38629427  0.011762800
#> [261,]  0.72119860  0.426974288
#> [262,] -0.22270826 -0.217032116
#> [263,]  0.35633411 -0.490525072
#> [264,] -0.11471560 -0.228023518
#> [265,] -0.23399006 -0.208893483
#> [266,] -0.43611340  0.127975861
#> [267,] -0.14123397 -0.237746210
#> [268,] -0.47658028  0.239166982
#> [269,] -0.40368227  0.049802570
#> [270,] -0.36297862 -0.035083996
#> [271,]  0.28272277 -0.511821232
#> [272,] -0.48445260  0.262626775
#> [273,] -0.33223367 -0.089654732
#> [274,]  0.16593477 -0.445264244
#> [275,] -0.25978966 -0.185864599
#> [276,] -0.12742171 -0.233836242
#> [277,] -0.30895536 -0.125548504
#> [278,] -0.47453756  0.233178991
#> [279,] -0.36035712 -0.040055574
#> [280,] -0.28458308 -0.158099427
#> [281,] -0.17016962 -0.238509627
#> [282,] -0.30833165 -0.126445840
#> [283,]  0.51692870 -0.276581551
#> [284,] -0.52719816  0.400986696
#> [285,] -0.20899055 -0.225297696
#> [286,] -0.47406761  0.231807132
#> [287,] -0.26007125 -0.185579946
#> [288,]  0.40178293 -0.453935529
#> [289,]  0.63444853  0.052322508
#> [290,]  0.36447674 -0.485293900
#> [291,] -0.27758997 -0.166481174
#> [292,] -0.38409755  0.007146514
#> [293,] -0.08493610 -0.205014881
#> [294,] -0.36861165 -0.024199539
#> [295,] -0.29004421 -0.151255609
#> [296,]  0.48535615 -0.337983352
#> [297,] -0.20932299 -0.225118890
#> [298,]  0.16097598 -0.439412701
#> [299,] -0.55873351  0.515667971
#> [300,]  0.50529989 -0.300388099
#> 
#> $grads_hat
#>               [,1]         [,2]
#>   [1,] -0.16720832 -0.238867804
#>   [2,]  0.55918836 -0.177716877
#>   [3,]  0.47130316 -0.362091929
#>   [4,]  0.68883918  0.270851673
#>   [5,] -0.28883430 -0.152794344
#>   [6,] -0.21026140 -0.224608348
#>   [7,] -0.27523160 -0.169210771
#>   [8,]  0.47508337 -0.355796918
#>   [9,]  0.40998966 -0.445390471
#>  [10,] -0.53252722  0.419586652
#>  [11,] -0.44392837  0.148247895
#>  [12,] -0.17551201 -0.237622398
#>  [13,] -0.27678238 -0.167421414
#>  [14,]  0.59966885 -0.063307529
#>  [15,]  0.39136779 -0.463919307
#>  [16,]  0.32109234 -0.506564481
#>  [17,]  0.24943438 -0.505905983
#>  [18,] -0.27835053 -0.165590434
#>  [19,]  0.54493520 -0.213295564
#>  [20,]  0.69796502  0.312698774
#>  [21,]  0.51790029 -0.274528215
#>  [22,] -0.34844334 -0.061900450
#>  [23,] -0.18238969 -0.236031141
#>  [24,] -0.53960781  0.444782710
#>  [25,] -0.43354375  0.121433335
#>  [26,] -0.44000370  0.137996642
#>  [27,]  0.41317652 -0.441910049
#>  [28,] -0.34686794 -0.064697277
#>  [29,] -0.57312097  0.571810711
#>  [30,] -0.28921328 -0.152313747
#>  [31,]  0.25526151 -0.507657390
#>  [32,] -0.42060663  0.089411308
#>  [33,]  0.41300581 -0.442098792
#>  [34,] -0.23399665 -0.208888386
#>  [35,]  0.21855748 -0.491410696
#>  [36,] -0.09152391 -0.211341801
#>  [37,]  0.77605784  0.749229084
#>  [38,] -0.31227564 -0.120714937
#>  [39,]  0.67491820  0.210064822
#>  [40,] -0.46124384  0.195197235
#>  [41,] -0.44662265  0.155368246
#>  [42,] -0.30435968 -0.132081321
#>  [43,] -0.23558674 -0.207645498
#>  [44,]  0.62302495  0.012445606
#>  [45,] -0.09753506 -0.216473276
#>  [46,] -0.33865647 -0.078928571
#>  [47,] -0.39754535  0.036071689
#>  [48,] -0.22886210 -0.212740907
#>  [49,]  0.52056643 -0.268842133
#>  [50,] -0.28633827 -0.155928302
#>  [51,] -0.41018205  0.064710696
#>  [52,]  0.48896836 -0.331471482
#>  [53,]  0.47334086 -0.358715899
#>  [54,] -0.40195856  0.045912150
#>  [55,] -0.20529629 -0.227211852
#>  [56,] -0.11833712 -0.229906175
#>  [57,]  0.36004718 -0.488210853
#>  [58,] -0.48866734  0.275438880
#>  [59,] -0.33175270 -0.090443649
#>  [60,]  0.57899560 -0.124287756
#>  [61,]  0.35539978 -0.491088616
#>  [62,]  0.54062924 -0.223587975
#>  [63,] -0.33840571 -0.079354011
#>  [64,]  0.42293582 -0.430683559
#>  [65,] -0.26710121 -0.178243723
#>  [66,]  0.80440016  0.952423317
#>  [67,]  0.50999191 -0.290952551
#>  [68,] -0.32541581 -0.100651587
#>  [69,]  0.47317734 -0.358988306
#>  [70,] -0.45161335  0.168736746
#>  [71,] -0.49966277  0.309702516
#>  [72,]  0.49453374 -0.321182843
#>  [73,] -0.23418449 -0.208742785
#>  [74,] -0.39246729  0.024961988
#>  [75,] -0.21815286 -0.219977439
#>  [76,]  0.20502205 -0.482209684
#>  [77,] -0.37576757 -0.009974738
#>  [78,] -0.35160330 -0.056225949
#>  [79,]  0.49916330 -0.312385890
#>  [80,]  0.42584157 -0.427174542
#>  [81,] -0.21342898 -0.222821805
#>  [82,]  0.37920471 -0.474369606
#>  [83,] -0.10338752 -0.220910738
#>  [84,] -0.32772171 -0.096977138
#>  [85,] -0.30330398 -0.133556155
#>  [86,] -0.28704393 -0.155047826
#>  [87,]  0.21862929 -0.491454793
#>  [88,] -0.46192712  0.197108065
#>  [89,] -0.35597612 -0.048231163
#>  [90,]  0.49717671 -0.316187408
#>  [91,]  0.38867687 -0.466343163
#>  [92,] -0.14013209 -0.237522284
#>  [93,]  0.47765668 -0.351432014
#>  [94,] -0.42819836  0.108017477
#>  [95,] -0.26288077 -0.182701004
#>  [96,] -0.24776086 -0.197356753
#>  [97,]  0.68768022  0.265652843
#>  [98,] -0.24052876 -0.203633130
#>  [99,] -0.18262590 -0.235967630
#> [100,]  0.43357230 -0.417464167
#> [101,]  0.64666110  0.097130400
#> [102,]  0.58142745 -0.117397147
#> [103,] -0.41781370  0.082697651
#> [104,] -0.39974058  0.040944982
#> [105,] -0.42824499  0.108133395
#> [106,]  0.72304140  0.436544035
#> [107,] -0.17681295 -0.237359891
#> [108,]  0.41956611 -0.434657018
#> [109,]  0.83152989  1.178301469
#> [110,] -0.21997805 -0.218821110
#> [111,] -0.19760025 -0.230766448
#> [112,] -0.25070219 -0.194667566
#> [113,]  0.23509022 -0.500274672
#> [114,] -0.38617976  0.011521101
#> [115,] -0.37871084 -0.003994503
#> [116,] -0.37787533 -0.005699808
#> [117,] -0.44292032  0.145601232
#> [118,]  0.67970233  0.230551960
#> [119,]  0.41185830 -0.443360732
#> [120,] -0.20560658 -0.227056232
#> [121,] -0.31929439 -0.110183632
#> [122,] -0.30378982 -0.132878629
#> [123,] -0.13850433 -0.237164086
#> [124,] -0.42580137  0.102086253
#> [125,] -0.45660156  0.182332583
#> [126,] -0.34446107 -0.068928870
#> [127,]  0.63003157  0.036675608
#> [128,] -0.30721430 -0.128044969
#> [129,]  0.51054180 -0.289831762
#> [130,]  0.29400656 -0.511603853
#> [131,] -0.44415126  0.148834369
#> [132,]  0.39855494 -0.457132460
#> [133,]  0.45771814 -0.383577180
#> [134,] -0.20163370 -0.228976921
#> [135,] -0.11527526 -0.228326521
#> [136,] -0.31640384 -0.114572294
#> [137,] -0.18203414 -0.236125636
#> [138,]  0.50431236 -0.302344970
#> [139,] -0.08754039 -0.207606548
#> [140,] -0.14471235 -0.238355794
#> [141,] -0.13528364 -0.236358111
#> [142,] -0.40879465  0.061496853
#> [143,]  0.47797643 -0.350885113
#> [144,]  0.46161825 -0.377589778
#> [145,]  0.45391328 -0.389279271
#> [146,]  0.45905882 -0.381535294
#> [147,] -0.44726062  0.157064190
#> [148,]  0.45894336 -0.381711823
#> [149,]  0.34216557 -0.498261066
#> [150,] -0.15254314 -0.239197427
#> [151,]  0.59958256 -0.063573615
#> [152,] -0.31669257 -0.114137173
#> [153,]  0.35856894 -0.489146485
#> [154,] -0.39365732  0.027545105
#> [155,] -0.30899451 -0.125492067
#> [156,] -0.43016286  0.112917596
#> [157,] -0.35903065 -0.042548493
#> [158,] -0.52360432  0.388617105
#> [159,] -0.30074226 -0.137094750
#> [160,]  0.63490797  0.053966817
#> [161,] -0.48149290  0.253735132
#> [162,] -0.30453657 -0.131833259
#> [163,]  0.62430860  0.016831130
#> [164,]  0.59354666 -0.081941196
#> [165,]  0.54388553 -0.215823807
#> [166,] -0.20994590 -0.224780951
#> [167,] -0.26379247 -0.181751598
#> [168,]  0.54106453 -0.222556996
#> [169,]  0.40001740 -0.455695521
#> [170,] -0.57938406  0.597035146
#> [171,]  0.53881707 -0.227857389
#> [172,] -0.35746040 -0.045479878
#> [173,]  0.31026657 -0.509340671
#> [174,]  0.56318663 -0.167311563
#> [175,] -0.29757035 -0.141397254
#> [176,] -0.13150476 -0.235245337
#> [177,] -0.30907787 -0.125371851
#> [178,] -0.43995485  0.137869951
#> [179,] -0.37506820 -0.011384641
#> [180,]  0.47430102 -0.357111131
#> [181,] -0.19197678 -0.232988515
#> [182,] -0.45094266  0.166926602
#> [183,] -0.23934027 -0.204618649
#> [184,] -0.32620083 -0.099405807
#> [185,]  0.58973178 -0.093304822
#> [186,] -0.31471519 -0.117102750
#> [187,]  0.50589683 -0.299200330
#> [188,] -0.17348231 -0.237995754
#> [189,] -0.22201891 -0.217490508
#> [190,] -0.28768099 -0.154249197
#> [191,]  0.42846276 -0.423943293
#> [192,]  0.63570928  0.056842350
#> [193,]  0.61051636 -0.029062232
#> [194,]  0.53495875 -0.236825768
#> [195,] -0.44124419  0.141221378
#> [196,]  0.67801052  0.223259662
#> [197,] -0.27687598 -0.167312725
#> [198,] -0.39696822  0.034797571
#> [199,]  0.60543390 -0.045305770
#> [200,] -0.41423738  0.074203509
#> [201,] -0.17427169 -0.237855808
#> [202,] -0.34450033 -0.068860245
#> [203,]  0.44124769 -0.407280560
#> [204,]  0.38947087 -0.465634591
#> [205,] -0.45292262  0.172282577
#> [206,] -0.41878533  0.085025253
#> [207,]  0.57252708 -0.142259413
#> [208,]  0.65277830  0.120453983
#> [209,] -0.22978706 -0.212065103
#> [210,]  0.56123032 -0.172426381
#> [211,] -0.11729749 -0.229384534
#> [212,]  0.37929591 -0.474296065
#> [213,] -0.26105086 -0.184584154
#> [214,] -0.17582502 -0.237560889
#> [215,]  0.51523581 -0.280135423
#> [216,]  0.50971048 -0.291524958
#> [217,] -0.29463133 -0.145305755
#> [218,] -0.29919456 -0.139205034
#> [219,]  0.42628617 -0.426630872
#> [220,]  0.40957831 -0.445833087
#> [221,] -0.14061075 -0.237621392
#> [222,]  0.60591057 -0.043797277
#> [223,] -0.27172116 -0.173182784
#> [224,] -0.35935076 -0.041948291
#> [225,] -0.49909432  0.307901123
#> [226,] -0.26504187 -0.180438478
#> [227,] -0.29052581 -0.150639570
#> [228,] -0.34186352 -0.073439634
#> [229,]  0.52952520 -0.249177402
#> [230,] -0.26915522 -0.176017053
#> [231,]  0.38876193 -0.466267516
#> [232,] -0.31988630 -0.109276058
#> [233,]  0.50804263 -0.294900265
#> [234,]  0.18986396 -0.469758573
#> [235,] -0.23568285 -0.207569619
#> [236,] -0.25464600 -0.190938731
#> [237,] -0.34870416 -0.061435355
#> [238,]  0.59653452 -0.072908870
#> [239,]  0.44798857 -0.397886647
#> [240,] -0.19349298 -0.232420920
#> [241,]  0.45906049 -0.381532743
#> [242,] -0.32309376 -0.104305449
#> [243,] -0.18766712 -0.234473631
#> [244,]  0.39491634 -0.460625418
#> [245,] -0.35811071 -0.044268441
#> [246,] -0.44153622  0.141982599
#> [247,] -0.34437107 -0.069086140
#> [248,] -0.47116252  0.223374462
#> [249,] -0.21958141 -0.219075101
#> [250,] -0.23407054 -0.208831152
#> [251,]  0.62489390  0.018838706
#> [252,] -0.19156146 -0.233139909
#> [253,] -0.26831110 -0.176936670
#> [254,]  0.40471244 -0.450954126
#> [255,] -0.15839429 -0.239358340
#> [256,] -0.36121355 -0.038437930
#> [257,] -0.22926214 -0.212449611
#> [258,] -0.22802586 -0.213344988
#> [259,] -0.16495127 -0.239076010
#> [260,] -0.16093692 -0.239306269
#> [261,] -0.23521910 -0.207934949
#> [262,] -0.10379840 -0.221202237
#> [263,]  0.48761011 -0.333935315
#> [264,] -0.22662040 -0.214345450
#> [265,] -0.26845899 -0.176776012
#> [266,]  0.31921869 -0.507117387
#> [267,] -0.19366855 -0.232353684
#> [268,] -0.30776647 -0.127256066
#> [269,] -0.50926556  0.340635258
#> [270,]  0.22728030 -0.496408125
#> [271,] -0.55031598  0.483954044
#> [272,]  0.26789783 -0.510405988
#> [273,] -0.24660262 -0.198394092
#> [274,] -0.16278707 -0.239222542
#> [275,] -0.47705162  0.240554448
#> [276,] -0.14187437 -0.237869507
#> [277,] -0.41708202  0.080950455
#> [278,] -0.44967079  0.163505512
#> [279,]  0.46157851 -0.377651515
#> [280,] -0.21958365 -0.219073674
#> [281,]  0.56494778 -0.162668095
#> [282,] -0.33948748 -0.077514791
#> [283,] -0.18698272 -0.234691914
#> [284,] -0.31886223 -0.110844365
#> [285,] -0.19135499 -0.233214517
#> [286,]  0.53787025 -0.230073503
#> [287,] -0.33178043 -0.090398216
#> [288,] -0.28917327 -0.152364547
#> [289,]  0.50130164 -0.308249014
#> [290,]  0.47177342 -0.361316387
#> [291,] -0.26938850 -0.175761792
#> [292,] -0.35490242 -0.050209503
#> [293,] -0.58980291  0.640084071
#> [294,] -0.20370922 -0.227992977
#> [295,] -0.35143071 -0.056538093
#> [296,]  0.30894267 -0.509610664
#> [297,] -0.42517668  0.100549076
#> [298,]  0.55173179 -0.196621973
#> [299,] -0.28473704 -0.157910064
#> [300,]  0.67774310  0.222111798
#> 
#> $inv_hessian
#>          [,1]     [,2]
#> [1,] 4.821593 0.875322
#> [2,] 0.875322 6.278641
#>