PPI++ Logistic Regression

Helper function for PPI++ logistic regression

Usage

ppi_plusplus_logistic(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  opts = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see https://arxiv.org/abs/2311.01453). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

opts

(list, optional): Options to pass to the optimizer. See ?optim for details.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ logistic regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier logistic regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) https://arxiv.org/abs/2311.01453

Examples

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_plusplus_logistic(X_l, Y_l, f_l, X_u, f_u)
#> $est
#>           [,1]
#> [1,] 0.4865938
#> [2,] 0.7040716
#> 
#> $se
#> [1] 0.1268517 0.1347547
#> 
#> $lambda
#> [1] 0.081643
#> 
#> $rectifier_est
#>              [,1]
#> [1,] -0.006157121
#> [2,] -0.006461035
#> 
#> $var_u
#>               [,1]          [,2]
#> [1,]  0.0013050030 -0.0005640268
#> [2,] -0.0005640268  0.0011172474
#> 
#> $var_l
#>            [,1]       [,2]
#> [1,]  0.2178305 -0.0413854
#> [2,] -0.0413854  0.1667002
#> 
#> $grads
#>               [,1]         [,2]
#>   [1,]  0.64981379  0.121483408
#>   [2,] -0.34865053 -0.068525846
#>   [3,]  0.55159384 -0.218955189
#>   [4,] -0.67080193  1.141779389
#>   [5,] -0.29953318 -0.154400444
#>   [6,]  0.56397541 -0.183659285
#>   [7,]  0.54753800 -0.230087001
#>   [8,] -0.28019681 -0.181827539
#>   [9,] -0.36687570 -0.030769545
#>  [10,] -0.53069983  0.459451715
#>  [11,] -0.28516830 -0.175123236
#>  [12,]  0.41170112 -0.493249199
#>  [13,]  0.37593880 -0.530434067
#>  [14,]  0.50406934 -0.336715636
#>  [15,]  0.47235425 -0.400715374
#>  [16,]  0.62060624  0.004873018
#>  [17,] -0.21294580 -0.248209545
#>  [18,] -0.17319888 -0.264822150
#>  [19,] -0.47280342  0.253635917
#>  [20,]  0.58752019 -0.110875203
#>  [21,]  0.27273457 -0.568417965
#>  [22,] -0.55328453  0.550512664
#>  [23,]  0.44853982 -0.441592210
#>  [24,] -0.31170860 -0.135273833
#>  [25,] -0.34191820 -0.081665409
#>  [26,] -0.33408434 -0.096407560
#>  [27,]  0.73882457  0.580583978
#>  [28,] -0.38665732  0.013844019
#>  [29,] -0.14390616 -0.265018904
#>  [30,] -0.65956354  1.075376028
#>  [31,]  0.53629900 -0.259851493
#>  [32,]  0.36566398 -0.538810162
#>  [33,] -0.23376174 -0.232608103
#>  [34,] -0.38185370  0.002660886
#>  [35,]  0.70804096  0.401545717
#>  [36,] -0.40782411  0.065816400
#>  [37,]  0.38773758 -0.519551856
#>  [38,] -0.19255699 -0.258964580
#>  [39,]  0.49004503 -0.366395951
#>  [40,] -0.37440244 -0.014241907
#>  [41,] -0.37323324 -0.016845275
#>  [42,] -0.37737464 -0.007564277
#>  [43,]  0.47317933 -0.399190709
#>  [44,] -0.27284229 -0.191301755
#>  [45,] -0.27444274 -0.189285242
#>  [46,] -0.34057269 -0.084239323
#>  [47,] -0.45818541  0.207557627
#>  [48,]  0.15968778 -0.486986115
#>  [49,] -0.23638102 -0.230323878
#>  [50,] -0.37260416 -0.018240529
#>  [51,] -0.41869881  0.094243217
#>  [52,]  0.70125021  0.365195221
#>  [53,]  0.43045030 -0.468683339
#>  [54,] -0.44430724  0.165900673
#>  [55,] -0.18877292 -0.260450470
#>  [56,]  0.57769557 -0.142171872
#>  [57,] -0.50369016  0.358667240
#>  [58,] -0.14569671 -0.265324239
#>  [59,] -0.39765028  0.040285390
#>  [60,]  0.38480410 -0.522384106
#>  [61,] -0.36688923 -0.030740320
#>  [62,] -0.30050342 -0.152928717
#>  [63,]  0.38868949 -0.518614730
#>  [64,]  0.49317236 -0.359969437
#>  [65,] -0.11389706 -0.253159685
#>  [66,] -0.51348100  0.394210375
#>  [67,]  0.41565022 -0.488368515
#>  [68,] -0.26336085 -0.202729518
#>  [69,] -0.49157438  0.316201073
#>  [70,] -0.24103058 -0.226095128
#>  [71,] -0.20686027 -0.251897877
#>  [72,] -0.25371565 -0.213437745
#>  [73,]  0.50809014 -0.327793344
#>  [74,] -0.18238363 -0.262586825
#>  [75,] -0.39308637  0.029163980
#>  [76,]  0.58836060 -0.108134954
#>  [77,]  0.42479247 -0.476469698
#>  [78,]  0.71688530  0.450522602
#>  [79,] -0.14631699 -0.265420010
#>  [80,]  0.61443605 -0.017973930
#>  [81,] -0.21262401 -0.248414606
#>  [82,] -0.39866588  0.042788108
#>  [83,] -0.29954920 -0.154376221
#>  [84,] -0.36067404 -0.043976254
#>  [85,]  0.44982345 -0.439541324
#>  [86,]  0.42796296 -0.472146417
#>  [87,] -0.08646345 -0.229768931
#>  [88,] -0.39857539  0.042564691
#>  [89,]  0.71846942  0.459496369
#>  [90,]  0.68034657  0.259714369
#>  [91,] -0.25121153 -0.216065383
#>  [92,] -0.26292888 -0.203228951
#>  [93,] -0.33851627 -0.088139620
#>  [94,]  0.51939225 -0.301708065
#>  [95,] -0.58056018  0.669280865
#>  [96,] -0.44273106  0.161296392
#>  [97,]  0.65693971  0.152174527
#>  [98,]  0.66275415  0.177907030
#>  [99,] -0.41407472  0.082011373
#> [100,] -0.55265216  0.547875426
#> [101,] -0.12503347 -0.259100241
#> [102,] -0.56034047  0.580289531
#> [103,] -0.27401322 -0.189828891
#> [104,] -0.38694108  0.014511677
#> [105,] -0.59808372  0.750995038
#> [106,] -0.17385918 -0.264694851
#> [107,]  0.74647724  0.629052748
#> [108,]  0.61408899 -0.019241304
#> [109,]  0.47986005 -0.386573298
#> [110,]  0.72259351  0.483153554
#> [111,]  0.74094662  0.593855268
#> [112,]  0.52849703 -0.279596054
#> [113,] -0.11032917 -0.250846303
#> [114,]  0.69875228  0.352089213
#> [115,] -0.32233253 -0.117419275
#> [116,]  0.47353305 -0.398534817
#> [117,]  0.54100271 -0.247586392
#> [118,] -0.33841607 -0.088328625
#> [119,] -0.32344829 -0.115481251
#> [120,] -0.53239410  0.466063600
#> [121,] -0.43629188  0.142752507
#> [122,] -0.65468810  1.047305767
#> [123,]  0.59608577 -0.082471427
#> [124,] -0.21559247 -0.246480706
#> [125,]  0.59897741 -0.072642164
#> [126,]  0.64683064  0.108905436
#> [127,]  0.66842766  0.203629801
#> [128,] -0.20369820 -0.253655114
#> [129,]  0.24268172 -0.559981640
#> [130,]  0.50257979 -0.339973891
#> [131,]  0.86212376  1.648703378
#> [132,] -0.11388179 -0.253150218
#> [133,]  0.35985977 -0.543091029
#> [134,]  0.80482131  1.063205913
#> [135,] -0.58463148  0.687876332
#> [136,] -0.60770828  0.797780248
#> [137,] -0.25103346 -0.216249836
#> [138,]  0.51091890 -0.321404414
#> [139,] -0.36251580 -0.040092858
#> [140,] -0.42213669  0.103476302
#> [141,] -0.27057984 -0.194109336
#> [142,]  0.72790129  0.514241716
#> [143,] -0.24983776 -0.217480078
#> [144,] -0.33919524 -0.086856331
#> [145,] -0.51551722  0.401742273
#> [146,]  0.32901680 -0.560406785
#> [147,] -0.38961974  0.020852987
#> [148,] -0.15203039 -0.266062943
#> [149,]  0.54470045 -0.237750969
#> [150,] -0.50778685  0.373404410
#> [151,] -0.37879876 -0.004334323
#> [152,] -0.18648518 -0.261269500
#> [153,] -0.17875979 -0.263587726
#> [154,] -0.41962409  0.096716528
#> [155,]  0.48234777 -0.381750449
#> [156,] -0.36991803 -0.024155018
#> [157,] -0.14367857 -0.264977007
#> [158,] -0.35308189 -0.059639648
#> [159,] -0.33434752 -0.095921850
#> [160,] -0.43918904  0.151043120
#> [161,] -0.41833563  0.093274759
#> [162,]  0.80485517  1.063497083
#> [163,] -0.22866073 -0.236852544
#> [164,] -0.29594209 -0.159768508
#> [165,] -0.41778245  0.091802201
#> [166,] -0.27362095 -0.190323819
#> [167,]  0.48224806 -0.381945043
#> [168,] -0.31011931 -0.137851532
#> [169,]  0.31979636 -0.563811579
#> [170,]  0.65440178  0.141138265
#> [171,] -0.60995549  0.808904898
#> [172,]  0.68610723  0.287851823
#> [173,]  0.75741641  0.701370499
#> [174,] -0.29598632 -0.159703164
#> [175,]  0.82147547  1.213165834
#> [176,]  0.56106903 -0.192124283
#> [177,]  0.68121545  0.263914459
#> [178,] -0.23053175 -0.235327127
#> [179,]  0.71561434  0.443367731
#> [180,]  0.80844196  1.094643315
#> [181,] -0.59137209  0.719178308
#> [182,] -0.39792899  0.040971211
#> [183,]  0.38585442 -0.521379686
#> [184,] -0.31714229 -0.126277423
#> [185,] -0.24464094 -0.222658922
#> [186,]  0.58984631 -0.103266010
#> [187,]  0.53023581 -0.275259894
#> [188,] -0.13917820 -0.264004182
#> [189,] -0.44141153  0.157461569
#> [190,]  0.45128365 -0.437187439
#> [191,]  0.41728440 -0.486303226
#> [192,] -0.68369222  1.220990787
#> [193,] -0.57458088  0.642387764
#> [194,] -0.38928001  0.020044848
#> [195,] -0.10214638 -0.244750087
#> [196,] -0.25895505 -0.207736216
#> [197,]  0.49090436 -0.364640956
#> [198,] -0.52899202  0.452822300
#> [199,]  0.74025318  0.589504409
#> [200,]  0.70331693  0.376145698
#> [201,] -0.39554593  0.035132056
#> [202,] -0.67618080  1.174426673
#> [203,] -0.54321127  0.509109327
#> [204,]  0.55609332 -0.206359164
#> [205,] -0.22095903 -0.242745411
#> [206,]  0.74677383  0.630965518
#> [207,] -0.22268499 -0.241479106
#> [208,] -0.24428575 -0.223002865
#> [209,]  0.49154819 -0.363320678
#> [210,]  0.65562502  0.146442876
#> [211,] -0.34723896 -0.071316877
#> [212,] -0.47622592  0.264755847
#> [213,] -0.54508522  0.516714410
#> [214,] -0.34018333 -0.084980904
#> [215,] -0.44525458  0.168680375
#> [216,] -0.68736814  1.244202561
#> [217,]  0.37170464 -0.534009535
#> [218,]  0.25347506 -0.564053298
#> [219,] -0.72551104  1.502973046
#> [220,]  0.51967954 -0.301025454
#> [221,]  0.51476583 -0.312566552
#> [222,] -0.69638565  1.302364973
#> [223,] -0.24794139 -0.219401600
#> [224,] -0.15160669 -0.266029952
#> [225,]  0.60752569 -0.042859366
#> [226,]  0.73923852  0.583162782
#> [227,] -0.28792610 -0.171300248
#> [228,]  0.73639832  0.565565573
#> [229,] -0.19928784 -0.255922252
#> [230,] -0.40721240  0.064252372
#> [231,] -0.39031313  0.022505928
#> [232,]  0.58426208 -0.121403963
#> [233,] -0.36733815 -0.029769869
#> [234,]  0.43240209 -0.465921835
#> [235,]  0.25394992 -0.564205413
#> [236,]  0.31962163 -0.563868226
#> [237,] -0.47075024  0.247025812
#> [238,] -0.31391648 -0.131652536
#> [239,]  0.52782987 -0.281250090
#> [240,]  0.47776262 -0.390586886
#> [241,] -0.36548763 -0.033757718
#> [242,] -0.31066969 -0.136961609
#> [243,]  0.50968035 -0.324213179
#> [244,]  0.52658856 -0.284313290
#> [245,]  0.72355183  0.488712630
#> [246,]  0.69363076  0.325653024
#> [247,] -0.30596844 -0.144469194
#> [248,]  0.55923755 -0.197401520
#> [249,] -0.27324921 -0.190791434
#> [250,]  0.62855486  0.035199335
#> [251,]  0.71570287  0.443864855
#> [252,] -0.35042070 -0.064998759
#> [253,]  0.56657504 -0.175992939
#> [254,] -0.26510535 -0.200693718
#> [255,]  0.33138567 -0.559400397
#> [256,]  0.40757479 -0.498182855
#> [257,] -0.43466018  0.138121047
#> [258,]  0.51394173 -0.314474376
#> [259,] -0.12498879 -0.259080167
#> [260,] -0.50399080  0.359742197
#> [261,] -0.42263219  0.104816874
#> [262,] -0.54295223  0.508061553
#> [263,] -0.52616695  0.441933272
#> [264,] -0.22534902 -0.239462867
#> [265,]  0.53042433 -0.274787562
#> [266,]  0.57791630 -0.141483497
#> [267,] -0.41713013  0.090069666
#> [268,] -0.35535691 -0.055004221
#> [269,] -0.39034161  0.022573915
#> [270,] -0.53610961  0.480686160
#> [271,] -0.40568818  0.060371377
#> [272,]  0.51506095 -0.311881418
#> [273,] -0.36928737 -0.025533521
#> [274,] -0.36262860 -0.039853953
#> [275,] -0.27285122 -0.191290564
#> [276,]  0.48524828 -0.376041440
#> [277,]  0.62867610  0.035669867
#> [278,] -0.42543499  0.112446438
#> [279,] -0.13453565 -0.262708237
#> [280,] -0.42575751  0.113329488
#> [281,]  0.39795080 -0.509033893
#> [282,] -0.57319830  0.636239157
#> [283,]  0.49055524 -0.365354955
#> [284,] -0.45658854  0.202661665
#> [285,]  0.40676747 -0.499128342
#> [286,] -0.41088569  0.073700151
#> [287,]  0.38936244 -0.517946886
#> [288,] -0.49028842  0.311790725
#> [289,] -0.45337911  0.192902835
#> [290,] -0.29528596 -0.160735829
#> [291,] -0.50200962  0.352677462
#> [292,] -0.26626540 -0.199323247
#> [293,] -0.48481313  0.293218511
#> [294,] -0.13982851 -0.264161855
#> [295,] -0.55319116  0.550122950
#> [296,]  0.39246739 -0.514808090
#> [297,] -0.09351397 -0.237064475
#> [298,] -0.16823942 -0.265610292
#> [299,] -0.61397328  0.828987703
#> [300,]  0.36558493 -0.538870648
#> 
#> $grads_hat_unlabeled
#>               [,1]          [,2]
#>   [1,]  0.78114738  0.8717910571
#>   [2,]  0.36319112 -0.5406737221
#>   [3,] -0.38258628  0.0043518486
#>   [4,]  0.68812850  0.2978887822
#>   [5,] -0.14724436 -0.2655536387
#>   [6,]  0.44135589 -0.4527542750
#>   [7,] -0.33089892 -0.1022337770
#>   [8,] -0.25424883 -0.2128701149
#>   [9,]  0.33903353 -0.5557840624
#>  [10,] -0.29065159 -0.1674493726
#>  [11,] -0.65680298  1.0594284916
#>  [12,] -0.63219340  0.9232603850
#>  [13,] -0.31667260 -0.1270662840
#>  [14,] -0.28986663 -0.1685658504
#>  [15,] -0.48650959  0.2989373646
#>  [16,] -0.14501763 -0.2652134986
#>  [17,]  0.49846424 -0.3488447661
#>  [18,] -0.22188969 -0.2420665159
#>  [19,]  0.67481590  0.2333385516
#>  [20,]  0.61015837 -0.0334650368
#>  [21,]  0.36740728 -0.5374609276
#>  [22,] -0.33173011 -0.1007229015
#>  [23,] -0.34896557 -0.0679003232
#>  [24,] -0.44345201  0.1633992109
#>  [25,]  0.55198746 -0.2178636740
#>  [26,] -0.28115053 -0.1805601185
#>  [27,] -0.57068214  0.6251159409
#>  [28,]  0.62139175  0.0078247413
#>  [29,]  0.49006108 -0.3663632396
#>  [30,]  0.57654514 -0.1457488054
#>  [31,] -0.43383766  0.1357966931
#>  [32,] -0.10443327 -0.2465671743
#>  [33,]  0.37265445 -0.5332225874
#>  [34,] -0.49046908  0.3124092046
#>  [35,]  0.50537553 -0.3338375423
#>  [36,]  0.37660016 -0.5298598913
#>  [37,] -0.29400192 -0.1626167944
#>  [38,]  0.40652201 -0.4994145167
#>  [39,] -0.12680495 -0.2598723853
#>  [40,] -0.36278426 -0.0395240758
#>  [41,] -0.25515988 -0.2118935919
#>  [42,] -0.35664888 -0.0523496432
#>  [43,] -0.53402740  0.4724706918
#>  [44,] -0.20310542 -0.2539723273
#>  [45,] -0.47399884  0.2575054934
#>  [46,] -0.30371346 -0.1479946725
#>  [47,] -0.41685345  0.0893361139
#>  [48,] -0.66274322  1.0939203594
#>  [49,] -0.69348242  1.2834468202
#>  [50,] -0.30543822 -0.1453025588
#>  [51,] -0.30882461 -0.1399334521
#>  [52,]  0.30502612 -0.5675640842
#>  [53,] -0.50771923  0.3731595983
#>  [54,] -0.20862279 -0.2508709804
#>  [55,] -0.30774262 -0.1416609418
#>  [56,]  0.65305561  0.1353319638
#>  [57,]  0.61043096 -0.0324862701
#>  [58,] -0.29673195 -0.1585985064
#>  [59,] -0.59944728  0.7575395361
#>  [60,] -0.40994145  0.0712587405
#>  [61,] -0.08242722 -0.2251558889
#>  [62,]  0.15351001 -0.4783457339
#>  [63,]  0.41866911 -0.4845322647
#>  [64,] -0.46675016  0.2342780304
#>  [65,] -0.34798904 -0.0698361571
#>  [66,] -0.32073259 -0.1201774645
#>  [67,] -0.28188470 -0.1795784027
#>  [68,]  0.63867242  0.0753020589
#>  [69,] -0.24447191 -0.2228227624
#>  [70,] -0.28912873 -0.1696099402
#>  [71,] -0.30692905 -0.1429524627
#>  [72,] -0.29664424 -0.1587287359
#>  [73,] -0.50564155  0.3656629962
#>  [74,] -0.59552805  0.7388024976
#>  [75,] -0.48907527  0.3076470073
#>  [76,]  0.63816229  0.0732389294
#>  [77,] -0.57550854  0.6465278516
#>  [78,] -0.31828601 -0.1243476565
#>  [79,] -0.32256348 -0.1170191058
#>  [80,] -0.19586746 -0.2575317419
#>  [81,] -0.33176587 -0.1006577407
#>  [82,] -0.17551510 -0.2643526792
#>  [83,]  0.57938077 -0.1368992487
#>  [84,]  0.61331081 -0.0220762510
#>  [85,]  0.28833435 -0.5692718793
#>  [86,]  0.54347714 -0.2410237569
#>  [87,] -0.13320048 -0.2622795970
#>  [88,] -0.18565955 -0.2615503129
#>  [89,] -0.35566344 -0.0543758618
#>  [90,]  0.59157672 -0.0975552933
#>  [91,] -0.22974691 -0.2359714365
#>  [92,]  0.39635761 -0.5107420253
#>  [93,] -0.43243477  0.1318482550
#>  [94,] -0.28091783 -0.1808701758
#>  [95,] -0.39952138  0.0449041668
#>  [96,] -0.68865110  1.2523708916
#>  [97,]  0.66456997  0.1860729150
#>  [98,] -0.37732048 -0.0076867297
#>  [99,] -0.56029533  0.5800969833
#> [100,] -0.39275756  0.0283706267
#> [101,]  0.55316466 -0.2145873974
#> [102,] -0.64075796  0.9694566627
#> [103,] -0.24563191 -0.2216925187
#> [104,] -0.46367424  0.2245919407
#> [105,] -0.25408549 -0.2130443127
#> [106,] -0.41943759  0.0962173127
#> [107,]  0.60746364 -0.0430795155
#> [108,] -0.34451414 -0.0766504104
#> [109,] -0.55013001  0.5374081406
#> [110,] -0.39507968  0.0339961398
#> [111,]  0.51334126 -0.3158594758
#> [112,]  0.48703882 -0.3724707935
#> [113,] -0.19946487 -0.2558353997
#> [114,] -0.52415202  0.4342255440
#> [115,] -0.19000968 -0.2599827601
#> [116,]  0.57069555 -0.1636564453
#> [117,]  0.29460351 -0.5689467920
#> [118,]  0.68234919  0.2694181400
#> [119,] -0.22777343 -0.2375631762
#> [120,]  0.48103926 -0.3842956675
#> [121,] -0.45700461  0.2039347172
#> [122,]  0.66286210  0.1783907430
#> [123,] -0.34718260 -0.0714279119
#> [124,] -0.20667188 -0.2520056302
#> [125,]  0.70048902  0.3611865432
#> [126,]  0.54979043 -0.2239307657
#> [127,] -0.21467333 -0.2470896270
#> [128,] -0.45190184  0.1884472474
#> [129,]  0.56249587 -0.1879824508
#> [130,]  0.74566685  0.6238394309
#> [131,] -0.17233183 -0.2649813635
#> [132,] -0.54231174  0.5054744188
#> [133,]  0.36966299 -0.5356715322
#> [134,] -0.53536511  0.4777425943
#> [135,]  0.77259786  0.8081219173
#> [136,]  0.41657863 -0.4871984634
#> [137,]  0.47515388 -0.3955120159
#> [138,] -0.49099157  0.3141999629
#> [139,] -0.49613894  0.3320052651
#> [140,]  0.29073956 -0.5691922700
#> [141,] -0.29720040 -0.1579017532
#> [142,] -0.10940435 -0.2502130693
#> [143,] -0.32094076 -0.1198199777
#> [144,] -0.57845638  0.6597623608
#> [145,] -0.58883295  0.7073110710
#> [146,] -0.18746867 -0.2609247624
#> [147,] -0.46694702  0.2349013833
#> [148,] -0.34630979 -0.0731436236
#> [149,] -0.47789452  0.2702233582
#> [150,] -0.55909699  0.5749948732
#> [151,] -0.33686301 -0.0912458628
#> [152,] -0.32105038 -0.1196315666
#> [153,] -0.43942509  0.1517223993
#> [154,] -0.18180283 -0.2627575793
#> [155,] -0.62452376  0.8829264071
#> [156,]  0.61445150 -0.0179174668
#> [157,] -0.36972788 -0.0245710603
#> [158,]  0.40565232 -0.5004236950
#> [159,]  0.51221299 -0.3184506781
#> [160,] -0.13504522 -0.2628651946
#> [161,]  0.51733448 -0.3065689084
#> [162,] -0.54384135  0.5116614875
#> [163,] -0.33233435 -0.0996203792
#> [164,]  0.64209876  0.0892747655
#> [165,]  0.23408846 -0.5558906982
#> [166,]  0.29240526 -0.5691042072
#> [167,]  0.39659669 -0.5104872927
#> [168,]  0.56228744 -0.1885891580
#> [169,] -0.34394419 -0.0777570063
#> [170,]  0.41490491 -0.4893015747
#> [171,]  0.46889335 -0.4070308133
#> [172,] -0.27432440 -0.1894352048
#> [173,] -0.42323683  0.1064560643
#> [174,] -0.25472848 -0.2123570347
#> [175,] -0.38653526  0.0135570579
#> [176,] -0.11956676 -0.2564217553
#> [177,] -0.27708141 -0.1859056429
#> [178,] -0.63409077  0.9333880740
#> [179,] -0.55731405  0.5674384849
#> [180,] -0.48903315  0.3075034450
#> [181,] -0.63683867  0.9481626357
#> [182,] -0.43318686  0.1339625107
#> [183,] -0.54124081  0.5011600486
#> [184,] -0.66044105  1.0804749274
#> [185,] -0.48696773  0.3004872582
#> [186,]  0.51023365 -0.3229606198
#> [187,]  0.37414724 -0.5319681198
#> [188,]  0.57158590 -0.1609607768
#> [189,] -0.47405175  0.2576771255
#> [190,] -0.24435361 -0.2229372564
#> [191,] -0.53638249  0.4817667757
#> [192,] -0.47216037  0.2515607676
#> [193,]  0.45200497 -0.4360164286
#> [194,]  0.62565473  0.0240162203
#> [195,] -0.42117672  0.1008861464
#> [196,] -0.57861084  0.6604591281
#> [197,] -0.29185768 -0.1657222308
#> [198,] -0.13584513 -0.2631042242
#> [199,]  0.43892878 -0.4564048885
#> [200,] -0.12703769 -0.2599703936
#> [201,] -0.38033407 -0.0008301041
#> [202,] -0.28746981 -0.1719378858
#> [203,] -0.50317276  0.3568196809
#> [204,] -0.48357775  0.2890739676
#> [205,]  0.55277525 -0.2156731403
#> [206,] -0.25415452 -0.2129707278
#> [207,]  0.49989777 -0.3457767278
#> [208,]  0.51220515 -0.3184686344
#> [209,] -0.44242687  0.1604107590
#> [210,]  0.49711599 -0.3517090658
#> [211,]  0.53959966 -0.2512735948
#> [212,] -0.33067426 -0.1026410208
#> [213,] -0.75320612  1.7142032468
#> [214,] -0.34351779 -0.0785828691
#> [215,] -0.33281992 -0.0987318611
#> [216,]  0.65949695  0.1634144385
#> [217,]  0.59211397 -0.0957735054
#> [218,] -0.51669601  0.4061247973
#> [219,]  0.62636895  0.0267575967
#> [220,] -0.42970363  0.1242189241
#> [221,] -0.52627099  0.4423325816
#> [222,] -0.21221843 -0.2486714809
#> [223,] -0.11427178 -0.2533908325
#> [224,] -0.69711588  1.3071525052
#> [225,] -0.27098945 -0.1936047670
#> [226,] -0.51586000  0.4030150010
#> [227,] -0.41077458  0.0734124153
#> [228,] -0.22078711 -0.2428698253
#> [229,] -0.47229416  0.2519921287
#> [230,] -0.31311519 -0.1329722168
#> [231,]  0.74112100  0.5949515095
#> [232,] -0.44933497  0.1807596366
#> [233,] -0.20584191 -0.2524757546
#> [234,] -0.24105965 -0.2260679907
#> [235,]  0.50385759 -0.3371803496
#> [236,] -0.32229741 -0.1174800937
#> [237,] -0.44478351  0.1672969880
#> [238,] -0.76020150  1.7711523237
#> [239,] -0.30540808 -0.1453498458
#> [240,] -0.22946154 -0.2362041204
#> [241,]  0.39440897 -0.5127973264
#> [242,] -0.38772814  0.0163676294
#> [243,]  0.63175868  0.0477137311
#> [244,] -0.27461863 -0.1890620829
#> [245,] -0.39099590  0.0241381395
#> [246,] -0.27707300 -0.1859165249
#> [247,] -0.28405803 -0.1766414175
#> [248,] -0.57279205  0.6344374055
#> [249,] -0.54996915  0.5367432750
#> [250,] -0.54851854  0.5307627798
#> [251,] -0.23947934 -0.2275306322
#> [252,]  0.73605023  0.5634243873
#> [253,]  0.59983628 -0.0696990012
#> [254,] -0.39780097  0.0406560982
#> [255,]  0.48129007 -0.3838092490
#> [256,] -0.71677975  1.4406800490
#> [257,] -0.29353683 -0.1632941678
#> [258,] -0.38439508  0.0085493604
#> [259,] -0.33763005 -0.0898079646
#> [260,]  0.63937143  0.0781363411
#> [261,]  0.46611487 -0.4120078675
#> [262,] -0.43415420  0.1366903791
#> [263,]  0.48951157 -0.3674812707
#> [264,] -0.34281865 -0.0799331812
#> [265,] -0.35293470 -0.0599378392
#> [266,] -0.45492636  0.1975939991
#> [267,] -0.60901748  0.8042519832
#> [268,] -0.31491435 -0.1300004880
#> [269,] -0.45138805  0.1869029860
#> [270,] -0.26846798 -0.1966844651
#> [271,] -0.29601659 -0.1596584144
#> [272,]  0.64122581  0.0856957338
#> [273,] -0.41820287  0.0929210620
#> [274,] -0.28278773 -0.1783636700
#> [275,] -0.39453418  0.0326698533
#> [276,]  0.26571521 -0.5672539432
#> [277,] -0.49862680  0.3407179681
#> [278,]  0.67393993  0.2292173040
#> [279,]  0.56550373 -0.1791631221
#> [280,] -0.19609114 -0.2574304854
#> [281,] -0.55282113  0.5485795940
#> [282,]  0.43694541 -0.4593431022
#> [283,] -0.37648006 -0.0095831511
#> [284,] -0.20633261 -0.2521987182
#> [285,]  0.66453525  0.1859161890
#> [286,] -0.46634919  0.2330096326
#> [287,] -0.36913187 -0.0258728215
#> [288,] -0.57678462  0.6522422359
#> [289,] -0.28716929 -0.1723567492
#> [290,] -0.32740706 -0.1085087103
#> [291,] -0.53563721  0.4788176147
#> [292,] -0.23563671 -0.2309801729
#> [293,] -0.07721231 -0.2186998766
#> [294,]  0.65443669  0.1412892605
#> [295,] -0.59865474  0.7537323352
#> [296,]  0.60001493 -0.0690854607
#> [297,] -0.25516262 -0.2118906508
#> [298,] -0.25169724 -0.2155606547
#> [299,]  0.57409621 -0.1533025729
#> [300,] -0.26330764 -0.2027911363
#> 
#> $grads_hat
#>               [,1]         [,2]
#>   [1,] -0.35018621 -0.065467702
#>   [2,] -0.34865053 -0.068525846
#>   [3,]  0.55159384 -0.218955189
#>   [4,] -0.67080193  1.141779389
#>   [5,] -0.29953318 -0.154400444
#>   [6,] -0.43602459  0.141991943
#>   [7,]  0.54753800 -0.230087001
#>   [8,] -0.28019681 -0.181827539
#>   [9,] -0.36687570 -0.030769545
#>  [10,] -0.53069983  0.459451715
#>  [11,] -0.28516830 -0.175123236
#>  [12,]  0.41170112 -0.493249199
#>  [13,] -0.62406120  0.880524506
#>  [14,] -0.49593066  0.331279047
#>  [15,]  0.47235425 -0.400715374
#>  [16,]  0.62060624  0.004873018
#>  [17,] -0.21294580 -0.248209545
#>  [18,] -0.17319888 -0.264822150
#>  [19,] -0.47280342  0.253635917
#>  [20,]  0.58752019 -0.110875203
#>  [21,]  0.27273457 -0.568417965
#>  [22,]  0.44671547 -0.444477496
#>  [23,] -0.55146018  0.542918403
#>  [24,] -0.31170860 -0.135273833
#>  [25,]  0.65808180  0.157179463
#>  [26,] -0.33408434 -0.096407560
#>  [27,]  0.73882457  0.580583978
#>  [28,] -0.38665732  0.013844019
#>  [29,] -0.14390616 -0.265018904
#>  [30,] -0.65956354  1.075376028
#>  [31,] -0.46370100  0.224675780
#>  [32,]  0.36566398 -0.538810162
#>  [33,] -0.23376174 -0.232608103
#>  [34,] -0.38185370  0.002660886
#>  [35,] -0.29195904 -0.165576440
#>  [36,] -0.40782411  0.065816400
#>  [37,] -0.61226242  0.820405591
#>  [38,] -0.19255699 -0.258964580
#>  [39,] -0.50995497  0.381282181
#>  [40,]  0.62559756  0.023797126
#>  [41,] -0.37323324 -0.016845275
#>  [42,]  0.62262536  0.012480199
#>  [43,]  0.47317933 -0.399190709
#>  [44,]  0.72715771  0.509842328
#>  [45,] -0.27444274 -0.189285242
#>  [46,] -0.34057269 -0.084239323
#>  [47,] -0.45818541  0.207557627
#>  [48,] -0.84031222  2.562627966
#>  [49,] -0.23638102 -0.230323878
#>  [50,] -0.37260416 -0.018240529
#>  [51,]  0.58130119 -0.130842726
#>  [52,] -0.29874979 -0.155582121
#>  [53,]  0.43045030 -0.468683339
#>  [54,] -0.44430724  0.165900673
#>  [55,] -0.18877292 -0.260450470
#>  [56,] -0.42230443  0.103929847
#>  [57,] -0.50369016  0.358667240
#>  [58,] -0.14569671 -0.265324239
#>  [59,] -0.39765028  0.040285390
#>  [60,] -0.61519590  0.835148469
#>  [61,] -0.36688923 -0.030740320
#>  [62,]  0.69949658  0.355979698
#>  [63,] -0.61131051  0.815650136
#>  [64,]  0.49317236 -0.359969437
#>  [65,] -0.11389706 -0.253159685
#>  [66,] -0.51348100  0.394210375
#>  [67,] -0.58434978  0.686582168
#>  [68,] -0.26336085 -0.202729518
#>  [69,]  0.50842562 -0.327040492
#>  [70,] -0.24103058 -0.226095128
#>  [71,] -0.20686027 -0.251897877
#>  [72,] -0.25371565 -0.213437745
#>  [73,]  0.50809014 -0.327793344
#>  [74,] -0.18238363 -0.262586825
#>  [75,] -0.39308637  0.029163980
#>  [76,] -0.41163940  0.075655318
#>  [77,]  0.42479247 -0.476469698
#>  [78,]  0.71688530  0.450522602
#>  [79,] -0.14631699 -0.265420010
#>  [80,] -0.38556395  0.011278797
#>  [81,] -0.21262401 -0.248414606
#>  [82,] -0.39866588  0.042788108
#>  [83,] -0.29954920 -0.154376221
#>  [84,] -0.36067404 -0.043976254
#>  [85,]  0.44982345 -0.439541324
#>  [86,]  0.42796296 -0.472146417
#>  [87,] -0.08646345 -0.229768931
#>  [88,] -0.39857539  0.042564691
#>  [89,] -0.28153058 -0.180052586
#>  [90,] -0.31965343 -0.122023970
#>  [91,] -0.25121153 -0.216065383
#>  [92,] -0.26292888 -0.203228951
#>  [93,]  0.66148373  0.172230789
#>  [94,] -0.48060775  0.279178658
#>  [95,] -0.58056018  0.669280865
#>  [96,]  0.55726894 -0.203024990
#>  [97,]  0.65693971  0.152174527
#>  [98,]  0.66275415  0.177907030
#>  [99,]  0.58592528 -0.116047985
#> [100,]  0.44734784 -0.443481287
#> [101,] -0.12503347 -0.259100241
#> [102,] -0.56034047  0.580289531
#> [103,] -0.27401322 -0.189828891
#> [104,] -0.38694108  0.014511677
#> [105,] -0.59808372  0.750995038
#> [106,] -0.17385918 -0.264694851
#> [107,] -0.25352276 -0.213642394
#> [108,] -0.38591101  0.012091783
#> [109,] -0.52013995  0.419022628
#> [110,]  0.72259351  0.483153554
#> [111,]  0.74094662  0.593855268
#> [112,] -0.47150297  0.249443921
#> [113,] -0.11032917 -0.250846303
#> [114,] -0.30124772 -0.151793529
#> [115,]  0.67766747  0.246860662
#> [116,] -0.52646695  0.443085042
#> [117,] -0.45899729  0.210057143
#> [118,] -0.33841607 -0.088328625
#> [119,]  0.67655171  0.241550314
#> [120,]  0.46760590 -0.409347298
#> [121,]  0.56370812 -0.184442457
#> [122,] -0.65468810  1.047305767
#> [123,] -0.40391423  0.055883541
#> [124,] -0.21559247 -0.246480706
#> [125,] -0.40102259  0.048634803
#> [126,] -0.35316936 -0.059462340
#> [127,] -0.33157234 -0.101010195
#> [128,] -0.20369820 -0.253655114
#> [129,]  0.24268172 -0.559981640
#> [130,]  0.50257979 -0.339973891
#> [131,] -0.13787624 -0.263670990
#> [132,] -0.11388179 -0.253150218
#> [133,] -0.64014023  0.966083019
#> [134,]  0.80482131  1.063205913
#> [135,]  0.41536852 -0.488721838
#> [136,]  0.39229172 -0.514988193
#> [137,]  0.74896654  0.645188462
#> [138,] -0.48908110  0.307666876
#> [139,] -0.36251580 -0.040092858
#> [140,]  0.57786331 -0.141648806
#> [141,] -0.27057984 -0.194109336
#> [142,] -0.27209871 -0.192230059
#> [143,] -0.24983776 -0.217480078
#> [144,] -0.33919524 -0.086856331
#> [145,]  0.48448278 -0.377557155
#> [146,]  0.32901680 -0.560406785
#> [147,] -0.38961974  0.020852987
#> [148,] -0.15203039 -0.266062943
#> [149,]  0.54470045 -0.237750969
#> [150,]  0.49221315 -0.361952190
#> [151,] -0.37879876 -0.004334323
#> [152,] -0.18648518 -0.261269500
#> [153,] -0.17875979 -0.263587726
#> [154,]  0.58037591 -0.133767210
#> [155,] -0.51765223  0.409691902
#> [156,]  0.63008197  0.041143281
#> [157,] -0.14367857 -0.264977007
#> [158,]  0.64691811  0.109272012
#> [159,] -0.33434752 -0.095921850
#> [160,] -0.43918904  0.151043120
#> [161,]  0.58166437 -0.129691569
#> [162,]  0.80485517  1.063497083
#> [163,] -0.22866073 -0.236852544
#> [164,] -0.29594209 -0.159768508
#> [165,]  0.58221755 -0.127934652
#> [166,] -0.27362095 -0.190323819
#> [167,]  0.48224806 -0.381945043
#> [168,] -0.31011931 -0.137851532
#> [169,]  0.31979636 -0.563811579
#> [170,] -0.34559822 -0.074536980
#> [171,] -0.60995549  0.808904898
#> [172,] -0.31389277 -0.131691671
#> [173,] -0.24258359 -0.224633333
#> [174,]  0.70401368  0.379859496
#> [175,] -0.17852453 -0.263647377
#> [176,]  0.56106903 -0.192124283
#> [177,] -0.31878455 -0.123502561
#> [178,] -0.23053175 -0.235327127
#> [179,] -0.28438566 -0.176194660
#> [180,] -0.19155804 -0.259372646
#> [181,] -0.59137209  0.719178308
#> [182,]  0.60207101 -0.061989900
#> [183,] -0.61414558  0.829854507
#> [184,] -0.31714229 -0.126277423
#> [185,] -0.24464094 -0.222658922
#> [186,] -0.41015369  0.071806731
#> [187,]  0.53023581 -0.275259894
#> [188,] -0.13917820 -0.264004182
#> [189,] -0.44141153  0.157461569
#> [190,]  0.45128365 -0.437187439
#> [191,] -0.58271560  0.679096748
#> [192,] -0.68369222  1.220990787
#> [193,] -0.57458088  0.642387764
#> [194,] -0.38928001  0.020044848
#> [195,] -0.10214638 -0.244750087
#> [196,] -0.25895505 -0.207736216
#> [197,]  0.49090436 -0.364640956
#> [198,] -0.52899202  0.452822300
#> [199,]  0.74025318  0.589504409
#> [200,] -0.29668307 -0.158671087
#> [201,] -0.39554593  0.035132056
#> [202,]  0.32381920 -0.562426369
#> [203,] -0.54321127  0.509109327
#> [204,]  0.55609332 -0.206359164
#> [205,] -0.22095903 -0.242745411
#> [206,] -0.25322617 -0.213956325
#> [207,] -0.22268499 -0.241479106
#> [208,]  0.75571425  0.689874210
#> [209,] -0.50845181  0.375814748
#> [210,]  0.65562502  0.146442876
#> [211,] -0.34723896 -0.071316877
#> [212,] -0.47622592  0.264755847
#> [213,]  0.45491478 -0.431237203
#> [214,] -0.34018333 -0.084980904
#> [215,]  0.55474542 -0.210159919
#> [216,] -0.68736814  1.244202561
#> [217,]  0.37170464 -0.534009535
#> [218,] -0.74652494  1.661227957
#> [219,] -0.72551104  1.502973046
#> [220,] -0.48032046  0.278226622
#> [221,] -0.48523417  0.294634885
#> [222,]  0.30361435 -0.567812804
#> [223,] -0.24794139 -0.219401600
#> [224,] -0.15160669 -0.266029952
#> [225,] -0.39247431  0.027688047
#> [226,] -0.26076148 -0.205706801
#> [227,] -0.28792610 -0.171300248
#> [228,] -0.26360168 -0.202450269
#> [229,] -0.19928784 -0.255922252
#> [230,] -0.40721240  0.064252372
#> [231,] -0.39031313  0.022505928
#> [232,] -0.41573792  0.086386286
#> [233,]  0.63266185  0.051272271
#> [234,] -0.56759791  0.611598019
#> [235,]  0.25394992 -0.564205413
#> [236,]  0.31962163 -0.563868226
#> [237,]  0.52924976 -0.277723389
#> [238,] -0.31391648 -0.131652536
#> [239,] -0.47217013  0.251592225
#> [240,] -0.52223738  0.426946491
#> [241,] -0.36548763 -0.033757718
#> [242,] -0.31066969 -0.136961609
#> [243,] -0.49031965  0.311897631
#> [244,] -0.47341144  0.255602139
#> [245,] -0.27644817 -0.186722924
#> [246,] -0.30636924 -0.143837436
#> [247,] -0.30596844 -0.144469194
#> [248,]  0.55923755 -0.197401520
#> [249,] -0.27324921 -0.190791434
#> [250,]  0.62855486  0.035199335
#> [251,] -0.28429713 -0.176315489
#> [252,] -0.35042070 -0.064998759
#> [253,] -0.43342496  0.134633062
#> [254,]  0.73489465  0.556340101
#> [255,]  0.33138567 -0.559400397
#> [256,] -0.59242521  0.724127417
#> [257,] -0.43466018  0.138121047
#> [258,] -0.48605827  0.297412845
#> [259,] -0.12498879 -0.259080167
#> [260,] -0.50399080  0.359742197
#> [261,]  0.57736781 -0.143192804
#> [262,] -0.54295223  0.508061553
#> [263,] -0.52616695  0.441933272
#> [264,] -0.22534902 -0.239462867
#> [265,]  0.53042433 -0.274787562
#> [266,] -0.42208370  0.103333093
#> [267,]  0.58286987 -0.125857354
#> [268,] -0.35535691 -0.055004221
#> [269,] -0.39034161  0.022573915
#> [270,] -0.53610961  0.480686160
#> [271,] -0.40568818  0.060371377
#> [272,]  0.51506095 -0.311881418
#> [273,]  0.63071263  0.043609165
#> [274,] -0.36262860 -0.039853953
#> [275,] -0.27285122 -0.191290564
#> [276,]  0.48524828 -0.376041440
#> [277,] -0.37132390 -0.021068200
#> [278,]  0.57456501 -0.151862894
#> [279,] -0.13453565 -0.262708237
#> [280,]  0.57424249 -0.152853689
#> [281,] -0.60204920  0.770103871
#> [282,] -0.57319830  0.636239157
#> [283,]  0.49055524 -0.365354955
#> [284,]  0.54341146 -0.241198938
#> [285,] -0.59323253  0.727932284
#> [286,]  0.58911431 -0.105668839
#> [287,] -0.61063756  0.812296691
#> [288,] -0.49028842  0.311790725
#> [289,] -0.45337911  0.192902835
#> [290,]  0.70471404  0.383603722
#> [291,] -0.50200962  0.352677462
#> [292,] -0.26626540 -0.199323247
#> [293,]  0.51518687 -0.311588767
#> [294,] -0.13982851 -0.264161855
#> [295,] -0.55319116  0.550122950
#> [296,]  0.39246739 -0.514808090
#> [297,] -0.09351397 -0.237064475
#> [298,] -0.16823942 -0.265610292
#> [299,]  0.38602672 -0.521213889
#> [300,] -0.63441507  0.935125158
#> 
#> $inv_hessian
#>          [,1]     [,2]
#> [1,] 4.809733 1.142218
#> [2,] 1.142218 5.838868
#>