PPI++ OLS

Helper function for PPI++ OLS estimation

Usage

ppi_plusplus_ols(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see doi:10.48550/arXiv.2311.01453 ). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ OLS regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier OLS regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) doi:10.48550/arXiv.2311.01453

Examples

dat <- simdat(model = "ols")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_plusplus_ols(X_l, Y_l, f_l, X_u, f_u)
#> $est
#> X(Intercept)          XX1 
#>    0.8931122    1.0152501 
#> 
#> $se
#> [1] 0.11429185 0.09049167
#> 
#> $lambda
#> [1] 0.6452418
#> 
#> $rectifier_est
#> X(Intercept)          XX1 
#>   0.12340269  -0.02119519 
#> 
#> $var_u
#>             [,1]        [,2]
#> [1,]  0.29234541 -0.05370285
#> [2,] -0.05370285  0.34850964
#> 
#> $var_l
#>           [,1]      [,2]
#> [1,] 3.6410544 0.5427797
#> [2,] 0.5427797 2.3822518
#> 
#> $grads
#>                 [,1]          [,2]
#>   [1,]   0.185850086  -0.101969515
#>   [2,]  -0.537802143  -0.432387947
#>   [3,]  -0.953080666   0.058004109
#>   [4,]   0.261338046   0.327210752
#>   [5,]   0.023775774  -0.004887591
#>   [6,]   0.493784543   0.759891383
#>   [7,]   0.239432028  -0.224708704
#>   [8,]   1.364818674   1.695363881
#>   [9,]   0.524555528  -0.100341442
#>  [10,]  -0.243400272   0.126037166
#>  [11,]   4.744173124   4.872710145
#>  [12,]  -0.944300901  -1.274991799
#>  [13,]   0.577671476   0.236713403
#>  [14,]  -1.095856936   2.359528632
#>  [15,]  -2.711193687  -2.120147401
#>  [16,]   0.898116469   0.004999821
#>  [17,]  -1.539522098   1.318869921
#>  [18,]   1.754350534   0.458140661
#>  [19,]   1.491895152   0.262400714
#>  [20,]   0.756429061  -0.839782168
#>  [21,]  -0.090046196   0.088552736
#>  [22,]  -2.946640479   4.109307832
#>  [23,]   0.703466041   0.482777205
#>  [24,]  -1.174565541  -0.104650147
#>  [25,]   0.348149894   0.134725774
#>  [26,]   2.687420942   1.353262775
#>  [27,]  -5.382839180  -4.868683254
#>  [28,]  -0.132207575  -0.002548527
#>  [29,]   0.351674795  -0.061383051
#>  [30,]   0.695594258   0.246662051
#>  [31,]   1.085361196   0.238982985
#>  [32,]   1.280932721   0.822849006
#>  [33,]   2.382850454  -2.506475357
#>  [34,]   0.651040661   0.351413466
#>  [35,]   1.021211935   0.974091656
#>  [36,]   1.082851913  -1.481452730
#>  [37,]   1.746398788  -0.454186515
#>  [38,]   0.302316594   0.102163317
#>  [39,]   0.832509049   0.969883894
#>  [40,]   2.793158302   1.277259582
#>  [41,]  -0.612390566   0.742185935
#>  [42,]   0.227434527  -0.364038132
#>  [43,]  -0.377758803  -0.102829546
#>  [44,]  -7.404654168  -1.951268676
#>  [45,]   1.851203707   0.636476151
#>  [46,]  -1.599293204   0.621883693
#>  [47,]   1.388775255   0.142615751
#>  [48,]   0.249525405  -0.451786516
#>  [49,]   0.770369808   0.268171362
#>  [50,]   0.015057926   0.002585219
#>  [51,]  -0.503920480   0.312113075
#>  [52,]   0.085632176   0.009956438
#>  [53,]  -1.690025730   2.574940830
#>  [54,]  -3.476915004   3.149050535
#>  [55,]   1.121204605   1.555348747
#>  [56,]  -0.510950980   0.195740796
#>  [57,]   2.288742182   1.839591489
#>  [58,]   0.401716902   0.661268281
#>  [59,]  -0.056676983  -0.101696207
#>  [60,]   1.149537160  -0.378652115
#>  [61,]  -3.051467796   0.991746968
#>  [62,]   0.272174794  -0.063099722
#>  [63,]  -2.596954567  -0.519127137
#>  [64,]   0.228796054   0.033905952
#>  [65,]   3.242654911  -0.417316970
#>  [66,]   1.301109572   0.692851614
#>  [67,]   1.713082859  -0.253540358
#>  [68,]  -1.100327633  -0.608175973
#>  [69,]   0.200299376  -0.165053921
#>  [70,]  -0.220122325   0.145962352
#>  [71,]  -1.770384337   3.376883709
#>  [72,]   1.534064218   3.163199793
#>  [73,]   0.972545567   0.100460465
#>  [74,]   1.185142040  -0.352126574
#>  [75,]  -0.045083612   0.030534450
#>  [76,]  -1.059321657   1.227985998
#>  [77,]  -2.994058114  -0.403556014
#>  [78,]  -1.392605973   1.120543800
#>  [79,]  -2.160507406  -3.493498954
#>  [80,]  -0.326678043  -0.185306974
#>  [81,] -12.525413213   2.189122528
#>  [82,]   0.374122053  -0.227901350
#>  [83,]   3.353048956  -5.572303138
#>  [84,]   0.385868045  -0.318588832
#>  [85,]  -1.417126694  -0.910238741
#>  [86,]  -0.844940339   0.709717443
#>  [87,]   0.336494073   0.559098898
#>  [88,]   0.288826766  -0.258134906
#>  [89,]  -0.406689133   0.144817125
#>  [90,]  -0.349056694   0.141539562
#>  [91,]   1.675903105  -5.333529397
#>  [92,]  -5.661120308   1.350531878
#>  [93,]  -0.854453181  -0.521884994
#>  [94,]   0.996159579   0.287494231
#>  [95,]  -3.251956517  -4.895223807
#>  [96,]  -2.114812420   2.212350967
#>  [97,]   1.028723461   1.038477605
#>  [98,]  -0.232644227   0.247753883
#>  [99,]  -0.518103035   0.037548219
#> [100,]  -1.113934110   0.625712717
#> [101,]   2.912718103  -1.883575369
#> [102,]   0.705280504  -0.614336243
#> [103,]   1.497765669   0.338944892
#> [104,]  -7.982321006 -13.747906914
#> [105,]  -6.322993934   1.998898558
#> [106,]  -1.596576537  -0.899382237
#> [107,]   0.845551325  -0.416038880
#> [108,]  -1.413636978   0.801072577
#> [109,]   0.066015216   0.050496165
#> [110,]   0.718286259   1.962797768
#> [111,]  -0.089133438  -0.008935741
#> [112,]  -0.164393416   0.016083312
#> [113,]  -0.348430684   0.375164146
#> [114,]   0.775681051   0.601713078
#> [115,]  -2.036955805   0.399587320
#> [116,]  -1.661275340   1.995097730
#> [117,]   0.815295131   0.698748650
#> [118,]   0.793535683   0.399791387
#> [119,]   0.480541824  -0.404412661
#> [120,]  -0.158265887  -0.066440384
#> [121,]   1.527235865   0.151681065
#> [122,]   0.880152581   2.322107879
#> [123,]  -3.402830726   1.708993065
#> [124,]  -1.011497496   0.159943329
#> [125,]  -1.035339382   0.690873606
#> [126,]   1.389018578   2.113706933
#> [127,]  -0.643597880   0.483597490
#> [128,]   2.880225517  -0.476653482
#> [129,]   0.985343705   0.186001146
#> [130,]  -2.091828411   0.325558867
#> [131,]  -1.257992457   0.385948924
#> [132,]   1.256323760  -0.766514411
#> [133,]  -2.402229241   0.576080568
#> [134,]   0.185185505  -0.164975488
#> [135,]   0.808564207  -0.136581254
#> [136,]   2.193688639  -2.258136520
#> [137,]   0.048470763  -0.073153373
#> [138,]  -0.268891720  -0.151586056
#> [139,]  -0.389529899   0.227463963
#> [140,]   0.035486131   0.017716215
#> [141,]  -1.088454857   1.412305369
#> [142,]  -1.148611983  -1.360788629
#> [143,]  -1.478757993  -1.493537953
#> [144,]  -1.074676743   1.394385909
#> [145,]  -2.914384596   0.874338324
#> [146,]   0.756631291   0.825114014
#> [147,]   0.598885859   0.210177037
#> [148,]   1.629286805   2.101941179
#> [149,]  -0.708371445   0.343949913
#> [150,]  -1.336022099  -1.083415492
#> [151,]   0.645727191   0.996513367
#> [152,] -14.041094018 -11.307769991
#> [153,]   1.442881072   2.663705463
#> [154,]   3.025138246   0.567526665
#> [155,]  -1.991513140  -0.606967699
#> [156,]   1.382859347   0.087185304
#> [157,]  -0.009409962  -0.001627998
#> [158,]   2.430251864   0.313295444
#> [159,]   0.577660680  -0.114683144
#> [160,]   2.825699526  -3.486981273
#> [161,]  -1.023062475   0.624310570
#> [162,]   0.423139967  -0.858092980
#> [163,]   0.583797002   0.445480066
#> [164,]   1.072097236   0.360574903
#> [165,]   0.564208136   0.110448017
#> [166,]   2.904556239   0.610144524
#> [167,]   0.189480531   0.012966179
#> [168,]  -0.417533535  -0.020564541
#> [169,]  -0.606563057   0.299823564
#> [170,]  -2.130888379  -2.827064857
#> [171,]   1.689439608  -2.208380433
#> [172,]   1.176459080   0.879516208
#> [173,]   1.750023042   2.459585909
#> [174,]   1.581496918   1.838231934
#> [175,]  -0.538860084  -0.993805587
#> [176,]   2.702999360  -2.570572459
#> [177,]   2.157644239  -0.010812071
#> [178,]  -0.777934398   1.936441065
#> [179,]   0.749290148  -0.253020053
#> [180,]   1.600222107   0.403400628
#> [181,]  -0.501401568  -0.313714966
#> [182,]   0.170299449  -0.103109533
#> [183,]  -2.194009823  -0.322696747
#> [184,]  -0.251078147  -0.444539257
#> [185,]   0.879084671   1.162086646
#> [186,]   0.891545694  -0.917760950
#> [187,]  -0.207766802  -0.076138592
#> [188,]   1.710541761  -2.061639087
#> [189,]  -0.108973852  -0.236996383
#> [190,]  -1.776907155   1.242223815
#> [191,]   0.262192222   0.043372694
#> [192,]  12.546143699  -4.225695985
#> [193,]   3.692040246  -1.763328974
#> [194,]   0.733550624  -0.028618326
#> [195,]  -0.049409108   0.034759662
#> [196,]  -0.041497895   0.029599718
#> [197,]  -0.315092283   0.126065329
#> [198,]  -1.076229882  -0.135566654
#> [199,]  -1.875019434   3.319618125
#> [200,]  -1.201318798  -0.762818535
#> [201,]  -1.703601040  -0.923922681
#> [202,]   0.836634595   0.670001372
#> [203,]  -2.133968254   2.919329728
#> [204,]   1.412328738   0.188194154
#> [205,]  -1.487218261  -1.839931786
#> [206,]  -0.957723632  -2.451144021
#> [207,]  -0.040749175   0.056551286
#> [208,]   0.570573685   1.057740074
#> [209,]  -0.655224377  -0.946715486
#> [210,]  -0.783275814   0.141294036
#> [211,]   0.825855578   1.479365981
#> [212,]   0.068675889   0.010680496
#> [213,]  -0.088876449  -0.049099491
#> [214,]  -0.234896401   0.265132916
#> [215,]  -2.047572501  -0.705168440
#> [216,]   2.383363834  -5.771399143
#> [217,]   1.267705431  -0.718380135
#> [218,]   1.473897455   0.910677690
#> [219,]   0.715573591   1.764358125
#> [220,]   0.746186147   1.427791853
#> [221,]   1.280427937  -2.300515882
#> [222,]  -2.168156904  -2.324788172
#> [223,]   0.819382450  -0.436004591
#> [224,]   1.317854008  -0.216756640
#> [225,]  -0.403164580   0.523690476
#> [226,]  -1.625781684  -3.275907033
#> [227,]   0.561711550  -0.527118890
#> [228,]   1.047405749   0.849224967
#> [229,]   0.981919902   0.094546079
#> [230,]  -0.502713859  -0.679273316
#> [231,]   0.242791630  -0.236535172
#> [232,]  -0.136048262  -0.003671071
#> [233,]  -0.181409208   0.003518503
#> [234,]  -1.593115956  -0.941598903
#> [235,]  -1.012200464   0.662405452
#> [236,]   1.132332199  -0.183527314
#> [237,]  -1.601351180   0.377622868
#> [238,]   1.966044327   1.902764592
#> [239,]  -0.780170871  -0.419588583
#> [240,]  -1.107650699  -3.215633026
#> [241,]  -1.213090752  -0.580635092
#> [242,]  -0.518468756  -0.281671492
#> [243,]   0.873087373  -0.853822188
#> [244,]  -0.827962253  -0.825257668
#> [245,]  -1.286425629  -1.390288849
#> [246,]   0.333633712  -0.014427171
#> [247,]  -4.634420275   3.479440300
#> [248,]  -0.856205099   0.179034389
#> [249,]  -1.105035870  -0.905338799
#> [250,]  -0.033635571  -0.056372756
#> [251,]  -2.651809558  -1.279983513
#> [252,]  -0.973031086  -0.150430562
#> [253,]  -1.064433684  -1.081908191
#> [254,]   0.198859512   0.421968407
#> [255,]  -0.700298549   0.496395934
#> [256,]  -1.056357241   0.287395215
#> [257,]   2.487270764  -1.710029887
#> [258,]   2.866051920   4.364967748
#> [259,]  -1.283346907  -1.553713983
#> [260,]  -0.890839068  -0.101519123
#> [261,]   2.828579975   1.167141592
#> [262,]  -2.261533356  -2.419822278
#> [263,]   2.614397503  -1.403148921
#> [264,]   1.060558276  -0.529183672
#> [265,]  -0.144520466   0.043004335
#> [266,]  -0.401412764   0.204347051
#> [267,]  -1.050673352  -0.815987017
#> [268,]   1.329727348   0.235632790
#> [269,]   2.260292804   2.212344896
#> [270,]   0.545771451  -0.812620517
#> [271,]   4.314005154   4.195793007
#> [272,]   0.710121491   0.267955090
#> [273,]   1.185548109  -0.591369320
#> [274,]   2.485581724  -0.976797994
#> [275,]   0.329997318  -0.257599135
#> [276,]   1.600151675   0.934501987
#> [277,]  -1.171097480   0.609339558
#> [278,]  -0.706117449   0.292129255
#> [279,]   1.766229668  -1.128230245
#> [280,]   0.682092862   0.096511858
#> [281,]  -1.596116653  -1.706313456
#> [282,]  -1.426971180   0.038238768
#> [283,]   1.972862090   1.246062896
#> [284,]  -1.568172459  -2.428187572
#> [285,]   0.703438545  -0.744700535
#> [286,]   0.594974230   0.351037549
#> [287,]  -0.361205193   0.489808248
#> [288,]   0.164608557  -0.164004689
#> [289,]   0.691045778   0.143716100
#> [290,]  -0.944947108  -1.503952270
#> [291,]  -0.176996878   0.295325904
#> [292,]   0.249459179  -0.078464560
#> [293,]   1.221134348   0.276736542
#> [294,]   1.224063157   2.845524042
#> [295,]  -3.386574506   6.436071974
#> [296,]   0.310480131   0.062698935
#> [297,]   0.008780474   0.012941428
#> [298,]   3.200501853   0.617902380
#> [299,]   1.291808755  -0.399907201
#> [300,]  -2.946027889   2.633400349
#> 
#> $grads_hat_unlabeled
#>                [,1]         [,2]
#>   [1,]  1.221131630 -0.412186071
#>   [2,]  0.921552983  1.309609564
#>   [3,] -0.036402282  0.003897242
#>   [4,] -0.987798990  0.651738058
#>   [5,] -1.316067763 -1.824421252
#>   [6,] -0.581039266  0.703857072
#>   [7,] -0.251368405  0.113658168
#>   [8,] -1.607192211  0.231952206
#>   [9,]  1.648941600  2.467367056
#>  [10,] -0.579173834  0.337896274
#>  [11,] -0.525659907 -0.032571861
#>  [12,] -0.231985031 -0.078058514
#>  [13,] -1.011624618  1.749647467
#>  [14,]  0.393307148 -0.209739588
#>  [15,] -0.967698650  0.632227851
#>  [16,]  0.236593951  0.110801256
#>  [17,]  0.021169523  0.017361493
#>  [18,] -0.133132942 -0.012967288
#>  [19,]  1.450244335 -2.461305655
#>  [20,] -0.441228509  0.513757534
#>  [21,]  1.732932399 -1.120710458
#>  [22,]  0.907859717 -0.024320732
#>  [23,]  0.158787534 -0.134176805
#>  [24,]  1.340320767 -1.571754127
#>  [25,] -0.052771392 -0.081507749
#>  [26,] -0.019292621 -0.006668394
#>  [27,] -0.895920486 -1.133311427
#>  [28,] -0.428094052  0.305002211
#>  [29,]  1.199092752 -0.231968289
#>  [30,]  0.815567632  2.042791518
#>  [31,] -0.999041260  0.880037838
#>  [32,]  1.280252306 -0.086679595
#>  [33,]  0.052896996 -0.012462148
#>  [34,]  0.091361601 -0.018931567
#>  [35,] -0.058706657 -0.005510336
#>  [36,]  0.356134276 -0.523412912
#>  [37,] -0.359002257  0.160997199
#>  [38,] -0.417702656 -0.112103714
#>  [39,]  0.189309647  0.290560613
#>  [40,] -0.931263675  0.548259727
#>  [41,] -1.692775391  2.223069008
#>  [42,]  0.742468639  0.549848265
#>  [43,]  1.533804983 -1.375987132
#>  [44,]  0.589035902 -0.898261161
#>  [45,]  0.008295018  0.007855553
#>  [46,]  0.460049797  0.555346533
#>  [47,]  0.047960100 -0.004266489
#>  [48,] -0.469091240 -0.452931975
#>  [49,] -0.941120787 -1.004931571
#>  [50,]  1.277287869  0.152955886
#>  [51,] -0.136949562 -0.046672290
#>  [52,] -0.534215841  0.190574557
#>  [53,]  0.982062198  0.902705939
#>  [54,]  0.500739675 -0.809066917
#>  [55,]  0.101753357 -0.072788680
#>  [56,] -0.318709773 -0.091300458
#>  [57,] -0.348259719  0.226681253
#>  [58,]  0.461328658 -0.708852275
#>  [59,] -1.525705459 -0.046906952
#>  [60,]  0.262253570  0.094743737
#>  [61,]  0.247784521 -0.098404699
#>  [62,] -0.223577776  0.139275223
#>  [63,]  0.423804782 -0.228017985
#>  [64,]  1.089747578  0.770151053
#>  [65,]  0.518117143  1.244149718
#>  [66,] -1.104323298 -2.228453491
#>  [67,]  0.652522332  0.439048588
#>  [68,] -1.653112431  1.094307308
#>  [69,] -0.885510719  0.215584946
#>  [70,]  0.211017318  0.171502797
#>  [71,]  0.328987050 -0.232361203
#>  [72,]  0.505832823  0.475547014
#>  [73,] -0.490339622  0.072287095
#>  [74,]  0.728758833  0.333916221
#>  [75,]  0.371847058 -0.651656689
#>  [76,] -0.352439715  0.327523608
#>  [77,] -0.642894023  0.592230341
#>  [78,]  0.271311224  0.483310628
#>  [79,]  1.961289138 -0.585522480
#>  [80,]  1.324785922 -1.053143143
#>  [81,]  0.690759361  1.021309852
#>  [82,]  0.403718644 -0.700921489
#>  [83,]  0.728837615 -0.603394783
#>  [84,] -0.185924602  0.111223264
#>  [85,]  0.001286426  0.001813696
#>  [86,]  0.340401544 -0.016274764
#>  [87,]  0.646144934 -0.533113103
#>  [88,] -0.249421560  0.001220306
#>  [89,]  0.728167773 -1.665273871
#>  [90,]  1.654967755 -0.422708842
#>  [91,]  0.485672258  0.536588813
#>  [92,]  0.501518321 -0.209947114
#>  [93,]  0.679924846  1.134686602
#>  [94,] -0.078663588 -0.105977134
#>  [95,]  0.888957585  1.066902304
#>  [96,]  1.089232628 -0.502277591
#>  [97,] -0.514932181 -0.175059030
#>  [98,] -0.847477949  1.608922063
#>  [99,]  0.166667846 -0.165497113
#> [100,]  1.422266331  0.061545454
#> [101,] -0.861859852  1.660453685
#> [102,]  0.999110766 -1.065865123
#> [103,]  0.824525474 -0.052130407
#> [104,] -0.557937557  0.039566017
#> [105,] -0.061067904 -0.007210787
#> [106,] -0.250216764 -0.182824349
#> [107,]  0.418894985 -0.500384422
#> [108,] -0.534012782  0.061275978
#> [109,]  0.379813486  0.202044079
#> [110,]  0.647041185  0.089016852
#> [111,]  0.205055958 -0.237140932
#> [112,] -0.962719210  0.940461928
#> [113,]  1.757885205  0.737176472
#> [114,] -1.400836331 -0.128395394
#> [115,] -0.107463679 -0.239080646
#> [116,]  1.228125702  0.420689377
#> [117,] -0.889127775  1.266302998
#> [118,]  0.751900134  0.675289153
#> [119,]  0.436717911  0.679400483
#> [120,] -1.723484681  3.371816333
#> [121,]  0.465746053 -0.119858216
#> [122,]  0.848928221  0.785102960
#> [123,]  1.144049356 -0.790408767
#> [124,]  0.826149391  0.843473993
#> [125,] -0.212523036 -0.180184785
#> [126,]  0.085700861 -0.051714744
#> [127,]  0.004183594  0.006867658
#> [128,] -0.134392137  0.022748411
#> [129,]  0.364770506 -0.063914254
#> [130,]  0.192682203 -0.128158609
#> [131,]  0.308959839  0.200715830
#> [132,]  0.884470483  1.393522534
#> [133,] -0.631389241  0.180562919
#> [134,] -0.741337631 -0.249275539
#> [135,]  1.077614385 -0.564087040
#> [136,]  1.946024426  0.365353751
#> [137,]  1.808295413 -1.623731969
#> [138,]  1.090787079 -1.831873306
#> [139,] -0.271792203 -0.112049342
#> [140,]  1.516351743  0.183429436
#> [141,]  0.648015005 -0.566182683
#> [142,] -0.073554146 -0.059220003
#> [143,]  1.691509501 -0.492275440
#> [144,] -0.524032302 -0.450428914
#> [145,]  0.873103855  0.861176768
#> [146,]  0.455018198 -0.809587933
#> [147,]  0.619387742 -0.029831436
#> [148,]  0.221294646  0.225234934
#> [149,]  0.733325846  1.156240305
#> [150,]  0.417995591  0.263972641
#> [151,]  0.141682727  0.302853583
#> [152,]  0.675851333 -0.935384778
#> [153,] -0.570408952 -0.368147807
#> [154,] -0.750318565  0.606952142
#> [155,]  0.943558905  0.130478410
#> [156,]  0.015855818 -0.012006073
#> [157,] -0.464789131 -0.333485072
#> [158,]  0.742211240 -0.064162558
#> [159,] -0.790026192 -1.805086076
#> [160,] -0.889602809  1.246707776
#> [161,]  0.038823595 -0.045613024
#> [162,]  0.481703681  0.120780425
#> [163,]  0.431437413 -0.230383553
#> [164,]  0.113397757  0.116799531
#> [165,] -0.075574658 -0.042000505
#> [166,] -0.305789554  0.414871915
#> [167,]  2.074087469 -1.178989240
#> [168,]  0.431829476  0.110573509
#> [169,] -1.037853156 -0.396904145
#> [170,] -1.120985628 -1.426230534
#> [171,] -0.143655925 -0.212093528
#> [172,]  0.337494634  0.020773799
#> [173,] -0.315864684 -0.098196380
#> [174,]  0.533981396 -0.230830084
#> [175,] -0.326398902 -0.433594992
#> [176,]  0.342670988  0.028581100
#> [177,]  1.049403175 -1.120529573
#> [178,] -1.426938611 -2.463945722
#> [179,]  1.302082040 -0.721476042
#> [180,]  0.677588893 -0.110433424
#> [181,] -0.247227617  0.315047983
#> [182,]  0.329649679 -0.219526763
#> [183,] -1.250106191 -0.630955021
#> [184,] -0.702338773 -0.164637434
#> [185,]  0.670071305 -0.174063203
#> [186,]  0.233063876 -0.124176063
#> [187,]  0.944973257 -0.112651169
#> [188,] -0.615109830  0.612490653
#> [189,] -0.727941409  0.918637536
#> [190,] -1.098313260  2.233513015
#> [191,]  0.207442855 -0.137400772
#> [192,]  0.743237762  0.493009491
#> [193,] -0.975678053  0.682839018
#> [194,]  1.874945841 -2.787902951
#> [195,]  0.894892958 -0.216155166
#> [196,]  0.715502431  0.562584087
#> [197,] -0.438097358 -0.114674127
#> [198,]  0.483045234  0.122845304
#> [199,]  0.607932519  0.337491730
#> [200,] -1.435804781  0.389782759
#> [201,]  0.312112543  0.704892639
#> [202,] -0.234128990 -0.124497050
#> [203,]  0.890842884 -0.592507647
#> [204,] -0.545488687  0.078409726
#> [205,] -1.871825088 -0.075319179
#> [206,] -0.941181274 -1.363021893
#> [207,]  0.964933527  1.469071651
#> [208,]  0.839414504 -0.369532212
#> [209,] -0.623808667  0.330902525
#> [210,]  0.116868999  0.047785558
#> [211,] -0.491568422  0.765944427
#> [212,] -0.378601283  0.287402735
#> [213,]  0.631012053  0.549206463
#> [214,]  0.355696842 -0.275531450
#> [215,] -0.142184900  0.037694039
#> [216,] -0.873104331  0.474326099
#> [217,]  0.029041151  0.020751844
#> [218,]  0.771887132  0.765646328
#> [219,] -0.622779192 -0.251094747
#> [220,] -1.019216368  1.802776031
#> [221,]  1.146641723  1.011736093
#> [222,] -0.686253275 -1.362175336
#> [223,]  0.574613754  0.080875404
#> [224,]  1.010735445 -1.167646702
#> [225,]  0.852289922  0.568106588
#> [226,]  1.510292700 -0.191848617
#> [227,]  0.326164475 -0.406296686
#> [228,]  0.346536678  0.144574102
#> [229,] -1.533334160 -2.415951214
#> [230,] -0.155088332  0.032757970
#> [231,] -0.082025195  0.033122151
#> [232,]  1.335379195  0.299827345
#> [233,] -0.430988524 -0.073359810
#> [234,]  0.551841789  0.363835308
#> [235,] -0.356629766 -0.543046109
#> [236,] -0.307850899  0.644023806
#> [237,]  0.872587429 -1.160131500
#> [238,]  0.660640182  1.363811969
#> [239,]  0.834433127  0.958844621
#> [240,] -0.028350180 -0.008179543
#> [241,]  0.709523687 -0.813695524
#> [242,] -0.535448275  0.695348942
#> [243,]  0.292484639 -0.263606963
#> [244,]  0.070662181  0.075317860
#> [245,] -1.494781617  0.195648013
#> [246,]  1.348617875  0.261099149
#> [247,]  0.464069545 -0.394464331
#> [248,]  0.141910902  0.143572033
#> [249,]  1.338874083 -0.090771513
#> [250,]  0.685951576  0.391147531
#> [251,]  0.111465680  0.011549084
#> [252,] -0.378028098  0.249992197
#> [253,]  0.448178079  0.108191920
#> [254,] -0.883415483  1.061310868
#> [255,]  0.472595879 -0.554282038
#> [256,]  1.543002322 -1.950492298
#> [257,]  0.912818481 -0.433637582
#> [258,]  0.304517948  0.075210820
#> [259,] -0.957291884 -0.450625210
#> [260,] -1.352936308  1.364751755
#> [261,]  0.935147845 -0.179133190
#> [262,] -1.828994013 -1.552237115
#> [263,]  1.159548447 -1.821239431
#> [264,]  0.573598167 -0.069531373
#> [265,]  0.662620001 -0.935778703
#> [266,]  0.203040314  0.103291774
#> [267,]  1.306969341 -0.307278513
#> [268,] -0.592945142  0.623787289
#> [269,]  0.005285423 -0.014087230
#> [270,] -0.397981094 -0.711376033
#> [271,]  1.464352147 -3.928857695
#> [272,] -1.875368764  0.112419123
#> [273,]  0.656781187 -0.153392139
#> [274,]  0.766906944  0.313550666
#> [275,]  1.067406734 -2.271061842
#> [276,] -0.117730538 -0.048239372
#> [277,]  0.578994328  0.285471554
#> [278,]  1.078197604  0.045804685
#> [279,] -1.293099723 -1.097666634
#> [280,]  1.000823133  0.251538040
#> [281,] -0.803330171 -2.073786760
#> [282,]  1.789307189 -0.729264638
#> [283,]  0.076639247  0.082236074
#> [284,]  0.042294184 -0.010955591
#> [285,]  1.632775890 -3.763919595
#> [286,] -1.090336560  1.097844602
#> [287,] -0.251674021 -0.071430738
#> [288,]  1.748540269  4.188146712
#> [289,] -1.061928391 -1.782032236
#> [290,] -1.484220370  2.849011429
#> [291,] -0.800791932  0.239529991
#> [292,] -1.316228969  0.471553746
#> [293,] -0.056494587 -0.066246350
#> [294,]  0.627026129  0.341407334
#> [295,]  0.228360513  0.339989209
#> [296,] -0.015088233  0.015759561
#> [297,]  0.175723264 -0.502053399
#> [298,]  0.616229737  0.072661607
#> [299,]  0.791377585  0.024013557
#> [300,] -0.259083486 -0.059484306
#> 
#> $grads_hat
#>                 [,1]          [,2]
#>   [1,] -0.5413647935  0.2970281411
#>   [2,] -1.2961803859 -1.0421170393
#>   [3,]  0.4273962134 -0.0260111628
#>   [4,] -0.2249547827 -0.2816567460
#>   [5,] -0.2770255650  0.0569482001
#>   [6,]  0.7107179245  1.0937329548
#>   [7,] -0.7490750971  0.7030124415
#>   [8,]  1.0539203496  1.3091691429
#>   [9,]  0.3552140618 -0.0679483665
#>  [10,]  0.1196447319 -0.0619542567
#>  [11,]  1.7739481164  1.8220108663
#>  [12,] -0.0053053023 -0.0071632007
#>  [13,] -0.3838953663 -0.1573094435
#>  [14,] -0.1499689769  0.3229035499
#>  [15,]  0.2881853209  0.2253602765
#>  [16,]  0.0431372279  0.0002401452
#>  [17,]  0.5335477364 -0.4570769472
#>  [18,]  1.1925614629  0.3114320011
#>  [19,] -0.5137258625 -0.0903562379
#>  [20,]  0.9385888165 -1.0420146335
#>  [21,] -0.0009563219  0.0009404608
#>  [22,] -1.2155638334  1.6951935656
#>  [23,] -0.4745650039 -0.3256861781
#>  [24,]  1.8760014605  0.1671459125
#>  [25,]  0.9273029559  0.3588443092
#>  [26,]  0.9309433202  0.4687806518
#>  [27,] -1.7391533582 -1.5730335883
#>  [28,] -0.4004701624 -0.0077197474
#>  [29,]  0.4715629002 -0.0823089111
#>  [30,]  1.3528973833  0.4797458277
#>  [31,]  0.4952000573  0.1090368702
#>  [32,]  0.7103398948  0.4563100519
#>  [33,]  0.0121723463 -0.0128038610
#>  [34,]  0.4290777683  0.2316041300
#>  [35,]  0.4645671202  0.4431312832
#>  [36,] -0.0231942012  0.0317320516
#>  [37,] -0.7118381873  0.1851279947
#>  [38,]  0.9774780958  0.3303239285
#>  [39,]  0.1448642875  0.1687687832
#>  [40,]  0.3287277558  0.1503211170
#>  [41,] -0.4977958732  0.6033030485
#>  [42,]  0.9384928239 -1.5021781395
#>  [43,] -0.4205862378 -0.1144875819
#>  [44,] -0.1888495968 -0.0497654981
#>  [45,] -0.6845024003 -0.2353438747
#>  [46,]  0.2203296086 -0.0856749658
#>  [47,] -0.0414121115 -0.0042526819
#>  [48,] -0.3096143335  0.5605825242
#>  [49,]  0.9505617113  0.3308974812
#>  [50,] -0.4598797337 -0.0789544333
#>  [51,] -0.4189594892  0.2594908117
#>  [52,] -0.0589826096 -0.0068578975
#>  [53,] -0.7109505798  1.0832117191
#>  [54,]  0.0101162125 -0.0091622788
#>  [55,] -0.7455688918 -1.0342622897
#>  [56,]  0.0192808816 -0.0073863350
#>  [57,] -0.0687094278 -0.0552256517
#>  [58,]  0.3337638575  0.5494104215
#>  [59,] -0.2817240065 -0.5055008434
#>  [60,] -0.2745197530  0.0904255109
#>  [61,]  0.3803281219 -0.1236091241
#>  [62,]  0.3877158212 -0.0898862090
#>  [63,] -1.3315961686 -0.2661839816
#>  [64,]  0.5402629063  0.0800631300
#>  [65,]  1.5638063121 -0.2012557396
#>  [66,] -0.5025786110 -0.2676272694
#>  [67,]  0.9118423527 -0.1349548481
#>  [68,]  0.6609573636  0.3653260860
#>  [69,] -0.1213289576  0.0999794438
#>  [70,] -0.9986885250  0.6622269047
#>  [71,] -0.7579977241  1.4458273903
#>  [72,]  0.1942020566  0.4004394977
#>  [73,] -0.4346764275 -0.0449005144
#>  [74,] -0.4994973870  0.1484094713
#>  [75,]  0.4340507404 -0.2939760123
#>  [76,] -0.0687217843  0.0796636113
#>  [77,]  0.2277367986  0.0306956482
#>  [78,] -0.9220616676  0.7419259324
#>  [79,]  0.7187053397  1.1621327220
#>  [80,]  0.1510683596  0.0856929971
#>  [81,] -2.4496053024  0.4281284825
#>  [82,]  0.8831872056 -0.5380050569
#>  [83,]  1.0348842477 -1.7198343409
#>  [84,]  0.8488993428 -0.7008868793
#>  [85,] -0.5140936055 -0.3302089491
#>  [86,] -0.9557647086  0.8028056581
#>  [87,]  0.9343624741  1.5524821130
#>  [88,]  0.9573498358 -0.8556181030
#>  [89,]  0.3205093331 -0.1141295317
#>  [90,] -0.3974871802  0.1611777184
#>  [91,]  1.4293128194 -4.5487605545
#>  [92,] -0.5957602335  0.1421261417
#>  [93,]  0.1106353196  0.0675741098
#>  [94,]  0.4168850071  0.1203140912
#>  [95,] -1.4353732654 -2.1606910619
#>  [96,] -1.3795619058  1.4431895176
#>  [97,]  0.4753668712  0.4798742020
#>  [98,]  0.0215270546 -0.0229251825
#>  [99,]  0.2263971193 -0.0164075637
#> [100,]  0.4056479030 -0.2278582272
#> [101,]  1.1874260539 -0.7678760486
#> [102,]  0.9940335048 -0.8658552240
#> [103,]  0.6280751165  0.1421336175
#> [104,] -2.1655023402 -3.7296325932
#> [105,] -2.0409370838  0.6452048251
#> [106,] -0.1312646497 -0.0739438990
#> [107,]  0.2552703008 -0.1256013289
#> [108,]  0.3097658549 -0.1755365313
#> [109,]  1.0337496198  0.7907327240
#> [110,]  0.2966606935  0.8106586195
#> [111,]  0.8291128277  0.0831196227
#> [112,]  1.2204597388 -0.1194028074
#> [113,]  0.1213203419 -0.1306286863
#> [114,]  1.1768500473  0.9129088093
#> [115,] -1.0411643420  0.2042440333
#> [116,]  0.1129705273 -0.1356712142
#> [117,] -1.3187050367 -1.1301960845
#> [118,] -0.7886915045 -0.3973508404
#> [119,]  1.4703578664 -1.2374184863
#> [120,] -1.0932396228 -0.4589445137
#> [121,]  0.2957194920  0.0293700852
#> [122,] -0.0404160594 -0.1066297502
#> [123,] -1.2735627047  0.6396174261
#> [124,] -0.1145097393  0.0181068850
#> [125,]  0.6416976643 -0.4281996678
#> [126,]  0.9774811973  1.4874594311
#> [127,] -0.1499364415  0.1126617862
#> [128,] -0.8425578993  0.1394363581
#> [129,] -0.3877317064 -0.0731912545
#> [130,] -0.4658134934  0.0724962489
#> [131,]  0.4185232209 -0.1284018724
#> [132,] -0.0814710294  0.0497075038
#> [133,] -0.2627070954  0.0630000044
#> [134,]  0.5337534028 -0.4755028114
#> [135,] -0.5463186180  0.0922831873
#> [136,]  1.3037101931 -1.3420115991
#> [137,]  0.6763767852 -1.0208059547
#> [138,] -0.0055240746 -0.0031141631
#> [139,]  0.7165245211 -0.4184107758
#> [140,]  0.3244475553  0.1619782852
#> [141,] -0.4209288999  0.5461688568
#> [142,] -0.2803679502 -0.3321587483
#> [143,]  0.4494484885  0.4539406576
#> [144,] -0.1227278047  0.1592385084
#> [145,] -1.2779255183  0.3833877165
#> [146,]  0.5886008945  0.6418751812
#> [147,]  0.5647127437  0.1981840933
#> [148,]  0.8947164079  1.1542726891
#> [149,] -0.2432279171  0.1180993694
#> [150,]  0.2365030075  0.1917865150
#> [151,]  0.0099417606  0.0153425431
#> [152,] -2.6130687539 -2.1043930339
#> [153,]  1.4471688831  2.6716211989
#> [154,]  1.1364261566  0.2131975778
#> [155,]  0.4110499706  0.1252786386
#> [156,] -0.1378239741 -0.0086894051
#> [157,] -0.4873034866 -0.0843073904
#> [158,]  1.0324907591  0.1331033445
#> [159,] -0.4029494238  0.0799976672
#> [160,]  1.8163710494 -2.2414456235
#> [161,]  0.8906511803 -0.5435082992
#> [162,]  0.1286018192 -0.2607938908
#> [163,]  1.1391943927  0.8692891384
#> [164,]  0.6875357156  0.2312366042
#> [165,]  1.3411190653  0.2625342171
#> [166,] -0.3370294974 -0.0707979758
#> [167,]  0.0837891010  0.0057336998
#> [168,] -0.8079918954 -0.0397955641
#> [169,] -0.0879670579  0.0434820362
#> [170,]  0.8319721278  1.1037833741
#> [171,] -0.7541503744  0.9858008079
#> [172,]  0.9356056344  0.6994551135
#> [173,]  0.0500103289  0.0702874747
#> [174,] -0.1946546085 -0.2262541984
#> [175,] -0.7920862166 -1.4608239319
#> [176,]  1.5777157683 -1.5004194087
#> [177,]  1.1581790309 -0.0058036974
#> [178,]  0.9811894981 -2.4423854286
#> [179,] -0.5399481826  0.1823295262
#> [180,]  0.7672073338  0.1934056022
#> [181,]  0.9141055386  0.5719339680
#> [182,] -0.0609494993  0.0369024940
#> [183,] -0.4003343047 -0.0588814947
#> [184,]  0.7031438785  1.2449313532
#> [185,] -0.6607247008 -0.8734304868
#> [186,]  0.4801114942 -0.4942288252
#> [187,] -0.0486519989 -0.0178290981
#> [188,]  0.1791038059 -0.2158657656
#> [189,] -0.4267434982 -0.9280819537
#> [190,]  0.1876543466 -0.1311878889
#> [191,]  0.0380807856  0.0062994480
#> [192,]  2.6092790624 -0.8788373801
#> [193,]  1.3816371654 -0.6598738593
#> [194,] -0.3984015491  0.0155430110
#> [195,] -0.2790375205  0.1963048980
#> [196,] -1.0697851730  0.7630589394
#> [197,]  1.0128764642 -0.4052419303
#> [198,] -0.9569853333 -0.1205460857
#> [199,] -0.7596591653  1.3449345050
#> [200,]  0.5496132229  0.3489957492
#> [201,]  0.3954490150  0.2144658905
#> [202,]  1.6177309219  1.2955260804
#> [203,] -1.0186807342  1.3935844386
#> [204,] -0.5272599978 -0.0702578986
#> [205,] -1.6706674918 -2.0668884330
#> [206,] -0.1050592856 -0.2688828293
#> [207,]  0.4770569993 -0.6620547956
#> [208,]  0.9971009910  1.8484443005
#> [209,] -1.3883636722 -2.0060080691
#> [210,]  0.4905626751 -0.0884919197
#> [211,]  0.3933275518  0.7045728275
#> [212,] -0.7702387924 -0.1197877840
#> [213,] -0.1472966578 -0.0813735358
#> [214,] -0.1427856448  0.1611654085
#> [215,] -0.7091499702 -0.2442258714
#> [216,]  1.3517027146 -3.2731955463
#> [217,] -0.2132928217  0.1208682414
#> [218,]  1.1944755368  0.7380311426
#> [219,] -0.7887779584 -1.9448548933
#> [220,] -0.2022720645 -0.3870380159
#> [221,]  1.5622000955 -2.8067695394
#> [222,] -0.4712233235 -0.5052652819
#> [223,] -0.5074609951  0.2700269253
#> [224,] -0.0697662942  0.0114749490
#> [225,]  0.6485718130 -0.8424621063
#> [226,]  0.0041748334  0.0084121788
#> [227,]  0.4317843385 -0.4051931662
#> [228,]  0.6350713262  0.5149087892
#> [229,] -0.6688318928 -0.0643997873
#> [230,] -0.1011401139 -0.1366617994
#> [231,] -0.0463793619  0.0451842197
#> [232,] -0.2823600088 -0.0076190886
#> [233,]  0.4505455867 -0.0087385093
#> [234,] -0.5603260033 -0.3311763643
#> [235,]  0.5384638648 -0.3523821738
#> [236,]  0.7484194725 -0.1213031082
#> [237,]  0.1565896175 -0.0369262040
#> [238,]  0.8702067343  0.8421979806
#> [239,]  0.5632750104  0.3029384618
#> [240,] -0.5691978581 -1.6524446137
#> [241,]  0.7547751532  0.3612664098
#> [242,] -0.4602397825 -0.2500371036
#> [243,] -0.8692932477  0.8501117818
#> [244,] -0.8963702584 -0.8934422154
#> [245,]  0.0485563335  0.0524766668
#> [246,]  0.3901199621 -0.0168697797
#> [247,]  0.6090594745 -0.4572710187
#> [248,] -0.1896761249  0.0396616993
#> [249,]  0.8215836452  0.6731107747
#> [250,]  0.0287735035  0.0482239985
#> [251,] -1.2278969851 -0.5926850561
#> [252,] -0.6181002871 -0.0955582767
#> [253,]  1.1623468555  1.1814287752
#> [254,] -0.3719782495 -0.7893163747
#> [255,] -0.8881710570  0.6295664926
#> [256,] -0.4293471579  0.1168092706
#> [257,] -0.3703486370  0.2546193391
#> [258,] -0.3364561099 -0.5124192126
#> [259,] -0.3172406171 -0.3840747816
#> [260,]  0.5558494089  0.0633440387
#> [261,] -0.4929321466 -0.2033959144
#> [262,] -0.7597874768 -0.8129664141
#> [263,]  0.1622893388 -0.0871007987
#> [264,]  0.0631440104 -0.0315067829
#> [265,]  1.0609131035 -0.3156913570
#> [266,] -0.4302519428  0.2190282020
#> [267,] -0.5897364744 -0.4580084811
#> [268,] -0.4447106170 -0.0788044282
#> [269,]  0.8194293104  0.8020466416
#> [270,] -0.7841767618  1.1675915336
#> [271,]  1.9562465693  1.9026416018
#> [272,] -0.1493266423 -0.0563464623
#> [273,]  1.7507014158 -0.8732763335
#> [274,]  0.7473524573 -0.2936988046
#> [275,]  0.1909145646 -0.1490297773
#> [276,]  0.0728041579  0.0425182383
#> [277,] -0.9978909321  0.5192175974
#> [278,]  0.3888725122 -0.1608812210
#> [279,]  0.0223319400 -0.0142651721
#> [280,]  1.1049853907  0.1563484958
#> [281,] -1.1763626203 -1.2575793661
#> [282,] -0.9831725022  0.0263462257
#> [283,]  0.1215149652  0.0767490491
#> [284,] -1.5005243044 -2.3234399038
#> [285,]  0.6394252938 -0.6769324231
#> [286,]  0.4401706044  0.2597026936
#> [287,] -0.7328901751  0.9938274964
#> [288,] -0.4301601026  0.4285820574
#> [289,]  0.2260091415  0.0470028952
#> [290,] -0.3599338724 -0.5728610205
#> [291,]  0.6728396747 -1.1226581338
#> [292,]  0.4792842726 -0.1507534419
#> [293,]  0.0757014771  0.0171556594
#> [294,]  1.2117489191  2.8168976913
#> [295,] -1.5903269163  3.0223632985
#> [296,] -0.1029681625 -0.0207935820
#> [297,] -0.7171842573 -1.0570486693
#> [298,]  1.3391923380  0.2585501184
#> [299,] -0.2706262255  0.0837781722
#> [300,] -0.7360494135  0.6579410840
#> 
#> $inv_hessian
#>             [,1]        [,2]
#> [1,]  1.00037870 -0.01898342
#> [2,] -0.01898342  0.95161054
#>