PPI++ OLS

Helper function for PPI++ OLS estimation

Usage

ppi_plusplus_ols(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see https://arxiv.org/abs/2311.01453). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ OLS regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier OLS regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) https://arxiv.org/abs/2311.01453`

Examples

dat <- simdat(model = "ols")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>

  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>

  matrix(ncol = 1)

ppi_plusplus_ols(X_l, Y_l, f_l, X_u, f_u)
#> $est
#> X(Intercept)          XX1 
#>    0.6544417    0.9111490 
#> 
#> $se
#> [1] 0.09422607 0.09331478
#> 
#> $lambda
#> [1] 0.5148807
#> 
#> $rectifier_est
#> X(Intercept)          XX1 
#>  -0.02792686  -0.12413267 
#> 
#> $var_u
#>            [,1]       [,2]
#> [1,]  0.2706170 -0.0613664
#> [2,] -0.0613664  0.3545745
#> 
#> $var_l
#>           [,1]      [,2]
#> [1,] 2.3769054 0.1485426
#> [2,] 0.1485426 2.7148554
#> 
#> $grads
#>                [,1]          [,2]
#>   [1,]  0.379868053 -0.8578402283
#>   [2,] -0.247440619 -0.2555719703
#>   [3,]  1.098087654 -0.3213743264
#>   [4,] -0.832634591 -0.0302422846
#>   [5,]  0.482010050  0.3545457105
#>   [6,] -1.697369132  1.8925094610
#>   [7,]  1.683521846  0.4344736388
#>   [8,] -0.276875655 -0.4146972375
#>   [9,] -1.676266828 -0.9483263109
#>  [10,]  1.615360135  1.0177743507
#>  [11,]  0.535657377 -0.7406377653
#>  [12,]  1.275485727  0.3430037750
#>  [13,] -0.053737786 -0.0405191043
#>  [14,] -0.095195165 -0.1630775571
#>  [15,]  4.469080965 -9.4397634519
#>  [16,]  0.108908919 -0.0752808928
#>  [17,]  1.185625682 -0.1191158265
#>  [18,]  1.258399192  1.5045637398
#>  [19,] -1.250471993  0.4093011194
#>  [20,] -0.766697498 -0.0868170738
#>  [21,]  0.650175534  0.4088765804
#>  [22,]  1.026471362 -1.7288022647
#>  [23,] -0.188502165 -0.0637660404
#>  [24,]  1.189194389 -0.8456633088
#>  [25,] -0.030183721 -0.0174548324
#>  [26,] -1.803833110  0.8124805498
#>  [27,] -0.379359998 -0.0859655935
#>  [28,] -0.047662072 -0.1252966845
#>  [29,]  2.751281185  2.3593842947
#>  [30,]  0.936718201  0.5791464795
#>  [31,]  0.129967890  0.1598360013
#>  [32,]  1.483073477 -2.8050531505
#>  [33,] -1.447952165  0.9192325385
#>  [34,] -1.461022042 -1.0967095429
#>  [35,] -1.360424539  0.5274867603
#>  [36,]  2.985862959  2.4095812862
#>  [37,] -0.572410970 -0.5214373001
#>  [38,]  2.931228586  2.9626238569
#>  [39,] -0.362851664  0.3492626008
#>  [40,]  1.848497203 -2.0165552456
#>  [41,] -0.356204932  0.1827133034
#>  [42,]  0.409298018  0.0319884741
#>  [43,] -1.405381318 -0.2539441194
#>  [44,] -0.814548874  0.2871992728
#>  [45,] -1.646016652  0.9646552248
#>  [46,]  0.137162181 -0.1089119167
#>  [47,] -2.590139377 -6.7558435195
#>  [48,] -3.394286085 -4.6528412916
#>  [49,] -1.018398986 -1.5529330091
#>  [50,]  1.041757619 -0.1997309172
#>  [51,] -0.526170991  0.7207630297
#>  [52,] -2.871811140 -0.1222609453
#>  [53,]  3.150625220  7.7620813721
#>  [54,]  0.370897634 -0.5243634732
#>  [55,]  0.816674627  0.2143042795
#>  [56,]  0.479033955 -0.3290304869
#>  [57,] -0.069625187 -0.0482740192
#>  [58,] -0.587295426 -0.1229168483
#>  [59,] -0.969874083 -1.2597517834
#>  [60,] -1.239664929 -1.0357885005
#>  [61,] -1.291613949  2.3697490855
#>  [62,] -5.366052454 -2.5538996917
#>  [63,] -0.004094896  0.0004328726
#>  [64,]  3.633200455  6.3069639691
#>  [65,] -0.731905691  0.4963351676
#>  [66,] -2.801582709 -1.9046186091
#>  [67,]  0.638237468 -0.1021512846
#>  [68,]  0.056800382  0.0721820777
#>  [69,]  1.659892571 -0.2600714493
#>  [70,]  0.508703917 -0.4747920473
#>  [71,] -0.650254103 -0.2047470439
#>  [72,] -0.403970221 -0.8810034954
#>  [73,] -1.066571672 -0.4225534759
#>  [74,] -0.299042453 -0.1446475301
#>  [75,] -0.753302434  0.8227743819
#>  [76,]  2.163424541 -1.0604217384
#>  [77,] -2.002198031  2.8670148588
#>  [78,]  0.021006878 -0.0023676770
#>  [79,]  1.040523197 -0.1257365151
#>  [80,] -0.112670281  0.0839088844
#>  [81,]  1.792735562 -1.0890736150
#>  [82,]  0.763653521  0.1583078694
#>  [83,]  0.771497212  0.9823834372
#>  [84,]  2.977255537 -2.0230597770
#>  [85,] -1.794756046  0.9051777974
#>  [86,]  0.150735800  0.0802457422
#>  [87,] -2.287740253 -0.6328976136
#>  [88,] -0.756791839  0.4180230853
#>  [89,] -1.266118987 -0.0944397613
#>  [90,]  2.723568755  0.8452315495
#>  [91,]  0.463069788  0.2276029715
#>  [92,] -1.555485976  0.2339515693
#>  [93,] -0.714452286  1.1024993352
#>  [94,]  1.723817047  1.4679573460
#>  [95,]  0.056697408 -0.0137536409
#>  [96,]  1.683263394  0.3989202593
#>  [97,]  1.204343196 -2.0775018868
#>  [98,]  1.773120949 -1.9043203327
#>  [99,]  0.061583525 -0.0139895803
#> [100,] -1.170029077  3.0515005895
#> [101,]  0.990442686 -0.4114379040
#> [102,] -1.105594195  1.5580605477
#> [103,] -0.202765980  0.0067244131
#> [104,]  0.770685894  0.1583927539
#> [105,]  5.584282933  6.9041986828
#> [106,] -2.356572304 -1.7351001061
#> [107,] -2.469005678 -2.5226506376
#> [108,] -2.952117839  2.5802385257
#> [109,]  0.650130366  0.7102491644
#> [110,] -3.083043266 -9.6295099734
#> [111,]  2.339718777 -0.2363766669
#> [112,]  0.205483335 -0.1954353583
#> [113,] -0.130870542  0.0943251610
#> [114,]  3.153692605  0.7396122617
#> [115,]  0.711622704  0.6094816696
#> [116,] -2.182344641  1.3054640978
#> [117,] -2.615959151  1.3990735231
#> [118,]  0.895877712  0.6548668325
#> [119,]  0.683652567 -1.1082945084
#> [120,] -0.413866212 -0.2458835600
#> [121,] -0.918752869  1.2140519772
#> [122,] -3.633182972 -1.4449624373
#> [123,]  0.207331300  0.0206628613
#> [124,]  0.906552710  1.7253037768
#> [125,]  3.927021233 -0.7514045272
#> [126,] -0.702390101 -0.4341252373
#> [127,] -0.373791248 -0.3016043183
#> [128,] -1.997224818 -0.5540765900
#> [129,]  0.358813015  0.1677765628
#> [130,] -1.270120597  1.3932406481
#> [131,] -0.640829769 -0.5941367253
#> [132,] -0.350134767 -0.5808079658
#> [133,]  0.853634981  0.3520312523
#> [134,]  0.078315442  0.1108900956
#> [135,]  2.085483719 -2.9326243417
#> [136,] -0.841035246  0.5690570875
#> [137,] -2.163061307 -4.8219385788
#> [138,]  0.034948057 -0.0163876304
#> [139,] -1.042014854 -0.3926622831
#> [140,]  1.093351120 -0.5821207367
#> [141,] -0.812144781 -0.6016563927
#> [142,] -0.156906088 -0.1567559698
#> [143,] -0.548840822  0.0699086463
#> [144,]  0.803460815 -0.9317074270
#> [145,]  0.044301448 -0.0222029974
#> [146,]  0.853680691  1.0001477976
#> [147,] -0.394606924  0.5007588108
#> [148,]  3.115431876  2.5373379734
#> [149,]  1.093502964 -0.5718515554
#> [150,] -1.635894165  1.2726447784
#> [151,]  0.368596352  0.7958645113
#> [152,] -1.914395627  0.3899563408
#> [153,] -0.447790937  0.8081022036
#> [154,]  2.546802175 -4.1361718115
#> [155,]  2.015175112  1.0609575916
#> [156,] -1.031027064 -0.0248413414
#> [157,]  0.610367058 -0.6936175496
#> [158,] -0.455172512  0.2190857795
#> [159,]  0.296087942  0.3144693626
#> [160,]  1.219722749 -0.5731384371
#> [161,]  2.312551053 -1.8278738131
#> [162,]  0.135876017  0.0531426256
#> [163,]  5.728887565  9.3165843934
#> [164,]  0.116208707  0.1629669625
#> [165,]  1.004154364  1.1010014497
#> [166,] -5.195732898  0.7558618108
#> [167,] -0.044290276 -0.0073771976
#> [168,]  0.163409209  0.1638469169
#> [169,]  1.094208968 -0.5936587080
#> [170,]  0.986967244  2.3845587948
#> [171,]  4.495095759 -1.8358906701
#> [172,] -1.785484990 -3.3696133426
#> [173,] -0.453622366  0.4424508817
#> [174,]  0.450631691  0.3399256762
#> [175,] -0.091145693  0.0704224827
#> [176,] -2.392134313  3.5165958415
#> [177,] -1.356381547 -0.7327037407
#> [178,]  0.803167041 -0.8757466765
#> [179,] -1.856941601  1.0386411142
#> [180,]  1.721382379  0.6782931706
#> [181,]  1.908416548 -1.0098282994
#> [182,] -0.558111405 -0.2584594830
#> [183,]  2.937944782 -2.4258494266
#> [184,]  1.013156726 -0.1705396075
#> [185,] -0.196416881  0.1092233448
#> [186,]  1.233514990 -0.9121702131
#> [187,] -1.670076156  0.0569475290
#> [188,] -1.389085378  1.1158622903
#> [189,]  0.477963088  0.1493941168
#> [190,] -1.875969756 -3.1771842131
#> [191,] -0.135627034  0.1939013701
#> [192,] -1.203572768  2.8248436220
#> [193,]  0.041644967  0.0649559427
#> [194,]  0.003729231 -0.0007100393
#> [195,] -2.922331370 -2.5619983104
#> [196,] -0.010158337 -0.0110338473
#> [197,]  1.320916632 -0.8429989862
#> [198,]  1.520019924 -1.0957875959
#> [199,]  0.989600162 -0.4754438857
#> [200,]  0.123999979  0.0442453839
#> [201,]  0.995994367 -0.5569338864
#> [202,]  1.315955988  2.2040200595
#> [203,]  0.419548860 -0.2026320944
#> [204,] -2.072415314  0.2946345616
#> [205,] -0.305130338 -0.6243533659
#> [206,]  0.046631528 -0.0131668254
#> [207,] -3.161071177  2.4030334494
#> [208,]  3.571124588  3.5801135785
#> [209,]  0.279572412 -0.2399954902
#> [210,] -1.103346490 -0.4780123395
#> [211,] -1.159600614 -3.7864040724
#> [212,]  0.579176631 -0.3195383801
#> [213,] -0.978986746  0.1751548393
#> [214,]  1.211400412  2.2648091377
#> [215,] -1.940936801 -1.2525262267
#> [216,]  0.565604216  0.7040780783
#> [217,]  1.326807464  0.8836519804
#> [218,] -1.404355965  0.8532030162
#> [219,] -0.414296450 -0.2618776923
#> [220,] -3.312654753  0.6479682418
#> [221,] -1.404439286 -1.7863749536
#> [222,] -0.037471166 -0.0252637303
#> [223,] -0.769144693 -1.9542503052
#> [224,] -0.820877204  0.4072500593
#> [225,]  1.409172308  1.3044369357
#> [226,]  1.626803595 -0.4874897912
#> [227,]  0.473848201 -0.5772489205
#> [228,]  0.746095637 -0.5831758435
#> [229,] -0.668016314 -0.0502741067
#> [230,]  0.623046841  0.0948806973
#> [231,]  0.980033073  1.7043085059
#> [232,] -0.171468003 -0.0308058380
#> [233,]  0.351149103 -0.1612242129
#> [234,]  1.131854397 -0.8251913390
#> [235,]  1.006992774 -0.7010010017
#> [236,] -0.341828506 -0.1225670230
#> [237,] -1.923131259  0.3859212105
#> [238,] -5.565484427  0.2215817032
#> [239,]  1.223937781  1.1564606222
#> [240,] -0.402685085 -0.4655516615
#> [241,]  0.064195919  0.0180347783
#> [242,] -0.413204920 -0.1938476223
#> [243,] -0.042329542 -0.0516280547
#> [244,]  2.043960572 -2.2033239485
#> [245,] -1.726774655  1.9994688664
#> [246,] -0.991582891 -0.1069203415
#> [247,] -1.369634574 -1.1233162860
#> [248,]  1.709498840  1.4424068068
#> [249,] -0.756459611  0.3709866138
#> [250,] -2.291185115 -3.3183755772
#> [251,]  0.330182552 -0.3342089569
#> [252,] -0.673354167  0.7667508709
#> [253,]  1.626811854 -0.2412241892
#> [254,]  0.613265643  0.5828391059
#> [255,]  0.041668714  0.0189042125
#> [256,]  1.419648195  1.3571990982
#> [257,] -2.059827738 -4.3525605142
#> [258,]  0.047238290  0.0207549321
#> [259,] -1.263770506 -1.6879306274
#> [260,]  1.183903024  0.1388306582
#> [261,]  1.683962536 -0.1238054913
#> [262,]  1.145420753  1.8095117280
#> [263,] -0.544473794  0.3425930833
#> [264,]  1.029986747 -1.2256676105
#> [265,] -1.346614905 -0.5312215537
#> [266,]  0.698322476 -0.4750619260
#> [267,] -1.345749455 -0.1344520657
#> [268,]  1.346596804  0.8960220614
#> [269,] -0.173173892  0.0189528661
#> [270,] -0.906703990  0.3299221066
#> [271,]  2.458711363  4.4775317383
#> [272,] -0.908167138  0.2611813173
#> [273,]  1.709002724 -1.0788246143
#> [274,]  1.816457098 -3.4832197948
#> [275,] -0.064736257  0.1437904842
#> [276,]  1.223748796  0.2645470238
#> [277,] -1.563878705 -0.4985993841
#> [278,]  0.929231046  1.8258979204
#> [279,]  0.324571883 -0.2859815030
#> [280,] -2.016216269 -3.5532560696
#> [281,]  0.670503503 -1.0159961719
#> [282,]  1.171763731  0.3212482042
#> [283,]  0.697694963 -0.1377215678
#> [284,] -6.672623397 13.0192501186
#> [285,] -0.059812745  0.0523776365
#> [286,]  3.578860759  1.9405211722
#> [287,] -3.259445573  2.7774074288
#> [288,] -1.857966123 -2.1250533688
#> [289,]  3.126634670  0.1903127610
#> [290,] -0.796155137 -0.0239903362
#> [291,] 10.522541876 -5.5854585017
#> [292,] -0.143259110  0.0374545320
#> [293,] -0.726028078  0.3768531716
#> [294,]  0.013337179  0.0134419385
#> [295,]  1.506159314  1.0125537428
#> [296,]  0.620302577  0.7599263569
#> [297,] -1.405620549  3.5230808891
#> [298,] -2.286704856 -0.0778841634
#> [299,] -3.615727792  2.2937944913
#> [300,]  3.019675290 -1.7289470389
#> 
#> $grads_hat_unlabeled
#>                [,1]          [,2]
#>   [1,]  0.183157343  0.1336966511
#>   [2,] -0.840917483 -0.3264281527
#>   [3,]  0.852378483  0.7100090438
#>   [4,] -2.276547082 -2.1045784849
#>   [5,] -0.116942788  0.0442690606
#>   [6,]  0.555681364 -0.4615951381
#>   [7,] -1.070813607 -0.0014801812
#>   [8,]  0.566661862 -1.0068621104
#>   [9,]  0.495801593 -0.1463274333
#>  [10,]  0.845202712  0.3770070061
#>  [11,] -1.971748454  0.6048374105
#>  [12,]  0.080928621  0.0765922751
#>  [13,]  1.817533429  2.2195696232
#>  [14,] -0.216987032  0.0587415867
#>  [15,]  0.657139431  1.1353205272
#>  [16,] -0.596752584  0.4746963634
#>  [17,]  0.662807745 -0.0058994388
#>  [18,] -0.438100986  0.5972101151
#>  [19,]  0.340404340  0.0986944498
#>  [20,]  1.019335796 -0.9689978049
#>  [21,]  0.608947174  0.3494751762
#>  [22,] -1.423596862 -1.0035757237
#>  [23,] -1.139455423  0.3552356769
#>  [24,] -1.273239151 -1.5529942543
#>  [25,]  0.199294893  0.0231694755
#>  [26,] -0.037520002 -0.0164920263
#>  [27,] -0.786796418 -0.4788113692
#>  [28,] -1.190741904  2.0564978447
#>  [29,] -0.576121423 -0.3240397672
#>  [30,]  1.487410002 -2.3311698585
#>  [31,]  0.869239877 -0.7828958867
#>  [32,] -2.460757983  0.9783126714
#>  [33,] -1.459496787 -0.2418011211
#>  [34,] -0.242806380  0.1972581388
#>  [35,] -0.603670059 -0.6618173110
#>  [36,] -0.761513902 -1.4824032256
#>  [37,] -1.346618412 -1.2862982511
#>  [38,]  1.057592843 -1.0458072171
#>  [39,] -1.361496639  0.2633144696
#>  [40,] -1.284973036 -2.3152346912
#>  [41,]  0.212936864 -0.0958479268
#>  [42,]  0.189146197  0.0573557572
#>  [43,]  0.841413629 -0.4866945987
#>  [44,]  0.602131662 -0.2876527285
#>  [45,]  0.137029646 -0.1295255799
#>  [46,] -0.755859119  0.0977946408
#>  [47,] -1.353508797 -3.0815670226
#>  [48,] -0.612820075  0.5847256346
#>  [49,]  0.204722698  0.0704219912
#>  [50,] -1.204010525 -0.2776054358
#>  [51,]  0.038871182 -0.0008558238
#>  [52,] -1.885959131  3.1902301323
#>  [53,] -0.309065380  0.1914323470
#>  [54,] -0.348236855  0.1798824868
#>  [55,] -0.342612734 -0.5699825376
#>  [56,]  0.110650434 -0.0195531612
#>  [57,]  1.376424089 -1.5292410453
#>  [58,] -0.717529329  0.0525965750
#>  [59,]  0.392344910 -0.2187852427
#>  [60,]  0.936827281  1.5103790720
#>  [61,]  1.422933468 -0.4887909774
#>  [62,] -0.480389351 -0.7534409226
#>  [63,] -1.112198872 -1.9688049363
#>  [64,] -0.676923643  1.2946047858
#>  [65,]  0.086490343 -0.1427454360
#>  [66,]  0.406170898  0.1148851358
#>  [67,] -0.715730380 -1.7648413292
#>  [68,] -0.284437829  0.0857614301
#>  [69,]  0.196054794 -0.1003652253
#>  [70,]  1.515461313 -2.2348535171
#>  [71,]  2.097439751 -3.2259576931
#>  [72,] -0.540171773 -0.7211582902
#>  [73,] -0.024528464 -0.0147956717
#>  [74,] -0.751664310 -0.0442345455
#>  [75,] -0.087618123 -0.0440283815
#>  [76,] -0.730311681 -0.6824360480
#>  [77,]  1.646802215 -0.1154836306
#>  [78,]  0.081400181  0.0230210409
#>  [79,] -0.839187366  0.2687343149
#>  [80,] -1.314530644  1.1532588262
#>  [81,]  0.720373416 -0.1876152667
#>  [82,]  1.480035191 -0.4503246811
#>  [83,]  0.978712262  0.9768164760
#>  [84,] -1.185520927 -1.0904368805
#>  [85,]  0.390356845 -0.0679198782
#>  [86,]  0.413413603 -0.2490551224
#>  [87,] -0.740508200 -0.4058216449
#>  [88,]  0.204308042  0.2296672928
#>  [89,] -0.227284643  0.0931479849
#>  [90,] -0.038368413  0.0073610395
#>  [91,]  0.172970180  0.0954560798
#>  [92,] -0.572077895 -0.0968356243
#>  [93,]  1.786234992 -2.1406237810
#>  [94,]  0.448086097  0.1504024466
#>  [95,] -0.526798217 -0.2407084932
#>  [96,] -1.757819073  0.2191334094
#>  [97,]  1.029323262  2.0906191236
#>  [98,] -1.026650668 -0.4127166744
#>  [99,] -0.624693424 -0.7112039927
#> [100,] -1.739754117 -0.1456524420
#> [101,] -0.231701505  0.0335787930
#> [102,]  0.430185572 -0.9604049143
#> [103,] -0.499092738  0.0835262535
#> [104,]  1.009451710  0.1060638889
#> [105,] -1.113472762 -1.0777790492
#> [106,] -0.734609879  0.3072090388
#> [107,]  1.001547251  0.0883912949
#> [108,] -0.823690822  0.8485469715
#> [109,]  0.733619128  0.0720205035
#> [110,]  0.173201198  0.0989259275
#> [111,]  1.681821611  1.0999368354
#> [112,]  0.968324810 -0.7350883176
#> [113,]  1.800198189 -1.5099306469
#> [114,]  0.961552924 -2.5662074048
#> [115,]  1.073656964 -0.9933555899
#> [116,]  0.510043732 -0.0832152312
#> [117,] -0.602415934 -0.3864673491
#> [118,]  0.496101054  0.5232294075
#> [119,] -0.906256605 -0.1326292389
#> [120,] -0.505195254 -0.2854096924
#> [121,] -0.495398021  0.3652018978
#> [122,]  0.396758660  0.1882644767
#> [123,]  0.827571129  0.2435088923
#> [124,]  0.894555518  0.0518535917
#> [125,]  1.131239666  1.3250338465
#> [126,] -0.694611302 -0.4360725617
#> [127,] -1.206967713 -0.5330525609
#> [128,] -0.940543606 -0.4566951846
#> [129,]  0.826986325  0.1479312482
#> [130,] -0.445976232 -0.1585934196
#> [131,] -1.704533009 -0.8213375017
#> [132,] -0.590393677  0.3623558615
#> [133,] -1.165369111 -0.8817123356
#> [134,] -0.498052167  0.7739060188
#> [135,]  0.547191216 -0.6293547724
#> [136,] -0.106645182 -0.1246577065
#> [137,] -1.840553648 -0.0571151591
#> [138,] -0.250640571  0.3422473281
#> [139,] -1.294846380 -1.5294752883
#> [140,]  0.259882371  0.5433762387
#> [141,]  2.481426295 -5.3161070918
#> [142,] -0.646344348  0.8301675231
#> [143,]  0.395956465  0.6659807499
#> [144,]  0.851563115 -0.1229363808
#> [145,] -0.634858553 -1.1675005168
#> [146,] -0.607907108  0.2822787779
#> [147,]  0.269755360  0.0126159773
#> [148,]  0.785307361 -0.3363056928
#> [149,] -0.707825080 -2.0221999704
#> [150,]  0.701694460  0.1343380654
#> [151,] -0.483452954 -0.0839570083
#> [152,] -0.334325424 -0.1911918993
#> [153,] -0.032532714 -0.0069325788
#> [154,] -0.061481678  0.0226659468
#> [155,]  0.241878080  0.2498404087
#> [156,] -0.582158995 -0.4221530417
#> [157,]  0.239495715 -0.0011270883
#> [158,] -0.006305324 -0.0072443939
#> [159,]  0.428708349  0.6573940578
#> [160,]  0.141311705 -0.0063262474
#> [161,] -0.524584988 -0.5622407125
#> [162,]  1.036230055 -1.0766582529
#> [163,] -0.180964802  0.2381260861
#> [164,] -1.263008679  0.5418811393
#> [165,] -1.414988810  0.0794523816
#> [166,] -0.124518934  0.0388088767
#> [167,]  1.610283681 -1.5884573850
#> [168,] -1.104705556 -0.7804455721
#> [169,]  0.204857640  0.1665820694
#> [170,] -0.078258885 -0.1332604927
#> [171,] -0.780999652  0.3496859313
#> [172,] -0.456690492 -0.1308984420
#> [173,]  0.289778026  0.4177058435
#> [174,] -0.086377844 -0.0898095645
#> [175,]  0.608314295 -0.4459250487
#> [176,] -0.578195649  1.3151153598
#> [177,]  0.276027186  0.0593480694
#> [178,]  1.181333546 -0.9049869727
#> [179,]  1.108652910  0.2659837946
#> [180,] -0.851771061 -0.8203631989
#> [181,]  0.598460400  0.1153879915
#> [182,] -1.599690823 -0.4523781141
#> [183,]  0.800839672 -0.3383932932
#> [184,] -0.860295060  0.5314132869
#> [185,] -1.319291477  2.5461687917
#> [186,]  0.852590419 -0.4063705007
#> [187,] -2.611801718 -0.8130863258
#> [188,] -0.190557432 -0.0760335132
#> [189,]  0.347802369  0.2557180928
#> [190,] -1.240529758 -1.0132860580
#> [191,]  2.259540834  2.3844147979
#> [192,] -1.061763528 -0.9023738069
#> [193,] -0.124768516 -0.1963880305
#> [194,]  1.240492265 -0.5581303747
#> [195,] -0.400057348  0.2402342841
#> [196,] -0.644357761  0.4641943544
#> [197,]  0.715266132 -0.1889326417
#> [198,] -0.986718973 -0.6763751164
#> [199,]  0.695494204  0.0116925824
#> [200,]  1.491968314  0.9610875799
#> [201,]  0.801862574 -0.5028251583
#> [202,] -1.744407991  3.4982313830
#> [203,] -1.715979698  2.1809412341
#> [204,]  0.322007000  0.0624815528
#> [205,] -1.071430545  1.4450015700
#> [206,] -1.318764296  2.8298138583
#> [207,]  0.956375066 -1.4909070341
#> [208,]  1.898715808 -4.0025337511
#> [209,] -1.279980267  1.5600911481
#> [210,] -1.092241244 -1.0899966552
#> [211,] -1.374020047 -0.5620167553
#> [212,] -0.450234395 -0.5117264719
#> [213,]  1.268476866 -0.6870347337
#> [214,] -0.591236164  0.8589855420
#> [215,] -0.346921506 -0.1904386159
#> [216,]  0.415972119 -0.0374380335
#> [217,]  0.357424339  0.2574933045
#> [218,] -0.701620424 -0.3782729782
#> [219,] -2.043753334  2.9809504774
#> [220,]  1.117434716 -0.2431978457
#> [221,]  1.186594512 -0.7043854455
#> [222,] -0.182837980  0.0742077438
#> [223,] -0.479187879  0.1107029723
#> [224,] -2.186028012  2.5957661441
#> [225,] -0.604512647 -0.4464631998
#> [226,] -0.681999287 -0.1408686279
#> [227,]  0.679627145 -0.1031038832
#> [228,]  1.208781992 -2.5729795694
#> [229,]  0.869924672 -0.4862823424
#> [230,]  1.299704552 -1.3738773788
#> [231,] -0.140625929 -0.0243617899
#> [232,]  1.077617340  1.1339437172
#> [233,]  0.178700226 -0.0483084824
#> [234,]  0.095084205 -0.0306279432
#> [235,] -0.257137118 -0.4781908119
#> [236,] -1.166779636  1.4247385899
#> [237,] -0.211474646 -0.0777230399
#> [238,] -0.979815542 -1.0111078148
#> [239,]  0.841247249  0.0847631632
#> [240,] -1.678900939 -2.3866625878
#> [241,] -0.308861792 -0.0806195877
#> [242,] -0.656094433  0.6855939204
#> [243,] -0.628496847 -0.0479084855
#> [244,] -0.822474477 -0.7100890398
#> [245,] -1.723123243 -2.1154618103
#> [246,]  1.002150744 -2.0313506290
#> [247,] -0.323546293 -0.6717041920
#> [248,]  1.916327146 -0.1785483636
#> [249,] -0.234019756 -0.2656904929
#> [250,]  1.562086334  1.8269517032
#> [251,]  0.781450698 -0.7906077071
#> [252,] -0.337411631  0.0386211336
#> [253,]  3.048703090 -8.2227290789
#> [254,] -0.551832507  0.6638281866
#> [255,]  1.907657945  1.8256458438
#> [256,] -1.293206984 -0.8644598327
#> [257,] -0.035503240 -0.0467297992
#> [258,] -1.554438046  2.4898572794
#> [259,] -0.076302056 -0.0046271839
#> [260,] -1.709243023 -1.1900378402
#> [261,] -1.544911579  1.6651659521
#> [262,] -2.752111595 -0.4844487926
#> [263,] -0.328369079 -0.1243307025
#> [264,] -0.336742244 -0.1752120761
#> [265,] -0.025203475  0.0450013679
#> [266,] -0.018674984  0.0336146214
#> [267,]  0.259444014 -0.2305589827
#> [268,] -0.750983231  0.8210842647
#> [269,]  1.536253076 -2.1339945960
#> [270,]  1.603867398  0.8422943662
#> [271,]  0.325173764  0.6892817021
#> [272,]  1.575966749 -0.7147329408
#> [273,]  0.747337125 -0.3338090637
#> [274,] -0.866280177  0.9731128084
#> [275,]  0.025765448 -0.0137581928
#> [276,] -0.689083353 -1.0983575351
#> [277,]  1.142311244  0.8528223147
#> [278,]  0.528220048  0.1894524928
#> [279,] -0.728637599 -0.5024216193
#> [280,] -0.546089915 -0.8978908705
#> [281,]  1.221396321  0.1033604479
#> [282,]  1.382183573  1.2836856931
#> [283,]  1.345246054  1.3627240089
#> [284,] -0.265892868 -0.3009428518
#> [285,] -1.716412089  0.0961390995
#> [286,] -0.037718832  0.1021533981
#> [287,]  1.717572176  0.0589552510
#> [288,] -1.609991181  1.9245798469
#> [289,]  1.032767900 -0.0812115868
#> [290,]  0.315957699  0.3943468914
#> [291,] -0.841133287 -1.1213348607
#> [292,]  0.432015144 -0.7269336792
#> [293,]  0.805500753  1.8065400581
#> [294,] -0.962389844 -0.6312531267
#> [295,] -0.178035280 -0.1883119481
#> [296,]  1.126132416  0.7103560493
#> [297,] -0.866825900 -1.2441509602
#> [298,]  1.328341587  0.6272447749
#> [299,]  1.283350304 -2.0190542822
#> [300,] -0.824154534 -0.5573438239
#> 
#> $grads_hat
#>                [,1]        [,2]
#>   [1,] -0.005839403  0.01318688
#>   [2,] -0.897842063 -0.92734680
#>   [3,]  0.622608561 -0.18221715
#>   [4,] -0.560900199 -0.02037257
#>   [5,] -1.174772249 -0.86411157
#>   [6,] -2.140623709  2.38672340
#>   [7,]  0.255418712  0.06591699
#>   [8,] -0.346658000 -0.51921544
#>   [9,] -2.398903625 -1.35714875
#>  [10,] -1.074650175 -0.67709445
#>  [11,]  0.897085899 -1.24037440
#>  [12,] -0.776610834 -0.20884628
#>  [13,] -0.245822801 -0.18535411
#>  [14,]  0.865667526  1.48296340
#>  [15,]  2.834558083 -5.98726181
#>  [16,]  1.688363711 -1.16704425
#>  [17,]  1.981747029 -0.19909946
#>  [18,]  1.151994806  1.37734482
#>  [19,]  0.231845999 -0.07588721
#>  [20,] -0.838352403 -0.09493092
#>  [21,]  1.255349241  0.78945281
#>  [22,] -0.069568687  0.11716888
#>  [23,]  1.275787893  0.43157033
#>  [24,] -1.045239507  0.74329370
#>  [25,] -0.389777806 -0.22540317
#>  [26,] -0.802183338  0.36131855
#>  [27,]  0.227775084  0.05161541
#>  [28,]  0.272444932  0.71621826
#>  [29,] -0.596668821 -0.51167836
#>  [30,] -0.462442089 -0.28591492
#>  [31,] -0.206921818 -0.25447483
#>  [32,] -0.211205313  0.39946917
#>  [33,] -0.335756972  0.21315534
#>  [34,] -0.192113855 -0.14420939
#>  [35,] -1.186788422  0.46016163
#>  [36,]  1.246089778  1.00559023
#>  [37,]  1.073698944  0.97808517
#>  [38,]  2.076449341  2.09868940
#>  [39,]  0.658822113 -0.63414874
#>  [40,]  1.946691173 -2.12367662
#>  [41,]  0.504986276 -0.25902985
#>  [42,] -0.459639100 -0.03592286
#>  [43,]  0.502094723  0.09072556
#>  [44,] -0.516839221  0.18223075
#>  [45,] -0.613798845  0.35971949
#>  [46,] -0.754297244  0.59894031
#>  [47,] -2.520289643 -6.57365492
#>  [48,] -0.520136331 -0.71299582
#>  [49,] -0.353295998 -0.53873288
#>  [50,] -1.155860710  0.22160733
#>  [51,]  1.167800366 -1.59968403
#>  [52,] -1.778918563 -0.07573348
#>  [53,]  0.853773388  2.10341061
#>  [54,]  0.506456452 -0.71601229
#>  [55,]  0.480077408  0.12597752
#>  [56,]  1.160298629 -0.79696568
#>  [57,] -1.035717336 -0.71810563
#>  [58,] -2.122687197 -0.44426367
#>  [59,]  0.139292719  0.18092478
#>  [60,]  1.415664732  1.18284322
#>  [61,]  0.506233299 -0.92879602
#>  [62,]  0.247316888  0.11770711
#>  [63,]  0.773142393 -0.08172910
#>  [64,]  2.042664692  3.54591297
#>  [65,]  0.654787518 -0.44403818
#>  [66,]  0.334875307  0.22766051
#>  [67,]  1.403065385 -0.22456364
#>  [68,]  0.329177031  0.41831906
#>  [69,]  1.366815835 -0.21415228
#>  [70,] -1.300557820  1.21385837
#>  [71,] -0.823919434 -0.25942946
#>  [72,]  0.643566048  1.40352905
#>  [73,] -1.471056669 -0.58280200
#>  [74,] -0.385465242 -0.18645043
#>  [75,]  1.161139411 -1.26822338
#>  [76,]  1.646721264 -0.80715504
#>  [77,] -1.699301635  2.43328730
#>  [78,]  1.148582973 -0.12945634
#>  [79,]  1.521286605 -0.18383182
#>  [80,]  1.133958075 -0.84449206
#>  [81,]  1.757966165 -1.06795146
#>  [82,]  1.573918550  0.32627846
#>  [83,] -0.494262685 -0.62936776
#>  [84,]  1.444216624 -0.98135230
#>  [85,] -1.965119345  0.99109982
#>  [86,]  0.397212091  0.21145991
#>  [87,] -2.402631904 -0.66468210
#>  [88,]  0.360975425 -0.19938912
#>  [89,]  1.081883941  0.08069768
#>  [90,] -0.556622418 -0.17274204
#>  [91,]  0.428646597  0.21068366
#>  [92,] -1.545448193  0.23244184
#>  [93,] -1.425917785  2.20038964
#>  [94,] -0.375676520 -0.31991626
#>  [95,]  1.277025735 -0.30978053
#>  [96,]  0.977453872  0.23164892
#>  [97,]  0.047127933 -0.08129607
#>  [98,]  1.519417211 -1.63184417
#>  [99,] -0.589548177  0.13392432
#> [100,] -0.263773365  0.68793553
#> [101,] -0.926093208  0.38470661
#> [102,] -1.951997865  2.75085640
#> [103,] -0.304933056  0.01011262
#> [104,]  1.051276217  0.21606018
#> [105,]  2.870734471  3.54926879
#> [106,] -0.771632480 -0.56813856
#> [107,] -1.710771053 -1.74794158
#> [108,] -0.972716060  0.85018268
#> [109,]  0.838335439  0.91585792
#> [110,]  0.077250431  0.24128231
#> [111,]  1.731556046 -0.17493532
#> [112,]  0.811598709 -0.77191216
#> [113,]  1.763768675 -1.27123921
#> [114,]  1.577841804  0.37003960
#> [115,]  1.265351330  1.08373220
#> [116,] -0.066540814  0.03980427
#> [117,] -1.844750682  0.98661397
#> [118,]  0.044987250  0.03288469
#> [119,]  2.024920168 -3.28267312
#> [120,]  0.890994077  0.52935173
#> [121,]  0.750608660 -0.99186404
#> [122,] -0.680555618 -0.27066551
#> [123,]  0.176337328  0.01757397
#> [124,] -1.097585682 -2.08886775
#> [125,]  2.068627637 -0.39581558
#> [126,]  0.528303248  0.32652763
#> [127,] -0.389319240 -0.31413353
#> [128,] -2.353910129 -0.65302939
#> [129,] -0.201521791 -0.09422912
#> [130,]  0.586385198 -0.64322687
#> [131,] -0.134241689 -0.12446038
#> [132,]  0.417168271  0.69200399
#> [133,]  0.312992029  0.12907505
#> [134,] -0.506746160 -0.71752299
#> [135,]  0.200225514 -0.28155876
#> [136,]  0.104626904 -0.07079214
#> [137,] -1.480895794 -3.30124187
#> [138,] -0.928941488  0.43559360
#> [139,]  0.470174172  0.17717565
#> [140,]  0.821046590 -0.43714067
#> [141,]  0.548062494  0.40601788
#> [142,]  1.152989470  1.15188636
#> [143,]  0.300468851 -0.03827225
#> [144,]  0.557425626 -0.64640065
#> [145,]  0.711961333 -0.35682074
#> [146,] -0.561863015 -0.65826258
#> [147,]  1.096173125 -1.39105098
#> [148,] -0.811519396 -0.66093533
#> [149,] -0.866002652  0.45287940
#> [150,]  0.353116257 -0.27470699
#> [151,] -0.753820998 -1.62763244
#> [152,] -1.807031433  0.36808659
#> [153,] -0.327336047  0.59072428
#> [154,]  2.500510083 -4.06099045
#> [155,] -0.075609907 -0.03980741
#> [156,]  1.018960767  0.02455062
#> [157,] -0.444224398  0.50481400
#> [158,] -1.150928571  0.55397037
#> [159,] -0.844571504 -0.89700330
#> [160,]  1.526844114 -0.71745243
#> [161,] -1.041019711  0.82283704
#> [162,] -1.181229906 -0.46199219
#> [163,]  2.530082554  4.11453836
#> [164,] -0.578954918 -0.81190581
#> [165,]  0.526648137  0.57744146
#> [166,] -2.516874557  0.36614841
#> [167,]  0.330929996  0.05512126
#> [168,] -0.909966725 -0.91240416
#> [169,] -0.364578501  0.19780061
#> [170,]  1.015069765  2.45245579
#> [171,]  2.286160569 -0.93371556
#> [172,] -2.020057061 -3.81230381
#> [173,] -0.435230332  0.42451179
#> [174,] -0.382645546 -0.28864159
#> [175,] -0.504011227  0.38941743
#> [176,]  0.328330818 -0.48266804
#> [177,] -0.767446408 -0.41456687
#> [178,]  0.968652569 -1.05618660
#> [179,]  0.046849389 -0.02620422
#> [180,]  1.173995471  0.46260094
#> [181,] -0.488488729  0.25848117
#> [182,]  0.189778747  0.08788589
#> [183,]  2.068989230 -1.70835625
#> [184,] -0.518359280  0.08725283
#> [185,] -0.597259851  0.33212379
#> [186,] -0.656290981  0.48531967
#> [187,] -1.222517174  0.04168632
#> [188,]  1.208682939 -0.97094371
#> [189,]  0.282997331  0.08845481
#> [190,]  0.292833623  0.49594956
#> [191,]  0.700641195 -1.00168295
#> [192,] -0.333838092  0.78353418
#> [193,] -0.949697830 -1.48129588
#> [194,] -0.600390533  0.11431334
#> [195,] -0.184540464 -0.16178602
#> [196,]  0.118661113  0.12888809
#> [197,]  0.778987371 -0.49714384
#> [198,] -0.432420235  0.31173324
#> [199,]  1.660123638 -0.79759045
#> [200,] -1.064219976 -0.37973250
#> [201,]  1.929745127 -1.07906278
#> [202,]  1.655902028  2.77337640
#> [203,]  2.290035970 -1.10603276
#> [204,] -2.530500366  0.35976035
#> [205,]  0.357248834  0.73099749
#> [206,] -0.432628207  0.12215641
#> [207,] -1.378704238  1.04808535
#> [208,]  0.065710655  0.06587606
#> [209,]  0.323357781 -0.27758250
#> [210,] -0.136264920 -0.05903523
#> [211,] -1.479778531 -4.83187003
#> [212,]  0.886911857 -0.48931943
#> [213,]  0.080191993 -0.01434750
#> [214,]  1.120094859  2.09410617
#> [215,] -0.917887294 -0.59233145
#> [216,] -1.401795708 -1.74498987
#> [217,]  1.542384331  1.02722588
#> [218,] -1.731174311  1.05175837
#> [219,] -1.815545354 -1.14761019
#> [220,] -1.260613465  0.24658093
#> [221,]  0.223060635  0.28372172
#> [222,] -0.647190331 -0.43634730
#> [223,] -2.015031462 -5.11981150
#> [224,] -1.403543868  0.69632013
#> [225,] -1.373151661 -1.27109349
#> [226,] -0.958983232  0.28736999
#> [227,] -1.492046627  1.81763338
#> [228,] -0.626417924  0.48963133
#> [229,]  0.671047230  0.05050221
#> [230,] -0.570683526 -0.08690655
#> [231,]  0.812890512  1.41364231
#> [232,] -0.726218634 -0.13047200
#> [233,] -0.649857586  0.29837120
#> [234,]  1.141917221 -0.83252776
#> [235,] -0.236639554  0.16473263
#> [236,] -0.470741586 -0.16879047
#> [237,] -0.858407742  0.17225957
#> [238,] -2.797345801  0.11137227
#> [239,] -1.634070017 -1.54398178
#> [240,] -1.299604680 -1.50249696
#> [241,] -1.203314457 -0.33805123
#> [242,] -1.798461697 -0.84371581
#> [243,]  0.637279782  0.77727075
#> [244,] -0.651253826  0.70203074
#> [245,] -0.595053590  0.68902513
#> [246,]  1.026882372  0.11072661
#> [247,]  0.060597112  0.04969919
#> [248,]  1.585049038  1.33740104
#> [249,]  0.525966502 -0.25794706
#> [250,]  0.255070432  0.36942432
#> [251,]  0.488872016 -0.49483356
#> [252,]  1.005478275 -1.14494182
#> [253,]  1.735172223 -0.25729190
#> [254,]  0.468948288  0.44568191
#> [255,] -0.786226306 -0.35669421
#> [256,] -0.188835726 -0.18052901
#> [257,] -1.591738195 -3.36345447
#> [258,]  0.340445932  0.14958061
#> [259,] -0.981822302 -1.31135196
#> [260,]  1.007744322  0.11817337
#> [261,] -1.013328904  0.07450028
#> [262,] -1.350357936 -2.13326720
#> [263,]  0.374712422 -0.23577606
#> [264,]  0.918579178 -1.09309440
#> [265,]  0.085289528  0.03364558
#> [266,]  1.701839206 -1.15774451
#> [267,] -2.197345755 -0.21953394
#> [268,] -0.280466812 -0.18662190
#> [269,] -1.161961726  0.12716989
#> [270,] -1.178943311  0.42898175
#> [271,]  1.744968819  3.17774318
#> [272,] -0.298763644  0.08592194
#> [273,]  0.097435452 -0.06150708
#> [274,]  0.941469214 -1.80535186
#> [275,]  0.541325996 -1.20237917
#> [276,]  0.995166190  0.21513259
#> [277,] -1.595219033 -0.50859138
#> [278,] -0.802784587 -1.57743622
#> [279,] -1.254116187  1.10500647
#> [280,] -2.461351651 -4.33773541
#> [281,]  0.839311560 -1.27178654
#> [282,]  0.698020929  0.19136790
#> [283,] -0.569787930  0.11247335
#> [284,] -2.777488108  5.41927968
#> [285,] -1.428182496  1.25065023
#> [286,]  2.168007820  1.17553192
#> [287,]  0.227948285 -0.19423710
#> [288,] -1.518611758 -1.73691597
#> [289,]  2.112961448  0.12861225
#> [290,]  0.405471488  0.01221797
#> [291,]  3.568261861 -1.89406503
#> [292,]  0.438482072 -0.11463942
#> [293,]  0.648251169 -0.33648218
#> [294,]  1.108711549  1.11742014
#> [295,]  0.952876371  0.64059527
#> [296,] -0.935534786 -1.14611412
#> [297,] -0.604457270  1.51502613
#> [298,]  1.496227557  0.05096085
#> [299,] -2.160024317  1.37030556
#> [300,]  1.800437180 -1.03085936
#> 
#> $inv_hessian
#>             [,1]        [,2]
#> [1,]  1.00312675 -0.05371302
#> [2,] -0.05371302  0.92271068
#>