PPI++ OLS

Helper function for PPI++ OLS estimation

Usage

ppi_plusplus_ols(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  w_l = NULL,
  w_u = NULL
)

Arguments

X_l

(matrix): n x p matrix of covariates in the labeled data.

Y_l

(vector): n-vector of labeled outcomes.

f_l

(vector): n-vector of predictions in the labeled data.

X_u

(matrix): N x p matrix of covariates in the unlabeled data.

f_u

(vector): N-vector of predictions in the unlabeled data.

lhat

(float, optional): Power-tuning parameter (see https://arxiv.org/abs/2311.01453). The default value, NULL, will estimate the optimal value from the data. Setting lhat = 1 recovers PPI with no power tuning, and setting lhat = 0 recovers the classical point estimate.

coord

(int, optional): Coordinate for which to optimize lhat = 1. If NULL, it optimizes the total variance over all coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.

w_l

(ndarray, optional): Sample weights for the labeled data set. Defaults to a vector of ones.

w_u

(ndarray, optional): Sample weights for the unlabeled data set. Defaults to a vector of ones.

Value

(list): A list containing the following:

est

(vector): vector of PPI++ OLS regression coefficient estimates.

se

(vector): vector of standard errors of the coefficients.

lambda

(float): estimated power-tuning parameter.

rectifier_est

(vector): vector of the rectifier OLS regression coefficient estimates.

var_u

(matrix): covariance matrix for the gradients in the unlabeled data.

var_l

(matrix): covariance matrix for the gradients in the labeled data.

grads

(matrix): matrix of gradients for the labeled data.

grads_hat_unlabeled

(matrix): matrix of predicted gradients for the unlabeled data.

grads_hat

(matrix): matrix of predicted gradients for the labeled data.

inv_hessian

(matrix): inverse Hessian matrix.

Details

PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023) https://arxiv.org/abs/2311.01453`

Examples

dat <- simdat(model = "ols")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set == "labeled",])

Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)

f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set == "unlabeled",])

f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)

ppi_plusplus_ols(X_l, Y_l, f_l, X_u, f_u)
#> $est
#> X(Intercept)          XX1 
#>    0.7776755    0.9294611 
#> 
#> $se
#> [1] 0.09557173 0.09476422
#> 
#> $lambda
#> [1] 0.4730261
#> 
#> $rectifier_est
#> X(Intercept)          XX1 
#>  -0.04204282  -0.06020154 
#> 
#> $var_u
#>            [,1]       [,2]
#> [1,]  0.2225795 -0.0147331
#> [2,] -0.0147331  0.2180297
#> 
#> $var_l
#>            [,1]       [,2]
#> [1,] 2.51264246 0.01134738
#> [2,] 0.01134738 2.63163397
#> 
#> $grads
#>                 [,1]          [,2]
#>   [1,]  -0.902855646  6.114056e-01
#>   [2,]  -0.471140491 -5.838350e-01
#>   [3,]   1.334473594 -7.148872e-01
#>   [4,]   2.867856294 -6.340545e-01
#>   [5,]   0.397035186 -7.634147e-01
#>   [6,]   1.494598189 -3.225684e-02
#>   [7,]  -5.585270969 -6.674540e+00
#>   [8,]   2.963988952 -2.019371e+00
#>   [9,]  -0.714560805 -6.728945e-01
#>  [10,]   0.240211143 -5.679459e-02
#>  [11,]  -4.003163670  3.798652e+00
#>  [12,]  -1.178452470 -1.205931e+00
#>  [13,]   0.919850843 -5.725607e-01
#>  [14,]   0.900253593 -3.090240e-01
#>  [15,]   1.813122247  4.119843e+00
#>  [16,]   0.474475535 -5.782018e-01
#>  [17,]  -1.485333731 -3.145548e+00
#>  [18,]   3.048689766 -1.676596e+00
#>  [19,]  -0.050076737 -1.817154e-02
#>  [20,]  -0.413663778 -2.961937e-01
#>  [21,]   0.164387567  1.206819e-01
#>  [22,]   0.132528294 -3.065350e-02
#>  [23,]  -0.306266900 -3.415679e-01
#>  [24,]  -0.582958352  1.193308e+00
#>  [25,]  -0.578757412 -1.350398e+00
#>  [26,]  -0.970943194 -2.379967e-01
#>  [27,]   1.129232265  1.541673e+00
#>  [28,]  -2.564811083 -8.294792e-02
#>  [29,]   0.850334277 -1.450537e-01
#>  [30,]  -0.249049695 -2.033830e-01
#>  [31,]  -0.529365184  4.770047e-01
#>  [32,]  -1.535823380 -1.847004e+00
#>  [33,]   0.802826785  4.490469e-01
#>  [34,]   0.521152999  4.498351e-01
#>  [35,]  -1.089756086 -1.273271e+00
#>  [36,]   0.120075745  1.113638e-01
#>  [37,]  -5.091635058  3.929724e+00
#>  [38,]  -0.938746369 -8.783234e-01
#>  [39,]   1.542996366 -1.664438e-01
#>  [40,]  -0.724955349  4.546202e-01
#>  [41,]   3.506225785 -1.337322e+00
#>  [42,]  -0.165153034  2.068035e-01
#>  [43,]   5.448204840  3.186648e+00
#>  [44,]  -0.807897328 -1.446038e+00
#>  [45,]   0.890302915  1.558333e+00
#>  [46,]   1.518464450 -1.733331e-01
#>  [47,]   2.089064459  1.977090e+00
#>  [48,]  -3.202229697  7.047847e-01
#>  [49,]  -0.087411162 -4.169204e-02
#>  [50,]  -0.224581731  2.122622e-01
#>  [51,]   1.466313001  6.264384e-01
#>  [52,]   0.928353332  3.466507e-01
#>  [53,]   1.026001293 -6.172357e-01
#>  [54,]   1.064456703  1.116203e+00
#>  [55,]   0.953967957  1.060336e+00
#>  [56,]   0.234843595 -4.548006e-02
#>  [57,]   0.355343548 -1.125259e-01
#>  [58,]  -0.700645041  8.634110e-01
#>  [59,]  -1.872994161  5.012151e-02
#>  [60,]   0.893404968 -1.749060e+00
#>  [61,]   2.499459354 -7.381661e-03
#>  [62,]  -0.815891750 -5.433177e-01
#>  [63,]  -0.185753600 -9.996290e-03
#>  [64,]   1.337114325  1.878338e+00
#>  [65,]   2.762923713  3.648273e+00
#>  [66,]  -3.882633740 -1.188419e+01
#>  [67,]   0.139427595  2.726380e-02
#>  [68,]  -6.582208422  4.652114e+00
#>  [69,]  -1.584785252 -3.660665e-01
#>  [70,]   0.054739244 -4.253634e-02
#>  [71,]   0.270391050  1.573632e-01
#>  [72,]  -0.818055046  3.521197e-01
#>  [73,]  -4.604878394 -3.722329e-01
#>  [74,]   0.772243451 -6.503191e-01
#>  [75,]   1.202724381  3.525894e-04
#>  [76,]   1.309946365  1.117065e+00
#>  [77,]   0.217210745  2.435528e-01
#>  [78,]   0.206827409 -1.110492e-01
#>  [79,]  -0.759858305 -1.143176e-01
#>  [80,]   1.099996052  7.017263e-01
#>  [81,]  -0.959213072  5.810668e-01
#>  [82,]  -0.513317062  8.467594e-02
#>  [83,]  -0.411215667 -4.498032e-02
#>  [84,]   0.795795814 -5.555848e-01
#>  [85,]  -0.478430524  1.584804e-01
#>  [86,]   0.291300456  6.155891e-01
#>  [87,]   1.575749870  2.925830e-01
#>  [88,]  -1.734828060 -1.240404e+00
#>  [89,]  -1.046070354  4.097279e+00
#>  [90,]   4.065347276 -1.449497e+00
#>  [91,]  -1.458821169 -1.522046e+00
#>  [92,]  -2.264234166 -1.909252e+00
#>  [93,]  -0.830547615  2.993846e-01
#>  [94,]  -2.164730795 -3.773986e+00
#>  [95,]   1.144447003 -1.555206e+00
#>  [96,]  -1.207963529  2.492817e-01
#>  [97,]  -0.908931880  9.879605e-01
#>  [98,]  -1.110321842 -2.544968e-01
#>  [99,]   0.597577756  2.037109e-01
#> [100,]  -1.751318684 -1.733308e+00
#> [101,]  -1.352063825  1.011172e+00
#> [102,]  -2.458400790 -1.117725e+00
#> [103,]  -0.493274954 -9.764612e-01
#> [104,]   0.766002107 -1.833428e-01
#> [105,]  -2.436202420 -2.267918e+00
#> [106,]   1.932812537  3.160919e-01
#> [107,]   1.367367955 -5.135052e-01
#> [108,]   0.849387853  6.851620e-01
#> [109,]   3.102519976  2.121882e-02
#> [110,]  -3.587446827  3.499459e+00
#> [111,]   1.501553569  2.406762e-01
#> [112,]   0.586232212 -3.158039e-01
#> [113,]   0.246172499 -1.014369e-02
#> [114,]   1.926025583  5.638783e-01
#> [115,]  -0.394500752  3.163908e-02
#> [116,]  -0.242960798  2.292559e-01
#> [117,]   0.992873479  1.640902e+00
#> [118,]  -1.695143617 -1.088487e-02
#> [119,]   0.722759480  6.401210e-01
#> [120,]   0.412572261 -2.091365e-01
#> [121,]  -2.296849300  1.713466e+00
#> [122,]  -1.074539270  8.137506e-01
#> [123,] -10.060151997  3.453647e-01
#> [124,]  -1.145498232 -2.243211e-02
#> [125,]  -1.311899433 -4.479359e-01
#> [126,]   1.214853535 -2.755551e-01
#> [127,]   0.182840346  2.078358e-01
#> [128,]  -1.280761934 -2.532955e+00
#> [129,]  -0.178419582  1.128394e-01
#> [130,]  -1.214949349  1.934967e+00
#> [131,]   0.473836737  5.972697e-02
#> [132,]  -0.465493272  6.082372e-01
#> [133,]   2.106392326  1.083241e+00
#> [134,]  -1.737478908 -6.645705e-01
#> [135,]   1.261401952 -5.146168e-01
#> [136,]   0.373328492  2.644960e-01
#> [137,]   2.114486038 -4.459456e-01
#> [138,]  -1.582021700  1.265509e+00
#> [139,]   0.589229660  5.626898e-01
#> [140,]  -1.520724765  1.602371e+00
#> [141,]  -0.131139409  8.705297e-02
#> [142,]   0.322532521 -1.304421e-01
#> [143,]   0.106171080 -1.571445e-01
#> [144,]  -0.534266416  3.684140e-01
#> [145,]  -1.983622454 -1.414859e-01
#> [146,]   0.647067003  1.885951e+00
#> [147,]  -1.318926213  1.203926e+00
#> [148,]  -0.691388866 -3.223849e-01
#> [149,]  -1.318118914  3.510479e-01
#> [150,]  -0.510830188  2.408151e-01
#> [151,]   0.709432991  1.271117e+00
#> [152,]   0.026341719 -2.843384e-02
#> [153,]  -0.162356913  2.450980e-02
#> [154,]   1.533650938 -1.427742e+00
#> [155,]  -1.137724100  1.907210e+00
#> [156,]   1.235019859  4.795421e-01
#> [157,]  -0.090294271 -5.280771e-03
#> [158,]   0.271695382  2.119037e-01
#> [159,]  -1.388028207  9.292247e-01
#> [160,]  -0.037623414  5.905377e-02
#> [161,]  -0.228425097  3.154660e-01
#> [162,]  -1.260956537 -5.486485e-02
#> [163,]  -1.035378380  1.333074e-01
#> [164,]   1.064513386 -1.999618e+00
#> [165,]   1.615139614 -1.084106e+00
#> [166,]  -0.845952520  8.216963e-02
#> [167,]   3.161350476  1.989463e+00
#> [168,]  -0.591709289  8.010457e-02
#> [169,]   2.024812379 -2.927035e+00
#> [170,]  -0.128267168  4.188740e-03
#> [171,]   1.158329383  6.480025e-01
#> [172,]   1.798016849  8.082848e-01
#> [173,]  -2.567564708 -9.284915e-02
#> [174,]  -0.808760930 -1.367525e+00
#> [175,]   0.432116439  5.387745e-01
#> [176,]   1.737886828 -7.857079e-02
#> [177,]   1.239439204 -5.932220e-01
#> [178,]   1.062766903 -7.795114e-01
#> [179,]   1.384349177  2.626654e+00
#> [180,]   0.658511365  7.290400e-01
#> [181,]  -0.079112044  8.177798e-02
#> [182,]   7.882839136  2.496855e+00
#> [183,]   1.418716693 -5.207099e-01
#> [184,]  -0.795622618  4.132705e-01
#> [185,]   1.648978442  5.385378e-01
#> [186,]  -2.858341080  2.647462e+00
#> [187,]  -0.005285383 -7.423771e-03
#> [188,]  -0.035396302  2.028736e-03
#> [189,]   1.431731254  1.538579e-01
#> [190,]   1.182456181 -1.430551e-01
#> [191,]  -2.814355308  1.686411e+00
#> [192,]   0.956925798  8.746875e-03
#> [193,]  -0.742854339  8.716333e-01
#> [194,]  -0.419370871 -4.583058e-01
#> [195,]  -0.220933655  3.367091e-01
#> [196,]   0.509965964  1.920087e-01
#> [197,]  -0.648229235  4.712979e-01
#> [198,]   0.597307804 -1.059883e+00
#> [199,]  -0.400331075  2.422264e-01
#> [200,]   0.357128552  2.472662e-01
#> [201,]  -0.141909054  5.955277e-02
#> [202,]  -1.496274215  6.620693e-01
#> [203,]  -0.346584023 -1.716995e-01
#> [204,]  -0.469908950 -8.926937e-02
#> [205,]   1.614568576 -4.888635e-01
#> [206,]   0.815853510  7.471921e-01
#> [207,]  -0.890523585 -6.704737e-01
#> [208,]   1.088811535 -5.624024e-01
#> [209,]  -0.588470110 -1.275980e-01
#> [210,]   1.796278025 -1.474258e+00
#> [211,]  -0.563197540  4.513137e-01
#> [212,]   0.961549386  6.678526e-01
#> [213,]   0.483561106 -5.637917e-02
#> [214,]   2.331417839  2.604374e+00
#> [215,]   0.496715361  2.629518e-02
#> [216,]  -2.538555701  4.671533e+00
#> [217,]   0.046069261  6.106023e-04
#> [218,]  -0.208244272 -6.501683e-01
#> [219,]   0.858410905 -9.640494e-01
#> [220,]  -0.068538123  4.100061e-03
#> [221,]  -0.555096883 -2.389209e-01
#> [222,]  -1.588888023  5.122022e+00
#> [223,]   1.064968616  2.255056e-02
#> [224,]  -1.792738521  5.349271e-01
#> [225,]  -0.528683202  4.664409e-01
#> [226,]   0.590887706 -9.844975e-01
#> [227,]  -0.502610606  3.714781e-01
#> [228,]   0.401183518 -1.560076e-01
#> [229,]  -0.407319312  1.567209e-01
#> [230,]   1.004534333  2.012632e+00
#> [231,]  -1.239798147 -2.550253e+00
#> [232,]   2.470972072  1.811256e+00
#> [233,]  -0.886598344 -5.012420e-01
#> [234,]   0.433799281 -3.848927e-02
#> [235,]  -0.687731293 -1.690562e+00
#> [236,]  -1.681429354 -7.755083e-01
#> [237,]  -0.613024317  5.501610e-01
#> [238,]   0.278102732 -1.136934e-01
#> [239,]  -0.331595870  6.240896e-02
#> [240,]   0.070089152 -4.074931e-02
#> [241,]   0.918273859  7.390863e-01
#> [242,]   0.293620655 -1.392897e-01
#> [243,]  -1.300015176  8.534762e-01
#> [244,]  -0.734091256 -7.360349e-01
#> [245,]   0.774623620  8.580835e-02
#> [246,]   0.386210350 -9.145060e-01
#> [247,]   2.440660562  1.549569e+00
#> [248,]  -1.555079642  1.297047e+00
#> [249,]   0.591716994  2.761750e-01
#> [250,]   2.197957992  1.442009e+00
#> [251,]   0.298507243  6.765830e-01
#> [252,]   1.779672528 -6.307539e-01
#> [253,]  -1.534926300  1.514100e+00
#> [254,]   0.357124847  9.138021e-02
#> [255,]   0.949216834  4.916260e-02
#> [256,]  -0.941107282  3.777974e-01
#> [257,]   0.658349743  7.610657e-01
#> [258,]   0.726983338 -1.594745e+00
#> [259,]  -0.197180528  6.788761e-02
#> [260,]   7.843581715 -1.653019e+01
#> [261,]  -0.698899695 -2.587408e-01
#> [262,]  -0.452460392  7.101409e-02
#> [263,]   1.571443591 -6.944026e-01
#> [264,]  -0.034673430 -3.270216e-02
#> [265,]   0.340020656  4.565734e-01
#> [266,]   0.409194876  4.765148e-01
#> [267,]   2.519812409 -1.658724e+00
#> [268,]   0.898442863 -2.264291e-01
#> [269,]  -0.851508191 -1.209455e-01
#> [270,]  -0.194081612  1.524446e-01
#> [271,]   0.513098691 -1.160589e+00
#> [272,]   1.350169047 -4.959419e-01
#> [273,]  -6.108175411 -3.305738e+00
#> [274,]   0.292333382 -1.056131e-01
#> [275,]   1.733291931  5.987292e-01
#> [276,]   0.453375876  4.188619e-01
#> [277,]   1.943581738 -9.860094e-01
#> [278,]   0.325253134  6.649674e-02
#> [279,]   0.588185841 -1.005581e-01
#> [280,]  -2.018163152 -2.466100e+00
#> [281,]  -0.501816669 -1.167445e+00
#> [282,]  -3.325995513 -7.287768e+00
#> [283,]   1.521982886  1.164055e+00
#> [284,]  -1.297804989  1.188212e+00
#> [285,]  -1.242609674 -1.771420e+00
#> [286,]   0.442731029 -9.521087e-01
#> [287,]  -3.353409256  7.100244e-02
#> [288,]   1.356627538  2.028155e+00
#> [289,]   1.321425494 -1.176106e+00
#> [290,]   0.335507283  1.624104e-01
#> [291,]  -3.684518760  2.071452e+00
#> [292,]   1.083138166 -4.494018e-01
#> [293,]   0.521641327  5.855491e-01
#> [294,]  -2.090614849  2.061038e-01
#> [295,]   1.549294891 -1.995358e+00
#> [296,]  -0.181185623 -2.338699e-01
#> [297,]   2.691334743 -3.288095e-01
#> [298,]  -1.541383322  1.271061e+00
#> [299,]   0.149813406 -9.712114e-02
#> [300,]   1.644469131 -5.002835e-01
#> 
#> $grads_hat_unlabeled
#>                 [,1]          [,2]
#>   [1,] -0.8675342631 -0.9686230172
#>   [2,] -1.2957332435  0.7818620340
#>   [3,] -1.0304206594  0.4936779847
#>   [4,] -1.9868688839  1.4965345955
#>   [5,] -1.0171596979 -0.2241327525
#>   [6,]  1.1404579897  0.5192694486
#>   [7,] -1.0197621678  0.7693006133
#>   [8,]  0.3302323415 -0.1498204394
#>   [9,] -1.0890407726  0.0105939339
#>  [10,] -1.6912975042 -2.6056799239
#>  [11,] -0.4395947274  0.2360411457
#>  [12,]  0.5023405650  0.5044115745
#>  [13,] -0.1547736872  0.4046551658
#>  [14,] -1.0205959804 -0.3413033443
#>  [15,]  0.8428515923 -0.1001647552
#>  [16,] -0.1290160769 -0.1243457197
#>  [17,]  0.5060008539 -0.9133988345
#>  [18,] -0.0308389439  0.0592089051
#>  [19,]  0.2495967742 -0.0247230112
#>  [20,]  0.5150288137  0.8802526506
#>  [21,] -0.9948711301  0.9261749560
#>  [22,] -0.3911628155 -0.0547370877
#>  [23,] -0.8779790040  0.7146835330
#>  [24,]  1.4066692528 -1.3759124531
#>  [25,]  0.6598975615 -1.1975854378
#>  [26,] -0.2922826142  0.5521419552
#>  [27,] -0.4171175194 -0.1555246716
#>  [28,]  2.1148629014 -0.3708061950
#>  [29,] -0.7798811853 -0.2556563525
#>  [30,]  0.3474902518  0.4276980385
#>  [31,]  2.3864085878  0.0467669161
#>  [32,]  0.3013450635 -0.0819212039
#>  [33,]  0.5925285749 -0.2211674504
#>  [34,]  0.1232730291 -0.0113759106
#>  [35,]  0.8987552310  0.6163306959
#>  [36,] -0.5272937253 -1.0459806161
#>  [37,] -0.2745193682  0.0210678545
#>  [38,] -2.2730201500  2.3019068802
#>  [39,]  0.6964339204  1.4441252771
#>  [40,] -1.2942539320  3.3752883215
#>  [41,] -0.5883911949 -0.3848943411
#>  [42,]  0.5914978522  0.6081355285
#>  [43,]  0.6173619298 -0.5468516032
#>  [44,]  1.1737445295 -1.6569798979
#>  [45,] -0.0012962263 -0.0006466135
#>  [46,] -0.9213146683  0.4387590925
#>  [47,]  1.3916163588  2.6318027356
#>  [48,]  0.6415686263  1.1482772795
#>  [49,] -0.5162909939 -0.2643708024
#>  [50,]  0.7901610091 -1.6315275805
#>  [51,]  1.9084487392  0.1680172099
#>  [52,]  0.5880142093  0.1866458830
#>  [53,]  0.0331540458  0.0110824855
#>  [54,]  0.7749500217  0.4281007616
#>  [55,]  0.4019433666 -0.1013582359
#>  [56,] -1.8942008326  2.9644628421
#>  [57,]  1.4721384850 -1.2341184411
#>  [58,] -0.3360462575 -0.2727253562
#>  [59,] -0.0914363086 -0.0692510802
#>  [60,] -0.6411116470  0.1134234439
#>  [61,]  1.0199693469  0.6868291822
#>  [62,]  0.5302811101 -0.7522253699
#>  [63,] -0.5511422835  0.2221860976
#>  [64,] -0.0686995450 -0.0253595204
#>  [65,] -0.2230915752  0.1571853929
#>  [66,] -0.1760126713  0.1338573903
#>  [67,] -1.8172042922 -1.0359557051
#>  [68,] -1.2191127821  1.4743395971
#>  [69,] -1.1509330792  0.7765056348
#>  [70,]  0.1705749123 -0.0516211074
#>  [71,]  1.0054220331  0.8719811092
#>  [72,]  0.3849921270  0.3561605608
#>  [73,] -0.1660412110 -0.0049316583
#>  [74,] -0.1429286769 -0.0292546723
#>  [75,]  1.1724965407  0.1070567709
#>  [76,]  0.7382213801  0.1273707515
#>  [77,]  1.1668556304  0.8850479313
#>  [78,] -0.1443547173  0.1539264437
#>  [79,] -0.7792626292 -0.1545770734
#>  [80,] -2.2981026298 -0.0532718122
#>  [81,]  1.3315059073  1.4515543567
#>  [82,]  0.2202515579  0.1072703212
#>  [83,] -0.0689663876 -0.0562356255
#>  [84,]  0.6495382202 -0.6860667598
#>  [85,]  0.4884122461 -0.7377501206
#>  [86,] -0.0550358962 -0.0056382460
#>  [87,] -0.0572679246  0.0013516510
#>  [88,]  0.5030590248  0.1243926928
#>  [89,]  0.7900721055 -0.7403218161
#>  [90,]  0.8887268861 -0.8584115970
#>  [91,]  1.9157336754 -2.8598951406
#>  [92,]  0.4756976713 -0.3098289695
#>  [93,]  0.3667851206 -0.2495069506
#>  [94,]  0.7940098254  0.7364083192
#>  [95,] -0.3305076750 -0.4395349194
#>  [96,]  0.4253550317  0.8512109024
#>  [97,]  0.7243577783 -0.7754312326
#>  [98,] -0.6165974490  0.3778988819
#>  [99,] -0.6147477690  0.8461177563
#> [100,] -0.4672910887  0.0183979717
#> [101,]  0.0277245373  0.0127532500
#> [102,]  1.0193787990  0.9047886769
#> [103,]  0.2382798673 -0.1617229808
#> [104,]  2.1402251877 -0.3127023468
#> [105,] -0.2030367732  0.2873773998
#> [106,] -0.5919115084 -0.0178671006
#> [107,] -0.6820095384  1.2130275847
#> [108,] -0.2391175100  0.0210517867
#> [109,]  0.3760791759  0.3296716428
#> [110,] -0.6098199526 -0.1570861591
#> [111,]  1.0851475911  0.6685958408
#> [112,]  0.0006791731 -0.0005416342
#> [113,] -1.4176454052  0.4851723168
#> [114,]  0.4364186625 -0.5000442081
#> [115,] -1.7394765035  1.8839426421
#> [116,]  1.4686830261 -0.7749904809
#> [117,] -1.0113399602 -0.5766349550
#> [118,]  1.2287495702 -0.8884747540
#> [119,] -1.9471397945 -0.5448994702
#> [120,]  0.9327058769 -0.3792313147
#> [121,] -1.1786849396 -0.7741483421
#> [122,] -0.6149742348  0.3162140930
#> [123,]  0.7562573526 -0.8059829698
#> [124,] -0.7908112249  0.6757954372
#> [125,] -1.2229367920  0.6884049455
#> [126,] -0.0934273926 -0.0424022151
#> [127,] -0.4042595530  0.0805428260
#> [128,] -0.5528717049 -0.3661707411
#> [129,]  0.3550591199 -0.5548568501
#> [130,] -1.4694859111  1.6944833146
#> [131,]  0.4314376705  0.2601025482
#> [132,] -1.1386027821 -0.6664524235
#> [133,]  0.7624736130  0.3779251765
#> [134,] -0.4823692111  1.0282426782
#> [135,]  0.3907375919 -0.3169464439
#> [136,]  1.5160774373  1.1787008497
#> [137,] -0.9311969142 -0.7543877173
#> [138,] -0.5822413877 -0.4628925178
#> [139,] -0.4671334925  0.0631902263
#> [140,] -0.4455079016 -0.2504169860
#> [141,] -0.6395210537 -0.5188661928
#> [142,]  0.7647449095  0.8202432289
#> [143,]  0.3599114222  0.0230298086
#> [144,] -0.4625211209  0.8703168687
#> [145,]  0.8426839593 -0.4474971058
#> [146,]  0.0589944784 -0.0824857096
#> [147,] -0.5771116435  0.0139074365
#> [148,]  0.3676640233 -0.3490778311
#> [149,] -1.2092638486 -0.6675817674
#> [150,] -1.5045145844 -0.7829160745
#> [151,] -1.3297441789  1.0544509259
#> [152,]  0.3752144365 -0.0473575591
#> [153,] -0.3813120840 -0.0184272637
#> [154,]  0.5398834132  0.1762945780
#> [155,] -1.2373791596  1.9103758287
#> [156,]  1.1266339265 -0.8962629564
#> [157,]  0.2296688032 -0.2725145990
#> [158,]  1.5754620460 -0.3123760361
#> [159,] -0.1213234474 -0.2312492021
#> [160,]  1.4235390976 -0.2948560401
#> [161,] -0.5518196108  0.7118245536
#> [162,] -0.1296523972  0.0593901374
#> [163,] -0.0987362827  0.0615276115
#> [164,]  2.0770946112 -4.8567494242
#> [165,]  0.9867326132 -0.1560091118
#> [166,]  1.6896438738 -2.2122034547
#> [167,]  1.2640771840  1.0357432859
#> [168,]  0.4421008216 -0.3228228594
#> [169,] -0.3669735687  0.2552670360
#> [170,] -0.1495310413  0.1514943195
#> [171,] -0.5339256247 -1.5157825474
#> [172,] -2.0001471724 -1.8758430479
#> [173,]  0.9748624959 -1.2867044736
#> [174,]  0.7851278186 -0.9245353464
#> [175,]  2.4917270175 -0.0826547509
#> [176,] -1.5748334366  1.2146706793
#> [177,]  0.2146664319  0.1517691248
#> [178,]  0.4171004638  0.7426898706
#> [179,]  0.0903058441 -0.0567387467
#> [180,]  2.2098494960 -1.4421654264
#> [181,] -1.3972653377  0.3949333176
#> [182,]  0.3659288014  0.1504689617
#> [183,] -1.0921788563 -1.0231927992
#> [184,]  1.6469495300  0.0859799333
#> [185,] -1.3098085971 -2.0367417462
#> [186,]  1.5847532442  1.3545542785
#> [187,] -0.7415401676  0.4656585622
#> [188,] -1.6240947337 -2.6423593442
#> [189,]  0.0070095670  0.0066656993
#> [190,] -1.4210025399 -0.7344665870
#> [191,] -0.8326939771 -0.2140479621
#> [192,]  0.8348540021  1.1135753670
#> [193,] -0.1122077495  0.0721504469
#> [194,] -1.5683885677 -2.0160742443
#> [195,]  0.6565565051 -0.4684988714
#> [196,] -1.0681758175  0.2324488493
#> [197,]  0.6068840390 -0.6209524080
#> [198,]  0.0750447955 -0.0171648761
#> [199,] -0.2801150869 -0.5521190629
#> [200,]  0.5405768681  0.7635137446
#> [201,] -0.5467012960 -0.4408550026
#> [202,] -0.8475315601 -1.1501314895
#> [203,]  0.3943677948 -0.0448688851
#> [204,]  0.4631888879  0.0018583084
#> [205,] -0.8446527005  1.9856260230
#> [206,]  0.2863805070  0.1270153112
#> [207,] -1.8295719822  1.4536425122
#> [208,] -2.1177922602  1.4032025693
#> [209,] -1.4713935950  0.0969689241
#> [210,] -0.7357805987  1.1038831779
#> [211,]  0.3772950768 -0.1704168102
#> [212,] -0.7391828056 -1.5219812437
#> [213,] -1.0375942582 -0.2332250915
#> [214,] -0.0484508703 -0.0320042507
#> [215,]  0.7209617663  1.2291720618
#> [216,] -3.0963084752 -4.3833427577
#> [217,]  0.5670717020  1.0722932106
#> [218,]  1.0774672487  0.7756675291
#> [219,] -1.9065544592  0.9760957501
#> [220,]  0.4926765975 -0.3376240290
#> [221,]  1.6431804204  1.3614942618
#> [222,] -0.6522615376 -0.3666569011
#> [223,]  0.5990687788 -0.7159199492
#> [224,]  1.1504102983 -2.1291874368
#> [225,] -0.1134315639  0.0497249362
#> [226,] -0.2265217800  0.0390819791
#> [227,] -0.6600287506 -0.2573924874
#> [228,] -0.2985356287  0.2432588341
#> [229,]  0.1299461109 -0.0176272207
#> [230,] -0.8374841422 -0.3906145632
#> [231,]  0.3193586935 -0.5057024898
#> [232,]  0.6770847256  0.7179541987
#> [233,] -1.0872787345 -0.2801846306
#> [234,] -0.0355280103  0.0268802671
#> [235,]  0.6486705151  0.6384211446
#> [236,] -1.3935051273  1.0956841482
#> [237,] -0.9499862587 -0.3267061741
#> [238,] -0.6628232472  0.2609415831
#> [239,]  0.0075626078  0.0049522254
#> [240,] -1.8492289620 -2.8578587519
#> [241,]  0.2261759373 -0.0967110924
#> [242,]  1.1966750980  0.5769139703
#> [243,]  1.1582664195 -0.4681848035
#> [244,] -1.0002482404  0.6833589528
#> [245,] -0.8551857375  0.1427084082
#> [246,] -0.8271689920  0.4603922573
#> [247,] -0.5624282740  0.5764987638
#> [248,] -1.0348725003 -1.8652888487
#> [249,]  1.2444190716  0.9179392611
#> [250,]  0.9707473107 -0.8929448533
#> [251,]  0.5675493431 -0.3645164671
#> [252,] -1.0496990067 -1.0534659875
#> [253,] -0.2905085677 -0.1198758057
#> [254,]  1.4362418477 -2.2023052969
#> [255,] -0.8249718053  1.3081838097
#> [256,] -1.7823415027  1.1057302444
#> [257,]  1.2641098337 -0.3298916079
#> [258,]  0.8982870177 -1.2068332382
#> [259,]  1.1232180324 -1.2559449995
#> [260,] -0.6524129007 -0.6855133451
#> [261,]  1.0748328735 -1.0226328512
#> [262,] -0.3882311110 -0.4693730042
#> [263,] -0.6163929196  0.7883836929
#> [264,] -0.7701061366 -0.2192475229
#> [265,] -0.1087859106 -0.0630019099
#> [266,]  1.4340788671 -1.4055702415
#> [267,]  0.3909407315 -0.1320247380
#> [268,] -0.1013132450 -0.0188188221
#> [269,] -1.6302032593  1.1218597259
#> [270,] -0.1756609878  0.3456772090
#> [271,]  0.5680824062 -1.0238324317
#> [272,] -2.7394311084 -0.0629648767
#> [273,] -0.1458463995  0.0310500042
#> [274,]  1.0042242731 -0.2788581010
#> [275,] -0.8931992161  0.2437078228
#> [276,]  2.3230454665  3.7385996882
#> [277,] -0.9053023551  0.0233845944
#> [278,]  0.5261836527 -0.8788280465
#> [279,]  0.0267943682  0.0271839002
#> [280,] -1.8680882668 -2.0869227138
#> [281,] -0.6125281040  0.7776852493
#> [282,]  1.5484303659  0.3796657526
#> [283,] -1.2838321979 -0.9110610686
#> [284,]  0.2955735566 -0.2289537564
#> [285,] -0.2378363212 -0.2491129612
#> [286,]  0.2288771638  0.0710035669
#> [287,]  0.8805354511  0.4934623342
#> [288,] -0.2503908325  0.3228304955
#> [289,]  0.8074080115  0.0650495015
#> [290,] -0.4243312848  0.1362060062
#> [291,]  1.6984032253 -2.4960125107
#> [292,]  0.3749012602 -0.6290526787
#> [293,]  0.0129321873  0.0115376284
#> [294,] -0.1797962613 -0.0510970867
#> [295,] -0.6124159360  0.4004025666
#> [296,]  0.2088770280  0.2238162087
#> [297,] -0.4645496967  0.0323739696
#> [298,]  1.0473562201  1.3250554830
#> [299,]  0.1369001671  0.1797925444
#> [300,]  1.2240353813  1.0617996289
#> 
#> $grads_hat
#>                 [,1]          [,2]
#>   [1,]  0.5519993207 -0.3738089220
#>   [2,] -0.0633674197 -0.0785245950
#>   [3,]  0.7286730495 -0.3903554720
#>   [4,] -0.1939572162  0.0428820137
#>   [5,]  1.5217286359 -2.9259624739
#>   [6,] -0.0755429150  0.0016303883
#>   [7,] -2.5103837457 -2.9999721572
#>   [8,] -0.2121276928  0.1445229713
#>   [9,]  0.6219995703  0.5857305857
#>  [10,]  0.7227507512 -0.1708843681
#>  [11,] -2.4399618077  2.3153101695
#>  [12,]  1.2711524405  1.3007922002
#>  [13,] -0.7253895916  0.4515183968
#>  [14,]  1.0943046563 -0.3756345460
#>  [15,] -0.0336895963 -0.0765507451
#>  [16,]  0.7937916532 -0.9673243254
#>  [17,]  0.8890136134  1.8826982379
#>  [18,]  0.8575496646 -0.4716006806
#>  [19,] -0.6711273447 -0.2435346436
#>  [20,] -1.0992502274 -0.7870908492
#>  [21,] -1.9250916744 -1.4132683492
#>  [22,] -0.0068973386  0.0015953393
#>  [23,] -1.1901165799 -1.3272921607
#>  [24,] -0.4937823188  1.0107656486
#>  [25,] -0.4786953779 -1.1169265044
#>  [26,]  0.5975722116  0.1464763741
#>  [27,] -0.2704358396 -0.3692097071
#>  [28,] -1.8738349863 -0.0606012336
#>  [29,]  1.6153536486 -0.2755540368
#>  [30,] -1.1568127508 -0.9446952725
#>  [31,]  0.3149395908 -0.2837883322
#>  [32,] -1.7034236663 -2.0485624082
#>  [33,]  1.3661322685  0.7641217572
#>  [34,] -0.1819292012 -0.1570328467
#>  [35,]  0.4804524661  0.5613606903
#>  [36,]  0.4362879866  0.4046337552
#>  [37,] -1.9385875552  1.4962018920
#>  [38,]  0.6545731014  0.6124411422
#>  [39,] -0.5415486130  0.0584171166
#>  [40,] -0.4685902687  0.2938534353
#>  [41,]  2.0229225302 -0.7715702930
#>  [42,]  0.6226033322 -0.7796196807
#>  [43,]  1.9684359764  1.1513358264
#>  [44,]  0.1957863253  0.3504337344
#>  [45,]  0.6639520832  1.1621419670
#>  [46,]  0.3648498323 -0.0416476962
#>  [47,] -1.0498120311 -0.9935415818
#>  [48,]  0.5583753083 -0.1228938696
#>  [49,] -0.8192256459 -0.3907416933
#>  [50,] -1.1274077200  1.0655631580
#>  [51,] -0.2992184225 -0.1278321305
#>  [52,]  0.5852582902  0.2185377231
#>  [53,] -1.3906497872  0.8366058021
#>  [54,]  0.1046711074  0.1097594487
#>  [55,] -0.8047446058 -0.8944741589
#>  [56,]  0.6783700559 -0.1313738743
#>  [57,] -0.5938301948  0.1880469862
#>  [58,] -0.0811560252  0.1000092767
#>  [59,] -1.1595428245  0.0310294834
#>  [60,] -0.2614206480  0.5117951681
#>  [61,]  0.2079760552 -0.0006142163
#>  [62,] -0.8184110693 -0.5449953310
#>  [63,] -0.3319045644 -0.0178613730
#>  [64,]  1.3800407574  1.9386399491
#>  [65,]  0.1108469243  0.1463666285
#>  [66,]  0.2757549941  0.8440467157
#>  [67,] -0.9138612548 -0.1786972685
#>  [68,] -2.5200910231  1.7811272583
#>  [69,] -0.0571795113 -0.0132077862
#>  [70,] -0.1296708742  0.1007636106
#>  [71,]  0.2270589202  0.1321445920
#>  [72,]  0.3288734189 -0.1415587197
#>  [73,] -2.2619918374 -0.1828469070
#>  [74,]  0.5925363750 -0.4989847475
#>  [75,]  1.4573787006  0.0004272436
#>  [76,]  1.2235928101  1.0434266697
#>  [77,] -0.6870168298 -0.7703343644
#>  [78,]  0.5496727514 -0.2951288769
#>  [79,] -0.3835663345 -0.0577059965
#>  [80,]  0.0643343417  0.0410411448
#>  [81,] -0.6401052975  0.3877594540
#>  [82,]  0.3443223480 -0.0567988480
#>  [83,]  0.0718487692  0.0078590900
#>  [84,] -1.0215939378  0.7132257352
#>  [85,] -1.3940749064  0.4617881797
#>  [86,] -0.5525545553 -1.1676829592
#>  [87,] -0.3150129814 -0.0584911701
#>  [88,] -0.0178670023 -0.0127749304
#>  [89,] -0.7465092012  2.9239490467
#>  [90,]  2.3409019937 -0.8346470144
#>  [91,]  0.4605648680  0.4805257178
#>  [92,] -1.5143362372 -1.2769214247
#>  [93,]  0.8115057335 -0.2925206324
#>  [94,] -1.7939744306 -3.1276103302
#>  [95,]  0.3302204679 -0.4487413590
#>  [96,] -0.2452012039  0.0506010161
#>  [97,] -1.5605388328  1.6962226069
#>  [98,] -0.4247636768 -0.0973600548
#>  [99,]  0.0357915000  0.0122011204
#> [100,] -2.1164917086 -2.0947256163
#> [101,]  0.4026848599 -0.3011571837
#> [102,] -1.4714922522 -0.6690220730
#> [103,] -1.4815151563 -2.9327295129
#> [104,] -0.5885808399  0.1408769917
#> [105,]  0.1150658195  0.1071174655
#> [106,]  1.6492600116  0.2697197578
#> [107,]  1.2317557530 -0.4625770222
#> [108,]  0.8436040322  0.6804964433
#> [109,]  1.1572912544  0.0079149724
#> [110,] -2.0000108484  1.9509575884
#> [111,] -0.2401369281 -0.0384902915
#> [112,]  0.2762000610 -0.1487892653
#> [113,] -1.1546918186  0.0475797846
#> [114,]  0.5545552702  0.1623559218
#> [115,] -0.0898088576  0.0072026985
#> [116,]  0.6334987466 -0.5977643624
#> [117,] -0.8347114654 -1.3795106329
#> [118,] -1.1928305011 -0.0076594110
#> [119,]  1.0488808347  0.9289544398
#> [120,]  0.4328386136 -0.2194096551
#> [121,] -1.6530125325  1.2331593456
#> [122,]  1.1034124570 -0.8356163682
#> [123,] -2.8699727516  0.0985260843
#> [124,] -0.2664784107 -0.0052184052
#> [125,]  0.2047381384  0.0699059406
#> [126,] -0.5873460093  0.1332228206
#> [127,] -0.1602317251 -0.1821364190
#> [128,] -0.5019129415 -0.9926303002
#> [129,]  0.3727796456 -0.2357601896
#> [130,]  0.6179750989 -0.9842066742
#> [131,]  0.4353456015  0.0548751727
#> [132,] -0.0596795843  0.0779803815
#> [133,] -0.3088160853 -0.1588128332
#> [134,] -0.4208083756 -0.1609555174
#> [135,] -0.0148432246  0.0060556220
#> [136,] -0.1562168395 -0.1106765910
#> [137,]  1.4531707913 -0.3064740739
#> [138,]  0.6755382404 -0.5403842547
#> [139,]  0.2483806409  0.2371931556
#> [140,]  0.6032909202 -0.6356812047
#> [141,] -0.5004686829  0.3322211328
#> [142,]  1.4906966365 -0.6028837966
#> [143,]  0.9357818307 -1.3850570565
#> [144,]  0.4545160389 -0.3134205260
#> [145,] -1.2643063234 -0.0901792350
#> [146,] -0.2975296057 -0.8671843791
#> [147,]  1.0860796941 -0.9913821257
#> [148,] -0.5157345667 -0.2404797929
#> [149,] -0.1104686239  0.0294205453
#> [150,] -0.3406197299  0.1605746418
#> [151,]  0.9490580243  1.7004624844
#> [152,] -0.0710304114  0.0766718143
#> [153,]  0.4537591090 -0.0685005826
#> [154,]  0.3183186554 -0.2963365537
#> [155,] -0.6425247766  1.0770886803
#> [156,] -0.6681540345 -0.2594355159
#> [157,]  0.7018590598  0.0410475318
#> [158,]  0.5660155336  0.4414531659
#> [159,] -0.8561294790  0.5731415528
#> [160,]  0.1923586148 -0.3019263643
#> [161,] -1.1588864719  1.6004778004
#> [162,]  0.0640222404  0.0027856397
#> [163,]  0.0298481377 -0.0038430191
#> [164,] -0.2910737183  0.5467628308
#> [165,]  0.8573978819 -0.5754983109
#> [166,] -0.7129925642  0.0692548775
#> [167,] -0.2809337337 -0.1767938369
#> [168,] -0.2226733713  0.0301451318
#> [169,] -0.0835361127  0.1207584285
#> [170,]  0.4514561620 -0.0147429175
#> [171,]  0.5010751475  0.2803157389
#> [172,] -0.2497552803 -0.1122755905
#> [173,] -0.7329551687 -0.0265053752
#> [174,] -0.9930393482 -1.6791189990
#> [175,]  0.8963572033  1.1176024491
#> [176,] -0.2743495899  0.0124034910
#> [177,]  0.1849479164 -0.0885200144
#> [178,] -0.7800607284  0.5721538929
#> [179,]  0.1274888416  0.2418963879
#> [180,]  0.6777453418  0.7503339971
#> [181,]  1.3349968979 -1.3799839864
#> [182,]  2.7555987078  0.8728240115
#> [183,]  1.8551150164 -0.6808806722
#> [184,]  0.0615295468 -0.0319603133
#> [185,]  0.0003968346  0.0001296017
#> [186,] -1.7197313881  1.5928554879
#> [187,]  0.2302137558  0.3233548177
#> [188,] -0.1000678510  0.0057353793
#> [189,] -0.8176964115 -0.0878719871
#> [190,] -0.7512654440  0.0908890767
#> [191,] -2.0038900865  1.2007660012
#> [192,]  1.4302104714  0.0130729797
#> [193,] -1.4317162886  1.6799141885
#> [194,] -0.8269082290 -0.9036795147
#> [195,]  0.5636085191 -0.8589552278
#> [196,]  1.0947013702  0.4121690106
#> [197,] -0.4291255694  0.3119975917
#> [198,]  0.9777056509 -1.7348736233
#> [199,] -0.5175246460  0.3131361156
#> [200,] -0.2300444387 -0.1592765775
#> [201,]  0.4557093615 -0.1912404706
#> [202,] -0.5151791976  0.2279557767
#> [203,]  0.7054862938  0.3495016052
#> [204,] -0.3020766059 -0.0573859836
#> [205,]  1.2266535048 -0.3714095263
#> [206,]  0.5146627508  0.4713492561
#> [207,]  0.9305779484  0.7006305162
#> [208,]  0.7983100600 -0.4123500707
#> [209,] -0.4998367077 -0.1083796245
#> [210,] -0.7551298746  0.6197572992
#> [211,] -0.6890215022  0.5521416718
#> [212,] -0.4165376031 -0.2893098534
#> [213,]  0.9204813985 -0.1073204169
#> [214,]  1.1017390443  1.2307277390
#> [215,]  0.9840570094  0.0520941315
#> [216,] -0.3084832526  0.5676809676
#> [217,] -0.1830765284 -0.0024264977
#> [218,]  0.5247811892  1.6384416643
#> [219,]  0.0838777759 -0.0942000167
#> [220,]  0.5738762066 -0.0343301982
#> [221,] -0.4618042888 -0.1987665920
#> [222,]  0.2614313864 -0.8427638783
#> [223,] -0.5197121317 -0.0110048288
#> [224,]  1.2420997977 -0.3706245422
#> [225,] -0.0246049042  0.0217081466
#> [226,] -0.3369390185  0.5613852208
#> [227,] -0.2428129543  0.1794623907
#> [228,]  1.1289079747 -0.4389965936
#> [229,] -0.3563336919  0.1371036123
#> [230,] -0.2426517256 -0.4861642413
#> [231,] -0.3352657971 -0.6896386209
#> [232,] -0.3786477330 -0.2775539554
#> [233,] -1.1191703822 -0.6327275886
#> [234,]  0.7575485460 -0.0672142410
#> [235,] -0.5459713043 -1.3420918264
#> [236,]  0.9662868639  0.4456705001
#> [237,] -1.0062424725  0.9030560922
#> [238,]  0.2674009504 -0.1093182798
#> [239,] -0.1415620770  0.0266431020
#> [240,] -0.9962490714  0.5792118096
#> [241,]  0.0784666140  0.0631550190
#> [242,] -0.1786415575  0.0847451409
#> [243,] -0.5590441598  0.3670194591
#> [244,] -0.2522958257 -0.2529638137
#> [245,]  0.7721034609  0.0855291837
#> [246,] -0.0596807062  0.1413177145
#> [247,]  2.4535617074  1.5577596220
#> [248,] -1.7542564089  1.4631741756
#> [249,] -1.1476951147 -0.5356693678
#> [250,]  0.3694343877  0.2423739730
#> [251,] -0.7857850270 -1.7810246751
#> [252,] -1.2507973747  0.4433092403
#> [253,] -0.7452429009  0.7351311619
#> [254,] -0.1236928852 -0.0316502254
#> [255,] -0.2609392161 -0.0135147724
#> [256,] -1.7066951389  0.6851344752
#> [257,] -0.3029294846 -0.3501926256
#> [258,]  1.5976073605 -3.5045866630
#> [259,]  1.0253975652 -0.3530358266
#> [260,]  3.0116266746 -6.3469442848
#> [261,] -0.7425386338 -0.2748964174
#> [262,]  0.4829533019 -0.0757999829
#> [263,]  0.2273505895 -0.1004635716
#> [264,] -0.1664656685 -0.1570017023
#> [265,]  0.1003161283  0.1347026325
#> [266,] -0.6342788469 -0.7386291405
#> [267,]  2.1101141279 -1.3890305000
#> [268,]  0.5459106988 -0.1375825408
#> [269,] -0.9030594136 -0.1282676523
#> [270,]  0.8625666968 -0.6775172626
#> [271,]  0.0427925440 -0.0967934183
#> [272,] -0.2045079275  0.0751195257
#> [273,] -2.0296264040 -1.0984317353
#> [274,]  0.4416658927 -0.1595634296
#> [275,]  0.7936924073  0.2741643080
#> [276,] -0.5616441075 -0.5188880257
#> [277,]  1.5434330998 -0.7830077429
#> [278,]  0.3080575441  0.0629811696
#> [279,] -1.3293239165  0.2272654644
#> [280,] -0.1367728714 -0.1671300008
#> [281,] -0.5940067554 -1.3819200271
#> [282,] -0.7927054937 -1.7369396237
#> [283,]  1.2822615461  0.9807094235
#> [284,] -1.3295636952  1.2172892592
#> [285,] -0.6904257238 -0.9842460457
#> [286,]  0.3981428140 -0.8562201299
#> [287,]  1.0743840681 -0.0227481584
#> [288,] -1.0063176812 -1.5044426714
#> [289,]  0.3748118356 -0.3335932461
#> [290,] -0.4594095091 -0.2223882684
#> [291,] -0.6689850499  0.3761061841
#> [292,]  0.4829046221 -0.2003606020
#> [293,] -0.6744704796 -0.7571017915
#> [294,] -0.2140479992  0.0211019754
#> [295,]  1.3169391199 -1.6961040180
#> [296,] -1.1687251812 -1.5085617193
#> [297,]  0.5756035686 -0.0703234319
#> [298,] -0.8278149712  0.6826355450
#> [299,]  1.1596362094 -0.7517697657
#> [300,] -1.1153759781  0.3393217811
#> 
#> $inv_hessian
#>            [,1]       [,2]
#> [1,] 1.00062255 0.02459982
#> [2,] 0.02459982 0.97204820
#>